%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSetTuple.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{AffymetrixCelSetTuple}
\docType{class}
\alias{AffymetrixCelSetTuple}


 \title{The AffymetrixCelSetTuple class}

 \description{
  Package:  aroma.affymetrix \cr
\bold{Class AffymetrixCelSetTuple}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{AffymetrixCelSetTuple}\cr

\bold{Directly known subclasses:}\cr
\link[aroma.affymetrix]{ChipEffectSetTuple}\cr

public static class \bold{AffymetrixCelSetTuple}\cr
extends \link[R.oo]{Object}\cr


 }

 \usage{AffymetrixCelSetTuple(csList=NULL, tags="*", ..., .setClass="AffymetrixCelSet")}

 \arguments{
   \item{csList}{A single or \code{\link[base]{list}} of \code{\link{AffymetrixCelSet}}:s.}
   \item{tags}{A \code{\link[base]{character}} \code{\link[base]{vector}} of tags.}
   \item{...}{Not used.}
   \item{.setClass}{The name of the class of the input set.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{byPath} \tab  -\cr
 \tab \code{clearCache} \tab  -\cr
 \tab \code{extract} \tab  -\cr
 \tab \code{getAlias} \tab  -\cr
 \tab \code{getArrays} \tab  -\cr
 \tab \code{getArrayTuple} \tab  -\cr
 \tab \code{getChipTypes} \tab  -\cr
 \tab \code{getFullName} \tab  -\cr
 \tab \code{getFullNames} \tab  -\cr
 \tab \code{getListOfSets} \tab  -\cr
 \tab \code{getName} \tab  -\cr
 \tab \code{getNames} \tab  -\cr
 \tab \code{getTags} \tab  -\cr
 \tab \code{nbrOfArrays} \tab  -\cr
 \tab \code{nbrOfChipTypes} \tab  -\cr
 \tab \code{setAlias} \tab  -\cr
 \tab \code{setTags} \tab  -\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstanciationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save


 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
