%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{AffymetrixModel}
\docType{class}
\alias{AffymetrixModel}


 \title{The AffymetrixModel class}

 \description{
  Package:  aroma.affymetrix \cr
\bold{Class AffymetrixModel}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\emph{\code{AffymetrixModel}}\cr

\bold{Directly known subclasses:}\cr
\cr

public abstract static class \bold{AffymetrixModel}\cr
extends \link[R.oo]{Object}\cr



 }

 \usage{AffymetrixModel(dataset=NULL, path=NULL, ...)}

 \arguments{
   \item{dataset}{}
   \item{path}{}
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{as.character} \tab  -\cr
 \tab \code{findUnitsDone} \tab  -\cr
 \tab \code{fit} \tab  -\cr
 \tab \code{fromAffymetrixDataset} \tab  -\cr
 \tab \code{getDataset} \tab  -\cr
 \tab \code{getFit} \tab  -\cr
 \tab \code{getPath} \tab  -\cr
 \tab \code{nbrOfArrays} \tab  -\cr
 \tab \code{nbrOfProbes} \tab  -\cr
 \tab \code{nbrOfProbesets} \tab  -\cr
}


\bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstanciationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save


 }

 \seealso{
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
