%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixParameterFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{AffymetrixParameterFile}
\docType{class}
\alias{AffymetrixParameterFile}


 \title{The AffymetrixParameterFile class}

 \description{
  Package:  aroma.affymetrix \cr
\bold{Class AffymetrixParameterFile}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[aroma.affymetrix]{AffymetrixFile}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{AffymetrixParameterFile}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{AffymetrixParameterFile}\cr
extends \link[aroma.affymetrix]{AffymetrixFile}\cr



  An AffymetrixParameterFile object represents parameter estimates for
  a single Affymetrix array.
 }

 \usage{AffymetrixParameterFile(...)}

 \arguments{
   \item{...}{Arguments passed to \code{\link{AffymetrixFile}}.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[aroma.affymetrix:fromFile.AffymetrixParameterFile]{fromFile}} \tab Defines an AffymetrixParameterFile object from a file.\cr
 \tab \code{getCdf} \tab  -\cr
}


\bold{Methods inherited from AffymetrixFile}:\cr
as.character, copyFile, fromFile, getFilename, getFileSize, getFileType, getLabel, getName, getPath, getPathname, setLabel, stextLabel, stextLabels

\bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstanciationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save


 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
