%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixPowerTools.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{AffymetrixPowerTools}
\docType{class}
\alias{AffymetrixPowerTools}


 \title{The AffymetrixPowerTools class}

 \description{
  Package:  aroma.affymetrix \cr
\bold{Class AffymetrixPowerTools}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{AffymetrixPowerTools}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{AffymetrixPowerTools}\cr
extends \link[R.oo]{Object}\cr



  This class represents an interface to the Affymetrix Power Tools (APT)
  applications [1].
 }

 \usage{AffymetrixPowerTools(...)}

 \arguments{
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{writeChrXFile} \tab  -\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstanciationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save


 }

 \details{
  Note that the methods of this class are calling the APT [1] binaries,
  which must be installed on the system.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
 }

 \references{
  [1] Affymetrix, Affymetrix Power Tools (APT) software, Dec 2006.
      \url{http://www.affymetrix.com/support/developer/powertools/index.affx}
      \cr
 }


\keyword{classes}
\keyword{internal}
