%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCnChpSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.AffymetrixCnChpSet.AffymetrixCnChpSet}
\alias{as.AffymetrixCnChpSet.AffymetrixCnChpSet}
\alias{AffymetrixCnChpSet.as.AffymetrixCnChpSet}
\alias{as.AffymetrixCnChpSet,AffymetrixCnChpSet-method}
\alias{as.AffymetrixCnChpSet.list}
\alias{as.AffymetrixCnChpSet.default}

\title{Coerce an object to an AffymetrixCnChpSet object}

\description{
  Coerce an object to an AffymetrixCnChpSet object.
}

\usage{
\method{as.AffymetrixCnChpSet}{AffymetrixCnChpSet}(object, ...)
}

\arguments{
 \item{...}{Other arguments passed to \code{\link[base]{list.files}}().}
}

\value{
  Returns an \code{\link{AffymetrixCnChpSet}} object.
}

\seealso{
  For more information see \code{\link{AffymetrixCnChpSet}}.
}
\keyword{internal}
\keyword{methods}
