%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.AffymetrixDataSet.AffymetrixDataSet}
\alias{as.AffymetrixDataSet.AffymetrixDataSet}
\alias{AffymetrixDataSet.as.AffymetrixDataSet}
\alias{as.AffymetrixDataSet.AffymetrixDataSet}
\alias{as.AffymetrixDataSet,AffymetrixDataSet-method}

\alias{as.AffymetrixDataSet.list}
\alias{as.AffymetrixDataSet.default}

\title{Coerce an object to an AffymetrixDataSet object}

\description{
  Coerce an object to an AffymetrixDataSet object.
}

\usage{\method{as.AffymetrixDataSet}{AffymetrixDataSet}(object, ...)}

\arguments{
 \item{...}{Other arguments passed to \code{\link[base]{list.files}}().}
}

\value{
  Returns an \code{\link{AffymetrixDataSet}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
