%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.AffymetrixFileSet.AffymetrixFileSet}
\alias{as.AffymetrixFileSet.AffymetrixFileSet}
\alias{AffymetrixFileSet.as.AffymetrixFileSet}
\alias{as.AffymetrixFileSet.AffymetrixFileSet}
\alias{as.AffymetrixFileSet,AffymetrixFileSet-method}

\alias{as.AffymetrixFileSet.list}
\alias{as.AffymetrixFileSet.default}

\title{Coerce an object to an AffymetrixFileSet object}

\description{
  Coerce an object to an AffymetrixFileSet object.
}

\usage{\method{as.AffymetrixFileSet}{AffymetrixFileSet}(object, ...)}

\arguments{
 \item{...}{Other arguments passed to \code{\link[base]{list.files}}().}
}

\value{
  Returns an \code{\link{AffymetrixFileSet}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFileSet}}.
}
\keyword{internal}
\keyword{methods}
