%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.AffymetrixCelSet}
\alias{as.character.AffymetrixCelSet}
\alias{AffymetrixCelSet.as.character}
\alias{as.character,AffymetrixCelSet-method}

\title{Returns a short string describing the Affymetrix CEL set}

\description{
 Returns a short string describing the Affymetrix CEL set.
}

\usage{
\method{as.character}{AffymetrixCelSet}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AffymetrixCelSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
