%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCnChpSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.AffymetrixCnChpSet}
\alias{as.character.AffymetrixCnChpSet}
\alias{AffymetrixCnChpSet.as.character}
\alias{as.character.AffymetrixCnChpSet}
\alias{as.character,AffymetrixCnChpSet-method}


\title{Returns a short string describing the set}

\description{
 Returns a short string describing the set.
}

\usage{\method{as.character}{AffymetrixCnChpSet}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixCnChpSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
