%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.AffymetrixDataFile}
\alias{as.character.AffymetrixDataFile}
\alias{AffymetrixDataFile.as.character}
\alias{as.character.AffymetrixDataFile}
\alias{as.character,AffymetrixDataFile-method}


\title{Returns a short string describing the Affymetrix data file}

\description{
 Returns a short string describing the Affymetrix data file.
}

\usage{\method{as.character}{AffymetrixDataFile}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
