%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.AffymetrixDataSet}
\alias{as.character.AffymetrixDataSet}
\alias{AffymetrixDataSet.as.character}
\alias{as.character.AffymetrixDataSet}
\alias{as.character,AffymetrixDataSet-method}


\title{Returns a short string describing the Affymetrix data set}

\description{
 Returns a short string describing the Affymetrix data set.
}

\usage{\method{as.character}{AffymetrixDataSet}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
