%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.AffymetrixFile}
\alias{as.character.AffymetrixFile}
\alias{AffymetrixFile.as.character}
\alias{as.character.AffymetrixFile}
\alias{as.character,AffymetrixFile-method}


\title{Returns a short string describing the Affymetrix file}

\description{
 Returns a short string describing the Affymetrix file.
}

\usage{\method{as.character}{AffymetrixFile}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
