%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.AffymetrixFileSet}
\alias{as.character.AffymetrixFileSet}
\alias{AffymetrixFileSet.as.character}
\alias{as.character.AffymetrixFileSet}
\alias{as.character,AffymetrixFileSet-method}


\title{Returns a short string describing the Affymetrix file set}

\description{
 Returns a short string describing the Affymetrix file set.
}

\usage{\method{as.character}{AffymetrixFileSet}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFileSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
