%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CnagCfhSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.CnagCfhSet}
\alias{as.character.CnagCfhSet}
\alias{CnagCfhSet.as.character}
\alias{as.character,CnagCfhSet-method}

\title{Returns a short string describing the CNAG CFH set}

\description{
 Returns a short string describing the CNAG CFH set.
}

\usage{
\method{as.character}{CnagCfhSet}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\seealso{
  For more information see \code{\link{CnagCfhSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
