%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  DChipDcpSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.DChipDcpSet}
\alias{as.character.DChipDcpSet}
\alias{DChipDcpSet.as.character}
\alias{as.character.DChipDcpSet}
\alias{as.character,DChipDcpSet-method}


\title{Returns a short string describing the DChip CHP set}

\description{
 Returns a short string describing the DChip CHP set.
}

\usage{\method{as.character}{DChipDcpSet}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{DChipDcpSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
