%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenotypeCallSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.GenotypeCallSet}
\alias{as.character.GenotypeCallSet}
\alias{GenotypeCallSet.as.character}
\alias{as.character.GenotypeCallSet}
\alias{as.character,GenotypeCallSet-method}


\title{Returns a short string describing the genotype call set}

\description{
 Returns a short string describing the genotype call set.
}

\usage{\method{as.character}{GenotypeCallSet}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenotypeCallSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
