%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParameterSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.ParameterSet}
\alias{as.character.ParameterSet}
\alias{ParameterSet.as.character}
\alias{as.character.ParameterSet}
\alias{as.character,ParameterSet-method}


\title{Returns a short string describing this parameter set}

\description{
 Returns a short string describing this parameter set.
}

\usage{\method{as.character}{ParameterSet}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ParameterSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
