%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CnagCfhSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.CnagCfhSet.CnagCfhSet}
\alias{as.CnagCfhSet.CnagCfhSet}
\alias{CnagCfhSet.as.CnagCfhSet}
\alias{as.CnagCfhSet,CnagCfhSet-method}
\alias{as.CnagCfhSet.list}
\alias{as.CnagCfhSet.default}

\title{Coerce an object to an CnagCfhSet object}

\description{
  Coerce an object to an CnagCfhSet object.
}

\usage{
\method{as.CnagCfhSet}{CnagCfhSet}(object, ...)
}

\arguments{
 \item{...}{Other arguments passed to \code{\link[base]{list.files}}().}
}

\value{
  Returns an \code{\link{CnagCfhSet}} object.
}

\seealso{
  For more information see \code{\link{CnagCfhSet}}.
}
\keyword{internal}
\keyword{methods}
