%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  DChipDcpSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.DChipDcpSet.DChipDcpSet}
\alias{as.DChipDcpSet.DChipDcpSet}
\alias{DChipDcpSet.as.DChipDcpSet}
\alias{as.DChipDcpSet,DChipDcpSet-method}
\alias{as.DChipDcpSet.list}
\alias{as.DChipDcpSet.default}

\title{Coerce an object to an DChipDcpSet object}

\description{
  Coerce an object to an DChipDcpSet object.
}

\usage{
\method{as.DChipDcpSet}{DChipDcpSet}(object, ...)
}

\arguments{
 \item{...}{Other arguments passed to \code{\link[base]{list.files}}().}
}

\value{
  Returns an \code{\link{DChipDcpSet}} object.
}

\seealso{
  For more information see \code{\link{DChipDcpSet}}.
}
\keyword{internal}
\keyword{methods}
