%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.GenericDataFileSet.GenericDataFileSet}
\alias{as.GenericDataFileSet.GenericDataFileSet}
\alias{GenericDataFileSet.as.GenericDataFileSet}
\alias{as.GenericDataFileSet.GenericDataFileSet}
\alias{as.GenericDataFileSet,GenericDataFileSet-method}

\alias{as.GenericDataFileSet.list}
\alias{as.GenericDataFileSet.default}

\title{Coerce an object to an GenericDataFileSet object}

\description{
  Coerce an object to an GenericDataFileSet object.
}

\usage{\method{as.GenericDataFileSet}{GenericDataFileSet}(object, ...)}

\arguments{
 \item{...}{Other arguments passed to \code{\link[base]{list.files}}().}
}

\value{
  Returns an \code{\link{GenericDataFileSet}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenericDataFileSet}}.
}
\keyword{internal}
\keyword{methods}
