%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenotypeCallSet.R
% 
% on Sat Jan 06 01:52:11 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.GenotypeCallSet.GenotypeCallSet}
\alias{as.GenotypeCallSet.GenotypeCallSet}
\alias{GenotypeCallSet.as.GenotypeCallSet}
\alias{as.GenotypeCallSet.GenotypeCallSet}
\alias{as.GenotypeCallSet,GenotypeCallSet-method}

\alias{as.GenotypeCallSet.list}
\alias{as.GenotypeCallSet.default}

\title{Coerce an object to an GenotypeCallSet object}

\description{
  Coerce an object to an GenotypeCallSet object.
}

\usage{\method{as.GenotypeCallSet}{GenotypeCallSet}(object, ...)}

\arguments{
 \item{...}{Other arguments passed to \code{\link[base]{list.files}}().}
}

\value{
  Returns an \code{\link{GenotypeCallSet}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenotypeCallSet}}.
}
\keyword{internal}
\keyword{methods}
