%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.list.AffymetrixDataSet}
\alias{as.list.AffymetrixDataSet}
\alias{AffymetrixDataSet.as.list}
\alias{as.list.AffymetrixDataSet}
\alias{as.list,AffymetrixDataSet-method}


\title{Returns the data files of the Affymetrix data set}

\description{
 Returns the data files of the Affymetrix data set.
}

\usage{\method{as.list}{AffymetrixDataSet}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} of \code{\link{AffymetrixDataFile}}s.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
