%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.list.AffymetrixFileSet}
\alias{as.list.AffymetrixFileSet}
\alias{AffymetrixFileSet.as.list}
\alias{as.list.AffymetrixFileSet}
\alias{as.list,AffymetrixFileSet-method}


\title{Returns the files of the Affymetrix file set}

\description{
 Returns the files of the Affymetrix file set.
}

\usage{\method{as.list}{AffymetrixFileSet}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} of \code{\link{AffymetrixFile}}s.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{AffymetrixFileSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
