%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParameterSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.list.ParameterSet}
\alias{as.list.ParameterSet}
\alias{ParameterSet.as.list}
\alias{as.list.ParameterSet}
\alias{as.list,ParameterSet-method}


\title{Returns the data files of the parameter set}

\description{
 Returns the data files of the parameter set.
}

\usage{\method{as.list}{ParameterSet}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} of \code{\link{ParameterFile}}s.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ParameterSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
