%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaChipTypeAnnotationFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{AromaChipTypeAnnotationFile$byChipType}
\alias{AromaChipTypeAnnotationFile$byChipType}
\alias{byChipType.AromaChipTypeAnnotationFile}
\alias{AromaChipTypeAnnotationFile.byChipType}
\alias{byChipType,AromaChipTypeAnnotationFile-method}
\alias{AromaChipTypeAnnotationFile.byName}
\alias{byName.AromaChipTypeAnnotationFile}
\alias{byName,AromaChipTypeAnnotationFile-method}


\title{Defines an AromaChipTypeAnnotationFile object by chip type}

\description{
 Defines an AromaChipTypeAnnotationFile object by chip type.
}

\usage{
## Static method (use this):
## AromaChipTypeAnnotationFile$byChipType(chipType, tags=NULL, ..., verbose=FALSE)

## Don't use the below:
\method{byChipType}{AromaChipTypeAnnotationFile}(static, chipType, tags=NULL, ..., verbose=FALSE)
}

\arguments{
 \item{chipType}{A \code{\link[base]{character}} string.}
 \item{tags}{An optional \code{\link[base]{character}} \code{\link[base]{vector}} of tags.}
 \item{...}{Not used.}
 \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns an \code{\link{AromaChipTypeAnnotationFile}} object.
}

\seealso{
  \code{\link[aroma.affymetrix:fromFile.AromaChipTypeAnnotationFile]{*fromFile}()}.
  For more information see \code{\link{AromaChipTypeAnnotationFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
