%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenomeInformation.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{byChipType.GenomeInformation}
\alias{byChipType.GenomeInformation}
\alias{GenomeInformation.byChipType}
\alias{byChipType.GenomeInformation}
\alias{byChipType,GenomeInformation-method}


\title{Static method to define a genome information set by chip type}

\description{
  Static method to define a genome information set by chip type.
}

\synopsis{byChipType.GenomeInformation(...)}
\usage{GenomeInformation$byChipType()}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link{GenomeInformation}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenomeInformation}}.
}
\keyword{internal}
\keyword{methods}
