%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SnpInformation.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{byChipType.SnpInformation}
\alias{byChipType.SnpInformation}
\alias{SnpInformation.byChipType}
\alias{byChipType,SnpInformation-method}

\title{Static method to define a genome information set by chip type}

\description{
  Static method to define a genome information set by chip type.
}

\usage{
\method{byChipType}{SnpInformation}(...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link{SnpInformation}} object.
}

\seealso{
  \code{\link[aroma.affymetrix:fromCdf.SnpInformation]{*fromCdf}()}.
  For more information see \code{\link{SnpInformation}}.
}
\keyword{internal}
\keyword{methods}
