%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UgpGenomeInformation.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{UgpGenomeInformation$byChipType}
\alias{UgpGenomeInformation$byChipType}
\alias{byChipType.UgpGenomeInformation}
\alias{UgpGenomeInformation.byChipType}
\alias{byChipType,UgpGenomeInformation-method}

\title{Defines a UgpGenomeInformation object by chip type}

\description{
 Defines a UgpGenomeInformation object by chip type.
}

\usage{
## Static method (use this):
## UgpGenomeInformation$byChipType(chipType, tags=NULL, ..., nbrOfUnits=NULL,
##   verbose=FALSE)

## Don't use the below:
\method{byChipType}{UgpGenomeInformation}(static, chipType, tags=NULL, ..., nbrOfUnits=NULL, verbose=FALSE)
}

\arguments{
 \item{chipType}{A \code{\link[base]{character}} string.}
 \item{tags}{...}
 \item{...}{Additional arguments passed to \code{\link{UgpGenomeInformation}}.}
 \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns an \code{\link{UgpGenomeInformation}} object.
 If no file was not found, an error is thrown.
}

\seealso{
  For more information see \code{\link{UgpGenomeInformation}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
