%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FirmaModel.R
% 
% on Fri Dec 14 14:17:46 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{calculateResiduals.FirmaModel}
\alias{calculateResiduals.FirmaModel}
\alias{FirmaModel.calculateResiduals}
\alias{calculateResiduals.FirmaModel}
\alias{calculateResiduals,FirmaModel-method}


\title{Gets the set of residuals corresponding to the PLM}

\description{
 Gets the set of residuals corresponding to the PLM.
}

\usage{\method{calculateResiduals}{FirmaModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or a \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link{ResidualSet}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{FirmaModel}}.
}
\keyword{internal}
\keyword{methods}
