%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCdfFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{compare.AffymetrixCdfFile}
\alias{compare.AffymetrixCdfFile}
\alias{AffymetrixCdfFile.compare}
\alias{compare,AffymetrixCdfFile-method}

\title{Checks if two AffymetrixCdfFile objects are equal}

\description{
 Checks if two AffymetrixCdfFile objects are equal.
}

\usage{
\method{compare}{AffymetrixCdfFile}(this, other, ...)
}

\arguments{
  \item{other}{The other \code{\link{AffymetrixCdfFile}} object to be compared to.}
  \item{...}{Additional arguments passed to \code{\link[affxparser]{compareCdfs}}.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the two objects are equal, otherwise \code{\link[base:logical]{FALSE}}.
}

\seealso{
  For more information see \code{\link{AffymetrixCdfFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
