%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ArrayExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{display.ArrayExplorer}
\alias{display.ArrayExplorer}
\alias{ArrayExplorer.display}
\alias{display.ArrayExplorer}
\alias{display,ArrayExplorer-method}


\title{Displays the explorer in the default browser}

\description{
 Displays the explorer in the default browser.
}

\usage{\method{display}{ArrayExplorer}(this, ..., verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ArrayExplorer}}.
}
\keyword{internal}
\keyword{methods}
