%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.TRANSFORM.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{doTransform.AffymetrixDataSet}
\alias{doTransform.AffymetrixDataSet}
\alias{AffymetrixDataSet.doTransform}
\alias{doTransform.AffymetrixDataSet}
\alias{doTransform,AffymetrixDataSet-method}


\title{Sets if transformation of probe signals should be done or not}

\description{
  Sets if transformation of probe signals should be done or not.
}

\usage{\method{doTransform}{AffymetrixDataSet}(this, status=TRUE, ...)}

\arguments{
 \item{status}{A \code{\link[base]{logical}}.}
 \item{...}{Not used.}
}

\value{
  Returns a nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
