%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParameterCelFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{encodeUnitGroup.ParameterCelFile}
\alias{encodeUnitGroup.ParameterCelFile}
\alias{ParameterCelFile.encodeUnitGroup}
\alias{encodeUnitGroup.ParameterCelFile}
\alias{encodeUnitGroup,ParameterCelFile-method}


\title{Returns a unit group with fields as in a CEL file}

\description{
 Returns a unit group with fields as in a CEL file.
}

\synopsis{encodeUnitGroup.ParameterCelFile(...)}
\usage{ParameterCelFile$encodeUnitGroup()}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}} structure with elements \code{intensities},
 \code{stdvs}, and \code{pixels}, all of equal length.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ParameterCelFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
