%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Explorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Explorer}
\docType{class}
\alias{Explorer}


 \title{The Explorer class}

 \description{
  Package:  aroma.affymetrix \cr
\bold{Class Explorer}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\emph{\code{Explorer}}\cr

\bold{Directly known subclasses:}\cr
\emph{\link[aroma.affymetrix]{ArrayExplorer}}, \link[aroma.affymetrix]{ChromosomeExplorer}\cr

public abstract static class \bold{Explorer}\cr
extends \link[R.oo]{Object}\cr


 }

 \usage{Explorer(tags="*", ...)}

 \arguments{
   \item{tags}{A \code{\link[base]{character}} \code{\link[base]{vector}} of tags to be added to the output path.}
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{addIncludes} \tab  -\cr
 \tab \code{display} \tab  -\cr
 \tab \code{getAlias} \tab  -\cr
 \tab \code{getArrays} \tab  -\cr
 \tab \code{getAsteriskTags} \tab  -\cr
 \tab \code{getFullName} \tab  -\cr
 \tab \code{getName} \tab  -\cr
 \tab \code{getPath} \tab  -\cr
 \tab \code{getTags} \tab  -\cr
 \tab \code{getTagsOfInput} \tab  -\cr
 \tab \code{nbrOfArrays} \tab  -\cr
 \tab \code{process} \tab  -\cr
 \tab \code{setAlias} \tab  -\cr
 \tab \code{setArrays} \tab  -\cr
 \tab \code{setup} \tab  -\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstanciationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save


 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
 }
\keyword{classes}
