%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{extract.AffymetrixDataSet}
\alias{extract.AffymetrixDataSet}
\alias{AffymetrixDataSet.extract}
\alias{extract.AffymetrixDataSet}
\alias{extract,AffymetrixDataSet-method}


\title{Extract a subset of the data set}

\description{
  Extract a subset of the data set.
}

\usage{\method{extract}{AffymetrixDataSet}(this, arrays, ...)}

\arguments{
 \item{arrays}{Array indices.}
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link{AffymetrixDataSet}} (or a subclass) object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
