%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{extract.AffymetrixFileSet}
\alias{extract.AffymetrixFileSet}
\alias{AffymetrixFileSet.extract}
\alias{extract.AffymetrixFileSet}
\alias{extract,AffymetrixFileSet-method}


\title{Extract a subset of the file set}

\description{
  Extract a subset of the file set.
}

\usage{\method{extract}{AffymetrixFileSet}(this, files, ...)}

\arguments{
 \item{files}{File indices.}
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link{AffymetrixFileSet}} (or a subclass) object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFileSet}}.
}
\keyword{internal}
\keyword{methods}
