%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParameterSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{extract.ParameterSet}
\alias{extract.ParameterSet}
\alias{ParameterSet.extract}
\alias{extract.ParameterSet}
\alias{extract,ParameterSet-method}


\title{Extract a subset of the parameter set}

\description{
  Extract a subset of the parameter set.
}

\usage{\method{extract}{ParameterSet}(this, arrays, ...)}

\arguments{
 \item{arrays}{Array indices.}
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link{ParameterSet}} (or a subclass) object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ParameterSet}}.
}
\keyword{internal}
\keyword{methods}
