%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FirmaModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{findUnitsTodo.FirmaModel}
\alias{findUnitsTodo.FirmaModel}
\alias{FirmaModel.findUnitsTodo}
\alias{findUnitsTodo,FirmaModel-method}

\title{Identifies non-fitted units}

\description{
 Identifies non-fitted units.
}

\usage{
\method{findUnitsTodo}{FirmaModel}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} of unit indices.
}

\seealso{
  Internally this methods calls the same method for the
  \code{\link{ChipEffectSet}} class.
  For more information see \code{\link{FirmaModel}}.
}
\keyword{internal}
\keyword{methods}
