%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ProbeLevelModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{findUnitsTodo.ProbeLevelModel}
\alias{findUnitsTodo.ProbeLevelModel}
\alias{ProbeLevelModel.findUnitsTodo}
\alias{findUnitsTodo.ProbeLevelModel}
\alias{findUnitsTodo,ProbeLevelModel-method}


\title{Identifies units for which the PLM has still not been fitted to}

\description{
 Identifies units for which the PLM has still not been fitted to.
}

\usage{\method{findUnitsTodo}{ProbeLevelModel}(this, ..., verbose=FALES)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  Internally this methods calls the same method for the
  \code{\link{ChipEffectSet}} class.
  For more information see \code{\link{ProbeLevelModel}}.
}
\keyword{internal}
\keyword{methods}
