%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{findUnitsTodo.UnitModel}
\alias{findUnitsTodo.UnitModel}
\alias{UnitModel.findUnitsTodo}
\alias{findUnitsTodo,UnitModel-method}

\title{Identifies non-fitted units}

\description{
 Identifies non-fitted units.
}

\usage{
\method{findUnitsTodo}{UnitModel}(...)
}

\arguments{
  \item{...}{Arguments specific to any subclass.}
}

\value{
 Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} of unit indices.
}

\seealso{
  For more information see \code{\link{UnitModel}}.
}
\keyword{internal}
\keyword{methods}
