%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixUnitGroupsModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fit.AffymetrixUnitGroupsModel}
\alias{fit.AffymetrixUnitGroupsModel}
\alias{AffymetrixUnitGroupsModel.fit}
\alias{fit.AffymetrixUnitGroupsModel}
\alias{fit,AffymetrixUnitGroupsModel-method}


\title{Estimates the model parameters}

\description{
 Estimates the model parameters for the given data set to all or a subset of the units.
}

\usage{\method{fit}{AffymetrixUnitGroupsModel}(...)}

\arguments{
  \item{...}{Arguments used by the method for a subclass.}
}

\value{
 Returns ???.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixUnitGroupsModel}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
