%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CopyNumberChromosomalModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fit.CopyNumberChromosomalModel}
\alias{fit.CopyNumberChromosomalModel}
\alias{CopyNumberChromosomalModel.fit}
\alias{fit.CopyNumberChromosomalModel}
\alias{fit,CopyNumberChromosomalModel-method}


\title{Fits the model}

\description{
 Fits the model.
}

\usage{\method{fit}{CopyNumberChromosomalModel}(...)}

\arguments{
  \item{...}{See subclasses.}
}

\value{
 See subclasses.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{CopyNumberChromosomalModel}}.
}
\keyword{internal}
\keyword{methods}
