%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CrlmmModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fit.CrlmmModel}
\alias{fit.CrlmmModel}
\alias{CrlmmModel.fit}
\alias{fit.CrlmmModel}
\alias{fit,CrlmmModel-method}


\title{Fits the model}

\description{
 Fits the model.
}

\usage{\method{fit}{CrlmmModel}(this, verbose=FALSE, ...)}

\arguments{
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[oligo]{SnpCallSetPlus-class}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[oligo]{crlmm}}.
  For more information see \code{\link{CrlmmModel}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
