%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenomeAnalyzer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fit.GenomeAnalyzer}
\alias{fit.GenomeAnalyzer}
\alias{GenomeAnalyzer.fit}
\alias{fit.GenomeAnalyzer}
\alias{fit,GenomeAnalyzer-method}


\title{Fits the model used by this analyzer}

\description{
 Fits the model used by this analyzer.
}

\usage{\method{fit}{GenomeAnalyzer}(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenomeAnalyzer}}.
}
\keyword{internal}
\keyword{methods}
