%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitGroupsModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fit.UnitGroupsModel}
\alias{fit.UnitGroupsModel}
\alias{UnitGroupsModel.fit}
\alias{fit.UnitGroupsModel}
\alias{fit,UnitGroupsModel-method}


\title{Estimates the model parameters}

\description{
 Estimates the model parameters for the given data set to all or a subset of the units.
}

\usage{\method{fit}{UnitGroupsModel}(...)}

\arguments{
  \item{...}{Arguments used by the method for a subclass.}
}

\value{
 Returns ???.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{UnitGroupsModel}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
