%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SnpInformation.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fromCdf.SnpInformation}
\alias{fromCdf.SnpInformation}
\alias{SnpInformation.fromCdf}
\alias{fromCdf.SnpInformation}
\alias{fromCdf,SnpInformation-method}


\title{Static method to define a genome information set from a CDF}

\description{
  Static method to define a genome information set from a CDF.
}

\synopsis{fromCdf.SnpInformation(static, cdf, ...)}
\usage{SnpInformation$fromCdf(cdf, ...)}

\arguments{
 \item{cdf}{An \code{\link{AffymetrixCdfFile}}.}
 \item{...}{Additional argument passed to \code{\link[aroma.affymetrix:byChipType.SnpInformation]{*byChipType}()}.}
}

\value{
  Returns a \code{\link{SnpInformation}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:byChipType.SnpInformation]{*byChipType}()}.
  For more information see \code{\link{SnpInformation}}.
}
\keyword{internal}
\keyword{methods}
