%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SnpInformation.R
% 
% on Tue Nov 20 12:41:08 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fromChipType.SnpInformation}
\alias{fromChipType.SnpInformation}
\alias{SnpInformation.fromChipType}
\alias{fromChipType.SnpInformation}
\alias{fromChipType,SnpInformation-method}


\title{Static method to define a genome information set by chip type}

\description{
  Static method to define a genome information set by chip type.
}

\usage{\method{fromChipType}{SnpInformation}(...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link{SnpInformation}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:fromCdf.SnpInformation]{*fromCdf}()}.
  For more information see \code{\link{SnpInformation}}.
}
\keyword{internal}
\keyword{methods}
