%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixPgfFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{AffymetrixPgfFile$fromFile}
\alias{AffymetrixPgfFile$fromFile}
\alias{fromFile.AffymetrixPgfFile}
\alias{AffymetrixPgfFile.fromFile}
\alias{fromFile,AffymetrixPgfFile-method}

\title{Defines an AffymetrixPgfFile object from a PGF file}

\description{
 Defines an AffymetrixPgfFile object from a PGF file.
}

\usage{
## Static method (use this):
## AffymetrixPgfFile$fromFile(filename, path=NULL, ...)

## Don't use the below:
\method{fromFile}{AffymetrixPgfFile}(static, filename, path=NULL, ...)
}

\arguments{
 \item{filename}{The filename of to the file.}
 \item{path}{The path to the file.}
 \item{...}{Not used.}
}

\value{
 Returns an instance of \code{\link{AffymetrixPgfFile}} or its subclasses.
 If the file is not found or if it is of the wrong file format, an
 error is thrown.
}

\seealso{
  For more information see \code{\link{AffymetrixPgfFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
