%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenotypeCallFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fromFile.GenotypeCallFile}
\alias{fromFile.GenotypeCallFile}
\alias{GenotypeCallFile.fromFile}
\alias{fromFile.GenotypeCallFile}
\alias{fromFile,GenotypeCallFile-method}


\title{Static method defining a genotype call file object}

\description{
 Static method defining a genotype call file object.
}

\synopsis{fromFile.GenotypeCallFile(static, filename, path=NULL, ..., .checkArgs=TRUE, verbose=FALSE)}
\usage{GenotypeCallFile$fromFile(filename, path=NULL, ..., .checkArgs=TRUE, verbose=FALSE)}

\arguments{
  \item{filename}{The filename of the file.}
  \item{path}{An optional path to the file.}
  \item{...}{Not used.}
  \item{.checkArgs}{If \code{\link[base:logical]{TRUE}}, arguments are validated and parsed,
    otherwise not.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link{GenotypeCallFile}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenotypeCallFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
