%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAlias.AffymetrixFileSet}
\alias{getAlias.AffymetrixFileSet}
\alias{AffymetrixFileSet.getAlias}
\alias{getAlias.AffymetrixFileSet}
\alias{getAlias,AffymetrixFileSet-method}


\title{Gets the alias of the file set}

\description{
  Gets the alias of the file set.
}

\usage{\method{getAlias}{AffymetrixFileSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}, or \code{\link[base]{NULL}} if no alias is set.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setAlias.AffymetrixFileSet]{*setAlias}()}.
  \code{\link[aroma.affymetrix:getName.AffymetrixFileSet]{*getName}()}.
  For more information see \code{\link{AffymetrixFileSet}}.
}
\keyword{internal}
\keyword{methods}
