%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSetReporter.R
% 
% on Tue Nov 20 12:39:31 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAlias.AffymetrixFileSetReporter}
\alias{getAlias.AffymetrixFileSetReporter}
\alias{AffymetrixFileSetReporter.getAlias}
\alias{getAlias.AffymetrixFileSetReporter}
\alias{getAlias,AffymetrixFileSetReporter-method}


\title{Gets the alias of the output set}

\description{
  Gets the alias of the output set.
}

\usage{\method{getAlias}{AffymetrixFileSetReporter}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}, or \code{\link[base]{NULL}} if no alias is set.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setAlias.AffymetrixFileSetReporter]{*setAlias}()}.
  \code{\link[aroma.affymetrix:getName.AffymetrixFileSetReporter]{*getName}()}.
  For more information see \code{\link{AffymetrixFileSetReporter}}.
}
\keyword{internal}
\keyword{methods}
