%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ArrayExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAlias.ArrayExplorer}
\alias{getAlias.ArrayExplorer}
\alias{ArrayExplorer.getAlias}
\alias{getAlias.ArrayExplorer}
\alias{getAlias,ArrayExplorer-method}


\title{Gets the alias of the output set}

\description{
  Gets the alias of the output set.
}

\usage{\method{getAlias}{ArrayExplorer}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}, or \code{\link[base]{NULL}} if no alias is set.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setAlias.ArrayExplorer]{*setAlias}()}.
  \code{\link[aroma.affymetrix:getName.ArrayExplorer]{*getName}()}.
  For more information see \code{\link{ArrayExplorer}}.
}
\keyword{internal}
\keyword{methods}
