%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomeExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAlias.ChromosomeExplorer}
\alias{getAlias.ChromosomeExplorer}
\alias{ChromosomeExplorer.getAlias}
\alias{getAlias.ChromosomeExplorer}
\alias{getAlias,ChromosomeExplorer-method}


\title{Gets the alias of the output set}

\description{
  Gets the alias of the output set.
}

\usage{\method{getAlias}{ChromosomeExplorer}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}, or \code{\link[base]{NULL}} if no alias is set.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setAlias.ChromosomeExplorer]{*setAlias}()}.
  \code{\link[aroma.affymetrix:getName.ChromosomeExplorer]{*getName}()}.
  For more information see \code{\link{ChromosomeExplorer}}.
}
\keyword{internal}
\keyword{methods}
