%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericReporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAlias.GenericReporter}
\alias{getAlias.GenericReporter}
\alias{GenericReporter.getAlias}
\alias{getAlias,GenericReporter-method}

\title{Gets the alias of the report}

\description{
  Gets the alias of the report.
}

\usage{
\method{getAlias}{GenericReporter}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}, or \code{\link[base]{NULL}} if no alias is set.
}

\seealso{
  \code{\link[aroma.affymetrix:setAlias.GenericReporter]{*setAlias}()}.
  \code{\link[aroma.affymetrix:getName.GenericReporter]{*getName}()}.
  For more information see \code{\link{GenericReporter}}.
}
\keyword{internal}
\keyword{methods}
