%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Model.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAlias.Model}
\alias{getAlias.Model}
\alias{Model.getAlias}
\alias{getAlias,Model-method}

\title{Gets the name alias for the model}

\description{
 Gets the name alias for the model, if set.
}

\usage{
\method{getAlias}{Model}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string, or \code{\link[base]{NULL}}.
}

\seealso{
  \code{\link[aroma.affymetrix:getName.Model]{*getName}()}
  \code{\link[aroma.affymetrix:setAlias.Model]{*setAlias}()}
  For more information see \code{\link{Model}}.
}
\keyword{internal}
\keyword{methods}
