%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCdfFile.getAllelePairs.R
% 
% on Sat Jan 06 01:46:23 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAllelePairUnitSets.AffymetrixCdfFile}
\alias{getAllelePairUnitSets.AffymetrixCdfFile}
\alias{AffymetrixCdfFile.getAllelePairUnitSets}
\alias{getAllelePairUnitSets.AffymetrixCdfFile}
\alias{getAllelePairUnitSets,AffymetrixCdfFile-method}


\title{Gets the indices of units for all possible allele pairs}

\description{
  Gets the indices of units for all possible allele pairs.
}

\usage{\method{getAllelePairUnitSets}{AffymetrixCdfFile}(this, ...)}

\arguments{
 \item{...}{Arguments passed to \code{\link[aroma.affymetrix:getAllelePairs.AffymetrixCdfFile]{*getAllelePairs}()}.}
}

\value{
  Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getAllelePairs.AffymetrixCdfFile]{*getAllelePairs}()}.
  For more information see \code{\link{AffymetrixCdfFile}}.
}
\keyword{internal}
\keyword{methods}
