%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSetTuple.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getArrays.AffymetrixCelSetTuple}
\alias{getArrays.AffymetrixCelSetTuple}
\alias{AffymetrixCelSetTuple.getArrays}
\alias{getArrays.AffymetrixCelSetTuple}
\alias{getArrays,AffymetrixCelSetTuple-method}


\title{Gets the names of the arrays}

\description{
 Gets the names of the arrays available in the model.
}

\usage{\method{getArrays}{AffymetrixCelSetTuple}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixCelSetTuple}}.
}
\keyword{internal}
\keyword{methods}
