%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomeExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getArrays.ChromosomeExplorer}
\alias{getArrays.ChromosomeExplorer}
\alias{ChromosomeExplorer.getArrays}
\alias{getArrays.ChromosomeExplorer}
\alias{getArrays,ChromosomeExplorer-method}


\title{Gets the names of the arrays}

\description{
 Gets the names of the arrays available in the model.
}

\usage{\method{getArrays}{ChromosomeExplorer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomeExplorer}}.
}
\keyword{internal}
\keyword{methods}
