%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChipEffectSet.getBaseline.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{getBaseline.ChipEffectSet}
\alias{getBaseline.ChipEffectSet}
\alias{ChipEffectSet.getBaseline}
\alias{getBaseline.ChipEffectSet}
\alias{getBaseline,ChipEffectSet-method}


\title{Gets the baseline signals across chromosomes}

\description{
 Gets the baseline signals across chromosomes.
}

\usage{\method{getBaseline}{ChipEffectSet}(this, force=FALSE, verbose=FALSE, ...)}

\arguments{
  \item{force}{If \code{\link[base:logical]{TRUE}}, the CEL file that stores the is recreated.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
  \item{...}{Not used.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChipEffectSet}}.
}
\keyword{internal}
\keyword{methods}
