%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelFile.R
% 
% on Sat Jan 06 01:46:59 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.AffymetrixCelFile}
\alias{getCdf.AffymetrixCelFile}
\alias{AffymetrixCelFile.getCdf}
\alias{getCdf.AffymetrixCelFile}
\alias{getCdf,AffymetrixCelFile-method}


\title{Gets the CDF structure for this CEL file}

\description{
 Gets the CDF structure for this CEL file.
}

\usage{\method{getCdf}{AffymetrixCelFile}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setCdf.AffymetrixCelFile]{*setCdf}()}.
  For more information see \code{\link{AffymetrixCelFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
