%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.AffymetrixCelSet}
\alias{getCdf.AffymetrixCelSet}
\alias{AffymetrixCelSet.getCdf}
\alias{getCdf.AffymetrixCelSet}
\alias{getCdf,AffymetrixCelSet-method}


\title{Gets the CDF structure for this CEL set}

\description{
 Gets the CDF structure for this CEL set.
}

\usage{\method{getCdf}{AffymetrixCelSet}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setCdf.AffymetrixCelSet]{*setCdf}()}.
  For more information see \code{\link{AffymetrixCelSet}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
