%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixUnitGroupsModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.AffymetrixUnitGroupsModel}
\alias{getCdf.AffymetrixUnitGroupsModel}
\alias{AffymetrixUnitGroupsModel.getCdf}
\alias{getCdf.AffymetrixUnitGroupsModel}
\alias{getCdf,AffymetrixUnitGroupsModel-method}


\title{Gets the CDF structure for this model}

\description{
 Gets the CDF structure for this model.
}

\usage{\method{getCdf}{AffymetrixUnitGroupsModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setCdf.AffymetrixUnitGroupsModel]{*setCdf}()}.
  For more information see \code{\link{AffymetrixUnitGroupsModel}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
