%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CnagCfhFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.CnagCfhFile}
\alias{getCdf.CnagCfhFile}
\alias{CnagCfhFile.getCdf}
\alias{getCdf.CnagCfhFile}
\alias{getCdf,CnagCfhFile-method}


\title{Gets the CDF structure for this CEL file}

\description{
 Gets the CDF structure for this CEL file.
}

\usage{\method{getCdf}{CnagCfhFile}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setCdf.CnagCfhFile]{*setCdf}()}.
  For more information see \code{\link{CnagCfhFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
