%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CnagCfhSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.CnagCfhSet}
\alias{getCdf.CnagCfhSet}
\alias{CnagCfhSet.getCdf}
\alias{getCdf,CnagCfhSet-method}

\title{Gets the CDF structure for this CFH set}

\description{
 Gets the CDF structure for this CFH set.
}

\usage{
\method{getCdf}{CnagCfhSet}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\seealso{
  \code{\link[aroma.affymetrix:setCdf.CnagCfhSet]{*setCdf}()}.
  For more information see \code{\link{CnagCfhSet}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
