%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Crlmm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.Crlmm}
\alias{getCdf.Crlmm}
\alias{Crlmm.getCdf}
\alias{getCdf.Crlmm}
\alias{getCdf,Crlmm-method}


\title{Gets the CDF structure for this model}

\description{
 Gets the CDF structure for this model.
}

\usage{\method{getCdf}{Crlmm}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setCdf.Crlmm]{*setCdf}()}.
  For more information see \code{\link{Crlmm}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
