%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenotypeCallSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.GenotypeCallSet}
\alias{getCdf.GenotypeCallSet}
\alias{GenotypeCallSet.getCdf}
\alias{getCdf.GenotypeCallSet}
\alias{getCdf,GenotypeCallSet-method}


\title{Gets the CDF structure for this CEL set}

\description{
 Gets the CDF structure for this CEL set.
}

\usage{\method{getCdf}{GenotypeCallSet}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[aroma.affymetrix:setCdf.GenotypeCallSet]{*setCdf}()}.
  For more information see \code{\link{GenotypeCallSet}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
