%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Model.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.Model}
\alias{getCdf.Model}
\alias{Model.getCdf}
\alias{getCdf.Model}
\alias{getCdf,Model-method}


\title{Gets the CDF structure for this model}

\description{
 Gets the CDF structure for this model.
}

\usage{\method{getCdf}{Model}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Model}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
