%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitGroupsModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.UnitGroupsModel}
\alias{getCdf.UnitGroupsModel}
\alias{UnitGroupsModel.getCdf}
\alias{getCdf.UnitGroupsModel}
\alias{getCdf,UnitGroupsModel-method}


\title{Gets the CDF structure for this model}

\description{
 Gets the CDF structure for this model.
}

\usage{\method{getCdf}{UnitGroupsModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[aroma.affymetrix:setCdf.UnitGroupsModel]{*setCdf}()}.
  For more information see \code{\link{UnitGroupsModel}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
