%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitModel.R
% 
% on Sat Jan 06 01:58:48 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCdf.UnitModel}
\alias{getCdf.UnitModel}
\alias{UnitModel.getCdf}
\alias{getCdf.UnitModel}
\alias{getCdf,UnitModel-method}


\title{Gets the CDF structure for this model}

\description{
 Gets the CDF structure for this model.
}

\usage{\method{getCdf}{UnitModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCdfFile}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setCdf.UnitModel]{*setCdf}()}.
  For more information see \code{\link{UnitModel}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
