%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Crlmm.R
% 
% on Sat Jan 06 01:50:49 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipEffects.Crlmm}
\alias{getChipEffects.Crlmm}
\alias{Crlmm.getChipEffects}
\alias{getChipEffects.Crlmm}
\alias{getChipEffects,Crlmm-method}


\title{Gets the chip effects for this model}

\description{
  Gets the chip effects for this model.
}

\usage{\method{getChipEffects}{Crlmm}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link{SnpChipEffectSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Crlmm}}.
}
\keyword{internal}
\keyword{methods}
