%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CrlmmModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipEffects.CrlmmModel}
\alias{getChipEffects.CrlmmModel}
\alias{CrlmmModel.getChipEffects}
\alias{getChipEffects.CrlmmModel}
\alias{getChipEffects,CrlmmModel-method}


\title{Gets the chip effects for this model}

\description{
  Gets the chip effects for this model.
}

\usage{\method{getChipEffects}{CrlmmModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link{SnpChipEffectSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{CrlmmModel}}.
}
\keyword{internal}
\keyword{methods}
