%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AlleleSummation.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipEffectSet.AlleleSummation}
\alias{getChipEffectSet.AlleleSummation}
\alias{AlleleSummation.getChipEffectSet}
\alias{getChipEffectSet,AlleleSummation-method}

\title{Gets the set of chip effects for this model}

\description{
 Gets the set of chip effects for this model.
 There is one chip-effect file per array.
}

\usage{
\method{getChipEffectSet}{AlleleSummation}(this, ..., verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or a \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link{ChipEffectSet}} object.
}

\seealso{
  For more information see \code{\link{AlleleSummation}}.
}
\keyword{internal}
\keyword{methods}
