%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipType.AffymetrixCelFile}
\alias{getChipType.AffymetrixCelFile}
\alias{AffymetrixCelFile.getChipType}
\alias{getChipType,AffymetrixCelFile-method}

\title{Gets the chip type for this CEL file}

\description{
 Gets the chip type for this CEL file \emph{according} to the \code{\link{AffymetrixCdfFile}} object.
}

\usage{
\method{getChipType}{AffymetrixCelFile}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AffymetrixCelFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
