%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipType.AffymetrixCelSet}
\alias{getChipType.AffymetrixCelSet}
\alias{AffymetrixCelSet.getChipType}
\alias{getChipType,AffymetrixCelSet-method}

\title{Gets the chip type for this CEL set}

\description{
 Gets the chip type for this CEL set.
 The chip type is inferred from the current CDF.
}

\usage{
\method{getChipType}{AffymetrixCelSet}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[aroma.affymetrix:getCdf.AffymetrixCelSet]{*getCdf}()}.
  For more information see \code{\link{AffymetrixCelSet}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
