%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomalModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipType.ChromosomalModel}
\alias{getChipType.ChromosomalModel}
\alias{ChromosomalModel.getChipType}
\alias{getChipType.ChromosomalModel}
\alias{getChipType,ChromosomalModel-method}


\title{Gets a label for all chip types merged}

\description{
 Gets a label for all chip types merged.
}

\usage{\method{getChipType}{ChromosomalModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomalModel}}.
}
\keyword{internal}
\keyword{methods}
