%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CnagCfhFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipType.CnagCfhFile}
\alias{getChipType.CnagCfhFile}
\alias{CnagCfhFile.getChipType}
\alias{getChipType,CnagCfhFile-method}

\title{Gets the chip type for this CEL file}

\description{
 Gets the chip type for this CEL file \emph{according} to the \code{\link{AffymetrixCdfFile}} object.
}

\usage{
\method{getChipType}{CnagCfhFile}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\seealso{
  For more information see \code{\link{CnagCfhFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
