%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CopyNumberSegmentationModel.R
% 
% on Sat Sep 15 13:21:17 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipType.CopyNumberSegmentationModel}
\alias{getChipType.CopyNumberSegmentationModel}
\alias{CopyNumberSegmentationModel.getChipType}
\alias{getChipType.CopyNumberSegmentationModel}
\alias{getChipType,CopyNumberSegmentationModel-method}


\title{Gets a label for all chip types merged}

\description{
 Gets a label for all chip types merged.
}

\usage{\method{getChipType}{CopyNumberSegmentationModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{CopyNumberSegmentationModel}}.
}
\keyword{internal}
\keyword{methods}
