%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenomeInformation.AFFX.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{getChipType.GenomeInformation}
\alias{getChipType.GenomeInformation}
\alias{GenomeInformation.getChipType}
\alias{getChipType.GenomeInformation}
\alias{getChipType,GenomeInformation-method}


\title{Gets the chip type of this genome information set}

\description{
  Gets the chip type of this genome information set.
}

\usage{\method{getChipType}{GenomeInformation}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenomeInformation}}.
}
\keyword{internal}
\keyword{methods}
