%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GladModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipType.GladModel}
\alias{getChipType.GladModel}
\alias{GladModel.getChipType}
\alias{getChipType.GladModel}
\alias{getChipType,GladModel-method}


\title{Gets a label for all chip types merged}

\description{
 Gets a label for all chip types merged.
}

\usage{\method{getChipType}{GladModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GladModel}}.
}
\keyword{internal}
\keyword{methods}
