%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SnpInformation.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipType.SnpInformation}
\alias{getChipType.SnpInformation}
\alias{SnpInformation.getChipType}
\alias{getChipType.SnpInformation}
\alias{getChipType,SnpInformation-method}


\title{Gets the chip type of this genome information set}

\description{
  Gets the chip type of this genome information set.
}

\usage{\method{getChipType}{SnpInformation}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SnpInformation}}.
}
\keyword{internal}
\keyword{methods}
