%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomeExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChromosomes.ChromosomeExplorer}
\alias{getChromosomes.ChromosomeExplorer}
\alias{ChromosomeExplorer.getChromosomes}
\alias{getChromosomes.ChromosomeExplorer}
\alias{getChromosomes,ChromosomeExplorer-method}


\title{Gets the chromosomes available}

\description{
 Gets the chromosomes available.
}

\usage{\method{getChromosomes}{ChromosomeExplorer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomeExplorer}}.
}
\keyword{internal}
\keyword{methods}
