%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CopyNumberSegmentationModel.R
% 
% on Sat Sep 15 13:21:17 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChromosomes.CopyNumberSegmentationModel}
\alias{getChromosomes.CopyNumberSegmentationModel}
\alias{CopyNumberSegmentationModel.getChromosomes}
\alias{getChromosomes.CopyNumberSegmentationModel}
\alias{getChromosomes,CopyNumberSegmentationModel-method}


\title{Gets the chromosomes available}

\description{
 Gets the chromosomes available.
}

\usage{\method{getChromosomes}{CopyNumberSegmentationModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{CopyNumberSegmentationModel}}.
}
\keyword{internal}
\keyword{methods}
