%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenomeAnalyzer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChromosomes.GenomeAnalyzer}
\alias{getChromosomes.GenomeAnalyzer}
\alias{GenomeAnalyzer.getChromosomes}
\alias{getChromosomes.GenomeAnalyzer}
\alias{getChromosomes,GenomeAnalyzer-method}


\title{Gets the chromosomes available}

\description{
 Gets the chromosomes available.
}

\synopsis{getChromosomes.GenomeAnalyzer(static, ...)}
\usage{GenomeAnalyzer$getChromosomes(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenomeAnalyzer}}.
}
\keyword{internal}
\keyword{methods}
