%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GladModel.R
% 
% on Sat Jan 06 01:52:50 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChromosomes.GladModel}
\alias{getChromosomes.GladModel}
\alias{GladModel.getChromosomes}
\alias{getChromosomes.GladModel}
\alias{getChromosomes,GladModel-method}


\title{Gets the chromosomes available}

\description{
 Gets the chromosomes available.
}

\synopsis{getChromosomes.GladModel(static, ...)}
\usage{GladModel$getChromosomes(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GladModel}}.
}
\keyword{internal}
\keyword{methods}
