%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiGladModel.R
% 
% on Sat Jan 06 01:53:53 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChromosomes.MultiGladModel}
\alias{getChromosomes.MultiGladModel}
\alias{MultiGladModel.getChromosomes}
\alias{getChromosomes.MultiGladModel}
\alias{getChromosomes,MultiGladModel-method}


\title{Gets the chromosomes available}

\description{
 Gets the chromosomes available.
}

\synopsis{getChromosomes.MultiGladModel(static, ...)}
\usage{MultiGladModel$getChromosomes(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MultiGladModel}}.
}
\keyword{internal}
\keyword{methods}
