%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSetReporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDataSet.AffymetrixCelSetReporter}
\alias{getDataSet.AffymetrixCelSetReporter}
\alias{AffymetrixCelSetReporter.getDataSet}
\alias{getDataSet.AffymetrixCelSetReporter}
\alias{getDataSet,AffymetrixCelSetReporter-method}


\title{Gets the data set}

\description{
 Gets the data set for which the reporter is generating output.
}

\usage{\method{getDataSet}{AffymetrixCelSetReporter}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{AffymetrixCelSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixCelSetReporter}}.
}
\keyword{internal}
\keyword{methods}
