%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixUnitGroupsModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDataSet.AffymetrixUnitGroupsModel}
\alias{getDataSet.AffymetrixUnitGroupsModel}
\alias{AffymetrixUnitGroupsModel.getDataSet}
\alias{getDataSet.AffymetrixUnitGroupsModel}
\alias{getDataSet,AffymetrixUnitGroupsModel-method}


\title{Gets the data set for this model}

\description{
  Gets the data set for this model.
}

\usage{\method{getDataSet}{AffymetrixUnitGroupsModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link{AffymetrixFileSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixUnitGroupsModel}}.
}
\keyword{internal}
\keyword{methods}
