%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ArrayExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDataSet.ArrayExplorer}
\alias{getDataSet.ArrayExplorer}
\alias{ArrayExplorer.getDataSet}
\alias{getDataSet,ArrayExplorer-method}

\title{Gets the data set}

\description{
 Gets the data set for which the explorer is displaying it results.
}

\usage{
\method{getDataSet}{ArrayExplorer}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{AffymetrixCelSet}}.
}

\seealso{
  For more information see \code{\link{ArrayExplorer}}.
}
\keyword{internal}
\keyword{methods}
