%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Model.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDataSet.Model}
\alias{getDataSet.Model}
\alias{Model.getDataSet}
\alias{getDataSet.Model}
\alias{getDataSet,Model-method}


\title{Gets the input data set for this model}

\description{
  Gets the input data set for this model.
}

\usage{\method{getDataSet}{Model}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link{AffymetrixCelSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Model}}.
}
\keyword{internal}
\keyword{methods}
