%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDataSet.UnitModel}
\alias{getDataSet.UnitModel}
\alias{UnitModel.getDataSet}
\alias{getDataSet.UnitModel}
\alias{getDataSet,UnitModel-method}


\title{Gets the data set for this model}

\description{
  Gets the data set for this model.
}

\usage{\method{getDataSet}{UnitModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link{AffymetrixFileSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{UnitModel}}.
}
\keyword{internal}
\keyword{methods}
