%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFilename.AffymetrixFile}
\alias{getFilename.AffymetrixFile}
\alias{AffymetrixFile.getFilename}
\alias{getFilename.AffymetrixFile}
\alias{getFilename,AffymetrixFile-method}


\title{Gets the filename of the file}

\description{
  Gets the filename of the file.
}

\usage{\method{getFilename}{AffymetrixFile}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFile}}.
}
\keyword{internal}
\keyword{methods}
