%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FirmaModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFirmaSet.FirmaModel}
\alias{getFirmaSet.FirmaModel}
\alias{FirmaModel.getFirmaSet}
\alias{getFirmaSet.FirmaModel}
\alias{getFirmaSet,FirmaModel-method}

\alias{FirmaModel.getFirmaScores}
\alias{getFirmaScores.FirmaModel}
\alias{getFirmaScores,FirmaModel-method}


\title{Gets the set of FIRMA results for this model}

\description{
 Gets the set of FIRMA results for this model.
 There is one chip-effect file per array.
}

\usage{\method{getFirmaSet}{FirmaModel}(this, ..., verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or a \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link{FirmaSet}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{FirmaModel}}.
}
\keyword{internal}
\keyword{methods}
