%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffinePlm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitFunction.AffinePlm}
\alias{getFitFunction.AffinePlm}
\alias{AffinePlm.getFitFunction}
\alias{getFitFunction.AffinePlm}
\alias{getFitFunction,AffinePlm-method}


\title{Gets the low-level function that fits the PLM}

\description{
 Gets the low-level function that fits the PLM.
}

\usage{\method{getFitFunction}{AffinePlm}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[aroma.light]{calibrateMultiscan}}.
  For more information see \code{\link{AffinePlm}}.
}
\keyword{internal}
\keyword{methods}
