%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ExonRmaPlm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitFunction.ExonRmaPlm}
\alias{getFitFunction.ExonRmaPlm}
\alias{ExonRmaPlm.getFitFunction}
\alias{getFitFunction.ExonRmaPlm}
\alias{getFitFunction,ExonRmaPlm-method}


\title{Gets the low-level function that fits the Exon PLM}

\description{
 Gets the low-level function that fits the Exon PLM.
}

\usage{\method{getFitFunction}{ExonRmaPlm}(this, ..., verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\author{
  Henrik Bengtsson and Ken Simpson (WEHI) utilizing Ben Bolstad's
  \pkg{affyPLM} package.
}

\seealso{
  For more information see \code{\link{ExonRmaPlm}}.
}
\keyword{internal}
\keyword{methods}
