%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HetLogAddPlm.R
% 
% on Sat May 17 11:54:55 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitFunction.HetLogAddPlm}
\alias{getFitFunction.HetLogAddPlm}
\alias{HetLogAddPlm.getFitFunction}
\alias{getFitFunction.HetLogAddPlm}
\alias{getFitFunction,HetLogAddPlm-method}


\title{Gets the low-level function that fits the PLM}

\description{
 Gets the low-level function that fits the PLM.
}

\usage{\method{getFitFunction}{HetLogAddPlm}(this, ..., verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\author{
  Henrik Bengtsson and Ken Simpson (WEHI) utilizing Ben Bolstad's
  \pkg{affyPLM} package.
}

\seealso{
  For more information see \code{\link{HetLogAddPlm}}.
}
\keyword{internal}
\keyword{methods}
