%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MbeiPlm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitFunction.MbeiPlm}
\alias{getFitFunction.MbeiPlm}
\alias{MbeiPlm.getFitFunction}
\alias{getFitFunction.MbeiPlm}
\alias{getFitFunction,MbeiPlm-method}


\title{Static method to get the low-level function that fits the PLM}

\description{
 Static method to get the low-level function that fits the PLM.
}

\synopsis{getFitFunction.MbeiPlm(static, ...)}
\usage{MbeiPlm$getFitFunction(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[affy]{fit.li.wong}}.
  For more information see \code{\link{MbeiPlm}}.
}
\keyword{internal}
\keyword{methods}
