%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffinePlm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitUnitGroupFunction.AffinePlm}
\alias{getFitUnitGroupFunction.AffinePlm}
\alias{AffinePlm.getFitUnitGroupFunction}
\alias{getFitUnitGroupFunction,AffinePlm-method}

\title{Gets the low-level function that fits the PLM}

\description{
 Gets the low-level function that fits the PLM.
}

\usage{
\method{getFitUnitGroupFunction}{AffinePlm}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\author{Henrik Bengtsson, James Bullard, Kasper Hansen, Pierre Neuvial, Elizabeth Purdom, Mark Robinson, Ken Simpson}

\seealso{
  \code{\link[aroma.light]{calibrateMultiscan}}.
  For more information see \code{\link{AffinePlm}}.
}
\keyword{internal}
\keyword{methods}
