%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FirmaModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitUnitGroupFunction.FirmaModel}
\alias{getFitUnitGroupFunction.FirmaModel}
\alias{FirmaModel.getFitUnitGroupFunction}
\alias{getFitUnitGroupFunction.FirmaModel}
\alias{getFitUnitGroupFunction,FirmaModel-method}


\title{Static method to get the low-level function that fits the PLM}

\description{
 Static method to get the low-level function that fits the PLM.
}

\usage{\method{getFitUnitGroupFunction}{FirmaModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{FirmaModel}}.
}
\keyword{internal}
\keyword{methods}
