%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RmaPlm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitUnitGroupFunction.RmaPlm}
\alias{getFitUnitGroupFunction.RmaPlm}
\alias{RmaPlm.getFitUnitGroupFunction}
\alias{getFitUnitGroupFunction.RmaPlm}
\alias{getFitUnitGroupFunction,RmaPlm-method}


\title{Gets the low-level function that fits the PLM}

\description{
 Gets the low-level function that fits the PLM.
}

\usage{\method{getFitUnitGroupFunction}{RmaPlm}(this, ..., verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\author{
  Henrik Bengtsson and Ken Simpson (WEHI) utilizing Ben Bolstad's
  \pkg{preprocessCore} package.
}

\seealso{
  For more information see \code{\link{RmaPlm}}.
}
\keyword{internal}
\keyword{methods}
