%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Model.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFullName.Model}
\alias{getFullName.Model}
\alias{Model.getFullName}
\alias{getFullName,Model-method}

\title{Gets the full name of the output set}

\description{
 Gets the full name of the output set, which consists of the name with appended
 comma-separated tags.
}

\usage{
\method{getFullName}{Model}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\seealso{
  For more information see \code{\link{Model}}.
}
\keyword{internal}
\keyword{methods}
