%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Preprocessing.R
% 
% on Sat Jan 06 01:55:07 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFullName.Preprocessing}
\alias{getFullName.Preprocessing}
\alias{Preprocessing.getFullName}
\alias{getFullName.Preprocessing}
\alias{getFullName,Preprocessing-method}


\title{Gets the full name of the output data set}

\description{
 Gets the full name of the output data set, which is the name with comma separated tags.
}

\usage{\method{getFullName}{Preprocessing}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Preprocessing}}.
}
\keyword{internal}
\keyword{methods}
