%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitGroupsModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFullName.UnitGroupsModel}
\alias{getFullName.UnitGroupsModel}
\alias{UnitGroupsModel.getFullName}
\alias{getFullName.UnitGroupsModel}
\alias{getFullName,UnitGroupsModel-method}


\title{Gets the full name of the output data set}

\description{
 Gets the full name of the output data set, which is the name with comma separated tags.
}

\usage{\method{getFullName}{UnitGroupsModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{UnitGroupsModel}}.
}
\keyword{internal}
\keyword{methods}
