%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitModel.R
% 
% on Sat Jan 06 01:58:47 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFullName.UnitModel}
\alias{getFullName.UnitModel}
\alias{UnitModel.getFullName}
\alias{getFullName.UnitModel}
\alias{getFullName,UnitModel-method}


\title{Gets the full name of the output data set}

\description{
 Gets the full name of the output data set, which is the name with comma separated tags.
}

\usage{\method{getFullName}{UnitModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{UnitModel}}.
}
\keyword{internal}
\keyword{methods}
