%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCdfFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getHeader.AffymetrixCdfFile}
\alias{getHeader.AffymetrixCdfFile}
\alias{AffymetrixCdfFile.getHeader}
\alias{getHeader.AffymetrixCdfFile}
\alias{getHeader,AffymetrixCdfFile-method}


\title{Gets the header of the CDF file}

\description{
 Gets the header of the CDF file.
}

\usage{\method{getHeader}{AffymetrixCdfFile}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} structure as returned by \code{\link[affxparser]{readCdfHeader}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixCdfFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
