%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixPgfFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getHeader.AffymetrixPgfFile}
\alias{getHeader.AffymetrixPgfFile}
\alias{AffymetrixPgfFile.getHeader}
\alias{getHeader,AffymetrixPgfFile-method}

\title{Gets the header of the PGF file}

\description{
 Gets the header of the PGF file.
}

\usage{
\method{getHeader}{AffymetrixPgfFile}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} structure as returned by \code{\link[affxparser]{readPgfHeader}}.
}

\seealso{
  For more information see \code{\link{AffymetrixPgfFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
