%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaChipTypeAnnotationFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getHeader.AromaChipTypeAnnotationFile}
\alias{getHeader.AromaChipTypeAnnotationFile}
\alias{AromaChipTypeAnnotationFile.getHeader}
\alias{getHeader.AromaChipTypeAnnotationFile}
\alias{getHeader,AromaChipTypeAnnotationFile-method}


\title{Gets the header of the annotation file}

\description{
 Gets the header of the annotation file.
}

\usage{\method{getHeader}{AromaChipTypeAnnotationFile}(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} structure.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AromaChipTypeAnnotationFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
