%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CnagCfhFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getHeader.CnagCfhFile}
\alias{getHeader.CnagCfhFile}
\alias{CnagCfhFile.getHeader}
\alias{getHeader,CnagCfhFile-method}

\title{Gets the header of the CEL file}

\description{
 Gets the header of the CEL file.
}

\usage{
\method{getHeader}{CnagCfhFile}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} structure as returned by \code{\link[affxparser]{readCelHeader}}.
}

\seealso{
  For more information see \code{\link{CnagCfhFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
