%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AllelicCrosstalkCalibrator.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInputDataSet.AllelicCrosstalkCalibrator}
\alias{getInputDataSet.AllelicCrosstalkCalibrator}
\alias{AllelicCrosstalkCalibrator.getInputDataSet}
\alias{getInputDataSet.AllelicCrosstalkCalibrator}
\alias{getInputDataSet,AllelicCrosstalkCalibrator-method}


\title{Gets the source data set}

\description{
 Gets the source data set that is to be normalized (or has been normalized).
}

\usage{\method{getInputDataSet}{AllelicCrosstalkCalibrator}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCelSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AllelicCrosstalkCalibrator}}.
}
\keyword{internal}
\keyword{methods}
