%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  QuantileNormalizer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInputDataSet.QuantileNormalizer}
\alias{getInputDataSet.QuantileNormalizer}
\alias{QuantileNormalizer.getInputDataSet}
\alias{getInputDataSet.QuantileNormalizer}
\alias{getInputDataSet,QuantileNormalizer-method}


\title{Gets the source data set}

\description{
 Gets the source data set that is to be normalized (or has been normalized).
}

\usage{\method{getInputDataSet}{QuantileNormalizer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCelSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{QuantileNormalizer}}.
}
\keyword{internal}
\keyword{methods}
