%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Transform.R
% 
% on Sat May 17 11:56:55 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInputDataSet.Transform}
\alias{getInputDataSet.Transform}
\alias{Transform.getInputDataSet}
\alias{getInputDataSet.Transform}
\alias{getInputDataSet,Transform-method}


\title{Gets the source data set}

\description{
 Gets the source data set that is to be (or has been) transformed.
}

\usage{\method{getInputDataSet}{Transform}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCelSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Transform}}.
}
\keyword{internal}
\keyword{methods}
