%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TransformReport.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInputDataSet.TransformReport}
\alias{getInputDataSet.TransformReport}
\alias{TransformReport.getInputDataSet}
\alias{getInputDataSet,TransformReport-method}

\title{Gets the source data set}

\description{
 Gets the source data set that is to be (or has been) transformed.
}

\usage{
\method{getInputDataSet}{TransformReport}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCelSet}}.
}

\seealso{
  For more information see \code{\link{TransformReport}}.
}
\keyword{internal}
\keyword{methods}
