%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataFile.MAP.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{getMapType.AffymetrixDataFile}
\alias{getMapType.AffymetrixDataFile}
\alias{AffymetrixDataFile.getMapType}
\alias{getMapType.AffymetrixDataFile}
\alias{getMapType,AffymetrixDataFile-method}


\title{Gets the map type of the data file}

\description{
  Gets the map type of the data file.
}

\usage{\method{getMapType}{AffymetrixDataFile}(...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataFile}}.
}
\keyword{internal}
\keyword{methods}
