%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomeExplorer.R
% 
% on Thu Jun 05 14:20:34 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getModel.ChromosomeExplorer}
\alias{getModel.ChromosomeExplorer}
\alias{ChromosomeExplorer.getModel}
\alias{getModel.ChromosomeExplorer}
\alias{getModel,ChromosomeExplorer-method}


\title{Gets the model}

\description{
 Gets the model for which the explorer is displaying it results.
}

\usage{\method{getModel}{ChromosomeExplorer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{CopyNumberChromosomalModel}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomeExplorer}}.
}
\keyword{internal}
\keyword{methods}
