%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenomeAnalyzer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getModel.GenomeAnalyzer}
\alias{getModel.GenomeAnalyzer}
\alias{GenomeAnalyzer.getModel}
\alias{getModel.GenomeAnalyzer}
\alias{getModel,GenomeAnalyzer-method}


\title{Gets the model used by this analyzer}

\description{
 Gets the model used by this analyzer.
}

\usage{\method{getModel}{GenomeAnalyzer}(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a model.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenomeAnalyzer}}.
}
\keyword{internal}
\keyword{methods}
