%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSetReporter.R
% 
% on Tue Nov 20 12:39:31 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.AffymetrixFileSetReporter}
\alias{getName.AffymetrixFileSetReporter}
\alias{AffymetrixFileSetReporter.getName}
\alias{getName.AffymetrixFileSetReporter}
\alias{getName,AffymetrixFileSetReporter-method}


\title{Gets the name of the explorer}

\description{
 Gets the name of the explorer, which is the same as the name of the data set.
}

\usage{\method{getName}{AffymetrixFileSetReporter}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\details{
 If a name alias has not been set explicitly, the name of the data set will
 used.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFileSetReporter}}.
}
\keyword{internal}
\keyword{methods}
