%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ArrayExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.ArrayExplorer}
\alias{getName.ArrayExplorer}
\alias{ArrayExplorer.getName}
\alias{getName.ArrayExplorer}
\alias{getName,ArrayExplorer-method}


\title{Gets the name of the explorer}

\description{
 Gets the name of the explorer, which is the same as the name of the data set.
}

\usage{\method{getName}{ArrayExplorer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\details{
 If a name alias has not been set explicitly, the name of the data set will
 used.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ArrayExplorer}}.
}
\keyword{internal}
\keyword{methods}
