%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomeExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.ChromosomeExplorer}
\alias{getName.ChromosomeExplorer}
\alias{ChromosomeExplorer.getName}
\alias{getName.ChromosomeExplorer}
\alias{getName,ChromosomeExplorer-method}


\title{Gets the name of the explorer}

\description{
 Gets the name of the explorer, which is the same as the name of the model.
}

\usage{\method{getName}{ChromosomeExplorer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomeExplorer}}.
}
\keyword{internal}
\keyword{methods}
