%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericReporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.GenericReporter}
\alias{getName.GenericReporter}
\alias{GenericReporter.getName}
\alias{getName,GenericReporter-method}

\title{Gets the name of the explorer}

\description{
 Gets the name of the explorer.
}

\usage{
\method{getName}{GenericReporter}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\details{
 If a name alias has not been set explicitly, the input name will be used.
}

\seealso{
  For more information see \code{\link{GenericReporter}}.
}
\keyword{internal}
\keyword{methods}
