%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Model.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.Model}
\alias{getName.Model}
\alias{Model.getName}
\alias{getName.Model}
\alias{getName,Model-method}


\title{Gets the name of the output data set}

\description{
 Gets the name of the output data set, which is the same as the name of the input data set, unless
 an alias is set.
}

\usage{\method{getName}{Model}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getAlias.Model]{*getAlias}()}
  For more information see \code{\link{Model}}.
}
\keyword{internal}
\keyword{methods}
