%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParameterSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.ParameterSet}
\alias{getName.ParameterSet}
\alias{ParameterSet.getName}
\alias{getName.ParameterSet}
\alias{getName,ParameterSet-method}


\title{Gets the name of this parameter set}

\description{
  Gets the name of this parameter set.
}

\usage{\method{getName}{ParameterSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ParameterSet}}.
}
\keyword{internal}
\keyword{methods}
