%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Preprocessing.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.Preprocessing}
\alias{getName.Preprocessing}
\alias{Preprocessing.getName}
\alias{getName.Preprocessing}
\alias{getName,Preprocessing-method}


\title{Gets the name of the output data set}

\description{
 Gets the name of the output data set, which is the same as the input data set.
}

\usage{\method{getName}{Preprocessing}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Preprocessing}}.
}
\keyword{internal}
\keyword{methods}
