%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  QuantileNormalizer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.QuantileNormalizer}
\alias{getName.QuantileNormalizer}
\alias{QuantileNormalizer.getName}
\alias{getName.QuantileNormalizer}
\alias{getName,QuantileNormalizer-method}


\title{Gets the name of the output data set}

\description{
 Gets the name of the output data set, which is the same as the input data set.
}

\usage{\method{getName}{QuantileNormalizer}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{QuantileNormalizer}}.
}
\keyword{internal}
\keyword{methods}
