%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitGroupsModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.UnitGroupsModel}
\alias{getName.UnitGroupsModel}
\alias{UnitGroupsModel.getName}
\alias{getName.UnitGroupsModel}
\alias{getName,UnitGroupsModel-method}


\title{Gets the name of the output summarized data set}

\description{
 Gets the name of the output summarized data set, which is the same as the input data set.
}

\usage{\method{getName}{UnitGroupsModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{UnitGroupsModel}}.
}
\keyword{internal}
\keyword{methods}
