%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitModel.R
% 
% on Sat Jan 06 01:58:47 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.UnitModel}
\alias{getName.UnitModel}
\alias{UnitModel.getName}
\alias{getName.UnitModel}
\alias{getName,UnitModel-method}


\title{Gets the name of the output summarized data set}

\description{
 Gets the name of the output summarized data set, which is the same as the input data set.
}

\usage{\method{getName}{UnitModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{UnitModel}}.
}
\keyword{internal}
\keyword{methods}
