%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AllelicCrosstalkCalibrator.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getOutputDataSet.AllelicCrosstalkCalibrator}
\alias{getOutputDataSet.AllelicCrosstalkCalibrator}
\alias{AllelicCrosstalkCalibrator.getOutputDataSet}
\alias{getOutputDataSet.AllelicCrosstalkCalibrator}
\alias{getOutputDataSet,AllelicCrosstalkCalibrator-method}


\title{Gets the normalized data set}

\description{
 Gets the normalized data set, if normalized.
}

\usage{\method{getOutputDataSet}{AllelicCrosstalkCalibrator}(this, ..., force=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{force}{If \code{\link[base:logical]{TRUE}}, any in-memory cached results are ignored.}
}

\value{
 Returns an \code{\link{AffymetrixCelSet}} or \code{\link[base]{NULL}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AllelicCrosstalkCalibrator}}.
}
\keyword{internal}
\keyword{methods}
