%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Preprocessing.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getOutputDataSet.Preprocessing}
\alias{getOutputDataSet.Preprocessing}
\alias{Preprocessing.getOutputDataSet}
\alias{getOutputDataSet.Preprocessing}
\alias{getOutputDataSet,Preprocessing-method}


\title{Gets the preprocessed data set}

\description{
 Gets the preprocessed data set, if processed.
}

\usage{\method{getOutputDataSet}{Preprocessing}(this, ..., force=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{force}{If \code{\link[base:logical]{TRUE}}, any in-memory cached results are ignored.}
}

\value{
 Returns an \code{\link{AffymetrixCelSet}} or \code{\link[base]{NULL}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Preprocessing}}.
}
\keyword{internal}
\keyword{methods}
