%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TransformReport.R
% 
% on Thu Sep 04 00:31:12 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getOutputDataSet.TransformReport}
\alias{getOutputDataSet.TransformReport}
\alias{TransformReport.getOutputDataSet}
\alias{getOutputDataSet.TransformReport}
\alias{getOutputDataSet,TransformReport-method}


\title{Gets the transformed data set}

\description{
 Gets the transformed data set, if processed.
}

\usage{\method{getOutputDataSet}{TransformReport}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AffymetrixCelSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{TransformReport}}.
}
\keyword{internal}
\keyword{methods}
