%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.AffymetrixFile}
\alias{getPath.AffymetrixFile}
\alias{AffymetrixFile.getPath}
\alias{getPath.AffymetrixFile}
\alias{getPath,AffymetrixFile-method}


\title{Gets the path (directory) of the file}

\description{
  Gets the path (directory) of the file.
}

\usage{\method{getPath}{AffymetrixFile}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{AffymetrixFile}}.
}
\keyword{internal}
\keyword{methods}
