%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixUnitGroupsModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.AffymetrixUnitGroupsModel}
\alias{getPath.AffymetrixUnitGroupsModel}
\alias{AffymetrixUnitGroupsModel.getPath}
\alias{getPath.AffymetrixUnitGroupsModel}
\alias{getPath,AffymetrixUnitGroupsModel-method}


\title{Gets the path (directory) for this model}

\description{
  Gets the path (directory) for this model.
}

\usage{\method{getPath}{AffymetrixUnitGroupsModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixUnitGroupsModel}}.
}
\keyword{internal}
\keyword{methods}
