%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AllelicCrosstalkCalibrator.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.AllelicCrosstalkCalibrator}
\alias{getPath.AllelicCrosstalkCalibrator}
\alias{AllelicCrosstalkCalibrator.getPath}
\alias{getPath.AllelicCrosstalkCalibrator}
\alias{getPath,AllelicCrosstalkCalibrator-method}


\title{Gets the path of the output data set}

\description{
 Gets the path of the output data set.
 If non-existing, then the directory is created.
}

\usage{\method{getPath}{AllelicCrosstalkCalibrator}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AllelicCrosstalkCalibrator}}.
}
\keyword{internal}
\keyword{methods}
