%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Crlmm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.Crlmm}
\alias{getPath.Crlmm}
\alias{Crlmm.getPath}
\alias{getPath.Crlmm}
\alias{getPath,Crlmm-method}


\title{Gets the path of this model}

\description{
 Gets the path of this model where the parameter files are stored.
 If non-existing, then the directory is created.
}

\usage{\method{getPath}{Crlmm}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Crlmm}}.
}
\keyword{internal}
\keyword{methods}
