%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Preprocessing.R
% 
% on Sat Jan 06 01:55:07 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.Preprocessing}
\alias{getPath.Preprocessing}
\alias{Preprocessing.getPath}
\alias{getPath.Preprocessing}
\alias{getPath,Preprocessing-method}


\title{Gets the path of the output data set}

\description{
 Gets the path of the output data set.
 If non-existing, then the directory is created.
}

\usage{\method{getPath}{Preprocessing}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Preprocessing}}.
}
\keyword{internal}
\keyword{methods}
