%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TransformReport.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.TransformReport}
\alias{getPath.TransformReport}
\alias{TransformReport.getPath}
\alias{getPath,TransformReport-method}

\title{Gets the path of the output data set}

\description{
 Gets the path of the output data set.
 If non-existing, then the directory is created.
}

\usage{
\method{getPath}{TransformReport}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\seealso{
  For more information see \code{\link{TransformReport}}.
}
\keyword{internal}
\keyword{methods}
