%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitModel.R
% 
% on Sat Jan 06 01:58:47 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.UnitModel}
\alias{getPath.UnitModel}
\alias{UnitModel.getPath}
\alias{getPath.UnitModel}
\alias{getPath,UnitModel-method}


\title{Gets the path of this model}

\description{
 Gets the path of this model where the parameter files are stored.
 If non-existing, then the directory is created.
}

\usage{\method{getPath}{UnitModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{UnitModel}}.
}
\keyword{internal}
\keyword{methods}
