%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPathname.AffymetrixFile}
\alias{getPathname.AffymetrixFile}
\alias{AffymetrixFile.getPathname}
\alias{getPathname.AffymetrixFile}
\alias{getPathname,AffymetrixFile-method}


\title{Gets the pathname of the file}

\description{
  Gets the pathname of the file.
}

\usage{\method{getPathname}{AffymetrixFile}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{AffymetrixFile}}.
}
\keyword{internal}
\keyword{methods}
