%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSet.R
% 
% on Sat Jan 06 01:47:52 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPathnames.AffymetrixFileSet}
\alias{getPathnames.AffymetrixFileSet}
\alias{AffymetrixFileSet.getPathnames}
\alias{getPathnames.AffymetrixFileSet}
\alias{getPathnames,AffymetrixFileSet-method}


\title{Gets the pathnames of the files in the file set}

\description{
  Gets the pathnames of the files in the file set.
}

\usage{\method{getPathnames}{AffymetrixFileSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFileSet}}.
}
\keyword{internal}
\keyword{methods}
