%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MbeiPlm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getProbeAffinityClass.MbeiPlm}
\alias{getProbeAffinityClass.MbeiPlm}
\alias{MbeiPlm.getProbeAffinityClass}
\alias{getProbeAffinityClass.MbeiPlm}
\alias{getProbeAffinityClass,MbeiPlm-method}


\title{Static method to get the ProbeAffinityFile Class object}

\description{
 Static method to get the ProbeAffinityFile Class object.
}

\synopsis{getProbeAffinityClass.MbeiPlm(static, ...)}
\usage{MbeiPlm$getProbeAffinityClass(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{Class}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MbeiPlm}}.
}
\keyword{internal}
\keyword{methods}
