%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ProbeLevelModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getProbeAffinityClass.ProbeLevelModel}
\alias{getProbeAffinityClass.ProbeLevelModel}
\alias{ProbeLevelModel.getProbeAffinityClass}
\alias{getProbeAffinityClass.ProbeLevelModel}
\alias{getProbeAffinityClass,ProbeLevelModel-method}


\title{Static method to get the ProbeAffinity Class object for this model}

\description{
 Static method to get the ProbeAffinity Class object for this model.
 Any subclass model must provide this method, which should return
 a \code{\link[R.oo]{Class}} that subsets the \code{\link{ProbeAffinityFile}} class.
}

\synopsis{getProbeAffinityClass.ProbeLevelModel(...)}
\usage{ProbeLevelModel$getProbeAffinityClass()}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{Class}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ProbeLevelModel}}.
}
\keyword{internal}
\keyword{methods}
