%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ProbeLevelModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getProbeAffinityFile.ProbeLevelModel}
\alias{getProbeAffinityFile.ProbeLevelModel}
\alias{ProbeLevelModel.getProbeAffinityFile}
\alias{getProbeAffinityFile,ProbeLevelModel-method}
\alias{ProbeLevelModel.getProbeAffinities}
\alias{getProbeAffinities.ProbeLevelModel}
\alias{getProbeAffinities,ProbeLevelModel-method}


\title{Gets the probe affinities for this model}

\description{
 Gets the probe affinities for this model.
}

\usage{
\method{getProbeAffinityFile}{ProbeLevelModel}(this, ..., .class=ProbeAffinityFile)
}

\arguments{
  \item{...}{Not used.}
  \item{.class}{A \code{\link{ProbeAffinityFile}} \emph{class}.}
}

\value{
 Returns a \code{\link{ProbeAffinityFile}} object.
}

\seealso{
  For more information see \code{\link{ProbeLevelModel}}.
}
\keyword{internal}
\keyword{methods}
