%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenomeAnalyzer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getProgress.GenomeAnalyzer}
\alias{getProgress.GenomeAnalyzer}
\alias{GenomeAnalyzer.getProgress}
\alias{getProgress.GenomeAnalyzer}
\alias{getProgress,GenomeAnalyzer-method}


\title{Gets the fraction of units process per chromosome}

\description{
 Gets the fraction of units process per chromosome.
}

\usage{\method{getProgress}{GenomeAnalyzer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenomeAnalyzer}}.
}
\keyword{internal}
\keyword{methods}
