%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.MAP.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{getReadMap.AffymetrixDataSet}
\alias{getReadMap.AffymetrixDataSet}
\alias{AffymetrixDataSet.getReadMap}
\alias{getReadMap.AffymetrixDataSet}
\alias{getReadMap,AffymetrixDataSet-method}


\title{Gets the read map for the data set}

\description{
  Gets the read map for the data set.
}

\usage{\method{getReadMap}{AffymetrixDataSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}} \code{\link[base]{vector}}, if a read map exists, otherwise \code{\link[base]{NULL}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:setApdMap.AffymetrixDataSet]{*setApdMap}()} and \code{\link[aroma.affymetrix:hasApdMap.AffymetrixDataSet]{*hasApdMap}()}.
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
