%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Crlmm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRootPath.Crlmm}
\alias{getRootPath.Crlmm}
\alias{Crlmm.getRootPath}
\alias{getRootPath.Crlmm}
\alias{getRootPath,Crlmm-method}


\title{Gets the root path of this model}

\description{
 Gets the root path of this model, which is the class name of the model preceeded by the
 string \code{"model"}.
}

\usage{\method{getRootPath}{Crlmm}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Crlmm}}.
}
\keyword{internal}
\keyword{methods}
