%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Model.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRootPath.Model}
\alias{getRootPath.Model}
\alias{Model.getRootPath}
\alias{getRootPath,Model-method}

\title{Gets the root path of this model}

\description{
 Gets the root path of this model.
 By default, this is the string \code{"model"} appended by the capitalized
 name of the model class, e.g. \code{"modelUnit"}.
}

\usage{
\method{getRootPath}{Model}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\seealso{
  For more information see \code{\link{Model}}.
}
\keyword{internal}
\keyword{methods}
