%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSampleNames.AffymetrixCelSet}
\alias{getSampleNames.AffymetrixCelSet}
\alias{AffymetrixCelSet.getSampleNames}
\alias{getSampleNames.AffymetrixCelSet}
\alias{getSampleNames,AffymetrixCelSet-method}


\title{Gets the names of the samples in the file set}

\description{
  Gets the names of the samples in the file set.
}

\usage{\method{getSampleNames}{AffymetrixCelSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{AffymetrixCelSet}}.
}
\keyword{internal}
\keyword{methods}
