%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSetReporter.R
% 
% on Tue Nov 20 12:39:31 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTags.AffymetrixFileSetReporter}
\alias{getTags.AffymetrixFileSetReporter}
\alias{AffymetrixFileSetReporter.getTags}
\alias{getTags.AffymetrixFileSetReporter}
\alias{getTags,AffymetrixFileSetReporter-method}


\title{Gets the tags of the explorer}

\description{
 Gets the tags of the explorer, which are the tags of the data set plus additional tags.
}

\usage{\method{getTags}{AffymetrixFileSetReporter}(this, collapse=NULL, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFileSetReporter}}.
}
\keyword{internal}
\keyword{methods}
