%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AllelicCrosstalkCalibrator.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTags.AllelicCrosstalkCalibrator}
\alias{getTags.AllelicCrosstalkCalibrator}
\alias{AllelicCrosstalkCalibrator.getTags}
\alias{getTags.AllelicCrosstalkCalibrator}
\alias{getTags,AllelicCrosstalkCalibrator-method}


\title{Gets the tags of the output data set}

\description{
 Gets the tags of the output data set, which equals the tags of the input data set plus the tags
 of this normalizer.
}

\usage{\method{getTags}{AllelicCrosstalkCalibrator}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AllelicCrosstalkCalibrator}}.
}
\keyword{internal}
\keyword{methods}
