%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ArrayExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTags.ArrayExplorer}
\alias{getTags.ArrayExplorer}
\alias{ArrayExplorer.getTags}
\alias{getTags.ArrayExplorer}
\alias{getTags,ArrayExplorer-method}


\title{Gets the tags of the explorer}

\description{
 Gets the tags of the explorer, which are the tags of the data set plus additional tags.
}

\usage{\method{getTags}{ArrayExplorer}(this, collapse=NULL, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ArrayExplorer}}.
}
\keyword{internal}
\keyword{methods}
