%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomeExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTags.ChromosomeExplorer}
\alias{getTags.ChromosomeExplorer}
\alias{ChromosomeExplorer.getTags}
\alias{getTags.ChromosomeExplorer}
\alias{getTags,ChromosomeExplorer-method}


\title{Gets the tags of the explorer}

\description{
 Gets the tags of the explorer, which are the tags of the model plus additional tags.
}

\usage{\method{getTags}{ChromosomeExplorer}(this, collapse=NULL, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomeExplorer}}.
}
\keyword{internal}
\keyword{methods}
