%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericReporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTags.GenericReporter}
\alias{getTags.GenericReporter}
\alias{GenericReporter.getTags}
\alias{getTags,GenericReporter-method}

\title{Gets the tags of the reporter}

\description{
 Gets the tags of the reporter, which are the input tags plus additional tags.
}

\usage{
\method{getTags}{GenericReporter}(this, collapse=NULL, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\seealso{
  For more information see \code{\link{GenericReporter}}.
}
\keyword{internal}
\keyword{methods}
