%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Preprocessing.R
% 
% on Sat Jan 06 01:55:06 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTags.Preprocessing}
\alias{getTags.Preprocessing}
\alias{Preprocessing.getTags}
\alias{getTags.Preprocessing}
\alias{getTags,Preprocessing-method}


\title{Gets the tags of the output data set}

\description{
 Gets the tags of the output data set, which equals the tags of the input data set plus the tags
 of this preprocessor.
}

\usage{\method{getTags}{Preprocessing}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Preprocessing}}.
}
\keyword{internal}
\keyword{methods}
