%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Transform.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTags.Transform}
\alias{getTags.Transform}
\alias{Transform.getTags}
\alias{getTags.Transform}
\alias{getTags,Transform-method}


\title{Gets the tags of the output data set}

\description{
 Gets the tags of the output data set, which equals the tags of the input data set plus the tags
 of this transformation.
}

\usage{\method{getTags}{Transform}(this, collapse=NULL, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Transform}}.
}
\keyword{internal}
\keyword{methods}
