%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TransformReport.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTags.TransformReport}
\alias{getTags.TransformReport}
\alias{TransformReport.getTags}
\alias{getTags,TransformReport-method}

\title{Gets the tags of the output data set}

\description{
 Gets the tags of the output data set, which equals the tags of the input data set plus the tags
 of this transformation.
}

\usage{
\method{getTags}{TransformReport}(this, collapse=NULL, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\seealso{
  For more information see \code{\link{TransformReport}}.
}
\keyword{internal}
\keyword{methods}
