%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSetTuple.R
% 
% on Tue Nov 20 12:39:20 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTuple.AffymetrixCelSetTuple}
\alias{getTuple.AffymetrixCelSetTuple}
\alias{AffymetrixCelSetTuple.getTuple}
\alias{getTuple.AffymetrixCelSetTuple}
\alias{getTuple,AffymetrixCelSetTuple-method}


\title{Gets arrays across chip types for one sample}

\description{
 Gets arrays across chip types for one sample.
}

\usage{\method{getTuple}{AffymetrixCelSetTuple}(this, array, ...)}

\arguments{
  \item{array}{Sample of interest.}
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}} of \code{\link{AffymetrixCelSet}}:s.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixCelSetTuple}}.
}
\keyword{internal}
\keyword{methods}
