%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenomeAnalyzer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getUnitsTodo.GenomeAnalyzer}
\alias{getUnitsTodo.GenomeAnalyzer}
\alias{GenomeAnalyzer.getUnitsTodo}
\alias{getUnitsTodo.GenomeAnalyzer}
\alias{getUnitsTodo,GenomeAnalyzer-method}


\title{Gets information about units per chromosome}

\description{
 Gets information about units per chromosome.
}

\usage{\method{getUnitsTodo}{GenomeAnalyzer}(this, chromosomes=NULL, ..., force=FALSE, verbose=FALSE)}

\arguments{
  \item{chromosomes}{The chromosomes of interest.}
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenomeAnalyzer}}.
}
\keyword{internal}
\keyword{methods}
