%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  QualityAssessmentModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getWeights.QualityAssessmentModel}
\alias{getWeights.QualityAssessmentModel}
\alias{QualityAssessmentModel.getWeights}
\alias{getWeights,QualityAssessmentModel-method}

\title{Calculates the weights from the robust fit to a probe-level model}

\description{
  Calculates the weights from the robust fit to a probe-level model.
}

\usage{
\method{getWeights}{QualityAssessmentModel}(this, path=NULL, name="qcData", tags="*", ..., ram=NULL, force=FALSE,
  verbose=FALSE)
}

\arguments{
 \item{path}{}
 \item{name}{}
 \item{tags}{}
 \item{...}{}
 \item{ram}{}
 \item{force}{}
 \item{verbose}{}
}

\value{
  Returns an \code{\link{QualityAssessmentSet}}.
}

\seealso{
  For more information see \code{\link{QualityAssessmentModel}}.
}
\keyword{internal}
\keyword{methods}
