%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataFile.MAP.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasApdMap.AffymetrixDataFile}
\alias{hasApdMap.AffymetrixDataFile}
\alias{AffymetrixDataFile.hasApdMap}
\alias{hasApdMap.AffymetrixDataFile}
\alias{hasApdMap,AffymetrixDataFile-method}


\title{Checks if a data file has a probe map}

\description{
  Checks if a data file has a probe map.
}

\usage{\method{hasApdMap}{AffymetrixDataFile}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns \code{\link[base:logical]{TRUE}} if a read map exists, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getReadMap.AffymetrixDataFile]{*getReadMap}()}.
  \code{\link[aroma.affymetrix:setApdMap.AffymetrixDataFile]{*setApdMap}()}.
  For more information see \code{\link{AffymetrixDataFile}}.
}
\keyword{internal}
\keyword{methods}
