%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.MAP.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasApdMap.AffymetrixDataSet}
\alias{hasApdMap.AffymetrixDataSet}
\alias{AffymetrixDataSet.hasApdMap}
\alias{hasApdMap.AffymetrixDataSet}
\alias{hasApdMap,AffymetrixDataSet-method}


\title{Checks if the data set has a read map}

\description{
  Checks if the data set has a read map.
}

\usage{\method{hasApdMap}{AffymetrixDataSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns \code{\link[base:logical]{TRUE}} if a read map exists, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getReadMap.AffymetrixDataSet]{*getReadMap}()} and \code{\link[aroma.affymetrix:setApdMap.AffymetrixDataSet]{*setApdMap}()}.
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
