%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.TRANSFORM.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasTransforms.AffymetrixDataSet}
\alias{hasTransforms.AffymetrixDataSet}
\alias{AffymetrixDataSet.hasTransforms}
\alias{hasTransforms.AffymetrixDataSet}
\alias{hasTransforms,AffymetrixDataSet-method}


\title{Checks if any of the data files has transform functions specified}

\description{
  Checks if any of the data files has transform functions specified.
}

\usage{\method{hasTransforms}{AffymetrixDataSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns \code{\link[base:logical]{TRUE}} or \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
