%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SnpChipEffectSet.importFromAPT.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{importFromAPT.SnpChipEffectSet}
\alias{importFromAPT.SnpChipEffectSet}
\alias{SnpChipEffectSet.importFromAPT}
\alias{importFromAPT.SnpChipEffectSet}
\alias{importFromAPT,SnpChipEffectSet-method}


\title{Imports probeset summaries from an Affymetrix Power Tool (APT) result file}

\description{
 Imports probeset summaries from an Affymetrix Power Tool (APT) result file named *.summary.txt and generated by the
 gtype-probeset-genotype command line tool.
}

\synopsis{importFromAPT.SnpChipEffectSet(static, filename, path=NULL, ..., skip=TRUE, verbose=FALSE)}
\usage{SnpChipEffectSet$importFromAPT(filename, path=NULL, ..., skip=TRUE, verbose=FALSE)}

\arguments{
  \item{filename}{The filename of the dChip result file.}
  \item{path}{An optional path to the file.}
  \item{...}{Not used.}
  \item{skip}{If \code{\link[base:logical]{TRUE}}, already imported arrays will be skipped.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns an \code{\link{SnpChipEffectSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SnpChipEffectSet}}.
}
\keyword{internal}
\keyword{methods}
