%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomalModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{indexOfArrays.ChromosomalModel}
\alias{indexOfArrays.ChromosomalModel}
\alias{ChromosomalModel.indexOfArrays}
\alias{indexOfArrays.ChromosomalModel}
\alias{indexOfArrays,ChromosomalModel-method}


\title{Gets the indices of the arrays}

\description{
 Gets the indices of the arrays.
}

\usage{\method{indexOfArrays}{ChromosomalModel}(this, arrays=NULL, ...)}

\arguments{
  \item{arrays}{A \code{\link[base]{character}} \code{\link[base]{vector}} of arrays names.
    If \code{\link[base]{NULL}}, all arrays are considered.}
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomalModel}}.
}
\keyword{internal}
\keyword{methods}
