%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CopyNumberSegmentationModel.R
% 
% on Sat Sep 15 13:21:17 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{indexOfArrays.CopyNumberSegmentationModel}
\alias{indexOfArrays.CopyNumberSegmentationModel}
\alias{CopyNumberSegmentationModel.indexOfArrays}
\alias{indexOfArrays.CopyNumberSegmentationModel}
\alias{indexOfArrays,CopyNumberSegmentationModel-method}


\title{Gets the indices of the arrays}

\description{
 Gets the indices of the arrays.
}

\usage{\method{indexOfArrays}{CopyNumberSegmentationModel}(this, arrays=NULL, ...)}

\arguments{
  \item{arrays}{A \code{\link[base]{character}} \code{\link[base]{vector}} of arrays names.
    If \code{\link[base]{NULL}}, all arrays are considered.}
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{CopyNumberSegmentationModel}}.
}
\keyword{internal}
\keyword{methods}
