%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AllelicCrosstalkCalibrator.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isDone.AllelicCrosstalkCalibrator}
\alias{isDone.AllelicCrosstalkCalibrator}
\alias{AllelicCrosstalkCalibrator.isDone}
\alias{isDone.AllelicCrosstalkCalibrator}
\alias{isDone,AllelicCrosstalkCalibrator-method}


\title{Checks if the data set is calibrated or not}

\description{
 Checks if the data set is calibrated or not.
}

\usage{\method{isDone}{AllelicCrosstalkCalibrator}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the data set is calibrated, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AllelicCrosstalkCalibrator}}.
}
\keyword{internal}
\keyword{methods}
