%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Preprocessing.R
% 
% on Sat Jan 06 01:55:07 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isDone.Preprocessing}
\alias{isDone.Preprocessing}
\alias{Preprocessing.isDone}
\alias{isDone.Preprocessing}
\alias{isDone,Preprocessing-method}


\title{Checks if the data set is processed or not}

\description{
 Checks if the data set is processed or not.
}

\usage{\method{isDone}{Preprocessing}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the data set is processed, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Preprocessing}}.
}
\keyword{internal}
\keyword{methods}
