%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  QuantileNormalizer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isDone.QuantileNormalizer}
\alias{isDone.QuantileNormalizer}
\alias{QuantileNormalizer.isDone}
\alias{isDone.QuantileNormalizer}
\alias{isDone,QuantileNormalizer-method}


\title{Checks if the data set is normalized or not}

\description{
 Checks if the data set is normalized or not.
}

\usage{\method{isDone}{QuantileNormalizer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the data set is normalized, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{QuantileNormalizer}}.
}
\keyword{internal}
\keyword{methods}
