%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCdfFile.getCellQuartets.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{isResequenceChip.AffymetrixCdfFile}
\alias{isResequenceChip.AffymetrixCdfFile}
\alias{AffymetrixCdfFile.isResequenceChip}
\alias{isResequenceChip.AffymetrixCdfFile}
\alias{isResequenceChip,AffymetrixCdfFile-method}


\title{Static method to check if a chip is a resequence (CustomSeq) chip}

\description{
  Static method to check if a chip is a resequence (CustomSeq) chip.
}

\usage{\method{isResequenceChip}{AffymetrixCdfFile}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns \code{\link[base:logical]{TRUE}} if the chip type refers to a resequence array,
  otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixCdfFile}}.
}
\keyword{internal}
\keyword{methods}
