%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCdfFile.SNPs.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{isSnpChip.AffymetrixCdfFile}
\alias{isSnpChip.AffymetrixCdfFile}
\alias{AffymetrixCdfFile.isSnpChip}
\alias{isSnpChip,AffymetrixCdfFile-method}

\title{Static method to check if a chip is a mapping (SNP) chip}

\description{
  Static method to check if a chip is a mapping (SNP) chip.
}

\usage{
\method{isSnpChip}{AffymetrixCdfFile}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns \code{\link[base:logical]{TRUE}} if the chip type refers to a SNP array, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AffymetrixCdfFile}}.
}
\keyword{internal}
\keyword{methods}
