%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.TRANSFORM.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isTransforming.AffymetrixDataSet}
\alias{isTransforming.AffymetrixDataSet}
\alias{AffymetrixDataSet.isTransforming}
\alias{isTransforming.AffymetrixDataSet}
\alias{isTransforming,AffymetrixDataSet-method}


\title{Checks if transformation is turned on for any of the data files}

\description{
  Checks if transformation is turned on for any of the data files.
}

\usage{\method{isTransforming}{AffymetrixDataSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns \code{\link[base:logical]{TRUE}} or \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
