%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParameterSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{lapply.ParameterSet}
\alias{lapply.ParameterSet}
\alias{ParameterSet.lapply}
\alias{lapply.ParameterSet}
\alias{lapply,ParameterSet-method}


\title{Applies a function to each of the parameter files}

\description{
  Applies a function to each of the parameter files.
}

\usage{\method{lapply}{ParameterSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{list}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{ParameterSet}}.
}
\keyword{internal}
\keyword{methods}
