%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{length.AffymetrixFileSet}
\alias{length.AffymetrixFileSet}
\alias{AffymetrixFileSet.length}
\alias{length.AffymetrixFileSet}
\alias{length,AffymetrixFileSet-method}

\alias{AffymetrixFileSet.nbrOfFiles}
\alias{nbrOfFiles.AffymetrixFileSet}
\alias{nbrOfFiles,AffymetrixFileSet-method}


\title{Gets the number of files in the file set}

\description{
  Gets the number of files in the file set.
}

\usage{\method{length}{AffymetrixFileSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{AffymetrixFileSet}}.
}
\keyword{internal}
\keyword{methods}
