%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfArrays.AffymetrixCelSet}
\alias{nbrOfArrays.AffymetrixCelSet}
\alias{AffymetrixCelSet.nbrOfArrays}
\alias{nbrOfArrays.AffymetrixCelSet}
\alias{nbrOfArrays,AffymetrixCelSet-method}


\title{Gets the number of arrays in the file set}

\description{
  Gets the number of arrays in the file set.
  This is just a wrapper for \code{nbrOfFiles()}.
}

\usage{\method{nbrOfArrays}{AffymetrixCelSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixCelSet}}.
}
\keyword{internal}
\keyword{methods}
