%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSetTuple.R
% 
% on Tue Nov 20 12:39:20 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfArrays.AffymetrixCelSetTuple}
\alias{nbrOfArrays.AffymetrixCelSetTuple}
\alias{AffymetrixCelSetTuple.nbrOfArrays}
\alias{nbrOfArrays.AffymetrixCelSetTuple}
\alias{nbrOfArrays,AffymetrixCelSetTuple-method}


\title{Gets the number of arrays}

\description{
 Gets the number of arrays used in the model.
}

\usage{\method{nbrOfArrays}{AffymetrixCelSetTuple}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixCelSetTuple}}.
}
\keyword{internal}
\keyword{methods}
