%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfArrays.AffymetrixDataSet}
\alias{nbrOfArrays.AffymetrixDataSet}
\alias{AffymetrixDataSet.nbrOfArrays}
\alias{nbrOfArrays.AffymetrixDataSet}
\alias{nbrOfArrays,AffymetrixDataSet-method}


\title{Gets the number of arrays in the data set}

\description{
  Gets the number of arrays in the data set.
}

\usage{\method{nbrOfArrays}{AffymetrixDataSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
