%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ArrayExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfArrays.ArrayExplorer}
\alias{nbrOfArrays.ArrayExplorer}
\alias{ArrayExplorer.nbrOfArrays}
\alias{nbrOfArrays.ArrayExplorer}
\alias{nbrOfArrays,ArrayExplorer-method}


\title{Gets the total number of arrays}

\description{
 Gets the total number of arrays considered by the explorer.
}

\usage{\method{nbrOfArrays}{ArrayExplorer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ArrayExplorer}}.
}
\keyword{internal}
\keyword{methods}
