%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomeExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfArrays.ChromosomeExplorer}
\alias{nbrOfArrays.ChromosomeExplorer}
\alias{ChromosomeExplorer.nbrOfArrays}
\alias{nbrOfArrays.ChromosomeExplorer}
\alias{nbrOfArrays,ChromosomeExplorer-method}


\title{Gets the total number of arrays}

\description{
 Gets the total number of arrays considered by the explorer.
}

\usage{\method{nbrOfArrays}{ChromosomeExplorer}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomeExplorer}}.
}
\keyword{internal}
\keyword{methods}
