%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenotypeCallSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfArrays.GenotypeCallSet}
\alias{nbrOfArrays.GenotypeCallSet}
\alias{GenotypeCallSet.nbrOfArrays}
\alias{nbrOfArrays.GenotypeCallSet}
\alias{nbrOfArrays,GenotypeCallSet-method}


\title{Gets the number of arrays in the file set}

\description{
  Gets the number of arrays in the file set.
  This is just a wrapper for \code{nbrOfFiles()}.
}

\usage{\method{nbrOfArrays}{GenotypeCallSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{GenotypeCallSet}}.
}
\keyword{internal}
\keyword{methods}
