%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GladModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfArrays.GladModel}
\alias{nbrOfArrays.GladModel}
\alias{GladModel.nbrOfArrays}
\alias{nbrOfArrays.GladModel}
\alias{nbrOfArrays,GladModel-method}


\title{Gets the number of arrays}

\description{
 Gets the number of arrays used in the model.
}

\usage{\method{nbrOfArrays}{GladModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GladModel}}.
}
\keyword{internal}
\keyword{methods}
