%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParameterSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfArrays.ParameterSet}
\alias{nbrOfArrays.ParameterSet}
\alias{ParameterSet.nbrOfArrays}
\alias{nbrOfArrays.ParameterSet}
\alias{nbrOfArrays,ParameterSet-method}


\title{Gets the number of arrays in this parameter set}

\description{
  Gets the number of arrays in this parameter set.
}

\usage{\method{nbrOfArrays}{ParameterSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{ParameterSet}}.
}
\keyword{internal}
\keyword{methods}
