%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CopyNumberSegmentationModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfChipTypes.CopyNumberSegmentationModel}
\alias{nbrOfChipTypes.CopyNumberSegmentationModel}
\alias{CopyNumberSegmentationModel.nbrOfChipTypes}
\alias{nbrOfChipTypes.CopyNumberSegmentationModel}
\alias{nbrOfChipTypes,CopyNumberSegmentationModel-method}


\title{Gets the number of chip types}

\description{
 Gets the number of chip types used in the model.
}

\usage{\method{nbrOfChipTypes}{CopyNumberSegmentationModel}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{CopyNumberSegmentationModel}}.
}
\keyword{internal}
\keyword{methods}
