%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% The AffymetrixCelSet class
\alias{getSiblings}

% The AffymetrixDataFile class
\alias{addTransform}
\alias{addTransform.AffymetrixDataFile}
\alias{createOptimalReadMap}
\alias{createOptimalReadMap.AffymetrixDataFile}
\alias{fromFile}
\alias{getApdMap}
\alias{getApdMap.AffymetrixDataFile}
\alias{getChipType}
\alias{getChipType.AffymetrixApdFile}
\alias{getProbes}
\alias{getProbes.AffymetrixDataFile}
\alias{hasApdMap}
\alias{readUnits}

% The AffymetrixApdFile class
\alias{getFileType.AffymetrixApdFile}
\alias{readUnits.AffymetrixApdFile}

% The AffymetrixCelFile class
\alias{getFileType.AffymetrixCelFile}

% The AffymetrixDataset class
\alias{as.AffymetrixDataset}
\alias{compactApdMap}
\alias{compactApdMap.AffymetrixDataset}
\alias{createOptimalReadMap.AffymetrixDataset}
\alias{getProbes.AffymetrixDataset}
\alias{findAverageQuantile}
\alias{findAverageQuantile.AffymetrixDataset}
\alias{getApdMap.AffymetrixDataset}
\alias{getFileType}
\alias{getFileType.AffymetrixDataset}

\alias{newInstance.Object}
\alias{getSnpNames}

% The ApdMap class
\alias{as.character.ApdMap}
\alias{equals.ApdMap}
\alias{findApdMap.ApdMap}
\alias{fromCdf}
\alias{fromCdf.ApdMap}
\alias{fromMapType}
\alias{fromMapType.ApdMap}
\alias{getChipType.ApdMap}
\alias{getMapType.ApdMap}
\alias{getReadMap}
\alias{getReadMap.ApdMap}
\alias{getWriteMap}
\alias{getWriteMap.ApdMap}
\alias{read}
\alias{read.ApdMap}
\alias{write}
\alias{write.ApdMap}

% The AffymetrixModel class
\alias{as.character.AffymetrixModel}
\alias{findUnitsDone}
\alias{findUnitsDone.AffymetrixModel}
\alias{fit}
\alias{fit.AffymetrixModel}
\alias{fromAffymetrixDataset}
\alias{fromAffymetrixDataset.AffymetrixModel}
\alias{getDataset}
\alias{getDataset.AffymetrixModel}
\alias{getFit}
\alias{getFit.AffymetrixModel}
\alias{getPath.AffymetrixModel}
\alias{nbrOfArrays}
\alias{nbrOfArrays.AffymetrixModel}
\alias{nbrOfProbes.AffymetrixModel}
\alias{nbrOfProbesets}
\alias{nbrOfProbesets.AffymetrixModel}

\alias{colMeans}
\alias{colSums}
\alias{doTransform}
\alias{findCdf}
\alias{findCdf.default}
\alias{fitQuantileNormFcn}
\alias{fromFiles}
\alias{getFilename}
\alias{getMapType}
\alias{getMapType.AffymetrixApdFile}
\alias{getMapType.AffymetrixCelFile}
\alias{getNames}
\alias{getProbeIndices}
\alias{getProbeIntensities}
\alias{getProbesetIntensities}
\alias{getProbesetNames}
\alias{getProbesets}
\alias{getProbesets.AffymetrixDataFile}
\alias{hasReadMap}
\alias{hasTransforms}
\alias{hasTransforms.AffymetrixDataFile}
\alias{isTransforming}
\alias{isTransforming.AffymetrixDataFile}
\alias{lapply}
\alias{lapply.default}
\alias{nbrOfProbes}
\alias{nbrOfProbes.AffymetrixApdFile}
\alias{nbrOfProbes.AffymetrixCelFile}
\alias{nbrOfSnps}
\alias{readCdfUnitsWriteMap}
\alias{readCdfUnitsWriteMap.default}
\alias{readIntensities}
\alias{readIntensities.AffymetrixApdFile}
\alias{readIntensities.AffymetrixCelFile}
\alias{setApdMap}
\alias{setTransform}
\alias{setTransform.AffymetrixDataFile}
\alias{setTransformStatus}
\alias{setTransformStatus.AffymetrixDataFile}
\alias{stringTree}
\alias{stringTree.character}
\alias{transformOff}
\alias{transformOff.AffymetrixDataFile}
\alias{transformOn}
\alias{transformOn.AffymetrixDataFile}
\alias{transformProbeSignals}
\alias{transformProbeSignals.AffymetrixDataFile}
\alias{write.default}
\alias{writeApd.AffymetrixDataFile}
% Calibration and normalization functions

% Miscellaneous statistical functions

% Other missing docs
\alias{[.AffymetrixCelFile}
\alias{[[.AffymetrixCelFile}
\alias{[.AffymetrixCelSet}
\alias{[[.AffymetrixCelSet}
\alias{annotateMvsA}
\alias{annotateMvsA.AffymetrixCelFile}
\alias{append}
\alias{append.AffymetrixCelSet}
\alias{append.AffymetrixFileSet}
\alias{appendDataSet}
\alias{appendDataSet.GenomeAnalyzer}
\alias{append.default}
\alias{appendFiles}
\alias{appendFiles.AffymetrixFileSet}
\alias{applyToUnitIntensities}
\alias{applyToUnitIntensities.AffymetrixCelSet}
\alias{as.AffymetrixCelSet}
\alias{as.AffymetrixFileSet}
\alias{as.character.AffymetrixCdfFile}
\alias{as.character.AffymetrixCelFile}
\alias{as.character.AffymetrixProbeTabFile}
\alias{as.character.CnChipEffectFile}
\alias{as.character.Crlmm}
\alias{as.character.GenomeAnalyzer}
\alias{as.character.GladModel}
\alias{as.character.Interface}
\alias{as.character.PlatformDesign}
\alias{as.character.Preprocessing}
\alias{as.character.SnpChipEffectFile}
\alias{as.character.UnitGroupsModel}
\alias{as.GenotypeCallSet}
\alias{backtransformGenotypeCone}
\alias{backtransformGenotypeCone.matrix}
\alias{bgAdjustGcrma}
\alias{bgAdjustOptical}
\alias{bgAdjustRma}
\alias{calcMvsA}
\alias{calcMvsA.AffymetrixCelFile}
\alias{calculateGsbParameters}
\alias{calculateTargetDistribution}
\alias{calculateTargetDistribution.QuantileNormalization}
\alias{calibrateAllelicCrosstalk}
\alias{cleanToRealChipType}
\alias{cleanToRealChipType.AffymetrixCdfFile}
\alias{clearCache.AffymetrixFileSet}
\alias{clearData}
\alias{clearFileCache}
\alias{clearFileCache.PlatformDesign}
\alias{clone.AffymetrixCelFile}
\alias{clone.AffymetrixCelSet}
\alias{clone.AffymetrixFile}
\alias{clone.AffymetrixFileSet}
\alias{CnPlmAnalyzer}
\alias{compare}
\alias{computeAffinities}
\alias{convert}
\alias{convertTable}
\alias{convertTable.default}
\alias{copyFile}
\alias{copyFile.AffymetrixFile}
\alias{create}
\alias{createBlankSet}
\alias{createBlankSet.AffymetrixCelSet}
\alias{createFrom}
\alias{createFrom.AffymetrixCelFile}
\alias{createFromCrlmmFile}
\alias{createFromCrlmmFile.GenotypeCallSet}
\alias{createFromFeatureSet}
\alias{createFromFeatureSet.AffymetrixCelSet}
\alias{create.GenotypeCallFile}
\alias{createMonoCell}
\alias{createMonoCell.AffymetrixCdfFile}
\alias{createParamCdf}
\alias{createParamCdf.ChipEffectFile}
\alias{dataFrame}
\alias{dataFrame.default}
\alias{decode}
\alias{decode.ParameterCelFile}
\alias{decodeUnit}
\alias{decodeUnit.ParameterCelFile}
\alias{encode}
\alias{encode.ParameterCelFile}
\alias{encodeUnit}
\alias{encodeUnit.ParameterCelFile}
\alias{equals.AffymetrixCdfFile}
\alias{extend.Interface}
\alias{extractAffyBatch}
\alias{extractSnpFeatureSet}
\alias{extractSnpQSet}
\alias{findUnitsTodo}
\alias{findUnitsTodo.ChipEffectFile}
\alias{findUnitsTodo.ChipEffectSet}
\alias{fit.CnPlmAnalyzer}
\alias{fitCrlmm}
\alias{fitGenotypeCone}
\alias{fitGenotypeCone.matrix}
\alias{fitGlad}
\alias{fitGlad.CnPlmAnalyzer}
\alias{fit.GladModel}
\alias{fit.PcrProbeCalibration}
\alias{forDataSet}
\alias{forDataSet.GdasAnnotationSet}
\alias{fromCdf.AffymetrixProbeTabFile}
\alias{fromChipType}
\alias{fromChipType.AffymetrixProbeTabFile}
\alias{fromChipType.DChipGenomeInformation}
\alias{fromChipType.DChipSnpInformation}
\alias{fromChipType.GdasAnnotationFile}
\alias{fromChipType.GdasAnnotationSet}
\alias{fromCleanChipType}
\alias{fromCleanChipType.AffymetrixCdfFile}
\alias{fromDataFile}
\alias{fromDataFile.ChipEffectFile}
\alias{fromDataSet}
\alias{fromDataSet.ChipEffectSet}
\alias{fromDataSet.GenomeInformation}
\alias{fromDataSet.SnpInformation}
\alias{fromFile.AffymetrixFile}
\alias{fromFiles.AffymetrixCelSet}
\alias{fromFiles.ChipEffectSet}
\alias{fromFiles.GdasAnnotationSet}
\alias{fromFiles.GenotypeCallSet}
\alias{gcrmaSummary}
\alias{[.GdasAnnotationFile}
\alias{GdasAnnotationFile}
\alias{[.GdasAnnotationSet}
\alias{GdasAnnotationSet}
\alias{GenericSummary}
\alias{GenomeCnAnalyzer}
\alias{[<-.GenotypeCallFile}
\alias{[.GenotypeCallFile}
\alias{[.GenotypeCallSet}
\alias{[[.GenotypeCallSet}
\alias{getAllelePairs}
\alias{getAllelePairUnitSets}
\alias{getAlleleProbePairs}
\alias{getAlleleProbePairs2}
\alias{getAlleleProbePairs2.AffymetrixCdfFile}
\alias{getAm}
\alias{getAM}
\alias{getAm.AffymetrixCelFile}
\alias{getAnnotations}
\alias{getAnnotations.PlatformDesign}
\alias{getAsteriskTag}
\alias{getAsteriskTag.Preprocessing}
\alias{getAverage}
\alias{getAverage.AffymetrixCelSet}
\alias{getAverageAsinh}
\alias{getAverageAsinh.AffymetrixCelSet}
\alias{getAverageFile}
\alias{getAverageFile.ChipEffectSet}
\alias{getAverageFile.CnChipEffectSet}
\alias{getAverageFile.SnpChipEffectSet}
\alias{getAverageLog}
\alias{getAverageLog.AffymetrixCelSet}
\alias{getBlockPositions}
\alias{getBlockPositions.GdasAnnotationFile}
\alias{getCachedFields}
\alias{getCachedFields.PlatformDesign}
\alias{getCachePath.PlatformDesign}
\alias{getCdf}
\alias{getCdf.AffymetrixProbeTabFile}
\alias{getCdf.FragmentLengthNormalization}
\alias{getCdf.GdasAnnotationSet}
\alias{getCdf.GenotypeCallFile}
\alias{getCdf.GladModel}
\alias{getCdf.PlatformDesign}
\alias{getCellIndices}
\alias{getCellIndices.AffymetrixSnpCelFile}
\alias{getCellIndices.ChipEffectFile}
\alias{getCellIndices.ChipEffectSet}
\alias{getCellIndices.CnChipEffectFile}
\alias{getCellIndices.CnPlm}
\alias{getCellIndices.CnProbeAffinityFile}
\alias{getCellIndices.ProbeAffinityFile}
\alias{getCellIndices.ProbeLevelModel}
\alias{getCellIndices.SnpChipEffectFile}
\alias{getCellIndices.SnpPlm}
\alias{getCellIndices.SnpProbeAffinityFile}
\alias{getCellMap}
\alias{getCellMap.ChipEffectFile}
\alias{getChipEffectFileClass}
\alias{getChipEffectFileClass.ChipEffectSet}
\alias{getChipEffectFileClass.CnChipEffectSet}
\alias{getChipEffectFileClass.SnpChipEffectSet}
\alias{getChipEffects}
\alias{getChipEffects.ChromosomeExplorer}
\alias{getChipEffects.CnPlm}
\alias{getChipEffects.CnPlmAnalyzer}
\alias{getChipEffectSetClass}
\alias{getChipEffectSetClass.CnPlm}
\alias{getChipEffectSetClass.ProbeLevelModel}
\alias{getChipEffectSetClass.SnpPlm}
\alias{getChipEffects.GladModel}
\alias{getChipEffects.SnpPlm}
\alias{getChipType.AffymetrixCdfFile}
\alias{getChipType.AffymetrixProbeTabFile}
\alias{getChipType.GdasAnnotationSet}
\alias{getChipType.GenomeAnalyzer}
\alias{getChipType.GladModel}
\alias{getChipType.PlatformDesign}
\alias{getChromosomes}
\alias{getChromosomes.GenomeInformation}
\alias{getChromosomeStats}
\alias{getChromosomeStats.GenomeInformation}
\alias{getCombineAlleles}
\alias{getCombineAlleles.CnChipEffectSet}
\alias{getCrlmmInfo}
\alias{getCrlmmInfo.PlatformDesign}
\alias{getData.AffymetrixCelSet}
\alias{getData.AffymetrixProbeTabFile}
\alias{getDataFile}
\alias{getDataFile.PlatformDesign}
\alias{getDataFiles}
\alias{getDataFiles.PlatformDesign}
\alias{getDataFlat}
\alias{getDataFlat.ChipEffectFile}
\alias{getData.RelativeCelSet}
\alias{getDataSet}
\alias{getDataSet.GenomeAnalyzer}
\alias{getDataSets}
\alias{getDataSets.GenomeAnalyzer}
\alias{getDescription.AffymetrixFileSet}
\alias{getDescription.PlatformDesign}
\alias{getDimension}
\alias{getDimension.AffymetrixCdfFile}
\alias{getFeatureInfo}
\alias{getFeatureInfo.PlatformDesign}
\alias{getFields.GenomeInformation}
\alias{getFields.SnpInformation}
\alias{getFile}
\alias{getFile.AffymetrixFileSet}
\alias{getFilename.Arguments}
\alias{getFiles}
\alias{getFiles.AffymetrixFileSet}
\alias{getFileSize.AffymetrixFile}
\alias{getFileSize.AffymetrixFileSet}
\alias{getFirstCellIndices}
\alias{getFirstCellIndices.AffymetrixCdfFile}
\alias{getFitFunction}
\alias{getFitUnitFunction}
\alias{getFitUnitFunction.CnPlm}
\alias{getFitUnitFunction.ProbeLevelModel}
\alias{getFragmentLengths}
\alias{getFragmentLengths.SnpInformation}
\alias{getFragmentStarts}
\alias{getFragmentStarts.SnpInformation}
\alias{getFragmentStops}
\alias{getFragmentStops.SnpInformation}
\alias{getFullName}
\alias{getFullName.AffymetrixCelSet}
\alias{getFullName.GladModel}
\alias{getGenomeInformation}
\alias{getGenomeInformation.GenomeAnalyzer}
\alias{getGenomeInformation.GladModel}
\alias{getHeader}
\alias{getHeaderV3}
\alias{getHeaderV3.AffymetrixCelFile}
\alias{getIdentifier}
\alias{getIdentifier.AffymetrixCelFile}
\alias{getIdentifier.AffymetrixCelSet}
\alias{getIdentifier.AffymetrixFileSet}
\alias{getIndexToRowMap}
\alias{getIndexToRowMap.AffymetrixProbeTabFile}
\alias{getInputDataSet}
\alias{getIntensities}
\alias{getLabel.AffymetrixFile}
\alias{getLabel.UnitGroupsModel}
\alias{getLog2Ratios}
\alias{getLog2Ratios.GladModel}
\alias{getMergeStrands}
\alias{getMergeStrands.SnpChipEffectSet}
\alias{getModel}
\alias{getModel.MbeiCnAnalyzer}
\alias{getModel.RmaCnAnalyzer}
\alias{getName.AffymetrixCelSet}
\alias{getName.GenotypeCallSet}
\alias{getName.GladModel}
\alias{getName.PlatformDesign}
\alias{getOligoToCelMap}
\alias{getOligoToCelMap.AffymetrixCdfFile}
\alias{getOutputDataSet}
\alias{getOutputFiles}
\alias{getOutputFiles.Preprocessing}
\alias{getOutputIdentifier}
\alias{getOutputIdentifier.QuantileNormalization}
\alias{getPackageName}
\alias{getPackageName.default}
\alias{getPackageName.PlatformDesign}
\alias{getParameter}
\alias{getParameter.ParameterFile}
\alias{getParameters}
\alias{getParametersAsString}
\alias{getParametersAsString.Preprocessing}
\alias{getParameterSet}
\alias{getParameterSet.UnitGroupsModel}
\alias{getParameters.FragmentLengthNormalization}
\alias{getParameters.ParameterFile}
\alias{getParameters.Preprocessing}
\alias{getParameters.QuantileNormalization}
\alias{getPath.GladModel}
\alias{getPathnames}
\alias{getPath.PlatformDesign}
\alias{getPlatformDesignObject}
\alias{getPlatformDesignObject.AffymetrixCdfFile}
\alias{getPlatformDesignObject.PlatformDesign}
\alias{getPmDesignMatrix}
\alias{getPmDesignMatrix.PcrProbeCalibration}
\alias{getPositions}
\alias{getProbeAffinities}
\alias{getProbeAffinities.AffinePlm}
\alias{getProbeAffinities.CnPlm}
\alias{getProbeAffinities.CnPlmAnalyzer}
\alias{getProbeAffinities.MbeiPlm}
\alias{getProbeAffinities.RmaPlm}
\alias{getProbeAffinities.SnpPlm}
\alias{getProbeEffects}
\alias{getProbeEffects.PlatformDesign}
\alias{getProgress}
\alias{getRectangle}
\alias{getRectangle.AffymetrixCelFile}
\alias{getReference}
\alias{getReferenceChipEffect}
\alias{getReferenceChipEffect.ChromosomeExplorer}
\alias{getReferenceFile}
\alias{getReferenceFile.RelativeCelSet}
\alias{getReference.GladModel}
\alias{getReferenceQuantiles}
\alias{getReferenceQuantiles.PlatformDesign}
\alias{getRegions}
\alias{getRegions.GladModel}
\alias{getRegions.profileCGH}
\alias{getRelativeFields}
\alias{getRelativeFields.RelativeCelSet}
\alias{getRestructor}
\alias{getRestructor.AffymetrixCdfFile}
\alias{getRootPath}
\alias{getRootPath.ChipEffectPreprocessing}
\alias{getRootPath.ChromosomeExplorer}
\alias{getRootPath.GladModel}
\alias{getRootPath.Preprocessing}
\alias{getRootPath.ProbeLevelModel}
\alias{getRootPath.ProbePreprocessing}
\alias{getSampleName}
\alias{getSampleName.AffymetrixCelFile}
\alias{getSampleName.ChipEffectFile}
\alias{getSampleNames}
\alias{getSnpInformation}
\alias{getSnpInformation.AffymetrixCdfFile}
\alias{getSubname}
\alias{getSubname.CnPlm}
\alias{getSubname.SnpPlm}
\alias{getSubsetToFit}
\alias{getSubsetToFit.FragmentLengthNormalization}
\alias{getSubsetToFit.PcrProbeCalibration}
\alias{getTags}
\alias{getTags.GladModel}
\alias{getTargetDistribution}
\alias{getTargetDistributionIdentifier}
\alias{getTargetDistributionIdentifier.QuantileNormalization}
\alias{getTargetDistributionPathname}
\alias{getTargetDistributionPathname.QuantileNormalization}
\alias{getTargetDistribution.QuantileNormalization}
\alias{getTargetFunction}
\alias{getTargetFunction.FragmentLengthNormalization}
\alias{getTimestamp}
\alias{getTimestamps}
\alias{getTimestamps.AffymetrixCelSet}
\alias{getUnitIndices}
\alias{getUnitIndices.GdasAnnotationSet}
\alias{getUnitIntensities}
\alias{getUnitNames}
\alias{getUnitSizes}
\alias{getUnitsOnChromosome}
\alias{getUnitsOnChromosome.GenomeInformation}
\alias{getUnitsTodo}
\alias{getXAM}
\alias{highlight}
\alias{highlight.AffymetrixCelFile}
\alias{highlight.ProbeLevelModel}
\alias{identifyCells}
\alias{identifyCells.AffymetrixCdfFile}
\alias{image270}
\alias{importFeatureSet}
\alias{importFeatureSet.AffymetrixCelFile}
\alias{importFromCrlmmFile}
\alias{importFromCrlmmFile.GenotypeCallFile}
\alias{indexOf}
\alias{isDuplicated}
\alias{isInstalled}
\alias{isInstalled.PlatformDesign}
\alias{isLoaded.PlatformDesign}
\alias{isPm}
\alias{isSnpChip}
\alias{join}
\alias{join.GdasAnnotationSet}
\alias{loadDataFile}
\alias{loadDataFile.PlatformDesign}
\alias{load.PlatformDesign}
\alias{MbeiCnAnalyzer}
\alias{nbrOfArrays.GenomeAnalyzer}
\alias{nbrOfArrays.GladModel}
\alias{nbrOfCells}
\alias{nbrOfCells.AffymetrixCdfFile}
\alias{nbrOfCells.AffymetrixCelFile}
\alias{nbrOfColumns}
\alias{nbrOfColumns.AffymetrixCdfFile}
\alias{nbrOfFiles}
\alias{nbrOfParameters}
\alias{nbrOfParameters.ParameterFile}
\alias{nbrOfQcUnits}
\alias{nbrOfQcUnits.AffymetrixCdfFile}
\alias{nbrOfRows}
\alias{nbrOfRows.AffymetrixCdfFile}
\alias{nbrOfUnits}
\alias{nbrOfUnits.AffymetrixCdfFile}
\alias{nbrOfUnits.GenomeInformation}
\alias{nbrOfUnits.SnpInformation}
\alias{normalizeAverage}
\alias{normalizeAverage.list}
\alias{normalizeAverage.matrix}
\alias{normalizeFragmentLength}
\alias{normalizeFragmentLength.default}
\alias{normalizeGenotypeCone}
\alias{normalizeGenotypeCone.matrix}
\alias{normalizeQuantiles}
\alias{plotFit}
\alias{plotFit.PcrProbeCalibration}
\alias{plotGlad}
\alias{plotGlad.CnPlmAnalyzer}
\alias{plot.GladModel}
\alias{plotMvsPosition}
\alias{plotMvsPosition.ProbeLevelModel}
\alias{plotMvsX}
\alias{plot.profileCGH}
\alias{plotProfile.profileCGH}
\alias{predict.lowess}
\alias{preprocess}
\alias{preprocess.MbeiCnAnalyzer}
\alias{preprocess.RmaCnAnalyzer}
\alias{print.GenericSummary}
\alias{print.Interface}
\alias{process}
\alias{process.PcrProbeCalibration}
\alias{range.AffymetrixCelFile}
\alias{range.AffymetrixCelSet}
\alias{read250K}
\alias{read250K.DChipSnpInformation}
\alias{read250KHg17}
\alias{read250KHg17.DChipGenomeInformation}
\alias{read50K}
\alias{read50K.DChipSnpInformation}
\alias{read50KHg17}
\alias{read50KHg17.DChipGenomeInformation}
\alias{readBlock}
\alias{readBlock.GdasAnnotationFile}
\alias{readColumnName}
\alias{readColumnName.GdasAnnotationFile}
\alias{readData}
\alias{readData.DChipGenomeInformation}
\alias{readData.DChipSnpInformation}
\alias{readData.GdasAnnotationFile}
\alias{readData.GenomeInformation}
\alias{readData.SnpInformation}
\alias{readHeader.GdasAnnotationFile}
\alias{readTableInternal}
\alias{readTableInternal.GenomeInformation}
\alias{readTableInternal.SnpInformation}
\alias{readUnits.AffymetrixCelSet}
\alias{readUnits.ChipEffectFile}
\alias{readUnits.ChipEffectSet}
\alias{readUnits.CnChipEffectFile}
\alias{readUnits.GenotypeCallFile}
\alias{readUnits.GenotypeCallSet}
\alias{readUnits.ParameterCelFile}
\alias{readUnits.ProbeAffinityFile}
\alias{readUnits.RelativeCelSet}
\alias{readUnits.SnpChipEffectFile}
\alias{reorder.AffymetrixFileSet}
\alias{restruct}
\alias{restruct.AffymetrixCdfFile}
\alias{RmaCnAnalyzer}
\alias{rmaSummary}
\alias{rowSds}
\alias{rowVars}
\alias{select}
\alias{select.GdasAnnotationSet}
\alias{setCdf}
\alias{setCombineAlleles}
\alias{setCombineAlleles.CnChipEffectSet}
\alias{setCombineAlleles.CnPlm}
\alias{setCombineAlleles.CnProbeAffinityFile}
\alias{setDecodeFunction}
\alias{setDecodeFunction.ParameterCelFile}
\alias{setEncodeFunction}
\alias{setEncodeFunction.ParameterCelFile}
\alias{setLabel.AffymetrixFile}
\alias{setLabel.UnitGroupsModel}
\alias{setMergeStrands}
\alias{setMergeStrands.SnpChipEffectSet}
\alias{setMergeStrands.SnpPlm}
\alias{setMergeStrands.SnpProbeAffinityFile}
\alias{setName}
\alias{setReference}
\alias{setReference.GladModel}
\alias{setRestructor}
\alias{setTags}
\alias{setTags.Preprocessing}
\alias{setTags.UnitGroupsModel}
\alias{setup}
\alias{setup.CnPlmAnalyzer}
\alias{setup.UnitGroupsModel}
\alias{smoothScatterMvsA}
\alias{stext}
\alias{stextChipType}
\alias{stextChipType.AffymetrixCdfFile}
\alias{stextChipType.AffymetrixCelFile}
\alias{stext.default}
\alias{stextLabel}
\alias{stextLabel.AffymetrixFile}
\alias{stextLabels}
\alias{stextLabels.AffymetrixFile}
\alias{stextSize}
\alias{stextSize.AffymetrixFile}
\alias{toCamelCase}
\alias{toCamelCase.default}
\alias{transformAffine}
\alias{transformAffine.AffymetrixCelFile}
\alias{transformAffine.AffymetrixCelSet}
\alias{unload.PlatformDesign}
\alias{updateDataFlat}
\alias{updateDataFlat.ChipEffectFile}
\alias{updateUnits}
\alias{updateUnits.ChipEffectFile}
\alias{updateUnits.ChipEffectSet}
\alias{updateUnits.GenotypeCallFile}
\alias{updateUnits.ParameterCelFile}
\alias{updateUnits.ProbeAffinityFile}
\alias{uses}
\alias{uses.Interface}
\alias{verify}
\alias{verify.DChipGenomeInformation}
\alias{verify.DChipSnpInformation}
\alias{whatDataType}
\alias{whatDataType.default}
\alias{writeRegions}
\alias{writeRegions.GladModel}
\alias{writeRegions.profileCGH}
\alias{writeSpatial}
\alias{writeSpatial.ProbeAffinityFile}
\alias{writeWig}
\alias{writeWig.CnChipEffectSet}
\alias{writeWig.GladModel}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
