%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelFile.PLOT.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{plotImage.AffymetrixCelFile}
\alias{plotImage.AffymetrixCelFile}
\alias{AffymetrixCelFile.plotImage}
\alias{plotImage.AffymetrixCelFile}
\alias{plotImage,AffymetrixCelFile-method}


\title{Opens an ImageMagick window with a spatial plot of a CEL file}

\description{
 Opens an ImageMagick window with a spatial plot of a CEL file.
}

\usage{\method{plotImage}{AffymetrixCelFile}(this, xrange=c(0, Inf), yrange=c(0, Inf), takeLog=TRUE, interleaved=FALSE, field=c("intensities", "stdvs", "pixels"), col=gray.colors(256), outlierCol="white", xSize=1000, ySize=1000, main=getName(this), ...)}

\arguments{
  \item{}{}
}

\value{
  Returns nothing.
}

\section{Details}{
}

\author{Ken Simpson (ksimpson[at]wehi.edu.au).}

\seealso{
  For more information see \code{\link{AffymetrixCelFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
