%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSetReporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.AffymetrixFileSetReporter}
\alias{process.AffymetrixFileSetReporter}
\alias{AffymetrixFileSetReporter.process}
\alias{process.AffymetrixFileSetReporter}
\alias{process,AffymetrixFileSetReporter-method}


\title{Generates report}

\description{
 Generates report.
}

\usage{\method{process}{AffymetrixFileSetReporter}(...)}

\arguments{
  \item{...}{See subclasses.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFileSetReporter}}.
}
\keyword{internal}
\keyword{methods}
