%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ArrayExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.ArrayExplorer}
\alias{process.ArrayExplorer}
\alias{ArrayExplorer.process}
\alias{process,ArrayExplorer-method}

\title{Generates image files, scripts and dynamic pages for the explorer}

\description{
 Generates image files, scripts and dynamic pages for the explorer.
}

\usage{
\method{process}{ArrayExplorer}(this, arrays=NULL, ..., verbose=FALSE)
}

\arguments{
  \item{arrays}{An optional \code{\link[base]{vector}} of arrays to be processed.
     If \code{\link[base]{NULL}}, all arrays are considered.}
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns nothing.
}

\seealso{
  For more information see \code{\link{ArrayExplorer}}.
}
\keyword{internal}
\keyword{methods}
