%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GcRmaBackgroundCorrection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.GcRmaBackgroundCorrection}
\alias{process.GcRmaBackgroundCorrection}
\alias{GcRmaBackgroundCorrection.process}
\alias{process,GcRmaBackgroundCorrection-method}

\title{Performs background correction}

\description{
 Performs background correction.
}

\usage{
\method{process}{GcRmaBackgroundCorrection}(this, ..., force=FALSE, verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns the output data set.
}

\seealso{
  For more information see \code{\link{GcRmaBackgroundCorrection}}.
}
\keyword{internal}
\keyword{methods}
