%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  LimmaBackgroundCorrection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.LimmaBackgroundCorrection}
\alias{process.LimmaBackgroundCorrection}
\alias{LimmaBackgroundCorrection.process}
\alias{process.LimmaBackgroundCorrection}
\alias{process,LimmaBackgroundCorrection-method}


\title{Performs background correction}

\description{
 Performs background correction.
}

\usage{\method{process}{LimmaBackgroundCorrection}(this, ..., force=FALSE, verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link[base]{double}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{LimmaBackgroundCorrection}}.
}
\keyword{internal}
\keyword{methods}
