%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MatNormalization.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.MatNormalization}
\alias{process.MatNormalization}
\alias{MatNormalization.process}
\alias{process,MatNormalization-method}

\title{Normalizes the data set}

\description{
 Normalizes the data set.
}

\usage{
\method{process}{MatNormalization}(this, ..., ram=NULL, force=FALSE, verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{force}{If \code{\link[base:logical]{TRUE}}, data already normalized is re-normalized,
      otherwise not.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link[base]{double}} \code{\link[base]{vector}}.
}

\seealso{
  For more information see \code{\link{MatNormalization}}.
}
\keyword{internal}
\keyword{methods}
