%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ReseqCrosstalkCalibration.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.ReseqCrosstalkCalibration}
\alias{process.ReseqCrosstalkCalibration}
\alias{ReseqCrosstalkCalibration.process}
\alias{process.ReseqCrosstalkCalibration}
\alias{process,ReseqCrosstalkCalibration-method}


\title{Calibrates the data set}

\description{
 Calibrates the data set.
}

\usage{\method{process}{ReseqCrosstalkCalibration}(this, ..., force=FALSE, verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{force}{If \code{\link[base:logical]{TRUE}}, data already calibrated is re-calibrated,
      otherwise not.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link[base]{double}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ReseqCrosstalkCalibration}}.
}
\keyword{internal}
\keyword{methods}
