%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RmaBackgroundCorrection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.RmaBackgroundCorrection}
\alias{process.RmaBackgroundCorrection}
\alias{RmaBackgroundCorrection.process}
\alias{process,RmaBackgroundCorrection-method}

\title{Performs background correction}

\description{
 Performs background correction.
}

\usage{
\method{process}{RmaBackgroundCorrection}(this, ..., force=FALSE, verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns the output data set.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RmaBackgroundCorrection}}.
}
\keyword{internal}
\keyword{methods}
