%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpatialReporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.SpatialReporter}
\alias{process.SpatialReporter}
\alias{SpatialReporter.process}
\alias{process,SpatialReporter-method}

\title{Generates image files, scripts and dynamic pages for the explorer}

\description{
 Generates image files, scripts and dynamic pages for the explorer.
}

\usage{
\method{process}{SpatialReporter}(this, ..., verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns nothing.
}

\seealso{
  For more information see \code{\link{SpatialReporter}}.
}
\keyword{internal}
\keyword{methods}
