%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitTypeScaleNormalization.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.UnitTypeScaleNormalization}
\alias{process.UnitTypeScaleNormalization}
\alias{UnitTypeScaleNormalization.process}
\alias{process,UnitTypeScaleNormalization-method}

\title{Normalizes the data set}

\description{
 Normalizes the data set.
}

\usage{
\method{process}{UnitTypeScaleNormalization}(this, ..., skip=FALSE, force=FALSE, verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{force}{If \code{\link[base:logical]{TRUE}}, data already normalized is re-normalized,
      otherwise not.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link[base]{double}} \code{\link[base]{vector}}.
}

\seealso{
  For more information see \code{\link{UnitTypeScaleNormalization}}.
}
\keyword{internal}
\keyword{methods}
