%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readIntensities.AffymetrixDataFile}
\alias{readIntensities.AffymetrixDataFile}
\alias{AffymetrixDataFile.readIntensities}
\alias{readIntensities.AffymetrixDataFile}
\alias{readIntensities,AffymetrixDataFile-method}


\title{Reads probe intensities from a data file}

\description{
  Reads probe intensities from a data file.
}

\usage{\method{readIntensities}{AffymetrixDataFile}(...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataFile}}.
}
\keyword{internal}
\keyword{methods}
