%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParameterSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{seq.ParameterSet}
\alias{seq.ParameterSet}
\alias{ParameterSet.seq}
\alias{seq.ParameterSet}
\alias{seq,ParameterSet-method}


\title{Gets an vector of file indices}

\description{
  Gets an vector of file indices.
}

\usage{\method{seq}{ParameterSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} in [1,N] where N is the number of arrays,
  or an empty vector if the data set is empty.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{ParameterSet}}.
}
\keyword{internal}
\keyword{methods}
