%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Model.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setAlias.Model}
\alias{setAlias.Model}
\alias{Model.setAlias}
\alias{setAlias.Model}
\alias{setAlias,Model-method}


\title{Sets the name alias for the model}

\description{
 Sets the name alias for the model, if set.
}

\usage{\method{setAlias}{Model}(this, alias=NULL, ...)}

\arguments{
 \item{alias}{A \code{\link[base]{character}} string, or \code{\link[base]{NULL}}.}
 \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getAlias.Model]{*getAlias}()}
  For more information see \code{\link{Model}}.
}
\keyword{internal}
\keyword{methods}
