%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataFile.MAP.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setApdMap.AffymetrixDataFile}
\alias{setApdMap.AffymetrixDataFile}
\alias{AffymetrixDataFile.setApdMap}
\alias{setApdMap.AffymetrixDataFile}
\alias{setApdMap,AffymetrixDataFile-method}


\title{Sets the probe read map for a data file}

\description{
  Sets the probe read map for a data file.
}

\usage{\method{setApdMap}{AffymetrixDataFile}(this, apdMap=NULL, ...)}

\arguments{
 \item{apdMap}{An \code{\link{ApdMap}} object or \code{\link[base]{NULL}}.}
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getReadMap.AffymetrixDataFile]{*getReadMap}()}.
  \code{\link[aroma.affymetrix:hasApdMap.AffymetrixDataFile]{*hasApdMap}()}.
  For more information see \code{\link{AffymetrixDataFile}}.
}
\keyword{internal}
\keyword{methods}
