%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.MAP.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setApdMap.AffymetrixDataSet}
\alias{setApdMap.AffymetrixDataSet}
\alias{AffymetrixDataSet.setApdMap}
\alias{setApdMap.AffymetrixDataSet}
\alias{setApdMap,AffymetrixDataSet-method}


\title{Sets the APD map for the data set}

\description{
  Sets the APD map for the data set.
}

\usage{\method{setApdMap}{AffymetrixDataSet}(this, apdMap, ...)}

\arguments{
 \item{apdMap}{An \code{\link{ApdMap}} object, or \code{\link[base]{NULL}}.}
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getReadMap.AffymetrixDataSet]{*getReadMap}()} and \code{\link[aroma.affymetrix:hasApdMap.AffymetrixDataSet]{*hasApdMap}()}.
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
