%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ArrayExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setArrays.ArrayExplorer}
\alias{setArrays.ArrayExplorer}
\alias{ArrayExplorer.setArrays}
\alias{setArrays.ArrayExplorer}
\alias{setArrays,ArrayExplorer-method}


\title{Sets the arrays}

\description{
 Sets the arrays to be processed by the explorer.
}

\usage{\method{setArrays}{ArrayExplorer}(this, arrays=NULL, ...)}

\arguments{
  \item{arrays}{A \code{\link[base]{character}} (or \code{\link[base]{integer}}) \code{\link[base]{vector}} of arrays to be
     considered. If \code{\link[base]{NULL}}, all arrays of the data set are considered.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ArrayExplorer}}.
}
\keyword{internal}
\keyword{methods}
