%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ChromosomeExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setArrays.ChromosomeExplorer}
\alias{setArrays.ChromosomeExplorer}
\alias{ChromosomeExplorer.setArrays}
\alias{setArrays.ChromosomeExplorer}
\alias{setArrays,ChromosomeExplorer-method}


\title{Sets the arrays}

\description{
 Sets the arrays to be processed by the explorer.
}

\usage{\method{setArrays}{ChromosomeExplorer}(this, arrays=NULL, ...)}

\arguments{
  \item{arrays}{A \code{\link[base]{character}} (or \code{\link[base]{integer}}) \code{\link[base]{vector}} of arrays used in the
     model.  If \code{\link[base]{NULL}}, all arrays of the model are considered.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ChromosomeExplorer}}.
}
\keyword{internal}
\keyword{methods}
