%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelFile.R
% 
% on Sat Jan 06 01:46:59 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setCdf.AffymetrixCelFile}
\alias{setCdf.AffymetrixCelFile}
\alias{AffymetrixCelFile.setCdf}
\alias{setCdf.AffymetrixCelFile}
\alias{setCdf,AffymetrixCelFile-method}


\title{Sets the CDF structure for this CEL file}

\description{
 Sets the CDF structure for this CEL file.
}

\usage{\method{setCdf}{AffymetrixCelFile}(this, cdf, ..., .checkArgs=TRUE)}

\arguments{
  \item{cdf}{An \code{\link{AffymetrixCdfFile}} object.}
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getCdf.AffymetrixCelFile]{*getCdf}()}.
  For more information see \code{\link{AffymetrixCelFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
