%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CnagCfhFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setCdf.CnagCfhFile}
\alias{setCdf.CnagCfhFile}
\alias{CnagCfhFile.setCdf}
\alias{setCdf.CnagCfhFile}
\alias{setCdf,CnagCfhFile-method}


\title{Sets the CDF structure for this CEL file}

\description{
 Sets the CDF structure for this CEL file.
}

\usage{\method{setCdf}{CnagCfhFile}(this, cdf, ..., .checkArgs=TRUE)}

\arguments{
  \item{cdf}{An \code{\link{AffymetrixCdfFile}} object.}
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getCdf.CnagCfhFile]{*getCdf}()}.
  For more information see \code{\link{CnagCfhFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
