%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  CnagCfhSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setCdf.CnagCfhSet}
\alias{setCdf.CnagCfhSet}
\alias{CnagCfhSet.setCdf}
\alias{setCdf.CnagCfhSet}
\alias{setCdf,CnagCfhSet-method}


\title{Sets the CDF structure for this CFH set}

\description{
 Sets the CDF structure for this CFH set.  This structures is assigned to all CFH files in the set.
}

\usage{\method{setCdf}{CnagCfhSet}(this, cdf, verbose=FALSE, ...)}

\arguments{
  \item{cdf}{An \code{\link{AffymetrixCdfFile}} object.}
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getCdf.CnagCfhSet]{*getCdf}()}.
  For more information see \code{\link{CnagCfhSet}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
