%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setTags.AffymetrixFileSet}
\alias{setTags.AffymetrixFileSet}
\alias{AffymetrixFileSet.setTags}
\alias{setTags.AffymetrixFileSet}
\alias{setTags,AffymetrixFileSet-method}


\title{Sets the tags of the file set}

\description{
  Sets the tags of the file set.
}

\usage{\method{setTags}{AffymetrixFileSet}(this, tags="*", ...)}

\arguments{
 \item{tags}{A \code{\link[base]{character}} \code{\link[base]{vector}} of tags.}
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\details{
  See \code{\link[aroma.affymetrix:getTags.AffymetrixFileSet]{*getTags}()} for so called \emph{special tags}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getTags.AffymetrixFileSet]{*getTags}()}.
  For more information see \code{\link{AffymetrixFileSet}}.
}
\keyword{internal}
\keyword{methods}
