%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Model.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setTags.Model}
\alias{setTags.Model}
\alias{Model.setTags}
\alias{setTags.Model}
\alias{setTags,Model-method}


\title{Sets the tags to be appended}

\description{
 Sets the tags to be appended to the tags of the input data set.
}

\usage{\method{setTags}{Model}(this, tags=NULL, ...)}

\arguments{
 \item{tags}{A \code{\link[base]{character}} \code{\link[base]{vector}} of tags.
   The tags may also be passed as comma-separated strings.}
 \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getTags.Model]{*getTags}()}
  For more information see \code{\link{Model}}.
}
\keyword{internal}
\keyword{methods}
