%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  setupExampleData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setupExampleData}
\alias{setupExampleData}


\title{Setups example data in the current directory}

\description{
 Setups example data in the current directory.
}

\usage{
setupExampleData(...)
}

\arguments{
  \item{dirs}{A \code{\link[base]{character}} \code{\link[base]{vector}} specifying which directories to setup.}
  \item{...}{Not used.}
}

\value{
  Returns (invisibly) \code{\link[base:logical]{TRUE}} if all requested data was installed,
  otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson}


\keyword{internal}
