%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.TRANSFORM.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{transformOn.AffymetrixDataSet}
\alias{transformOn.AffymetrixDataSet}
\alias{AffymetrixDataSet.transformOn}
\alias{transformOn.AffymetrixDataSet}
\alias{transformOn,AffymetrixDataSet-method}

\alias{AffymetrixDataSet.transformOff}
\alias{transformOff.AffymetrixDataSet}
\alias{transformOff,AffymetrixDataSet-method}


\title{Turns transformations on and off for all data files in the data set}

\description{
  Turns transformations on and off for all data files in the data set.
}

\usage{\method{transformOn}{AffymetrixDataSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
