%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ArrayExplorer.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{updateSamplesFile.ArrayExplorer}
\alias{updateSamplesFile.ArrayExplorer}
\alias{ArrayExplorer.updateSamplesFile}
\alias{updateSamplesFile.ArrayExplorer}
\alias{updateSamplesFile,ArrayExplorer-method}


\title{Updates the samples.js file}

\description{
 Updates the samples.js file.
}

\usage{\method{updateSamplesFile}{ArrayExplorer}(this, ..., verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns (invisibly) the pathname to the samples.js file.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ArrayExplorer}}.
}
\keyword{internal}
\keyword{methods}
