%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SnpInformation.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{verify.SnpInformation}
\alias{verify.SnpInformation}
\alias{SnpInformation.verify}
\alias{verify.SnpInformation}
\alias{verify,SnpInformation-method}


\title{Verifies the correctness of the underlying file}

\description{
  Verifies the correctness of the underlying file.
}

\usage{\method{verify}{SnpInformation}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns (visibly) \code{\link[base:logical]{TRUE}} if the file is valid, otherwise an error is
  thrown.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SnpInformation}}.
}
\keyword{internal}
\keyword{methods}
