%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixDataSet.WRITE.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{write.AffymetrixDataSet}
\alias{write.AffymetrixDataSet}
\alias{AffymetrixDataSet.write}
\alias{write.AffymetrixDataSet}
\alias{write,AffymetrixDataSet-method}


\title{Writes the data files in this data set to file}

\description{
  Writes the data files in this data set to file.
}

\usage{\method{write}{AffymetrixDataSet}(this, format="apd", ...)}

\arguments{
 \item{format}{A \code{\link[base]{character}} string specifying the file format to be written.}
 \item{...}{Arguments passed to the internal write function.}
}

\value{
  Returns a new \code{AffymetrixDataSet} instance.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixDataSet}}.
}
\keyword{internal}
\keyword{methods}
