%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSetTuple.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{AffymetrixCelSetTuple}
\docType{class}
\alias{AffymetrixCelSetTuple}


 \title{The AffymetrixCelSetTuple class}

 \description{
  Package:  aroma.affymetrix \cr
\bold{Class AffymetrixCelSetTuple}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[aroma.core]{AromaMicroarrayDataSetTuple}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{AffymetrixCelSetTuple}\cr

\bold{Directly known subclasses:}\cr
\link[aroma.affymetrix]{ChipEffectSetTuple}\cr

public static class \bold{AffymetrixCelSetTuple}\cr
extends \emph{\link[aroma.core]{AromaMicroarrayDataSetTuple}}\cr


 }

 \usage{AffymetrixCelSetTuple(..., .setClass="AffymetrixCelSet")}

 \arguments{
   \item{...}{Arguments passed to the constructor of
      \code{\link[aroma.core]{AromaMicroarrayDataSetTuple}}.}
   \item{.setClass}{The name of the class of the input set.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{byPath} \tab  -\cr
}


 \bold{Methods inherited from AromaMicroarrayDataSetTuple}:\cr
asMatrixOfFiles, byPath, extract, getAlias, getArrays, getArrayTuple, getAsteriskTags, getChipTypes, getFullName, getFullNames, getListOfSets, getName, getNames, getTableOfArrays, getTags, nbrOfArrays, nbrOfChipTypes, setAlias, setTags

 \bold{Methods inherited from Object}:\cr
asThis, $, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save


 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
