%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.AffymetrixCelSet.AffymetrixCelSet}
\alias{as.AffymetrixCelSet.AffymetrixCelSet}
\alias{AffymetrixCelSet.as.AffymetrixCelSet}
\alias{as.AffymetrixCelSet.AffymetrixCelSet}
\alias{as.AffymetrixCelSet,AffymetrixCelSet-method}

\alias{as.AffymetrixCelSet.list}
\alias{as.AffymetrixCelSet.default}

\title{Coerce an object to an AffymetrixCelSet object}

\description{
  Coerce an object to an AffymetrixCelSet object.
}

\usage{\method{as.AffymetrixCelSet}{AffymetrixCelSet}(object, ...)}

\arguments{
 \item{...}{Other arguments passed to \code{\link[base]{list.files}}().}
}

\value{
  Returns an \code{\link{AffymetrixCelSet}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixCelSet}}.
}
\keyword{internal}
\keyword{methods}
