%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AromaChipTypeAnnotationFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{byChipType.AromaChipTypeAnnotationFile}
\alias{byChipType.AromaChipTypeAnnotationFile}
\alias{AromaChipTypeAnnotationFile.byChipType}
\alias{byChipType.AromaChipTypeAnnotationFile}
\alias{byChipType,AromaChipTypeAnnotationFile-method}

\alias{AromaChipTypeAnnotationFile.byName}
\alias{byName.AromaChipTypeAnnotationFile}
\alias{byName,AromaChipTypeAnnotationFile-method}


\title{Defines an AromaChipTypeAnnotationFile object by chip type}

\description{
 Defines an AromaChipTypeAnnotationFile object by chip type.
}

\synopsis{byChipType.AromaChipTypeAnnotationFile(static, chipType, tags=NULL, ...)}
\usage{AromaChipTypeAnnotationFile$byChipType(chipType, tags=NULL, ...)}

\arguments{
 \item{chipType}{A \code{\link[base]{character}} string.}
 \item{tags}{An optional \code{\link[base]{character}} \code{\link[base]{vector}} of tags.}
 \item{...}{Not used.}
}

\value{
 Returns an \code{\link{AromaChipTypeAnnotationFile}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:fromFile.AromaChipTypeAnnotationFile]{*fromFile}()}.
  For more information see \code{\link{AromaChipTypeAnnotationFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
