%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenomeInformation.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{GenomeInformation$byChipType}
\alias{GenomeInformation$byChipType}
\alias{byChipType.GenomeInformation}
\alias{GenomeInformation.byChipType}
\alias{byChipType,GenomeInformation-method}

\title{Static method to define a genome information set by chip type}

\description{
  Static method to define a genome information set by chip type.
}

\usage{
## Static method (use this):
## GenomeInformation$byChipType()

## Don't use the below:
\method{byChipType}{GenomeInformation}(...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link{GenomeInformation}} object.
}

\seealso{
  For more information see \code{\link{GenomeInformation}}.
}
\keyword{internal}
\keyword{methods}
