%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FirmaModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{calculateResidualSet.FirmaModel}
\alias{calculateResidualSet.FirmaModel}
\alias{FirmaModel.calculateResidualSet}
\alias{calculateResidualSet,FirmaModel-method}
\alias{FirmaModel.calculateResiduals}
\alias{calculateResiduals.FirmaModel}
\alias{calculateResiduals,FirmaModel-method}


\title{Gets the set of residuals corresponding to the PLM}

\description{
 Gets the set of residuals corresponding to the PLM.
}

\usage{
\method{calculateResidualSet}{FirmaModel}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{ResidualSet}} object.
}

\seealso{
  For more information see \code{\link{FirmaModel}}.
}
\keyword{internal}
\keyword{methods}
