%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ProbeLevelModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{findUnitsTodo.ProbeLevelModel}
\alias{findUnitsTodo.ProbeLevelModel}
\alias{ProbeLevelModel.findUnitsTodo}
\alias{findUnitsTodo,ProbeLevelModel-method}

\title{Identifies non-fitted units}

\description{
 Identifies non-fitted units.
}

\usage{
\method{findUnitsTodo}{ProbeLevelModel}(this, verbose=FALSE, ...)
}

\arguments{
  \item{verbose}{A \code{\link[base]{logical}} or a \code{\link[R.utils]{Verbose}}.}
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} of unit indices.
}

\seealso{
  Internally this methods calls the same method for the
  \code{\link{ChipEffectSet}} class.
  For more information see \code{\link{ProbeLevelModel}}.
}
\keyword{internal}
\keyword{methods}
