%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Model.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fit.Model}
\alias{fit.Model}
\alias{Model.fit}
\alias{fit,Model-method}

\title{Estimates the model parameters}

\description{
 Estimates the model parameters.
}

\usage{
\method{fit}{Model}(...)
}

\arguments{
  \item{...}{Arguments specific to any subclass.}
}

\value{
 Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} specifying what units where fitted.
}

\seealso{
  For more information see \code{\link{Model}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
