%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenomeInformation.R
% 
% on Sat Jan 06 01:51:44 2007.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fromChipType.GenomeInformation}
\alias{fromChipType.GenomeInformation}
\alias{GenomeInformation.fromChipType}
\alias{fromChipType.GenomeInformation}
\alias{fromChipType,GenomeInformation-method}


\title{Static method to define a genome information set by chip type}

\description{
  Static method to define a genome information set by chip type.
}

\usage{\method{fromChipType}{GenomeInformation}(...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link{GenomeInformation}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:fromCdf.GenomeInformation]{*fromCdf}()}.
  For more information see \code{\link{GenomeInformation}}.
}
\keyword{internal}
\keyword{methods}
