%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ProbeLevelModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getChipEffects.ProbeLevelModel}
\alias{getChipEffects.ProbeLevelModel}
\alias{ProbeLevelModel.getChipEffects}
\alias{getChipEffects.ProbeLevelModel}
\alias{getChipEffects,ProbeLevelModel-method}


\title{Gets the chip-effect files for this model}

\description{
 Gets the chip-effect files for this model.  There is one chip-effect file per array.
}

\usage{\method{getChipEffects}{ProbeLevelModel}(this, ..., verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{A \code{\link[base]{logical}} or a \code{\link{Verbose}} object.}
}

\value{
 Returns a \code{\link{ChipEffectSet}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ProbeLevelModel}}.
}
\keyword{internal}
\keyword{methods}
