%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCdfFile.SNPs.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCnNames.AffymetrixCdfFile}
\alias{getCnNames.AffymetrixCdfFile}
\alias{AffymetrixCdfFile.getCnNames}
\alias{getCnNames,AffymetrixCdfFile-method}

\title{Gets the names of the CN units}

\description{
  Gets the names of the CN units.
}

\usage{
\method{getCnNames}{AffymetrixCdfFile}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[aroma.affymetrix:getSnpNames.AffymetrixCdfFile]{*getSnpNames}()}.
  Internally, \code{\link[aroma.affymetrix:getUnitTypes.AffymetrixCdfFile]{*getUnitTypes}()}.
  is used.

  For more information see \code{\link{AffymetrixCdfFile}}.
}
\keyword{internal}
\keyword{methods}
