%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MbeiPlm.R
% 
% on Sat May 17 11:55:05 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitFunction.MbeiPlm}
\alias{getFitFunction.MbeiPlm}
\alias{MbeiPlm.getFitFunction}
\alias{getFitFunction.MbeiPlm}
\alias{getFitFunction,MbeiPlm-method}


\title{Gets the low-level function that fits the PLM}

\description{
 Gets the low-level function that fits the PLM.
}

\usage{\method{getFitFunction}{MbeiPlm}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[affy]{fit.li.wong}}.
  For more information see \code{\link{MbeiPlm}}.
}
\keyword{internal}
\keyword{methods}
