%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RmaPlm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitFunction.RmaPlm}
\alias{getFitFunction.RmaPlm}
\alias{RmaPlm.getFitFunction}
\alias{getFitFunction.RmaPlm}
\alias{getFitFunction,RmaPlm-method}


\title{Gets the low-level function that fits the PLM}

\description{
 Gets the low-level function that fits the PLM.
}

\usage{\method{getFitFunction}{RmaPlm}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\author{
  Henrik Bengtsson and Ken Simpson (WEHI) utilizing Ben Bolstad's
  \pkg{affyPLM} package.
}

\seealso{
  For more information see \code{\link{RmaPlm}}.
}
\keyword{internal}
\keyword{methods}
