%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AvgPlm.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitUnitGroupFunction.AvgPlm}
\alias{getFitUnitGroupFunction.AvgPlm}
\alias{AvgPlm.getFitUnitGroupFunction}
\alias{getFitUnitGroupFunction,AvgPlm-method}

\title{Gets the low-level function that fits the PLM}

\description{
 Gets the low-level function that fits the PLM.
}

\usage{
\method{getFitUnitGroupFunction}{AvgPlm}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{function}}.
}

\seealso{
  For more information see \code{\link{AvgPlm}}.
}
\keyword{internal}
\keyword{methods}
