%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  UnitGroupsModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.UnitGroupsModel}
\alias{getPath.UnitGroupsModel}
\alias{UnitGroupsModel.getPath}
\alias{getPath.UnitGroupsModel}
\alias{getPath,UnitGroupsModel-method}


\title{Gets the path (directory) for this model}

\description{
  Gets the path (directory) for this model.
}

\usage{\method{getPath}{UnitGroupsModel}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{UnitGroupsModel}}.
}
\keyword{internal}
\keyword{methods}
