%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  QualityAssessmentModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getResiduals.QualityAssessmentModel}
\alias{getResiduals.QualityAssessmentModel}
\alias{QualityAssessmentModel.getResiduals}
\alias{getResiduals.QualityAssessmentModel}
\alias{getResiduals,QualityAssessmentModel-method}


\title{Calculates the residuals from a probe-level model}

\description{
  Calculates the residuals from a probe-level model.
}

\usage{\method{getResiduals}{QualityAssessmentModel}(this, units=NULL, path=NULL, name="qcData", tags="*", ram=NULL, force=FALSE, verbose=FALSE, ...)}

\arguments{
 \item{...}{Additional arguments passed \code{fromFiles()} of
    \code{\link{QualityAssessmentSet}}.}
}

\value{
  Returns an \code{\link{QualityAssessmentSet}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{QualityAssessmentModel}}.
}
\keyword{internal}
\keyword{methods}
