%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCdfFile.SNPs.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSnpNames.AffymetrixCdfFile}
\alias{getSnpNames.AffymetrixCdfFile}
\alias{AffymetrixCdfFile.getSnpNames}
\alias{getSnpNames.AffymetrixCdfFile}
\alias{getSnpNames,AffymetrixCdfFile-method}


\title{Gets the names of the SNP units}

\description{
  Gets the names of the SNP units.
}

\usage{\method{getSnpNames}{AffymetrixCdfFile}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma.affymetrix:getCnNames.AffymetrixCdfFile]{*getCnNames}()}.
  Internally, \code{\link[aroma.affymetrix:getUnitTypes.AffymetrixCdfFile]{*getUnitTypes}()}.
  is used.

  For more information see \code{\link{AffymetrixCdfFile}}.
}
\keyword{internal}
\keyword{methods}
