%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Transform.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isDone.Transform}
\alias{isDone.Transform}
\alias{Transform.isDone}
\alias{isDone.Transform}
\alias{isDone,Transform-method}


\title{Checks if the data set is processed or not}

\description{
 Checks if the data set is processed or not.
}

\usage{\method{isDone}{Transform}(this, ..., verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the data set is processed, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Transform}}.
}
\keyword{internal}
\keyword{methods}
