%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{lapply.AffymetrixFileSet}
\alias{lapply.AffymetrixFileSet}
\alias{AffymetrixFileSet.lapply}
\alias{lapply.AffymetrixFileSet}
\alias{lapply,AffymetrixFileSet-method}


\title{Applies a function to each of the Affymetrix file files}

\description{
  Applies a function to each of the Affymetrix file files.
}

\usage{\method{lapply}{AffymetrixFileSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{list}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AffymetrixFileSet}}.
}
\keyword{internal}
\keyword{methods}
