%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{[.AffymetrixCelFile}
\alias{[.AffymetrixCelSet}
\alias{[.CnagCfhFile}
\alias{[.CnagCfhSet}
\alias{[[.AffymetrixCelFile}
\alias{[[.AffymetrixCelSet}
\alias{[[.CnagCfhFile}
\alias{[[.CnagCfhSet}
\alias{addColorMap}
\alias{addColorMap.ArrayExplorer}
\alias{addColorMap.SpatialReporter}
\alias{addExclCells}
\alias{addExclCells.DChipQuantileNormalization}
\alias{AffymetrixAptSummaryFile}
\alias{AffymetrixCnChpFile}
\alias{AffymetrixCsvFile}
\alias{AffymetrixCsvGenomeInformation}
\alias{AffymetrixNetAffxCsvFile}
\alias{AffymetrixPlatform}
\alias{AffymetrixTabularFile}
\alias{AffymetrixTsvFile}
\alias{allocate.AromaCellMatchScoreFile}
\alias{allocate.CrlmmParametersFile}
\alias{allocateFromCdf}
\alias{allocateFromCdf.AromaCellMatchScoreFile}
\alias{allocateFromCdf.AromaCellPositionFile}
\alias{allocateFromCdf.AromaCellSequenceFile}
\alias{allocateFromCdf.AromaUfcFile}
\alias{allocateFromCdf.AromaUflFile}
\alias{allocateFromCdf.AromaUnitGcContentFile}
\alias{annotateMvsA}
\alias{annotateMvsA.AffymetrixCelFile}
\alias{append.AffymetrixCelSet}
\alias{append.CnagCfhSet}
\alias{applyAnyOrder}
\alias{applyAnyOrder.CopyNumberChromosomalModel}
\alias{applyCCF0}
\alias{applyCCF0.CopyNumberChromosomalModel}
\alias{applyCFC0}
\alias{applyCFC0.CopyNumberChromosomalModel}
\alias{applyToUnitIntensities}
\alias{applyToUnitIntensities.AffymetrixCelSet}
\alias{AromaAffymetrix}
\alias{AromaCellMatchScoreFile}
\alias{AromaPipeline}
\alias{AromaUfcFile}
\alias{AromaUgcFile}
\alias{AromaUnitGcContentFile}
\alias{as.AffymetrixCelSet}
\alias{as.AffymetrixCnChpSet}
\alias{as.AffymetrixFileSet}
\alias{as.character.AffymetrixCdfFile}
\alias{as.character.AffymetrixCelFile}
\alias{as.character.AffymetrixCelSetReporter}
\alias{as.character.AffymetrixCnChpFile}
\alias{as.character.AffymetrixProbeTabFile}
\alias{as.character.AromaChipTypeAnnotationFile}
\alias{as.character.ArrayExplorer}
\alias{as.character.ChipEffectFile}
\alias{as.character.ChipEffectSet}
\alias{as.character.CnagCfhFile}
\alias{as.character.DChipCdfBinFile}
\alias{as.character.DChipDcpFile}
\alias{as.character.DChipQuantileNormalization}
\alias{as.character.FirmaModel}
\alias{as.character.GenericReporter}
\alias{as.character.GenomeInformation}
\alias{as.character.Model}
\alias{as.character.ProbeAffinityFile}
\alias{as.character.QualityAssessmentModel}
\alias{as.character.ResidualFile}
\alias{as.character.ResidualSet}
\alias{as.character.SmoothMultiarrayModel}
\alias{as.character.SnpInformation}
\alias{as.character.SpatialReporter}
\alias{as.character.TransformReport}
\alias{as.character.WeightsFile}
\alias{as.character.WeightsSet}
\alias{as.ChipEffectSetTuple}
\alias{as.ChipEffectSetTuple.ChipEffectSetTuple}
\alias{as.ChipEffectSetTuple.default}
\alias{as.CnagCfhSet}
\alias{as.CnChipEffectSetTuple}
\alias{as.CnChipEffectSetTuple.CnChipEffectSetTuple}
\alias{as.CnChipEffectSetTuple.default}
\alias{as.CopyNumberDataSetTuple.CnChipEffectSet}
\alias{as.DChipDcpSet}
\alias{ASCRMAv2}
\alias{assertAnnotationData}
\alias{assertAnnotationData.ASCRMAv2}
\alias{bgAdjustGcrma}
\alias{bgAdjustOptical}
\alias{bgAdjustRma}
\alias{boxplotStats}
\alias{boxplotStats.ChipEffectSet}
\alias{byChipType.AffymetrixCsvGenomeInformation}
\alias{byChipType.AffymetrixProbeTabFile}
\alias{byChipType.AffymetrixTabularFile}
\alias{byChipType.AffymetrixTsvFile}
\alias{byChipType.AromaCellMatchScoreFile}
\alias{byChipType.DChipCdfBinFile}
\alias{byChipType.UflSnpInformation}
\alias{byName.AffymetrixCelSet}
\alias{byName.AffymetrixCnChpSet}
\alias{byName.CnagCfhSet}
\alias{byName.CrlmmParametersSet}
\alias{byName.DChipDcpSet}
\alias{byPath.AffymetrixCelSet}
\alias{byPath.AffymetrixCelSetTuple}
\alias{byPath.AffymetrixCnChpSet}
\alias{byPath.ChipEffectSet}
\alias{byPath.CnagCfhSet}
\alias{byPath.CnChipEffectSet}
\alias{byPath.CrlmmParametersSet}
\alias{byPath.DChipDcpSet}
\alias{byPath.ExonChipEffectSet}
\alias{byPath.FirmaSet}
\alias{byPath.ResidualSet}
\alias{byPath.SnpChipEffectSet}
\alias{byPath.WeightsSet}
\alias{calcRawCnStats}
\alias{calcRawCnStats.CopyNumberChromosomalModel}
\alias{calcRawCnStats.default}
\alias{calculateAverages}
\alias{calculateAverages.FragmentEquivalentClassNormalization}
\alias{calculateBaseline}
\alias{calculateConfidenceScores}
\alias{calculateConfidenceScores.CrlmmModel}
\alias{calculateFieldBoxplotStats}
\alias{calculateFieldBoxplotStats.ChipEffectSet}
\alias{calculateMargins}
\alias{calculateMargins.SpatialReporter}
\alias{calculateNuseBoxplotStats}
\alias{calculateNuseBoxplotStats.ChipEffectSet}
\alias{calculateParametersGsb}
\alias{calculateResiduals}
\alias{calculateResiduals.ProbeLevelModel}
\alias{calculateResidualSet}
\alias{calculateResidualSet.ProbeLevelModel}
\alias{calculateRleBoxplotStats}
\alias{calculateRleBoxplotStats.ChipEffectSet}
\alias{calculateTargetDistribution}
\alias{calculateTargetDistribution.QuantileNormalization}
\alias{calculateWeights}
\alias{calculateWeights.ExonRmaPlm}
\alias{calculateWeights.FirmaModel}
\alias{calculateWeights.ProbeLevelModel}
\alias{calibrateOne}
\alias{calibrateOne.ReseqCrosstalkCalibration}
\alias{cdfMergeGroups}
\alias{ChipEffectSetTuple}
\alias{clearCache.AdditiveCovariatesNormalization}
\alias{clearCache.AffymetrixCdfFile}
\alias{clearCache.AffymetrixCelFile}
\alias{clearCache.AffymetrixCelSet}
\alias{clearCache.AffymetrixCnChpFile}
\alias{clearCache.AffymetrixNetAffxCsvFile}
\alias{clearCache.AffymetrixProbeTabFile}
\alias{clearCache.AffymetrixTsvFile}
\alias{clearCache.AllelicCrosstalkCalibration}
\alias{clearCache.ChipEffectFile}
\alias{clearCache.ChipEffectSet}
\alias{clearCache.CnagCfhFile}
\alias{clearCache.CnagCfhSet}
\alias{clearCache.DChipCdfBinFile}
\alias{clearCache.ExonChipEffectFile}
\alias{clearCache.FirmaFile}
\alias{clearCache.FirmaModel}
\alias{clearCache.FirmaSet}
\alias{clearCache.FragmentEquivalentClassNormalization}
\alias{clearCache.FragmentLengthNormalization}
\alias{clearCache.GenomeInformation}
\alias{clearCache.ParameterCelFile}
\alias{clearCache.ProbeAffinityFile}
\alias{clearCache.ProbeLevelModel}
\alias{clearCache.ProbeLevelTransform3}
\alias{clearCache.QualityAssessmentFile}
\alias{clearCache.QuantileNormalization}
\alias{clearCache.ReseqCrosstalkCalibration}
\alias{clearCache.ResidualFile}
\alias{clearCache.ResidualSet}
\alias{clearCache.SmoothMultiarrayModel}
\alias{clearCache.SnpChipEffectFile}
\alias{clearCache.SnpInformation}
\alias{clearCache.TransformReport}
\alias{clearCache.UgpGenomeInformation}
\alias{clearCache.WeightsFile}
\alias{clearCache.WeightsSet}
\alias{clearData}
\alias{clone.AffymetrixCelFile}
\alias{clone.AffymetrixCelSet}
\alias{clone.AffymetrixCnChpFile}
\alias{clone.CnagCfhFile}
\alias{clone.CnagCfhSet}
\alias{CnChipEffectSetTuple}
\alias{compare}
\alias{computeAffinities}
\alias{convert}
\alias{convertToUnique}
\alias{convertToUnique.AffymetrixCelSet}
\alias{convertUnits}
\alias{createFrom}
\alias{createMonoCell}
\alias{createMonocellCdf}
\alias{createOutputTuple}
\alias{createOutputTuple.SmoothMultiarrayModel}
\alias{createParamCdf}
\alias{createParamCdf.ChipEffectFile}
\alias{createParamCdf.FirmaFile}
\alias{createUnique}
\alias{createUniqueCdf}
\alias{CrlmmModel}
\alias{decode}
\alias{decode.ParameterCelFile}
\alias{dim.DChipDcpFile}
\alias{doCRMA}
\alias{doCRMA.default}
\alias{encode}
\alias{encode.ParameterCelFile}
\alias{excludeChrXFromFit}
\alias{excludeChrXFromFit.DChipQuantileNormalization}
\alias{exportAromaSignalBinaryFileList}
\alias{exportAromaSignalBinaryFileList.CnChipEffectFile}
\alias{exportTotalAndFracB}
\alias{exportTotalAndFracB.CnChipEffectFile}
\alias{exportTotalAndFracB.CnChipEffectSet}
\alias{exportTotalAndFracB.DChipDcpSet}
\alias{exportTotalAndFracB.SnpChipEffectFile}
\alias{exportTotalAndFracB.SnpChipEffectSet}
\alias{exportTotalCnRatioSet.AffymetrixCnChpSet}
\alias{extractAffyBatch}
\alias{extractChromosomalDataFrame}
\alias{extractChromosomalDataFrame.ChipEffectFile}
\alias{extractChromosomalDataFrame.ChipEffectSet}
\alias{extractCNT}
\alias{extractCNT.SnpChipEffectFile}
\alias{extractCNT.SnpChipEffectSet}
\alias{extractDataFrame}
\alias{extractDataFrame.ParameterCelFile}
\alias{extractLogRatios}
\alias{extractLogRatios.AffymetrixCnChpFile}
\alias{extractLogRatios.AffymetrixCnChpSet}
\alias{extractMatrix.AffymetrixCelFile}
\alias{extractMatrix.ChipEffectFile}
\alias{extractMatrix.ChipEffectSet}
\alias{extractMatrix.FirmaFile}
\alias{extractMatrix.FirmaSet}
\alias{extractMatrix.ParameterCelFile}
\alias{extractSnpCnvQSet}
\alias{extractSnpCnvQSet.SnpChipEffectSet}
\alias{extractSnpFeatureSet}
\alias{extractSnpFeatureSet.AffymetrixCelSet}
\alias{extractSnpQSet}
\alias{extractSnpQSet.SnpChipEffectSet}
\alias{extractTheta}
\alias{extractTheta.ChipEffectFile}
\alias{extractTheta.ChipEffectSet}
\alias{extractTheta.CnChipEffectFile}
\alias{extractTheta.CnChipEffectSet}
\alias{extractTheta.DChipDcpFile}
\alias{extractTheta.DChipDcpSet}
\alias{extractTheta.SnpChipEffectFile}
\alias{extractTheta.SnpChipEffectSet}
\alias{extractTheta.SnpCnvQSet}
\alias{extractTheta.SnpQSet}
\alias{extractTotalAndFracB}
\alias{extractTotalAndFracB.SnpChipEffectFile}
\alias{extractTotalAndFreqB}
\alias{extractTotalAndFreqB.CnChipEffectFile}
\alias{extractTotalAndFreqB.CnChipEffectSet}
\alias{extractTotalAndFreqB.default}
\alias{extractTotalAndFreqB.SnpChipEffectSet}
\alias{findByCdf2}
\alias{findByCdf2.default}
\alias{findByChipType.AffymetrixCsvFile}
\alias{findByChipType.AffymetrixCsvGenomeInformation}
\alias{findByChipType.AffymetrixNetAffxCsvFile}
\alias{findByChipType.AffymetrixProbeTabFile}
\alias{findByChipType.AffymetrixTabularFile}
\alias{findByChipType.AffymetrixTsvFile}
\alias{findByChipType.DChipCdfBinFile}
\alias{findByChipType.DChipGenomeInformation}
\alias{findByChipType.DChipSnpInformation}
\alias{findByChipType.UflSnpInformation}
\alias{findByChipType.UgpGenomeInformation}
\alias{findByName.AffymetrixCelSet}
\alias{findByName.AffymetrixCnChpSet}
\alias{findByName.ChipEffectSet}
\alias{findByName.CnagCfhSet}
\alias{findByName.DChipDcpSet}
\alias{findUnitsTodo.AlleleSummation}
\alias{findUnitsTodo.ChipEffectFile}
\alias{findUnitsTodo.ChipEffectSet}
\alias{findUnitsTodo.CrlmmModel}
\alias{findUnitsTodo.CrlmmParametersFile}
\alias{findUnitsTodo.CrlmmParametersSet}
\alias{findUnitsTodo.ExonChipEffectSet}
\alias{findUnitsTodo.FirmaFile}
\alias{findUnitsTodo.FirmaSet}
\alias{findUnitsTodo.QualityAssessmentFile}
\alias{findUnitsTodo.ResidualFile}
\alias{findUnitsTodo.ResidualSet}
\alias{findUnitsTodo.WeightsFile}
\alias{findUnitsTodo.WeightsSet}
\alias{fit.CrlmmModel}
\alias{fitCnProbes}
\alias{fitCnProbes.UnitModel}
\alias{fitOne}
\alias{fitOne.AbstractProbeSequenceNormalization}
\alias{fitOne.BaseCountNormalization}
\alias{fitOne.LinearModelProbeSequenceNormalization}
\alias{fitOne.MatNormalization}
\alias{fitOne.ReseqCrosstalkCalibration}
\alias{fitOne.ScaleNormalization3}
\alias{fitOne.UnitTypeScaleNormalization}
\alias{fitOneArray}
\alias{fitOneArray.SingleArrayUnitModel}
\alias{fitOneChromosome}
\alias{fitOneChromosome.SmoothMultiarrayModel}
\alias{fitPlasqUnit}
\alias{fitPlasqUnit.matrix}
\alias{fitQuantileNormFcn}
\alias{fixSearchPath.AromaAffymetrix}
\alias{fromCdf}
\alias{fromCdf.AffymetrixProbeTabFile}
\alias{fromChipType}
\alias{fromChipType.AffymetrixCsvGenomeInformation}
\alias{fromChipType.AffymetrixProbeTabFile}
\alias{fromChipType.AffymetrixTabularFile}
\alias{fromChipType.AffymetrixTsvFile}
\alias{fromChipType.AromaChipTypeAnnotationFile}
\alias{fromChipType.DChipGenomeInformation}
\alias{fromChipType.DChipSnpInformation}
\alias{fromChipType.GenomeInformation}
\alias{fromChipType.SnpInformation}
\alias{fromChipType.UflSnpInformation}
\alias{fromChipType.UgpGenomeInformation}
\alias{fromDataFile}
\alias{fromDataFile.FirmaFile}
\alias{fromDataFile.ResidualFile}
\alias{fromDataFile.WeightsFile}
\alias{fromDataSet}
\alias{fromDataSet.ChipEffectSet}
\alias{fromDataSet.FirmaSet}
\alias{fromDataSet.GenomeInformation}
\alias{fromDataSet.ResidualSet}
\alias{fromDataSet.SnpInformation}
\alias{fromDataSet.WeightsSet}
\alias{fromFile.AffymetrixCnChpFile}
\alias{fromFile.DChipCdfBinFile}
\alias{fromFile.DChipDcpFile}
\alias{fromName}
\alias{fromName.AffymetrixCelSet}
\alias{fromName.CnagCfhSet}
\alias{fromName.DChipDcpSet}
\alias{getAlias.ArrayExplorer}
\alias{getAlias.TransformReport}
\alias{getAlleleCellPairs}
\alias{getAlleleProbePairs}
\alias{getAlleleProbePairs2}
\alias{getAlleleProbePairs2.AffymetrixCdfFile}
\alias{getAlleleProbePairs3}
\alias{getAm}
\alias{getAm.AffymetrixCelFile}
\alias{getAromaCellMatchScoreFile}
\alias{getAromaCellMatchScoreFile.MatNormalization}
\alias{getAromaCellPositionFile}
\alias{getAromaCellPositionFile.MatSmoothing}
\alias{getAromaCellSequenceFile}
\alias{getAromaCellSequenceFile.AbstractProbeSequenceNormalization}
\alias{getAromaUfcFile}
\alias{getAromaUfcFile.FragmentEquivalentClassNormalization}
\alias{getAromaUflFile.UflSnpInformation}
\alias{getAromaUgpFile.UgpGenomeInformation}
\alias{getAromaUnitFracBCnBinarySet}
\alias{getAromaUnitFracBCnBinarySet.default}
\alias{getAromaUnitTotalCnBinarySet}
\alias{getAromaUnitTotalCnBinarySet.default}
\alias{getArrayNames}
\alias{getArrayNames.AffymetrixAptSummaryFile}
\alias{getArraysOfInput.ArrayExplorer}
\alias{getAsFullCelFile}
\alias{getAsFullCelFile.ChipEffectFile}
\alias{getAsFullCelSet}
\alias{getAsFullCelSet.ChipEffectSet}
\alias{getAsteriskTags.AdditiveCovariatesNormalization}
\alias{getAsteriskTags.AffineCnPlm}
\alias{getAsteriskTags.AffinePlm}
\alias{getAsteriskTags.AffineSnpPlm}
\alias{getAsteriskTags.AlleleSummation}
\alias{getAsteriskTags.AllelicCrosstalkCalibration}
\alias{getAsteriskTags.ArrayExplorer}
\alias{getAsteriskTags.AvgCnPlm}
\alias{getAsteriskTags.AvgPlm}
\alias{getAsteriskTags.AvgSnpPlm}
\alias{getAsteriskTags.BaseCountNormalization}
\alias{getAsteriskTags.BasePositionNormalization}
\alias{getAsteriskTags.CrlmmModel}
\alias{getAsteriskTags.ExonRmaPlm}
\alias{getAsteriskTags.FirmaModel}
\alias{getAsteriskTags.FragmentEquivalentClassNormalization}
\alias{getAsteriskTags.FragmentLengthNormalization}
\alias{getAsteriskTags.GenericReporter}
\alias{getAsteriskTags.HetLogAddCnPlm}
\alias{getAsteriskTags.HetLogAddPlm}
\alias{getAsteriskTags.HetLogAddSnpPlm}
\alias{getAsteriskTags.LimmaBackgroundCorrection}
\alias{getAsteriskTags.MatNormalization}
\alias{getAsteriskTags.MbeiCnPlm}
\alias{getAsteriskTags.MbeiPlm}
\alias{getAsteriskTags.MbeiSnpPlm}
\alias{getAsteriskTags.Model}
\alias{getAsteriskTags.NormExpBackgroundCorrection}
\alias{getAsteriskTags.ProbeLevelModel}
\alias{getAsteriskTags.ProbeLevelTransform3}
\alias{getAsteriskTags.QualityAssessmentModel}
\alias{getAsteriskTags.ReseqCrosstalkCalibration}
\alias{getAsteriskTags.RmaCnPlm}
\alias{getAsteriskTags.RmaPlm}
\alias{getAsteriskTags.RmaSnpPlm}
\alias{getAsteriskTags.ScaleNormalization}
\alias{getAsteriskTags.SmoothMultiarrayModel}
\alias{getAsteriskTags.SmoothRmaModel}
\alias{getAsteriskTags.SmoothSaModel}
\alias{getAsteriskTags.SnpChipEffectGroupMerge}
\alias{getAsteriskTags.UnitModel}
\alias{getAverage}
\alias{getAverage.AffymetrixCelSet}
\alias{getAverage.CnagCfhSet}
\alias{getAverageAsinh}
\alias{getAverageAsinh.AffymetrixCelSet}
\alias{getAverageAsinh.CnagCfhSet}
\alias{getAverageFile.ChipEffectSet}
\alias{getAverageFile.CnChipEffectSet}
\alias{getAverageFile.ExonChipEffectSet}
\alias{getAverageFile.ResidualSet}
\alias{getAverageFile.SnpChipEffectSet}
\alias{getAverageFile.WeightsSet}
\alias{getAverageLog}
\alias{getAverageLog.AffymetrixCelSet}
\alias{getAverageLog.CnagCfhSet}
\alias{getBandwidth}
\alias{getBandwidth.SmoothMultiarrayModel}
\alias{getBaseline}
\alias{getBaseline.CnChipEffectSet}
\alias{getBaseline.SnpChipEffectSet}
\alias{getBlockSizes}
\alias{getBlockSizes.SpatialRowColumnNormalization}
\alias{getCalculateResidualsFunction}
\alias{getCalculateResidualsFunction.AvgPlm}
\alias{getCalculateResidualsFunction.ProbeLevelModel}
\alias{getCalculateResidualsFunction.RmaPlm}
\alias{getCalls}
\alias{getCalls.DChipDcpFile}
\alias{getCallSet}
\alias{getCallSet.CrlmmModel}
\alias{getCdf}
\alias{getCdf.AdditiveCovariatesNormalization}
\alias{getCdf.AffymetrixCnChpFile}
\alias{getCdf.AffymetrixCnChpSet}
\alias{getCdf.AffymetrixProbeTabFile}
\alias{getCdf.AffymetrixTsvFile}
\alias{getCdf.AromaUnitTabularBinaryFile}
\alias{getCdf.ChipEffectGroupMerge}
\alias{getCdf.FirmaModel}
\alias{getCdf.FragmentEquivalentClassNormalization}
\alias{getCdf.FragmentLengthNormalization}
\alias{getCdf.GcContentNormalization}
\alias{getCdf.QualityAssessmentModel}
\alias{getCdf.TransformReport}
\alias{getCdfBin}
\alias{getCdfBin.DChipDcpSet}
\alias{getCellDim}
\alias{getCellDim.DChipCdfBinFile}
\alias{getCellIndices}
\alias{getCellIndices.ChipEffectFile}
\alias{getCellIndices.ChipEffectSet}
\alias{getCellIndices.CnChipEffectFile}
\alias{getCellIndices.CnPlm}
\alias{getCellIndices.CnProbeAffinityFile}
\alias{getCellIndices.ExonChipEffectFile}
\alias{getCellIndices.ExonProbeAffinityFile}
\alias{getCellIndices.ExonRmaPlm}
\alias{getCellIndices.FirmaFile}
\alias{getCellIndices.FirmaSet}
\alias{getCellIndices.ProbeAffinityFile}
\alias{getCellIndices.ResidualFile}
\alias{getCellIndices.ResidualSet}
\alias{getCellIndices.SnpChipEffectFile}
\alias{getCellIndices.SnpPlm}
\alias{getCellIndices.SnpProbeAffinityFile}
\alias{getCellIndices.WeightsFile}
\alias{getCellIndices.WeightsSet}
\alias{getCellMap}
\alias{getCellMap.FirmaFile}
\alias{getCellMap.ResidualFile}
\alias{getCellMap.WeightsFile}
\alias{getCellMapForMainCdf}
\alias{getCellMapForMainCdf.ChipEffectFile}
\alias{getCellQuartets}
\alias{getCellQuartets.AffymetrixCdfFile}
\alias{getCellsTo}
\alias{getCellsTo.ProbeLevelTransform3}
\alias{getCellsToFit}
\alias{getCellsToFit.ProbeLevelTransform3}
\alias{getCellsToInternal}
\alias{getCellsToInternal.ProbeLevelTransform3}
\alias{getCellsToUpdate}
\alias{getCellsToUpdate.ProbeLevelTransform3}
\alias{getChipEffectFileClass}
\alias{getChipEffectFileClass.ChipEffectSet}
\alias{getChipEffectFileClass.CnChipEffectSet}
\alias{getChipEffectFileClass.ExonChipEffectSet}
\alias{getChipEffectFileClass.SnpChipEffectSet}
\alias{getChipEffects}
\alias{getChipEffects.ExonRmaPlm}
\alias{getChipEffects.QualityAssessmentModel}
\alias{getChipEffectSet}
\alias{getChipEffectSet.CnPlm}
\alias{getChipEffectSet.ExonRmaPlm}
\alias{getChipEffectSet.QualityAssessmentModel}
\alias{getChipEffectSet.SnpPlm}
\alias{getChipEffectSetClass}
\alias{getChipEffectSetClass.AlleleSummation}
\alias{getChipEffectSetClass.CnPlm}
\alias{getChipEffectSetClass.ExonRmaPlm}
\alias{getChipEffectSetClass.ProbeLevelModel}
\alias{getChipEffectSetClass.SnpPlm}
\alias{getChipType.AffymetrixAptSummaryFile}
\alias{getChipType.AffymetrixCdfFile}
\alias{getChipType.AffymetrixCelSetReporter}
\alias{getChipType.AffymetrixCnChpFile}
\alias{getChipType.AffymetrixProbeTabFile}
\alias{getChipType.AffymetrixTsvFile}
\alias{getChipType.AromaChipTypeAnnotationFile}
\alias{getChipType.CrlmmModel}
\alias{getChipType.DChipCdfBinFile}
\alias{getChipType.DChipDcpSet}
\alias{getChipType.UflSnpInformation}
\alias{getChipType.UgpGenomeInformation}
\alias{getChromosomes.GenomeInformation}
\alias{getChromosomes.UgpGenomeInformation}
\alias{getChromosomeStats}
\alias{getChromosomeStats.GenomeInformation}
\alias{getCnNames}
\alias{getColorMaps}
\alias{getColorMaps.ArrayExplorer}
\alias{getColorMaps.SpatialReporter}
\alias{getColumnNames.AffymetrixProbeTabFile}
\alias{getColumnNames.AromaCellMatchScoreFile}
\alias{getColumnNames.AromaUfcFile}
\alias{getColumnNames.AromaUnitGcContentFile}
\alias{getCombineAlleles}
\alias{getCombineAlleles.CnChipEffectSet}
\alias{getCombineAlleles.CnPlm}
\alias{getConfidenceScoreSet}
\alias{getConfidenceScoreSet.CrlmmModel}
\alias{getCovariates}
\alias{getCovariates.AdditiveCovariatesNormalization}
\alias{getCovariates.GcContentNormalization2}
\alias{getCrlmmParametersSet}
\alias{getCrlmmParametersSet.CrlmmModel}
\alias{getCrlmmPriors}
\alias{getCrlmmPriors.CrlmmModel}
\alias{getCrlmmSNPs}
\alias{getCrlmmSNPs.CrlmmModel}
\alias{getCrlmmSNPsOnChrX}
\alias{getCrlmmSNPsOnChrX.CrlmmModel}
\alias{getCrlmmSplineParameters}
\alias{getCrlmmSplineParameters.CrlmmModel}
\alias{getData}
\alias{getData.AffymetrixCelSet}
\alias{getData.AffymetrixCnChpFile}
\alias{getData.AffymetrixProbeTabFile}
\alias{getData.AffymetrixTsvFile}
\alias{getData.CnagCfhSet}
\alias{getData.UflSnpInformation}
\alias{getData.UgpGenomeInformation}
\alias{getDataColumns}
\alias{getDataColumns.GenomeInformation}
\alias{getDataColumns.UflSnpInformation}
\alias{getDataColumns.UgpGenomeInformation}
\alias{getDataFlat}
\alias{getDataFlat.ChipEffectFile}
\alias{getDataFlat.FirmaFile}
\alias{getDataFlat.ResidualFile}
\alias{getDataFlat.WeightsFile}
\alias{getDataPairs}
\alias{getDataPairs.AllelicCrosstalkCalibration}
\alias{getDataSet}
\alias{getDataSet.FirmaModel}
\alias{getDataSet.QualityAssessmentModel}
\alias{getDefaultFullName.AffymetrixFileSet}
\alias{getDefaultFullName.CnagCfhSet}
\alias{getDefaultSettings}
\alias{getDefaultSettings.AromaAffymetrix}
\alias{getDefaultSettings.Package}
\alias{getDesignMatrix}
\alias{getDesignMatrix.BaseCountNormalization}
\alias{getDesignMatrix.BasePositionNormalization}
\alias{getDesignMatrix.LinearModelProbeSequenceNormalization}
\alias{getDesignMatrix.MatNormalization}
\alias{getDimension}
\alias{getDimension.AffymetrixCdfFile}
\alias{getExclCells}
\alias{getExclCells.DChipQuantileNormalization}
\alias{getExclCells.QuantileNormalization}
\alias{getExcludes}
\alias{getExcludes.DChipDcpFile}
\alias{getExclusiveUnitSubsets}
\alias{getExclusiveUnitSubsets.FragmentEquivalentClassNormalization}
\alias{getExpandedCellMap}
\alias{getExpandedCellMap.ChipEffectFile}
\alias{getExpandedCellMap.CnChipEffectFile}
\alias{getExpandedCellMap.SnpChipEffectFile}
\alias{getExpectedOutputFiles.MatSmoothing}
\alias{getExpectedOutputFullnames.MatSmoothing}
\alias{getExtensionPattern.AffymetrixCdfFile}
\alias{getExtensionPattern.AffymetrixCelFile}
\alias{getExtensionPattern.AffymetrixCnChpFile}
\alias{getExtensionPattern.AffymetrixCsvFile}
\alias{getExtensionPattern.AffymetrixTsvFile}
\alias{getExtensionPattern.AromaCellMatchScoreFile}
\alias{getExtensionPattern.AromaUfcFile}
\alias{getExtensionPattern.AromaUnitGcContentFile}
\alias{getExtensionPattern.CnagCfhFile}
\alias{getExtensionPattern.DChipDcpFile}
\alias{getField}
\alias{getField.AffymetrixTsvFile}
\alias{getFields.SnpInformation}
\alias{getFields.UflSnpInformation}
\alias{getFileClass.FirmaSet}
\alias{getFileFormat}
\alias{getFileFormat.AffymetrixCdfFile}
\alias{getFileFormat.AffymetrixCelFile}
\alias{getFileFormat.AffymetrixCnChpFile}
\alias{getFileFormat.DChipCdfBinFile}
\alias{getFileFormat.DChipDcpFile}
\alias{getFilenameExtension.AromaCellMatchScoreFile}
\alias{getFilenameExtension.AromaUfcFile}
\alias{getFilenameExtension.AromaUnitGcContentFile}
\alias{getFileSet}
\alias{getFileSet.AffymetrixFileSetReporter}
\alias{getFileSetClass}
\alias{getFileSetClass.FirmaModel}
\alias{getFirmaScores}
\alias{getFirmaSet}
\alias{getFirstCellIndices}
\alias{getFirstCellIndices.AffymetrixCdfFile}
\alias{getFirstCellPerUnitIndices}
\alias{getFirstCellPerUnitIndices.ExonChipEffectSet}
\alias{getFit}
\alias{getFit.BasePositionNormalization}
\alias{getFitFunction.MultiArrayUnitModel}
\alias{getFitSingleCellUnitFunction}
\alias{getFitSingleCellUnitFunction.UnitModel}
\alias{getFitUnitFunction}
\alias{getFitUnitFunction.CnPlm}
\alias{getFitUnitFunction.FirmaModel}
\alias{getFitUnitFunction.MultiArrayUnitModel}
\alias{getFitUnitFunction.SingleArrayUnitModel}
\alias{getFitUnitFunction.UnitModel}
\alias{getFitUnitGroupFunction}
\alias{getFitUnitGroupFunction.SmoothMultiarrayModel}
\alias{getFitUnitGroupFunction.SmoothRmaModel}
\alias{getFitUnitGroupFunction.SmoothSaModel}
\alias{getFragmentLengths}
\alias{getFragmentLengths.AffymetrixTsvFile}
\alias{getFragmentLengths.SnpInformation}
\alias{getFragmentLengths.UflSnpInformation}
\alias{getFragmentStarts}
\alias{getFragmentStarts.SnpInformation}
\alias{getFragmentStarts.UflSnpInformation}
\alias{getFragmentStops}
\alias{getFragmentStops.SnpInformation}
\alias{getFragmentStops.UflSnpInformation}
\alias{getFullName.GenericReporter}
\alias{getFullName.QualityAssessmentModel}
\alias{getFullNames.ChipEffectSetTuple}
\alias{getGc}
\alias{getGc.AffymetrixTsvFile}
\alias{getGcContent}
\alias{getGcContent.GcContentNormalization}
\alias{getGenomeInformation}
\alias{getGroupDirections}
\alias{getGroupDirections.AffymetrixCdfFile}
\alias{getHeader.AffymetrixCnChpFile}
\alias{getHeader.DChipCdfBinFile}
\alias{getHeader.DChipDcpFile}
\alias{getHeaderLength}
\alias{getHeaderLength.CnagCfhFile}
\alias{getHeaderV3}
\alias{getHeaderV3.AffymetrixCelFile}
\alias{getIdentifier}
\alias{getIdentifier.AffymetrixCelFile}
\alias{getIdentifier.AffymetrixCelSet}
\alias{getIdentifier.AffymetrixFileSet}
\alias{getIdentifier.CnagCfhFile}
\alias{getIdentifier.CnagCfhSet}
\alias{getImage.AffymetrixCdfFile}
\alias{getImage.AromaUnitTabularBinaryFile}
\alias{getImage.ResidualFile}
\alias{getImage.WeightsFile}
\alias{getIndexToRowMap}
\alias{getIndexToRowMap.AffymetrixProbeTabFile}
\alias{getInputName}
\alias{getInputName.AffymetrixFileSetReporter}
\alias{getInputName.GenericReporter}
\alias{getInputTags}
\alias{getInputTags.AffymetrixFileSetReporter}
\alias{getInputTags.GenericReporter}
\alias{getIntensities}
\alias{getLabel.Model}
\alias{getListOfCdfs}
\alias{getListOfCdfs.AffymetrixCelSetTuple}
\alias{getListOfGenomeInformations}
\alias{getListOfGenomeInformations.ChromosomalModel}
\alias{getListOfPriors}
\alias{getListOfPriors.MultiArrayUnitModel}
\alias{getListOfReporters}
\alias{getListOfReporters.ArrayExplorer}
\alias{getListOfUnitNamesFiles.AffymetrixCelSetTuple}
\alias{getListOfUnitNamesFiles.ArrayExplorer}
\alias{getListOfUnitTypesFiles.AffymetrixCelSetTuple}
\alias{getListOfUnitTypesFiles.ArrayExplorer}
\alias{getMainCdf}
\alias{getMainCdf.AffymetrixCdfFile}
\alias{getMainPath.GenericReporter}
\alias{getMatrixChipEffectFiles}
\alias{getMatrixChipEffectFiles.CopyNumberChromosomalModel}
\alias{getMergeFunction}
\alias{getMergeFunction.ChipEffectGroupMerge}
\alias{getMergeFunction.SnpChipEffectGroupMerge}
\alias{getMergeGroups}
\alias{getMergeGroups.ExonChipEffectSet}
\alias{getMergeStrands}
\alias{getMergeStrands.SnpChipEffectSet}
\alias{getMergeStrands.SnpPlm}
\alias{getMonoCell}
\alias{getMonoCell.AffymetrixCdfFile}
\alias{getMonocellCdf}
\alias{getMonocellCdf.AffymetrixCdfFile}
\alias{getName.AffymetrixPlatform}
\alias{getName.FirmaModel}
\alias{getName.QualityAssessmentModel}
\alias{getNameOfInput.ArrayExplorer}
\alias{getNormalEquations}
\alias{getNormalEquations.LinearModelProbeSequenceNormalization}
\alias{getNormalizedIntensities}
\alias{getNormalizedIntensities.DChipDcpFile}
\alias{getNormalizeSignalsOne}
\alias{getNormalizeSignalsOne.ScaleNormalization3}
\alias{getNormalizeSignalsOne.UnitTypeScaleNormalization}
\alias{getNumberOfFilesAveraged.CnChipEffectFile}
\alias{getOrderedFragmentPairMap}
\alias{getOrderedFragmentPairMap.AromaUfcFile}
\alias{getOrderedFragmentPairs}
\alias{getOrderedFragmentPairs.AromaUfcFile}
\alias{getOutputDataSet00}
\alias{getOutputDataSet00.AdditiveCovariatesNormalization}
\alias{getOutputDataSet00.FragmentEquivalentClassNormalization}
\alias{getOutputDataSet00.FragmentLengthNormalization}
\alias{getOutputDataSet00.GcContentNormalization}
\alias{getOutputDataSetOLD20090509}
\alias{getOutputDataSetOLD20090509.Transform}
\alias{getOutputFiles.Transform}
\alias{getOutputIdentifier}
\alias{getOutputIdentifier.QuantileNormalization}
\alias{getOutputTuple}
\alias{getOutputTuple.SmoothMultiarrayModel}
\alias{getPackageName}
\alias{getParameter.AffymetrixAptSummaryFile}
\alias{getParameters.AbstractProbeSequenceNormalization}
\alias{getParameters.AdditiveCovariatesNormalization}
\alias{getParameters.AllelicCrosstalkCalibration}
\alias{getParameters.BackgroundCorrection}
\alias{getParameters.BaseCountNormalization}
\alias{getParameters.BasePositionNormalization}
\alias{getParameters.ChipEffectFile}
\alias{getParameters.ChipEffectGroupMerge}
\alias{getParameters.ChipEffectSet}
\alias{getParameters.CnChipEffectFile}
\alias{getParameters.DChipQuantileNormalization}
\alias{getParameters.ExonChipEffectFile}
\alias{getParameters.FragmentEquivalentClassNormalization}
\alias{getParameters.FragmentLengthNormalization}
\alias{getParameters.GcContentNormalization}
\alias{getParameters.GcRmaBackgroundCorrection}
\alias{getParameters.LimmaBackgroundCorrection}
\alias{getParameters.MatNormalization}
\alias{getParameters.MatSmoothing}
\alias{getParameters.Model}
\alias{getParameters.NormExpBackgroundCorrection}
\alias{getParameters.OpticalBackgroundCorrection}
\alias{getParameters.ParameterCelFile}
\alias{getParameters.ProbeAffinityFile}
\alias{getParameters.ProbeLevelTransform3}
\alias{getParameters.QuantileNormalization}
\alias{getParameters.ReseqCrosstalkCalibration}
\alias{getParameters.ResidualFile}
\alias{getParameters.ResidualSet}
\alias{getParameters.RmaBackgroundCorrection}
\alias{getParameters.ScaleNormalization}
\alias{getParameters.ScaleNormalization3}
\alias{getParameters.SnpChipEffectFile}
\alias{getParameters.SnpChipEffectGroupMerge}
\alias{getParameters.SpatialRowColumnNormalization}
\alias{getParameters.UnitTypeScaleNormalization}
\alias{getParameters.WeightsFile}
\alias{getParameters.WeightsSet}
\alias{getParametersAsString.ChipEffectFile}
\alias{getParametersAsString.ChipEffectSet}
\alias{getParametersAsString.Model}
\alias{getParametersAsString.ProbeAffinityFile}
\alias{getParametersAsString.ResidualFile}
\alias{getParametersAsString.ResidualSet}
\alias{getParametersAsString.WeightsFile}
\alias{getParametersAsString.WeightsSet}
\alias{getParameterSet}
\alias{getParameterSet.AvgPlm}
\alias{getParameterSet.CnPlm}
\alias{getParameterSet.CrlmmModel}
\alias{getParameterSet.ExonRmaPlm}
\alias{getParameterSet.Model}
\alias{getParameterSet.RmaPlm}
\alias{getParameterSet.SnpPlm}
\alias{getParameterSet.UnitModel}
\alias{getPath.AffymetrixCelSetReporter}
\alias{getPath.GenericReporter}
\alias{getPath.QualityAssessmentModel}
\alias{getPcuTheta}
\alias{getPcuTheta.ChromosomalModel}
\alias{getPlasqTypes}
\alias{getPlatform.AffymetrixCdfFile}
\alias{getPlatform.AffymetrixCelFile}
\alias{getPlatform.AffymetrixCelSet}
\alias{getPlatform.AromaChipTypeAnnotationFile}
\alias{getPlatform.DChipCdfBinFile}
\alias{getPlatformDesignDB}
\alias{getPlatformDesignDB.CrlmmModel}
\alias{getPlm}
\alias{getPlm.FirmaModel}
\alias{getPlm.QualityAssessmentModel}
\alias{getPosition.AffymetrixTsvFile}
\alias{getPositionChipTypeUnit}
\alias{getPositionChipTypeUnit.ChromosomalModel}
\alias{getPositionChipTypeUnit.CopyNumberSegmentationModel}
\alias{getProbeAffinities}
\alias{getProbeAffinities.ExonRmaPlm}
\alias{getProbeAffinityFile}
\alias{getProbeAffinityFile.AffinePlm}
\alias{getProbeAffinityFile.AvgPlm}
\alias{getProbeAffinityFile.CnPlm}
\alias{getProbeAffinityFile.ExonRmaPlm}
\alias{getProbeAffinityFile.MbeiPlm}
\alias{getProbeAffinityFile.RmaPlm}
\alias{getProbeAffinityFile.SnpPlm}
\alias{getProbeSequenceData}
\alias{getProbeSequenceData.AffymetrixCdfFile}
\alias{getProbesetIds}
\alias{getProbesetIds.AffymetrixAptSummaryFile}
\alias{getQuantificationScale}
\alias{getQuantificationScale.AffymetrixAptSummaryFile}
\alias{getRawIntensities}
\alias{getRawIntensities.DChipDcpFile}
\alias{getReadArguments.AffymetrixAptSummaryFile}
\alias{getRectangle}
\alias{getRectangle.AffymetrixCelFile}
\alias{getReference}
\alias{getReference.SpatialReporter}
\alias{getRegions.profileCGH}
\alias{getReportSet}
\alias{getReportSet.GenericReporter}
\alias{getReportSet.SpatialReporter}
\alias{getResidualFileClass}
\alias{getResidualFileClass.ResidualSet}
\alias{getResiduals}
\alias{getResidualSet}
\alias{getResidualSet.ProbeLevelModel}
\alias{getResidualSetClass}
\alias{getResidualSetClass.ProbeLevelModel}
\alias{getRestructor}
\alias{getRestructor.AffymetrixCdfFile}
\alias{getRlmFitFunctions}
\alias{getRlmFitFunctions.RmaPlm}
\alias{getRootPath.AlleleSummation}
\alias{getRootPath.ChipEffectTransform}
\alias{getRootPath.CrlmmModel}
\alias{getRootPath.FirmaModel}
\alias{getRootPath.GenericReporter}
\alias{getRootPath.ProbeLevelModel}
\alias{getRootPath.ProbeLevelTransform}
\alias{getRootPath.QualityAssessmentModel}
\alias{getRootPath.SmoothMultiarrayModel}
\alias{getRootPath.SmoothRmaModel}
\alias{getRootPath.TransformReport}
\alias{getSetsOfProbes}
\alias{getSetsOfProbes.AllelicCrosstalkCalibration}
\alias{getSetsOfProbes.ReseqCrosstalkCalibration}
\alias{getSettings}
\alias{getSettings.Package}
\alias{getSetTuple.ArrayExplorer}
\alias{getSignalTransform}
\alias{getSignalTransform.BasePositionNormalization}
\alias{getSignalTransform.LinearModelProbeSequenceNormalization}
\alias{getSnpInformation}
\alias{getSnpInformation.AffymetrixCdfFile}
\alias{getSnpNames}
\alias{getSpar}
\alias{getSpar.SpatialRowColumnNormalization}
\alias{getSteps}
\alias{getSteps.AromaPipeline}
\alias{getSteps.ASCRMAv2}
\alias{getSubsetOfCellIndices}
\alias{getSubsetOfCellIndices.AffymetrixCdfFile}
\alias{getSubsetOfUnits}
\alias{getSubsetOfUnits.AffymetrixCdfFile}
\alias{getSubsetToAvg}
\alias{getSubsetToAvg.AllelicCrosstalkCalibration}
\alias{getSubsetToAvg.QuantileNormalization}
\alias{getSubsetToAvg.ReseqCrosstalkCalibration}
\alias{getSubsetToAvg.ScaleNormalization}
\alias{getSubsetToFit}
\alias{getSubsetToFit.AdditiveCovariatesNormalization}
\alias{getSubsetToFit.FragmentEquivalentClassNormalization}
\alias{getSubsetToFit.FragmentLengthNormalization}
\alias{getSubsetToFit.GcContentNormalization}
\alias{getSubsetToUpdate}
\alias{getSubsetToUpdate.BackgroundCorrection}
\alias{getSubsetToUpdate.DChipQuantileNormalization}
\alias{getSubsetToUpdate.QuantileNormalization}
\alias{getSubsetToUpdate.ScaleNormalization}
\alias{getSubsetToUpdate0}
\alias{getSubsetToUpdate0.LimmaBackgroundCorrection}
\alias{getTags.FirmaModel}
\alias{getTags.QualityAssessmentModel}
\alias{getTagsOfInput.ArrayExplorer}
\alias{getTargetAverages}
\alias{getTargetAverages.FragmentEquivalentClassNormalization}
\alias{getTargetDistribution}
\alias{getTargetDistribution.QuantileNormalization}
\alias{getTargetDistributionIdentifier}
\alias{getTargetDistributionIdentifier.QuantileNormalization}
\alias{getTargetDistributionPathname}
\alias{getTargetDistributionPathname.QuantileNormalization}
\alias{getTargetFile}
\alias{getTargetFile.AbstractProbeSequenceNormalization}
\alias{getTargetFunction}
\alias{getTargetFunction.GcContentNormalization}
\alias{getTargetFunctions}
\alias{getTargetFunctions.AdditiveCovariatesNormalization}
\alias{getTargetFunctions.FragmentLengthNormalization}
\alias{getThetas}
\alias{getThetas.DChipDcpFile}
\alias{getThetasAB}
\alias{getThetasAB.DChipDcpFile}
\alias{getThetaStds}
\alias{getThetaStds.DChipDcpFile}
\alias{getTimestamp}
\alias{getTimeStamp}
\alias{getTimeStamp.AffymetrixAptSummaryFile}
\alias{getTimestamp.AffymetrixCnChpFile}
\alias{getTimestamp.CnagCfhFile}
\alias{getTimestamps}
\alias{getTimestamps.AffymetrixCelSet}
\alias{getTimestamps.CnagCfhSet}
\alias{getUnique}
\alias{getUnique.AffymetrixCdfFile}
\alias{getUniqueCdf}
\alias{getUniqueCdf.AffymetrixCdfFile}
\alias{getUnitGroupCellArrayMap}
\alias{getUnitGroupCellArrayMap.ChipEffectFile}
\alias{getUnitGroupCellChromosomePositionMap}
\alias{getUnitGroupCellChromosomePositionMap.AffymetrixCdfFile}
\alias{getUnitGroupCellChromosomePositionMap.ChipEffectFile}
\alias{getUnitGroupCellMap}
\alias{getUnitGroupCellMap.AffymetrixCdfFile}
\alias{getUnitGroupCellMap.AffymetrixCelSet}
\alias{getUnitGroupCellMap.FirmaFile}
\alias{getUnitGroupCellMap.ResidualFile}
\alias{getUnitGroupCellMap.WeightsFile}
\alias{getUnitGroupCellMapWithMonocell}
\alias{getUnitGroupCellMapWithMonocell.AffymetrixCdfFile}
\alias{getUnitGroupCellMapWithUnique}
\alias{getUnitGroupCellMapWithUnique.AffymetrixCdfFile}
\alias{getUnitGroupCellMatrixMap}
\alias{getUnitGroupCellMatrixMap.ChipEffectFile}
\alias{getUnitGroupNamesFromUgcMap}
\alias{getUnitGroupNamesFromUgcMap.AffymetrixCdfFile}
\alias{getUnitIndices}
\alias{getUnitIntensities}
\alias{getUnitNames.AffymetrixNetAffxCsvFile}
\alias{getUnitNames.DChipCdfBinFile}
\alias{getUnitNamesFile.AffymetrixCdfFile}
\alias{getUnitNamesFile.AffymetrixCelFile}
\alias{getUnitNamesFile.AffymetrixCelSet}
\alias{getUnitNamesFile.AffymetrixCnChpFile}
\alias{getUnitNamesFile.AffymetrixPlatform}
\alias{getUnitNamesFile.TransformReport}
\alias{getUnitReadMap}
\alias{getUnitReadMap.AffymetrixCnChpFile}
\alias{getUnitSizes}
\alias{getUnitSizes.DChipCdfBinFile}
\alias{getUnitsOnChromosome.GenomeInformation}
\alias{getUnitsOnChromosome.UgpGenomeInformation}
\alias{getUnitsOnChromosomes.GenomeInformation}
\alias{getUnitsTo}
\alias{getUnitsTo.ProbeLevelTransform3}
\alias{getUnitsToFit}
\alias{getUnitsToFit.CrlmmModel}
\alias{getUnitsToFit.ProbeLevelTransform3}
\alias{getUnitsToUpdate}
\alias{getUnitsToUpdate.ProbeLevelTransform3}
\alias{getUnitTypesFile.AffymetrixCdfFile}
\alias{getUnitTypesFile.AffymetrixCelFile}
\alias{getUnitTypesFile.AffymetrixCelSet}
\alias{getUnitTypesFile.AffymetrixCnChpFile}
\alias{getUnitTypesFile.AffymetrixPlatform}
\alias{getUnitTypesFile.TransformReport}
\alias{getWeightsFileClass}
\alias{getWeightsFileClass.WeightsSet}
\alias{getWeightsSet}
\alias{getWeightsSet.ProbeLevelModel}
\alias{getWeightsSetClass}
\alias{getWeightsSetClass.ProbeLevelModel}
\alias{getYY}
\alias{getYY.TransformReport}
\alias{hasAlleleBFractions.CnChipEffectFile}
\alias{hasAlleleBFractions.CnChipEffectSet}
\alias{hasAlleleBFractions.CnChipEffectSetTuple}
\alias{hasColumnHeader.AffymetrixProbeTabFile}
\alias{hasMbeiData}
\alias{hasMbeiData.DChipDcpFile}
\alias{hasNormalizedData}
\alias{hasNormalizedData.DChipDcpFile}
\alias{hasStrandiness.CnChipEffectFile}
\alias{hasStrandiness.CnChipEffectSet}
\alias{hasStrandiness.CnChipEffectSetTuple}
\alias{hasUnitReadMap}
\alias{hasUnitReadMap.AffymetrixCnChpFile}
\alias{hasUnitTypes}
\alias{hasUnitTypes.AffymetrixCdfFile}
\alias{highlight}
\alias{highlight.AffymetrixCelFile}
\alias{identifyCells}
\alias{identifyCells.AffymetrixCdfFile}
\alias{image270}
\alias{importFrom.AromaUnitTabularBinaryFile}
\alias{importFromAffymetrixNetAffxCsvFile}
\alias{importFromAffymetrixNetAffxCsvFile.AromaUflFile}
\alias{importFromAffymetrixNetAffxCsvFile.AromaUgpFile}
\alias{importFromAffymetrixNetAffxCsvFile.AromaUnitGcContentFile}
\alias{importFromAffymetrixNetAffxCsvFile.AromaUnitTabularBinaryFile}
\alias{importFromAffymetrixProbeTabFile}
\alias{importFromAffymetrixProbeTabFile.AromaCellSequenceFile}
\alias{importFromAffymetrixTabularFile}
\alias{importFromAffymetrixTabularFile.AromaUfcFile}
\alias{importFromAffymetrixTabularFile.AromaUgpFile}
\alias{importFromAffymetrixTabularFile.AromaUnitTabularBinaryFile}
\alias{importFromApt}
\alias{importFromBpMap}
\alias{importFromBpMap.AromaCellMatchScoreFile}
\alias{importFromBpMap.AromaCellSequenceFile}
\alias{importFromDChip}
\alias{importFromDChipGenomeInformation}
\alias{importFromDChipGenomeInformation.AromaUnitTabularBinaryFile}
\alias{importFromGenericTabularFile.AromaUnitTabularBinaryFile}
\alias{importFromGenomeInformation}
\alias{importFromGenomeInformation.AromaUgpFile}
\alias{importFromGenomeInformation.AromaUnitTabularBinaryFile}
\alias{indexOfMissingSequences}
\alias{indexOfMissingSequences.AbstractProbeSequenceNormalization}
\alias{inferMmFromPm}
\alias{inferMmFromPm.AromaCellSequenceFile}
\alias{inferParameters}
\alias{inferParameters.CnChipEffectSet}
\alias{inferParameters.SnpChipEffectSet}
\alias{isCompatibleWithCdf}
\alias{isCompatibleWithCdf.GenomeInformation}
\alias{isCompatibleWithCdf.SnpInformation}
\alias{isCompatibleWithCdf.UflSnpInformation}
\alias{isCompatibleWithCdf.UgpGenomeInformation}
\alias{isDuplicated}
\alias{isMissing.AromaCellMatchScoreFile}
\alias{isMonocellCdf}
\alias{isMonocellCdf.AffymetrixCdfFile}
\alias{isPm}
\alias{isResequenceChip}
\alias{isSnpChip}
\alias{isUniqueCdf}
\alias{isUniqueCdf.AffymetrixCdfFile}
\alias{isUnitGroupCellMap}
\alias{isUnitGroupCellMap.data.frame}
\alias{isUnitGroupCellMap.default}
\alias{isUnitGroupCellMap.matrix}
\alias{justSNPRMA}
\alias{justSNPRMA.AffymetrixCelSet}
\alias{justSNPRMA.character}
\alias{mapToUnitNamesFile}
\alias{mapToUnitNamesFile.DChipCdfBinFile}
\alias{mergeGroups}
\alias{mergeGroups.ChipEffectFile}
\alias{mergeStrands}
\alias{mergeStrands.CnChipEffectFile}
\alias{mergeStrands.SnpChipEffectFile}
\alias{migrateTool}
\alias{migrateTool.CopyNumberSegmentationModel}
\alias{nbrOfArrays.AffymetrixAptSummaryFile}
\alias{nbrOfArrays.AffymetrixCelSetReporter}
\alias{nbrOfArrays.QualityAssessmentModel}
\alias{nbrOfArrays.TransformReport}
\alias{nbrOfCells.AffymetrixCdfFile}
\alias{nbrOfCells.AffymetrixCelFile}
\alias{nbrOfCells.CnagCfhFile}
\alias{nbrOfCells.DChipCdfBinFile}
\alias{nbrOfCells.DChipDcpFile}
\alias{nbrOfCellsPerUnit}
\alias{nbrOfCellsPerUnit.AffymetrixCdfFile}
\alias{nbrOfCellsPerUnitGroup}
\alias{nbrOfCellsPerUnitGroup.AffymetrixCdfFile}
\alias{nbrOfChipTypes.ArrayExplorer}
\alias{nbrOfColumns.AffymetrixCdfFile}
\alias{nbrOfEnzymes.AromaUfcFile}
\alias{nbrOfEnzymes.SnpInformation}
\alias{nbrOfEnzymes.UflSnpInformation}
\alias{nbrOfGroupsPerUnit}
\alias{nbrOfQcUnits}
\alias{nbrOfQcUnits.AffymetrixCdfFile}
\alias{nbrOfRows.AffymetrixCdfFile}
\alias{nbrOfSnps}
\alias{nbrOfSnps.CnagCfhFile}
\alias{nbrOfSteps}
\alias{nbrOfSteps.AromaPipeline}
\alias{nbrOfUnits.AffymetrixCdfFile}
\alias{nbrOfUnits.DChipCdfBinFile}
\alias{nbrOfUnits.DChipDcpFile}
\alias{nbrOfUnits.GenomeInformation}
\alias{nbrOfUnits.SnpInformation}
\alias{nbrOfUnits.UflSnpInformation}
\alias{nbrOfUnits.UgpGenomeInformation}
\alias{normalizeOneArrayVector}
\alias{normalizeOneArrayVector.FragmentEquivalentClassNormalization}
\alias{patch.AromaAffymetrix}
\alias{plasqGaussian}
\alias{plotAllelePairs}
\alias{plotAllelePairs.AllelicCrosstalkCalibration}
\alias{plotBasepair}
\alias{plotBasepair.AllelicCrosstalkCalibration}
\alias{plotBoxplot}
\alias{plotBoxplot.ChipEffectSet}
\alias{plotBoxplotStats}
\alias{plotCovariateEffects}
\alias{plotCovariateEffects.GcContentNormalization2}
\alias{plotImage}
\alias{plotMargins}
\alias{plotMargins.SpatialReporter}
\alias{plotMvsX}
\alias{plotNuse}
\alias{plotNuse.QualityAssessmentModel}
\alias{plotRle}
\alias{plotRle.QualityAssessmentModel}
\alias{plotXYCurve.TransformReport}
\alias{plotXYCurveLog2}
\alias{plotXYCurveLog2.TransformReport}
\alias{predictOne}
\alias{predictOne.AbstractProbeSequenceNormalization}
\alias{predictOne.BaseCountNormalization}
\alias{predictOne.LinearModelProbeSequenceNormalization}
\alias{predictOne.MatNormalization}
\alias{process.AlleleSummation}
\alias{process.AromaPipeline}
\alias{process.SpatialRowColumnNormalization}
\alias{range.AffymetrixCelFile}
\alias{range.AffymetrixCelSet}
\alias{range.CnagCfhFile}
\alias{range.CnagCfhSet}
\alias{read10K}
\alias{read10K.DChipSnpInformation}
\alias{read250K}
\alias{read250K.DChipSnpInformation}
\alias{read250KHg17}
\alias{read250KHg17.DChipGenomeInformation}
\alias{read50K}
\alias{read50K.DChipSnpInformation}
\alias{read50KHg17}
\alias{read50KHg17.DChipGenomeInformation}
\alias{readArrays}
\alias{readArrays.AffymetrixAptSummaryFile}
\alias{readCdfGroupStrands}
\alias{readCfhHeader}
\alias{readCfhHeader.default}
\alias{readCfhUnits}
\alias{readCfhUnits.default}
\alias{readCfnHeader}
\alias{readCfnHeader.default}
\alias{readCfnUnits}
\alias{readCfnUnits.default}
\alias{readData.AffymetrixTsvFile}
\alias{readData.GenomeInformation}
\alias{readData.SnpInformation}
\alias{readDataFrame.AffymetrixCdfFile}
\alias{readDataFrame.AffymetrixCsvGenomeInformation}
\alias{readDataFrame.AffymetrixTsvFile}
\alias{readDataFrame.AromaUfcFile}
\alias{readDataFrame.DChipCdfBinFile}
\alias{readDataFrame.DChipGenomeInformation}
\alias{readDataFrame.DChipSnpInformation}
\alias{readDataFrame.GenomeInformation}
\alias{readDataFrame.SnpInformation}
\alias{readDataFrame.UflSnpInformation}
\alias{readDataFrame.UgpGenomeInformation}
\alias{readDataFrame2}
\alias{readDataFrame2.AffymetrixProbeTabFile}
\alias{readDataUnitChromosomePosition}
\alias{readDataUnitChromosomePosition.AffymetrixNetAffxCsvFile}
\alias{readDataUnitFragmentLength}
\alias{readDataUnitFragmentLength.AffymetrixNetAffxCsvFile}
\alias{readGenomeWideSNP}
\alias{readGenomeWideSNP.DChipGenomeInformation}
\alias{readHeader.AffymetrixAptSummaryFile}
\alias{readMatchScores}
\alias{readMatchScores.AromaCellMatchScoreFile}
\alias{readMouse430}
\alias{readMouse430.DChipGenomeInformation}
\alias{readMouse430KenHardwired}
\alias{readMouse430KenHardwired.DChipGenomeInformation}
\alias{readParameter}
\alias{readParameter.CrlmmParametersFile}
\alias{readPlasqTypes}
\alias{readPlasqTypes.AffymetrixCdfFile}
\alias{readPriorsByUnits}
\alias{readPriorsByUnits.MultiArrayUnitModel}
\alias{readProbesetIds}
\alias{readProbesetIds.AffymetrixAptSummaryFile}
\alias{readRawData}
\alias{readRawData.AffymetrixCnChpFile}
\alias{readRawDataRectangle}
\alias{readRawDataRectangle.AffymetrixCelFile}
\alias{readRawDataRectangle.SpatialReporter}
\alias{readSequenceDataFrame}
\alias{readSequenceDataFrame.AffymetrixProbeTabFile}
\alias{readTableInternal}
\alias{readTableInternal.GenomeInformation}
\alias{readTableInternal.SnpInformation}
\alias{readUnitNames}
\alias{readUnitNames.AffymetrixNetAffxCsvFile}
\alias{readUnits}
\alias{readUnits.AffymetrixCelSet}
\alias{readUnits.ChipEffectFile}
\alias{readUnits.ChipEffectSet}
\alias{readUnits.CnagCfhSet}
\alias{readUnits.CnChipEffectFile}
\alias{readUnits.ExonChipEffectFile}
\alias{readUnits.FirmaFile}
\alias{readUnits.FirmaSet}
\alias{readUnits.ParameterCelFile}
\alias{readUnits.ProbeAffinityFile}
\alias{readUnits.ResidualFile}
\alias{readUnits.ResidualSet}
\alias{readUnits.SnpChipEffectFile}
\alias{readUnits.WeightsFile}
\alias{readUnits.WeightsSet}
\alias{readUnitsByQuartets}
\alias{rescale.AllelicCrosstalkCalibration}
\alias{rescaleByAll}
\alias{rescaleByAll.AllelicCrosstalkCalibration}
\alias{rescaleByAll.ReseqCrosstalkCalibration}
\alias{rescaleByGroups}
\alias{rescaleByGroups.AllelicCrosstalkCalibration}
\alias{restruct}
\alias{restruct.AffymetrixCdfFile}
\alias{seq.TransformReport}
\alias{setAlias.ArrayExplorer}
\alias{setAlias.TransformReport}
\alias{setBandwidth}
\alias{setBandwidth.SmoothMultiarrayModel}
\alias{setCdf}
\alias{setCdf.AffymetrixCnChpFile}
\alias{setCdf.AffymetrixCnChpSet}
\alias{setColorMaps}
\alias{setColorMaps.ArrayExplorer}
\alias{setColorMaps.SpatialReporter}
\alias{setCombineAlleles}
\alias{setCombineAlleles.CnChipEffectSet}
\alias{setCombineAlleles.CnPlm}
\alias{setCombineAlleles.CnProbeAffinityFile}
\alias{setCustomFindCdf}
\alias{setDecodeFunction}
\alias{setDecodeFunction.ParameterCelFile}
\alias{setEncodeFunction}
\alias{setEncodeFunction.ParameterCelFile}
\alias{setGenomeInformation}
\alias{setGenomeInformation.AffymetrixCdfFile}
\alias{setLabel.Model}
\alias{setListOfPriors}
\alias{setListOfPriors.MultiArrayUnitModel}
\alias{setMergeFunction}
\alias{setMergeFunction.ChipEffectGroupMerge}
\alias{setMergeGroups}
\alias{setMergeGroups.ExonChipEffectSet}
\alias{setMergeGroups.ExonProbeAffinityFile}
\alias{setMergeGroups.ExonRmaPlm}
\alias{setMergeStrands}
\alias{setMergeStrands.SnpChipEffectSet}
\alias{setMergeStrands.SnpPlm}
\alias{setMergeStrands.SnpProbeAffinityFile}
\alias{setReference}
\alias{setReference.SpatialReporter}
\alias{setRestructor}
\alias{setSettings}
\alias{setSettings.Package}
\alias{setSnpInformation}
\alias{setSnpInformation.AffymetrixCdfFile}
\alias{setUnitReadMap}
\alias{setUnitReadMap.AffymetrixCnChpFile}
\alias{setup.ArrayExplorer}
\alias{setup.GenericReporter}
\alias{SmoothSaModel}
\alias{smoothScatterMvsA}
\alias{SpatialRowColumnNormalization}
\alias{stextChipType.AffymetrixCdfFile}
\alias{Transform}
\alias{transformAffine}
\alias{transformAffine.AffymetrixCelFile}
\alias{transformAffine.AffymetrixCelSet}
\alias{translateColumnNames.AffymetrixAptSummaryFile}
\alias{translateColumnNames.AffymetrixProbeTabFile}
\alias{translateColumnNames.AffymetrixTabularFile}
\alias{translateFullName.AffymetrixProbeTabFile}
\alias{UflSnpInformation}
\alias{update.AromaAffymetrix}
\alias{update2.AffymetrixCelSet}
\alias{updateDataFlat}
\alias{updateDataFlat.ChipEffectFile}
\alias{updateDataFlat.FirmaFile}
\alias{updateDataFlat.ResidualFile}
\alias{updateDataFlat.WeightsFile}
\alias{updateMatchScores}
\alias{updateMatchScores.AromaCellMatchScoreFile}
\alias{updateOnChipTypeJS}
\alias{updateOnChipTypeJS.ArrayExplorer}
\alias{updateOnLoadJS}
\alias{updateOnLoadJS.ArrayExplorer}
\alias{updateParameter}
\alias{updateParameter.CrlmmParametersFile}
\alias{updateSampleAnnotationSet}
\alias{updateSampleAnnotationSet.AffymetrixCelSet}
\alias{updateSettings}
\alias{updateSettings.Package}
\alias{updateUnits}
\alias{updateUnits.ChipEffectFile}
\alias{updateUnits.ChipEffectSet}
\alias{updateUnits.FirmaFile}
\alias{updateUnits.FirmaSet}
\alias{updateUnits.ParameterCelFile}
\alias{updateUnits.ProbeAffinityFile}
\alias{updateUnits.ResidualFile}
\alias{updateUnits.ResidualSet}
\alias{updateUnits.WeightsFile}
\alias{updateUnits.WeightsSet}
\alias{validate.AvgPlm}
\alias{validate.MultiArrayUnitModel}
\alias{verify.AffymetrixCsvGenomeInformation}
\alias{verify.AffymetrixTsvFile}
\alias{verify.DChipGenomeInformation}
\alias{verify.DChipSnpInformation}
\alias{verify.UflSnpInformation}
\alias{verify.UgpGenomeInformation}
\alias{write}
\alias{writeAsFullCelFile}
\alias{writeAsFullCelFile.ChipEffectFile}
\alias{writeCNT}
\alias{writeCNT.SnpChipEffectFile}
\alias{writeCNT.SnpChipEffectSet}
\alias{writeImage}
\alias{writeImage.ResidualFile}
\alias{writeImage.WeightsFile}
\alias{writeImageCombined}
\alias{writeImageCombined.TransformReport}
\alias{writeImages}
\alias{writeImages.SpatialReporter}
\alias{writeImages.TransformReport}
\alias{writeRegions.profileCGH}
\alias{writeSgr}
\alias{writeSgr.AffymetrixCelSet}
\alias{writeSignals}
\alias{writeSignals.ProbeLevelTransform3}
\alias{writeWig}
\alias{writeWig.CnChipEffectSet}
\alias{boxplot.stats}
\alias{doCRMAv2}
\alias{doCRMAv2.AffymetrixCelSet}
\alias{doCRMAv2.character}
\alias{allocateFromCdf.AromaCellCpgFile}
\alias{doCRMAv1}
\alias{doCRMAv1.AffymetrixCelSet}
\alias{doCRMAv1.character}
\alias{groupUnitsByDimension}
\alias{extractAlleleSet}
\alias{extractAlleleSet.SnpChipEffectSet}
\alias{env2Cdf}
\alias{Env2Cdf}
\alias{doRMA}
\alias{doRMA.AffymetrixCelSet}
\alias{doRMA.character}
\alias{doASCRMAv1}
\alias{doASCRMAv2}
\alias{doASCRMAv1.default}
\alias{doASCRMAv2.default}
\alias{doGCRMA}
\alias{doGCRMA.AffymetrixCelSet}
\alias{doGCRMA.character}
\alias{extractAffyBatch.ChipEffectSet}
\alias{computeAffinitiesByACS}
\alias{computeAffinitiesByACS.AffymetrixCdfFile}
\alias{getACSFile}
\alias{getACSFile.AffymetrixCdfFile}
\alias{getAromaCellSequenceFile.AffymetrixCdfFile}
\alias{writeCdf}
\alias{writeCdf.default}
\alias{getFilteredFragmentLengths}
\alias{getFilteredFragmentLengths.FragmentLengthNormalization}
\alias{findTargetDistributionFile}
\alias{findTargetDistributionFile.QuantileNormalization}
\alias{readGeneAssignments}
\alias{readGeneAssignments.AffymetrixNetAffxCsvFile}
\alias{extractExpressionSet}
\alias{cat}
\alias{getOption}
\alias{append}
\alias{lapply}
\alias{sapply}
\alias{apply}
\alias{colMeans}
\alias{colSums}
\alias{library}
\alias{require}
\alias{createExonByTranscriptCdf}
\alias{getGenomeBuild}
\alias{getGenomeBuild.AffymetrixNetAffxCsvFile}
\alias{getHeaderAttributes}
\alias{getHeaderAttributes.AffymetrixNetAffxCsvFile}
\alias{getNetAffxBuild}
\alias{getNetAffxBuild.AffymetrixNetAffxCsvFile}
\alias{getNetAffxDate}
\alias{getNetAffxDate.AffymetrixNetAffxCsvFile}
\alias{dropCellsFromCdfList}
\alias{writeCdfByExcludingCells.AffymetrixCdfFile}
\alias{writeCdfByExcludingCells}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
