%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{[.AffymetrixCelFile}
\alias{[.AffymetrixCelSet}
\alias{[.GdasAnnotationFile}
\alias{[.GdasAnnotationSet}
\alias{[.GenotypeCallFile}
\alias{[.GenotypeCallSet}
\alias{[[.AffymetrixCelFile}
\alias{[[.AffymetrixCelSet}
\alias{[[.GenotypeCallSet}
\alias{[<-.GenotypeCallFile}
\alias{annotateMvsA}
\alias{annotateMvsA.AffymetrixCelFile}
\alias{append}
\alias{append.AffymetrixCelSet}
\alias{append.AffymetrixFileSet}
\alias{append.default}
\alias{appendFiles}
\alias{appendFiles.AffymetrixFileSet}
\alias{applyToUnitIntensities}
\alias{applyToUnitIntensities.AffymetrixCelSet}
\alias{AptProbesetGenotype}
\alias{as.AffymetrixCelSet}
\alias{as.AffymetrixFileSet}
\alias{as.character.AffymetrixCdfFile}
\alias{as.character.AffymetrixCelFile}
\alias{as.character.AffymetrixPowerTools}
\alias{as.character.AffymetrixProbeTabFile}
\alias{as.character.AptProbesetGenotype}
\alias{as.character.ChromosomeExplorer}
\alias{as.character.CnChipEffectFile}
\alias{as.character.CrlmmModel}
\alias{as.character.Interface}
\alias{as.character.Model}
\alias{as.character.GladModel}
\alias{as.character.PlatformDesign}
\alias{as.character.QcInfo}
\alias{as.character.SnpChipEffectFile}
\alias{as.character.Transform}
\alias{as.GenotypeCallSet}
\alias{backtransformGenotypeCone}
\alias{backtransformGenotypeCone.matrix}
\alias{bgAdjustGcrma}
\alias{bgAdjustOptical}
\alias{bgAdjustRma}
\alias{calculateGsbParameters}
\alias{calculateTargetDistribution}
\alias{calculateTargetDistribution.QuantileNormalization}
\alias{calibrateAllelicCrosstalk}
\alias{classifyAsXorXX}
\alias{cleanToRealChipType}
\alias{cleanToRealChipType.AffymetrixCdfFile}
\alias{clearCache.AffymetrixFileSet}
\alias{clearData}
\alias{clearFileCache}
\alias{clearFileCache.PlatformDesign}
\alias{clone.AffymetrixCelFile}
\alias{clone.AffymetrixCelSet}
\alias{clone.AffymetrixFile}
\alias{clone.AffymetrixFileSet}
\alias{cna}
\alias{cna.AffymetrixCelSet}
\alias{colMeans}
\alias{colSums}
\alias{compare}
\alias{computeAffinities}
\alias{convert}
\alias{convertTable}
\alias{convertTable.default}
\alias{convertTxtFilesToXdr}
\alias{convertTxtFilesToXdr.AptProbesetGenotype}
\alias{convertUnits}
\alias{copyFile}
\alias{copyFile.AffymetrixFile}
\alias{create}
\alias{create.GenotypeCallFile}
\alias{createBlankSet}
\alias{createBlankSet.AffymetrixCelSet}
\alias{createFrom}
\alias{createFrom.AffymetrixCelFile}
\alias{createFromCrlmmFile}
\alias{createFromCrlmmFile.GenotypeCallSet}
\alias{createFromFeatureSet}
\alias{createFromFeatureSet.AffymetrixCelSet}
\alias{createMonoCell}
\alias{createMonoCell.AffymetrixCdfFile}
\alias{createParamCdf}
\alias{createParamCdf.ChipEffectFile}
\alias{dataFrame}
\alias{dataFrame.default}
\alias{decode}
\alias{decode.ParameterCelFile}
\alias{drawCytoBand}
\alias{drawCytoBand.default}
\alias{encode}
\alias{encode.ParameterCelFile}
\alias{extend.Interface}
\alias{extractAffyBatch}
\alias{extractSnpFeatureSet}
\alias{extractSnpQSet}
\alias{findUnitsTodo}
\alias{findUnitsTodo.ChipEffectFile}
\alias{findUnitsTodo.ChipEffectSet}
\alias{fit}
\alias{fit.GladModel}
\alias{fit.PcrProbeCalibration}
\alias{fitGenotypeCone}
\alias{fitGenotypeCone.matrix}
\alias{fitGlad}
\alias{fitGlad.CnChipEffectSet}
\alias{fitOne}
\alias{fitPlasqUnit}
\alias{fitPlasqUnit.matrix}
\alias{fitQuantileNormFcn}
\alias{forDataSet}
\alias{forDataSet.GdasAnnotationSet}
\alias{fromCdf}
\alias{fromCdf.AffymetrixProbeTabFile}
\alias{fromChipType}
\alias{fromChipType.AffymetrixProbeTabFile}
\alias{fromChipType.DChipGenomeInformation}
\alias{fromChipType.DChipSnpInformation}
\alias{fromChipType.GdasAnnotationFile}
\alias{fromChipType.GdasAnnotationSet}
\alias{fromCleanChipType}
\alias{fromCleanChipType.AffymetrixCdfFile}
\alias{fromDataFile}
\alias{fromDataFile.ChipEffectFile}
\alias{fromDataSet}
\alias{fromDataSet.ChipEffectSet}
\alias{fromDataSet.GenomeInformation}
\alias{fromDataSet.SnpInformation}
\alias{fromFile}
\alias{fromFile.AffymetrixFile}
\alias{fromFile.GenotypeCallXdrFile}
\alias{fromFiles}
\alias{fromFiles.AffymetrixCelSet}
\alias{fromFiles.ChipEffectSet}
\alias{fromFiles.GdasAnnotationSet}
\alias{fromFiles.GenotypeCallSet}
\alias{fromFiles.GenotypeCallXdrSet}
\alias{gcrmaSummary}
\alias{GdasAnnotationFile}
\alias{GdasAnnotationSet}
\alias{GenericSummary}
\alias{getAlleleAB}
\alias{getAlleleAB.PlatformDesign}
\alias{getAllelePairs}
\alias{getAllelePairUnitSets}
\alias{getAlleleProbePairs}
\alias{getAlleleProbePairs2}
\alias{getAlleleProbePairs2.AffymetrixCdfFile}
\alias{getAm}
\alias{getAM}
\alias{getAm.AffymetrixCelFile}
\alias{getAnnotations}
\alias{getAnnotations.PlatformDesign}
\alias{getAptPathnames}
\alias{getAptPathnames.AptProbesetGenotype}
\alias{getArrays}
\alias{getArrays.ChromosomeExplorer}
\alias{getArrays.GladModel}
\alias{getAsteriskTag}
\alias{getAsteriskTag.Transform}
\alias{getAverage}
\alias{getAverage.AffymetrixCelSet}
\alias{getAverageAsinh}
\alias{getAverageAsinh.AffymetrixCelSet}
\alias{getAverageFile}
\alias{getAverageFile.ChipEffectSet}
\alias{getAverageFile.CnChipEffectSet}
\alias{getAverageFile.SnpChipEffectSet}
\alias{getAverageLog}
\alias{getAverageLog.AffymetrixCelSet}
\alias{getBlockPositions}
\alias{getBlockPositions.GdasAnnotationFile}
\alias{getCachedFields}
\alias{getCachedFields.PlatformDesign}
\alias{getCachePath.PlatformDesign}
\alias{getCdf}
\alias{getCdf.AffymetrixProbeTabFile}
\alias{getCdf.AptProbesetGenotype}
\alias{getCdf.FragmentLengthNormalization}
\alias{getCdf.GdasAnnotationSet}
\alias{getCdf.GenotypeCallFile}
\alias{getCdf.GenotypeCallXdrFile}
\alias{getCdf.GladModel}
\alias{getCdf.PlatformDesign}
\alias{getCdf.QcInfo}
\alias{getCellIndices}
\alias{getCellIndices.ChipEffectFile}
\alias{getCellIndices.ChipEffectSet}
\alias{getCellIndices.CnChipEffectFile}
\alias{getCellIndices.CnPlm}
\alias{getCellIndices.CnProbeAffinityFile}
\alias{getCellIndices.PlatformDesign}
\alias{getCellIndices.ProbeAffinityFile}
\alias{getCellIndices.ProbeLevelModel}
\alias{getCellIndices.SnpChipEffectFile}
\alias{getCellIndices.SnpPlm}
\alias{getCellIndices.SnpProbeAffinityFile}
\alias{getCellMap}
\alias{getCellMap.ChipEffectFile}
\alias{getChipEffectFileClass}
\alias{getChipEffectFileClass.ChipEffectSet}
\alias{getChipEffectFileClass.CnChipEffectSet}
\alias{getChipEffectFileClass.SnpChipEffectSet}
\alias{getChipEffectFiles}
\alias{getChipEffectFiles.GladModel}
\alias{getChipEffects}
\alias{getChipEffects.CnPlm}
\alias{getChipEffects.GladModel}
\alias{getChipEffects.PlasqModel}
\alias{getChipEffects.QcInfo}
\alias{getChipEffects.SnpPlm}
\alias{getChipEffectSetClass}
\alias{getChipEffectSetClass.CnPlm}
\alias{getChipEffectSetClass.ProbeLevelModel}
\alias{getChipEffectSetClass.SnpPlm}
\alias{getChipType}
\alias{getChipType.AffymetrixCdfFile}
\alias{getChipType.AffymetrixProbeTabFile}
\alias{getChipType.GdasAnnotationSet}
\alias{getChipType.GenotypeCallFile}
\alias{getChipType.GenotypeCallSet}
\alias{getChipType.GladModel}
\alias{getChipType.MultiGladModel}
\alias{getChipType.PlatformDesign}
\alias{getChipTypes}
\alias{getChipTypes.MultiGladModel}
\alias{getChromosomes}
\alias{getChromosomes.GenomeInformation}
\alias{getChromosomeStats}
\alias{getChromosomeStats.GenomeInformation}
\alias{getChrXFile}
\alias{getChrXFile.AptProbesetGenotype}
\alias{getCombineAlleles}
\alias{getCombineAlleles.CnChipEffectSet}
\alias{getCommonPrefix}
\alias{getCrlmmInfo}
\alias{getCrlmmInfo.PlatformDesign}
\alias{getData.AffymetrixCelSet}
\alias{getData.AffymetrixProbeTabFile}
\alias{getDataFile}
\alias{getDataFile.PlatformDesign}
\alias{getDataFiles}
\alias{getDataFiles.PlatformDesign}
\alias{getDataFlat}
\alias{getDataFlat.ChipEffectFile}
\alias{getDataSet}
\alias{getDataSet.AptProbesetGenotype}
\alias{getDescription.AffymetrixFileSet}
\alias{getDescription.PlatformDesign}
\alias{getDimension}
\alias{getDimension.AffymetrixCdfFile}
\alias{getExternalCommand}
\alias{getExternalCommand.AptProbesetGenotype}
\alias{getFeatureInfo}
\alias{getFeatureInfo.PlatformDesign}
\alias{getFields.GenomeInformation}
\alias{getFields.SnpInformation}
\alias{getFile}
\alias{getFile.AffymetrixFileSet}
\alias{getFilename}
\alias{getFilename.Arguments}
\alias{getFiles}
\alias{getFiles.AffymetrixFileSet}
\alias{getFileSize.AffymetrixFile}
\alias{getFileSize.AffymetrixFileSet}
\alias{getFileType}
\alias{getFirstCellIndices}
\alias{getFirstCellIndices.AffymetrixCdfFile}
\alias{getFitFunction}
\alias{getFitUnitFunction}
\alias{getFitUnitFunction.CnPlm}
\alias{getFitUnitFunction.PlasqModel}
\alias{getFitUnitFunction.ProbeLevelModel}
\alias{getFitUnitFunction.UnitModel}
\alias{getFragmentLengths}
\alias{getFragmentLengths.SnpInformation}
\alias{getFragmentStarts}
\alias{getFragmentStarts.SnpInformation}
\alias{getFragmentStops}
\alias{getFragmentStops.SnpInformation}
\alias{getFullName}
\alias{getFullName.AffymetrixCelSet}
\alias{getFullName.AptProbesetGenotype}
\alias{getFullName.ChromosomeExplorer}
\alias{getFullName.GenotypeCallSet}
\alias{getFullName.GladModel}
\alias{getFullName.MultiGladModel}
\alias{getFullName.QcInfo}
\alias{getGenomeInformation}
\alias{getGenomeInformation.GladModel}
\alias{getGenotypeCallSet}
\alias{getGenotypeCallSet.AptProbesetGenotype}
\alias{getHeader}
\alias{getHeaderV3}
\alias{getHeaderV3.AffymetrixCelFile}
\alias{getIdentifier}
\alias{getIdentifier.AffymetrixCelFile}
\alias{getIdentifier.AffymetrixCelSet}
\alias{getIdentifier.AffymetrixFileSet}
\alias{getIndexToRowMap}
\alias{getIndexToRowMap.AffymetrixProbeTabFile}
\alias{getInputDataSet}
\alias{getIntensities}
\alias{getLabel.AffymetrixFile}
\alias{getLabel.Model}
\alias{getListOfCdfs}
\alias{getListOfCdfs.MultiGladModel}
\alias{getListOfChipEffects}
\alias{getListOfChipEffects.MultiGladModel}
\alias{getListOfGenomeInformations}
\alias{getListOfGenomeInformations.MultiGladModel}
\alias{getListOfReferences}
\alias{getListOfReferences.MultiGladModel}
\alias{getLog2Ratios}
\alias{getLog2Ratios.GladModel}
\alias{getLog2Ratios.MultiGladModel}
\alias{getMergeStrands}
\alias{getMergeStrands.SnpChipEffectSet}
\alias{getModel}
\alias{getModel.ChromosomeExplorer}
\alias{getName.AffymetrixCelSet}
\alias{getName.AptProbesetGenotype}
\alias{getName.ChromosomeExplorer}
\alias{getName.GladModel}
\alias{getName.MultiGladModel}
\alias{getName.PlatformDesign}
\alias{getName.QcInfo}
\alias{getNames}
\alias{getNames.MultiGladModel}
\alias{getOligoSequences}
\alias{getOligoSequences.PlatformDesign}
\alias{getOligoToCelMap}
\alias{getOligoToCelMap.AffymetrixCdfFile}
\alias{getOutputDataSet}
\alias{getOutputFiles}
\alias{getOutputFiles.Transform}
\alias{getOutputIdentifier}
\alias{getOutputIdentifier.QuantileNormalization}
\alias{getPackageName}
\alias{getPackageName.default}
\alias{getPackageName.PlatformDesign}
\alias{getParameters}
\alias{getParameters.FragmentLengthNormalization}
\alias{getParameters.PcrProbeCalibration}
\alias{getParameters.QuantileNormalization}
\alias{getParameters.Transform}
\alias{getParametersAsString}
\alias{getParametersAsString.Transform}
\alias{getParameterSet}
\alias{getParameterSet.Model}
\alias{getPath.AptProbesetGenotype}
\alias{getPath.ChromosomeExplorer}
\alias{getPath.GladModel}
\alias{getPath.MultiGladModel}
\alias{getPath.PlatformDesign}
\alias{getPath.QcInfo}
\alias{getPathnames}
\alias{getPathnamesFile}
\alias{getPathnamesFile.AptProbesetGenotype}
\alias{getPlasqTypes}
\alias{getPlatformDesignObject}
\alias{getPlatformDesignObject.AffymetrixCdfFile}
\alias{getPlatformDesignObject.PlatformDesign}
\alias{getPlm}
\alias{getPlm.QcInfo}
\alias{getPmDesignMatrix}
\alias{getPmDesignMatrix.PcrProbeCalibration}
\alias{getPositions}
\alias{getProbeAffinities}
\alias{getProbeAffinities.AffinePlm}
\alias{getProbeAffinities.CnPlm}
\alias{getProbeAffinities.MbeiPlm}
\alias{getProbeAffinities.RmaPlm}
\alias{getProbeAffinities.SnpPlm}
\alias{getProbeEffects}
\alias{getProbeEffects.PlatformDesign}
\alias{getProbeParameters}
\alias{getProbeParameters.PlasqModel}
\alias{getReadMap}
\alias{getReadMap.PlatformDesign}
\alias{getRectangle}
\alias{getRectangle.AffymetrixCelFile}
\alias{getReference}
\alias{getReference.GladModel}
\alias{getReferenceFiles}
\alias{getReferenceFiles.MultiGladModel}
\alias{getReferenceName}
\alias{getReferenceName.MultiGladModel}
\alias{getReferenceQuantiles}
\alias{getReferenceQuantiles.PlatformDesign}
\alias{getRegions}
\alias{getRegions.GladModel}
\alias{getRegions.MultiGladModel}
\alias{getRegions.profileCGH}
\alias{getResiduals}
\alias{getResiduals.QcInfo}
\alias{getRestructor}
\alias{getRestructor.AffymetrixCdfFile}
\alias{getRootPath}
\alias{getRootPath.AptProbesetGenotype}
\alias{getRootPath.ChipEffectTransform}
\alias{getRootPath.ChromosomeExplorer}
\alias{getRootPath.GladModel}
\alias{getRootPath.MultiGladModel}
\alias{getRootPath.PlasqModel}
\alias{getRootPath.ProbeLevelModel}
\alias{getRootPath.ProbeLevelTransform}
\alias{getRootPath.QcInfo}
\alias{getRootPath.Transform}
\alias{getSiblings}
\alias{getSnpInformation}
\alias{getSnpInformation.AffymetrixCdfFile}
\alias{getSnpNames}
\alias{getSubsetToFit}
\alias{getSubsetToFit.FragmentLengthNormalization}
\alias{getSubsetToFit.PcrProbeCalibration}
\alias{getSubsetToUpdate}
\alias{getSubsetToUpdate.OligoQuantileNormalization}
\alias{getSubsetToUpdate.QuantileNormalization}
\alias{getTableOfArrays}
\alias{getTableOfArrays.MultiGladModel}
\alias{getTags}
\alias{getTags.AptProbesetGenotype}
\alias{getTags.ChromosomeExplorer}
\alias{getTags.GladModel}
\alias{getTags.MultiGladModel}
\alias{getTags.QcInfo}
\alias{getTargetDistribution}
\alias{getTargetDistribution.OligoQuantileNormalization}
\alias{getTargetDistribution.QuantileNormalization}
\alias{getTargetDistributionIdentifier}
\alias{getTargetDistributionIdentifier.QuantileNormalization}
\alias{getTargetDistributionPathname}
\alias{getTargetDistributionPathname.QuantileNormalization}
\alias{getTargetFunction}
\alias{getTargetFunction.FragmentLengthNormalization}
\alias{getTimestamp}
\alias{getTimestamps}
\alias{getTimestamps.AffymetrixCelSet}
\alias{getUnitIndices}
\alias{getUnitIndices.GdasAnnotationSet}
\alias{getUnitIntensities}
\alias{getUnitNames}
\alias{getUnitNames.PlatformDesign}
\alias{getUnitSizes}
\alias{getUnitsOnChromosome}
\alias{getUnitsOnChromosome.GenomeInformation}
\alias{getUnitsOnChromosome.PlatformDesign}
\alias{getWriteMap}
\alias{getWriteMap.PlatformDesign}
\alias{getX}
\alias{getX.PlatformDesign}
\alias{getXAM}
\alias{getY}
\alias{getY.PlatformDesign}
\alias{highlight}
\alias{highlight.AffymetrixCelFile}
\alias{highlight.ProbeLevelModel}
\alias{identifyCells}
\alias{identifyCells.AffymetrixCdfFile}
\alias{image270}
\alias{importFeatureSet}
\alias{importFeatureSet.AffymetrixCelFile}
\alias{importFromCrlmmFile}
\alias{importFromCrlmmFile.GenotypeCallFile}
\alias{importFromDChip}
\alias{importFromDChip.CnChipEffectSet}
\alias{indexOf}
\alias{isDone.AptProbesetGenotype}
\alias{isDuplicated}
\alias{isInstalled}
\alias{isInstalled.PlatformDesign}
\alias{isLoaded.PlatformDesign}
\alias{isPm}
\alias{isPm.PlatformDesign}
\alias{isSnpChip}
\alias{join}
\alias{join.GdasAnnotationSet}
\alias{lapply}
\alias{lapply.default}
\alias{load.PlatformDesign}
\alias{loadDataFile}
\alias{loadDataFile.PlatformDesign}
\alias{mergeByCommonTails}
\alias{nbrOfArrays}
\alias{nbrOfArrays.AptProbesetGenotype}
\alias{nbrOfArrays.ChromosomeExplorer}
\alias{nbrOfArrays.GladModel}
\alias{nbrOfArrays.MultiGladModel}
\alias{nbrOfArrays.QcInfo}
\alias{nbrOfCells}
\alias{nbrOfCells.AffymetrixCdfFile}
\alias{nbrOfCells.AffymetrixCelFile}
\alias{nbrOfChipTypes}
\alias{nbrOfChipTypes.MultiGladModel}
\alias{nbrOfColumns}
\alias{nbrOfColumns.AffymetrixCdfFile}
\alias{nbrOfFiles}
\alias{nbrOfQcUnits}
\alias{nbrOfQcUnits.AffymetrixCdfFile}
\alias{nbrOfRows}
\alias{nbrOfRows.AffymetrixCdfFile}
\alias{nbrOfSnps}
\alias{nbrOfUnits}
\alias{nbrOfUnits.AffymetrixCdfFile}
\alias{nbrOfUnits.GenomeInformation}
\alias{nbrOfUnits.SnpInformation}
\alias{normalizeAverage}
\alias{normalizeAverage.list}
\alias{normalizeAverage.matrix}
\alias{normalizeFragmentLength}
\alias{normalizeFragmentLength.default}
\alias{normalizeGenotypeCone}
\alias{normalizeGenotypeCone.matrix}
\alias{plot.GladModel}
\alias{plot.MultiGladModel}
\alias{plot.profileCGH}
\alias{plotFit}
\alias{plotFit.PcrProbeCalibration}
\alias{plotMvsPosition}
\alias{plotMvsPosition.ProbeLevelModel}
\alias{plotMvsX}
\alias{plotProfile2}
\alias{plotProfile2.profileCGH}
\alias{predict.lowess}
\alias{print.GenericSummary}
\alias{print.Interface}
\alias{process}
\alias{process.AptProbesetGenotype}
\alias{process.ChromosomeExplorer}
\alias{process.PcrProbeCalibration}
\alias{range.AffymetrixCelFile}
\alias{range.AffymetrixCelSet}
\alias{read250K}
\alias{read250K.DChipSnpInformation}
\alias{read250KHg17}
\alias{read250KHg17.DChipGenomeInformation}
\alias{read50K}
\alias{read50K.DChipSnpInformation}
\alias{read50KHg17}
\alias{read50KHg17.DChipGenomeInformation}
\alias{readBlock}
\alias{readBlock.GdasAnnotationFile}
\alias{readCdfGroupStrands}
\alias{readColumnName}
\alias{readColumnName.GdasAnnotationFile}
\alias{readData}
\alias{readData.DChipGenomeInformation}
\alias{readData.DChipSnpInformation}
\alias{readData.GdasAnnotationFile}
\alias{readData.GenomeInformation}
\alias{readData.SnpInformation}
\alias{readHeader.GdasAnnotationFile}
\alias{readLog}
\alias{readLog.AptProbesetGenotype}
\alias{readPlasqTypes}
\alias{readPlasqTypes.AffymetrixCdfFile}
\alias{readTableInternal}
\alias{readTableInternal.GenomeInformation}
\alias{readTableInternal.SnpInformation}
\alias{readUnits}
\alias{readUnits.AffymetrixCelSet}
\alias{readUnits.ChipEffectFile}
\alias{readUnits.ChipEffectSet}
\alias{readUnits.CnChipEffectFile}
\alias{readUnits.GenotypeCallFile}
\alias{readUnits.GenotypeCallSet}
\alias{readUnits.GenotypeCallXdrFile}
\alias{readUnits.ParameterCelFile}
\alias{readUnits.ProbeAffinityFile}
\alias{readUnits.SnpChipEffectFile}
\alias{reorder.AffymetrixFileSet}
\alias{restruct}
\alias{restruct.AffymetrixCdfFile}
\alias{rmaSummary}
\alias{rowSds}
\alias{rowVars}
\alias{select}
\alias{select.GdasAnnotationSet}
\alias{setCdf}
\alias{setCombineAlleles}
\alias{setCombineAlleles.CnChipEffectSet}
\alias{setCombineAlleles.CnPlm}
\alias{setCombineAlleles.CnProbeAffinityFile}
\alias{setDecodeFunction}
\alias{setDecodeFunction.ParameterCelFile}
\alias{setEncodeFunction}
\alias{setEncodeFunction.ParameterCelFile}
\alias{setLabel.AffymetrixFile}
\alias{setLabel.Model}
\alias{setMergeStrands}
\alias{setMergeStrands.SnpChipEffectSet}
\alias{setMergeStrands.SnpPlm}
\alias{setMergeStrands.SnpProbeAffinityFile}
\alias{setName}
\alias{setReference}
\alias{setReference.GladModel}
\alias{setRestructor}
\alias{setTags}
\alias{setTags.Transform}
\alias{showAptVersion}
\alias{showAptVersion.AptProbesetGenotype}
\alias{showLog}
\alias{showLog.AptProbesetGenotype}
\alias{smoothScatterMvsA}
\alias{splitByCommonTails}
\alias{stext}
\alias{stext.default}
\alias{stextChipType}
\alias{stextChipType.AffymetrixCdfFile}
\alias{stextLabel}
\alias{stextLabel.AffymetrixFile}
\alias{stextLabels}
\alias{stextLabels.AffymetrixFile}
\alias{stextSize}
\alias{stextSize.AffymetrixFile}
\alias{stringTree}
\alias{stringTree.character}
\alias{toCamelCase}
\alias{toCamelCase.default}
\alias{transformAffine}
\alias{transformAffine.AffymetrixCelFile}
\alias{transformAffine.AffymetrixCelSet}
\alias{unload.PlatformDesign}
\alias{updateDataFlat}
\alias{updateDataFlat.ChipEffectFile}
\alias{updateUnits}
\alias{updateUnits.ChipEffectFile}
\alias{updateUnits.ChipEffectSet}
\alias{updateUnits.GenotypeCallFile}
\alias{updateUnits.GenotypeCallXdrFile}
\alias{updateUnits.ParameterCelFile}
\alias{updateUnits.ProbeAffinityFile}
\alias{uses}
\alias{uses.Interface}
\alias{validate.PlatformDesign}
\alias{verify}
\alias{verify.DChipGenomeInformation}
\alias{verify.DChipSnpInformation}
\alias{whatDataType}
\alias{whatDataType.default}
\alias{write}
\alias{writeChrXFile}
\alias{writeChrXFile.AffymetrixPowerTools}
\alias{writeGraphs}
\alias{writeGraphs.ChromosomeExplorer}
\alias{writeRegions}
\alias{writeRegions.ChromosomeExplorer}
\alias{writeRegions.GladModel}
\alias{writeRegions.MultiGladModel}
\alias{writeRegions.profileCGH}
\alias{writeSpatial}
\alias{writeSpatial.ProbeAffinityFile}
\alias{writeWig}
\alias{writeWig.CnChipEffectSet}
\alias{as.character.ChipEffectFile}
\alias{as.character.ChipEffectSet}
\alias{digest}
\alias{getChipTypes.GladModel}
\alias{getDataColumns}
\alias{getDataColumns.GenomeInformation}
\alias{getListOfCdfs.GladModel}
\alias{getListOfChipEffects.GladModel}
\alias{getListOfGenomeInformations.GladModel}
\alias{getListOfReferences.GladModel}
\alias{getNames.GladModel}
\alias{getOutputDataSet.FragmentLengthNormalization}
\alias{getParameters.ChipEffectFile}
\alias{getParameters.ChipEffectSet}
\alias{getParameters.CnChipEffectFile}
\alias{getParameters.SnpChipEffectFile}
\alias{getParametersAsString.ChipEffectFile}
\alias{getParametersAsString.ChipEffectSet}
\alias{getReferenceFiles.GladModel}
\alias{getReferenceName.GladModel}
\alias{getTableOfArrays.GladModel}
\alias{nbrOfChipTypes.GladModel}
\alias{addIncludes}
\alias{addIncludes.ChromosomeExplorer}
\alias{getAlias}
\alias{getTemplatePath}
\alias{getTemplatePath.ChromosomeExplorer}
\alias{setAlias}
\alias{setup}
\alias{setup.ChromosomeExplorer}
\alias{updateSamplesFile}
\alias{updateSamplesFile.ChromosomeExplorer}
\alias{hasTag}
\alias{hasTag.AffymetrixFile}
\alias{hasTag.AffymetrixFileSet}
\alias{hasTags}
\alias{hasTags.AffymetrixFile}
\alias{hasTags.AffymetrixFileSet}
\alias{display}
\alias{display.ChromosomeExplorer}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
