%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{[.AffymetrixCelFile}
\alias{[.AffymetrixCelSet}
\alias{[.AromaTabularBinaryFile}
\alias{[.CnagCfhFile}
\alias{[.CnagCfhSet}
\alias{[[.AffymetrixCelFile}
\alias{[[.AffymetrixCelSet}
\alias{[[.AromaTabularBinaryFile}
\alias{[[.CnagCfhFile}
\alias{[[.CnagCfhSet}
\alias{[<-.AromaTabularBinaryFile}
\alias{addColorMap}
\alias{addColorMap.ArrayExplorer}
\alias{addColorMap.SpatialReporter}
\alias{addExclCells}
\alias{addExclCells.DChipQuantileNormalization}
\alias{addIncludes}
\alias{addIncludes.Explorer}
\alias{addIndexFile}
\alias{addIndexFile.ChromosomeExplorer}
\alias{addIndexFile.Explorer}
\alias{AffymetrixCsvFile}
\alias{AffymetrixCsvGenomeInformation}
\alias{AffymetrixNetAffxCsvFile}
\alias{AffymetrixTabularFile}
\alias{AffymetrixTsvFile}
\alias{allocate}
\alias{allocate.AromaTabularBinaryFile}
\alias{allocateFromCdf}
\alias{allocateFromCdf.AromaUflFile}
\alias{allocateFromCdf.AromaUgpFile}
\alias{allocateFromCdf.AromaUnitTabularBinaryFile}
\alias{annotateMvsA}
\alias{annotateMvsA.AffymetrixCelFile}
\alias{append}
\alias{append.AffymetrixCelSet}
\alias{append.CnagCfhSet}
\alias{append.default}
\alias{append.GenericDataFileSet}
\alias{appendFiles}
\alias{appendFiles.GenericDataFileSet}
\alias{apply}
\alias{apply.default}
\alias{apply.SampleAnnotationFile}
\alias{applyRegions}
\alias{applyRegions.CopyNumberOutliers}
\alias{applyRows}
\alias{applyRows.CopyNumberRegions}
\alias{applyToUnitIntensities}
\alias{applyToUnitIntensities.AffymetrixCelSet}
\alias{AromaAffymetrix}
\alias{AromaTabularBinaryFile}
\alias{AromaUflFile}
\alias{AromaUgpFile}
\alias{AromaUnitTabularBinaryFile}
\alias{as.AffymetrixCelSet}
\alias{as.AffymetrixFileSet}
\alias{as.character.AffymetrixCdfFile}
\alias{as.character.AffymetrixCelFile}
\alias{as.character.AffymetrixCelSetReporter}
\alias{as.character.AffymetrixCelSetTuple}
\alias{as.character.AffymetrixFileSetReporter}
\alias{as.character.AffymetrixProbeTabFile}
\alias{as.character.AromaTabularBinaryFile}
\alias{as.character.ArrayExplorer}
\alias{as.character.ChipEffectFile}
\alias{as.character.ChipEffectSet}
\alias{as.character.ChromosomalModel}
\alias{as.character.ChromosomeExplorer}
\alias{as.character.CnagCfhFile}
\alias{as.character.CopyNumberOutliers}
\alias{as.character.CopyNumberRegions}
\alias{as.character.CopyNumberSegmentationModel}
\alias{as.character.DChipQuantileNormalization}
\alias{as.character.Explorer}
\alias{as.character.FirmaModel}
\alias{as.character.GenericTabularFile}
\alias{as.character.GenomeInformation}
\alias{as.character.Interface}
\alias{as.character.Model}
\alias{as.character.ProbeAffinityFile}
\alias{as.character.QualityAssessmentModel}
\alias{as.character.RawCopyNumbers}
\alias{as.character.ResidualFile}
\alias{as.character.ResidualSet}
\alias{as.character.SmoothMultiarrayModel}
\alias{as.character.SnpInformation}
\alias{as.character.SpatialReporter}
\alias{as.character.Transform}
\alias{as.character.TransformReport}
\alias{as.character.WeightsFile}
\alias{as.character.WeightsSet}
\alias{as.CnagCfhSet}
\alias{as.data.frame.CopyNumberOutliers}
\alias{as.data.frame.CopyNumberRegions}
\alias{as.data.frame.RawCopyNumbers}
\alias{as.GenericDataFileSet}
\alias{backtransformGenotypeCone}
\alias{backtransformGenotypeCone.matrix}
\alias{bgAdjustGcrma}
\alias{bgAdjustOptical}
\alias{bgAdjustRma}
\alias{byChipType}
\alias{byName}
\alias{byName.AffymetrixCelSet}
\alias{byPath}
\alias{byPath.AffymetrixCelSetTuple}
\alias{calculateBaseline}
\alias{calculateChromosomeStatistics}
\alias{calculateChromosomeStatistics.CopyNumberSegmentationModel}
\alias{calculateParametersGsb}
\alias{calculateResiduals}
\alias{calculateResiduals.ProbeLevelModel}
\alias{calculateResidualSet}
\alias{calculateResidualSet.ProbeLevelModel}
\alias{calculateTargetDistribution}
\alias{calculateTargetDistribution.QuantileNormalization}
\alias{calculateWeights}
\alias{calculateWeights.ExonRmaPlm}
\alias{calculateWeights.FirmaModel}
\alias{calculateWeights.ProbeLevelModel}
\alias{cdfMergeGroups}
\alias{ChipEffectSetTuple}
\alias{cleanToRealChipType}
\alias{cleanToRealChipType.AffymetrixCdfFile}
\alias{clearCache.AffymetrixCdfFile}
\alias{clearCache.AffymetrixCelFile}
\alias{clearCache.AffymetrixCelSet}
\alias{clearCache.AffymetrixCelSetTuple}
\alias{clearCache.AffymetrixFileSet}
\alias{clearCache.AffymetrixProbeTabFile}
\alias{clearCache.AffymetrixTsvFile}
\alias{clearCache.ChipEffectFile}
\alias{clearCache.ChipEffectSet}
\alias{clearCache.ChromosomalModel}
\alias{clearCache.CnagCfhFile}
\alias{clearCache.CnagCfhSet}
\alias{clearCache.CopyNumberSegmentationModel}
\alias{clearCache.ExonChipEffectFile}
\alias{clearCache.FirmaFile}
\alias{clearCache.FirmaModel}
\alias{clearCache.FirmaSet}
\alias{clearCache.FragmentLengthNormalization}
\alias{clearCache.GenericDataFileSet}
\alias{clearCache.GenomeInformation}
\alias{clearCache.ParameterCelFile}
\alias{clearCache.ProbeAffinityFile}
\alias{clearCache.ProbeLevelModel}
\alias{clearCache.QualityAssessmentFile}
\alias{clearCache.QuantileNormalization}
\alias{clearCache.ResidualFile}
\alias{clearCache.ResidualSet}
\alias{clearCache.SmoothMultiarrayModel}
\alias{clearCache.SnpChipEffectFile}
\alias{clearCache.SnpInformation}
\alias{clearCache.Transform}
\alias{clearCache.TransformReport}
\alias{clearCache.WeightsFile}
\alias{clearCache.WeightsSet}
\alias{clearData}
\alias{clone.AffymetrixCelFile}
\alias{clone.AffymetrixCelSet}
\alias{clone.CnagCfhFile}
\alias{clone.CnagCfhSet}
\alias{clone.GenericDataFile}
\alias{clone.GenericDataFileSet}
\alias{colMeans}
\alias{colMeans.AromaTabularBinaryFile}
\alias{colMeans.default}
\alias{colMedians}
\alias{colMedians.AromaTabularBinaryFile}
\alias{colnames.AromaTabularBinaryFile}
\alias{colnames.AromaUflFile}
\alias{colnames.AromaUgpFile}
\alias{colnames.GenericTabularFile}
\alias{colStats}
\alias{colStats.AromaTabularBinaryFile}
\alias{colSums}
\alias{colSums.AromaTabularBinaryFile}
\alias{colSums.default}
\alias{compare}
\alias{compareChecksum}
\alias{compareChecksum.GenericDataFile}
\alias{computeAffinities}
\alias{convert}
\alias{convertTable}
\alias{convertTable.default}
\alias{convertUnits}
\alias{CopyNumberOutliers}
\alias{CopyNumberRegions}
\alias{copyTo}
\alias{copyTo.GenericDataFile}
\alias{copyTo.GenericDataFileSet}
\alias{createBlankSet}
\alias{createBlankSet.AffymetrixCelSet}
\alias{createFrom}
\alias{createMonoCell}
\alias{createOutputTuple}
\alias{createOutputTuple.SmoothMultiarrayModel}
\alias{createParamCdf}
\alias{createParamCdf.ChipEffectFile}
\alias{createParamCdf.FirmaFile}
\alias{decode}
\alias{decode.ParameterCelFile}
\alias{digest2}
\alias{dim.AromaTabularBinaryFile}
\alias{dimnames<-.AromaTabularBinaryFile}
\alias{display}
\alias{display.ChromosomeExplorer}
\alias{drawCnRegions}
\alias{drawCnRegions.DNAcopy}
\alias{drawCnRegions.profileCGH}
\alias{drawCytoband}
\alias{drawCytoband.ChromosomalModel}
\alias{drawCytoband.default}
\alias{drawCytoband.profileCGH}
\alias{drawCytoband2}
\alias{drawCytoband2.default}
\alias{drawExtraAnnotations}
\alias{drawExtraAnnotations.default}
\alias{drawExtraAnnotations.profileCGH}
\alias{drawLevels}
\alias{drawLevels.CopyNumberOutliers}
\alias{drawLevels.CopyNumberRegions}
\alias{encode}
\alias{encode.ParameterCelFile}
\alias{equals.GenericDataFile}
\alias{excludeChrXFromFit}
\alias{excludeChrXFromFit.DChipQuantileNormalization}
\alias{extend.Interface}
\alias{extractAffyBatch}
\alias{extractCNRs}
\alias{extractCNRs.default}
\alias{extractCopyNumberOutliers}
\alias{extractCopyNumberOutliers.DNAcopy}
\alias{extractCopyNumberOutliers.profileCGH}
\alias{extractCopyNumberRegions}
\alias{extractCopyNumberRegions.DNAcopy}
\alias{extractCopyNumberRegions.profileCGH}
\alias{extractMatrix}
\alias{extractRawCNs}
\alias{extractRawCNs.default}
\alias{extractRawCopyNumbers}
\alias{extractRawCopyNumbers.DNAcopy}
\alias{extractRawCopyNumbers.profileCGH}
\alias{findByChipType}
\alias{findByChipType.AffymetrixCsvFile}
\alias{findByChipType.AffymetrixCsvGenomeInformation}
\alias{findByChipType.AffymetrixNetAffxCsvFile}
\alias{findByChipType.AffymetrixProbeTabFile}
\alias{findByChipType.AffymetrixTabularFile}
\alias{findByChipType.AffymetrixTsvFile}
\alias{findByChipType.AromaUnitTabularBinaryFile}
\alias{findByChipType.DChipGenomeInformation}
\alias{findByChipType.DChipSnpInformation}
\alias{findByChipType.UflSnpInformation}
\alias{findByChipType.UgpGenomeInformation}
\alias{findByName}
\alias{findByName.AffymetrixCelSet}
\alias{findByName.ChipEffectSet}
\alias{findByName.CnagCfhSet}
\alias{findCdf.patch}
\alias{findSAFs}
\alias{findSAFs.SampleAnnotationSet}
\alias{findUnitsTodo}
\alias{findUnitsTodo.ChipEffectFile}
\alias{findUnitsTodo.ChipEffectSet}
\alias{findUnitsTodo.ExonChipEffectSet}
\alias{findUnitsTodo.FirmaFile}
\alias{findUnitsTodo.FirmaSet}
\alias{findUnitsTodo.QualityAssessmentFile}
\alias{findUnitsTodo.ResidualFile}
\alias{findUnitsTodo.ResidualSet}
\alias{findUnitsTodo.WeightsFile}
\alias{findUnitsTodo.WeightsSet}
\alias{fit}
\alias{fit.ChromosomalModel}
\alias{fitGenotypeCone}
\alias{fitGlad}
\alias{fitGlad.CnChipEffectSet}
\alias{fitOne}
\alias{fitOneChromosome}
\alias{fitOneChromosome.SmoothMultiarrayModel}
\alias{fitPlasqUnit}
\alias{fitPlasqUnit.matrix}
\alias{fitQuantileNormFcn}
\alias{fitWRMA}
\alias{fitWRMA.matrix}
\alias{fixSearchPath}
\alias{fixSearchPath.AromaAffymetrix}
\alias{fromCdf}
\alias{fromCdf.AffymetrixProbeTabFile}
\alias{fromChipType}
\alias{fromChipType.AffymetrixCdfFile}
\alias{fromChipType.AffymetrixCsvGenomeInformation}
\alias{fromChipType.AffymetrixProbeTabFile}
\alias{fromChipType.AffymetrixTabularFile}
\alias{fromChipType.AffymetrixTsvFile}
\alias{fromChipType.AromaUnitTabularBinaryFile}
\alias{fromChipType.UflSnpInformation}
\alias{fromCleanChipType}
\alias{fromCleanChipType.AffymetrixCdfFile}
\alias{fromDataFile}
\alias{fromDataFile.ChipEffectFile}
\alias{fromDataFile.FirmaFile}
\alias{fromDataFile.ResidualFile}
\alias{fromDataFile.WeightsFile}
\alias{fromDataSet}
\alias{fromDataSet.ChipEffectSet}
\alias{fromDataSet.FirmaSet}
\alias{fromDataSet.GenomeInformation}
\alias{fromDataSet.ResidualSet}
\alias{fromDataSet.SnpInformation}
\alias{fromDataSet.WeightsSet}
\alias{fromFile}
\alias{fromFile.GenericDataFile}
\alias{fromFiles}
\alias{fromFiles.AffymetrixCelSet}
\alias{fromFiles.ChipEffectSet}
\alias{fromFiles.CnagCfhSet}
\alias{fromFiles.CnChipEffectSet}
\alias{fromFiles.ExonChipEffectSet}
\alias{fromFiles.FirmaSet}
\alias{fromFiles.ResidualSet}
\alias{fromFiles.SnpChipEffectSet}
\alias{fromFiles.WeightsSet}
\alias{fromName}
\alias{fromName.AffymetrixCelSet}
\alias{fromName.CnagCfhSet}
\alias{fromPath}
\alias{fromPath.SampleAnnotationFile}
\alias{fromPath.SampleAnnotationSet}
\alias{gaussianSmoothing}
\alias{gaussianSmoothing.matrix}
\alias{gaussianSmoothing.numeric}
\alias{gaussianSmoothing.RawCopyNumbers}
\alias{gaussKernel}
\alias{gcrmaSummary}
\alias{GenericSummary}
\alias{GenericTabularFile}
\alias{getAlias}
\alias{getAlias.AffymetrixCelSetTuple}
\alias{getAlias.ArrayExplorer}
\alias{getAlias.ChromosomalModel}
\alias{getAlias.CopyNumberSegmentationModel}
\alias{getAlias.GenericDataFile}
\alias{getAlias.TransformReport}
\alias{getAllelePairs}
\alias{getAllelePairUnitSets}
\alias{getAlleleProbePairs}
\alias{getAlleleProbePairs2}
\alias{getAlleleProbePairs2.AffymetrixCdfFile}
\alias{getAm}
\alias{getAM}
\alias{getAm.AffymetrixCelFile}
\alias{getAromaUflFile}
\alias{getAromaUflFile.UflSnpInformation}
\alias{getAromaUgpFile}
\alias{getAromaUgpFile.UgpGenomeInformation}
\alias{getArrays}
\alias{getArraysOfInput}
\alias{getArraysOfInput.ArrayExplorer}
\alias{getArraysOfInput.ChromosomeExplorer}
\alias{getAsteriskTags}
\alias{getAsteriskTags.AffinePlm}
\alias{getAsteriskTags.AffymetrixCelSetTuple}
\alias{getAsteriskTags.AvgPlm}
\alias{getAsteriskTags.CbsModel}
\alias{getAsteriskTags.ChromosomalModel}
\alias{getAsteriskTags.CopyNumberSegmentationModel}
\alias{getAsteriskTags.GladModel}
\alias{getAsteriskTags.MbeiPlm}
\alias{getAsteriskTags.Model}
\alias{getAsteriskTags.ProbeLevelModel}
\alias{getAsteriskTags.RmaPlm}
\alias{getAsteriskTags.SmoothMultiarrayModel}
\alias{getAsteriskTags.SmoothRmaModel}
\alias{getAsteriskTags.SmoothSaModel}
\alias{getAsteriskTags.SnpChipEffectGroupMerge}
\alias{getAsteriskTags.Transform}
\alias{getAsteriskTagss}
\alias{getAsteriskTagss.AffymetrixFileSetReporter}
\alias{getAsteriskTagss.ArrayExplorer}
\alias{getAsteriskTagss.Explorer}
\alias{getAttribute}
\alias{getAttribute.AffymetrixFile}
\alias{getAttributes}
\alias{getAttributes.AffymetrixFile}
\alias{getAttributeXY}
\alias{getAttributeXY.AffymetrixCelFile}
\alias{getAverage}
\alias{getAverage.AffymetrixCelSet}
\alias{getAverage.CnagCfhSet}
\alias{getAverageAsinh}
\alias{getAverageAsinh.AffymetrixCelSet}
\alias{getAverageAsinh.CnagCfhSet}
\alias{getAverageFile}
\alias{getAverageFile.ChipEffectSet}
\alias{getAverageFile.CnChipEffectSet}
\alias{getAverageFile.ExonChipEffectSet}
\alias{getAverageFile.ResidualSet}
\alias{getAverageFile.SnpChipEffectSet}
\alias{getAverageFile.WeightsSet}
\alias{getAverageLog}
\alias{getAverageLog.AffymetrixCelSet}
\alias{getAverageLog.CnagCfhSet}
\alias{getBandwidth}
\alias{getBandwidth.SmoothMultiarrayModel}
\alias{getBaseline}
\alias{getBaseline.CnChipEffectSet}
\alias{getBaseline.SnpChipEffectSet}
\alias{getBytesPerColumn}
\alias{getBytesPerColumn.AromaTabularBinaryFile}
\alias{getCalculateResidualsFunction}
\alias{getCalculateResidualsFunction.AvgPlm}
\alias{getCalculateResidualsFunction.ProbeLevelModel}
\alias{getCalculateResidualsFunction.RmaPlm}
\alias{getCdf}
\alias{getCdf.AffymetrixProbeTabFile}
\alias{getCdf.AffymetrixTsvFile}
\alias{getCdf.AromaUnitTabularBinaryFile}
\alias{getCdf.ChipEffectGroupMerge}
\alias{getCdf.FirmaModel}
\alias{getCdf.FragmentLengthNormalization}
\alias{getCdf.GcContentNormalization}
\alias{getCdf.QualityAssessmentModel}
\alias{getCdf.TransformReport}
\alias{getCellIndices}
\alias{getCellIndices.ChipEffectFile}
\alias{getCellIndices.ChipEffectSet}
\alias{getCellIndices.CnChipEffectFile}
\alias{getCellIndices.CnPlm}
\alias{getCellIndices.CnProbeAffinityFile}
\alias{getCellIndices.ExonChipEffectFile}
\alias{getCellIndices.ExonProbeAffinityFile}
\alias{getCellIndices.ExonRmaPlm}
\alias{getCellIndices.FirmaFile}
\alias{getCellIndices.FirmaSet}
\alias{getCellIndices.ProbeAffinityFile}
\alias{getCellIndices.ProbeLevelModel}
\alias{getCellIndices.ResidualFile}
\alias{getCellIndices.ResidualSet}
\alias{getCellIndices.SnpChipEffectFile}
\alias{getCellIndices.SnpPlm}
\alias{getCellIndices.SnpProbeAffinityFile}
\alias{getCellIndices.WeightsFile}
\alias{getCellIndices.WeightsSet}
\alias{getCellMap}
\alias{getCellMap.FirmaFile}
\alias{getCellMap.ResidualFile}
\alias{getCellMap.WeightsFile}
\alias{getChecksum}
\alias{getChecksum.GenericDataFile}
\alias{getChipEffectFileClass}
\alias{getChipEffectFileClass.ChipEffectSet}
\alias{getChipEffectFileClass.CnChipEffectSet}
\alias{getChipEffectFileClass.ExonChipEffectSet}
\alias{getChipEffectFileClass.SnpChipEffectSet}
\alias{getChipEffectFiles}
\alias{getChipEffectFiles.ChromosomalModel}
\alias{getChipEffects}
\alias{getChipEffects.ExonRmaPlm}
\alias{getChipEffects.QualityAssessmentModel}
\alias{getChipEffectSet}
\alias{getChipEffectSet.CnPlm}
\alias{getChipEffectSet.ExonRmaPlm}
\alias{getChipEffectSet.SnpPlm}
\alias{getChipEffectSetClass}
\alias{getChipEffectSetClass.CnPlm}
\alias{getChipEffectSetClass.ExonRmaPlm}
\alias{getChipEffectSetClass.ProbeLevelModel}
\alias{getChipEffectSetClass.SnpPlm}
\alias{getChipType}
\alias{getChipType.AffymetrixCdfFile}
\alias{getChipType.AffymetrixCelSetReporter}
\alias{getChipType.AffymetrixProbeTabFile}
\alias{getChipType.AffymetrixTsvFile}
\alias{getChipType.AromaUnitTabularBinaryFile}
\alias{getChipTypes}
\alias{getChipTypes.ChromosomalModel}
\alias{getChipTypes.CopyNumberSegmentationModel}
\alias{getChromosome}
\alias{getChromosome.Arguments}
\alias{getChromosomes}
\alias{getChromosomes.Arguments}
\alias{getChromosomes.AromaUgpFile}
\alias{getChromosomes.GenomeInformation}
\alias{getChromosomes.UgpGenomeInformation}
\alias{getChromosomeStats}
\alias{getChromosomeStats.GenomeInformation}
\alias{getCnNames}
\alias{getCNs}
\alias{getCNs.RawCopyNumbers}
\alias{getColClasses}
\alias{getColClasses.AromaTabularBinaryFile}
\alias{getColorMaps}
\alias{getColorMaps.ArrayExplorer}
\alias{getColorMaps.SpatialReporter}
\alias{getCombineAlleles}
\alias{getCombineAlleles.CnChipEffectSet}
\alias{getCommonListElements}
\alias{getCommonPrefix}
\alias{getData}
\alias{getData.AffymetrixCelSet}
\alias{getData.AffymetrixProbeTabFile}
\alias{getData.AffymetrixTsvFile}
\alias{getData.CnagCfhSet}
\alias{getData.UflSnpInformation}
\alias{getData.UgpGenomeInformation}
\alias{getDataColumns}
\alias{getDataColumns.GenomeInformation}
\alias{getDataColumns.UflSnpInformation}
\alias{getDataColumns.UgpGenomeInformation}
\alias{getDataFlat}
\alias{getDataFlat.ChipEffectFile}
\alias{getDataFlat.FirmaFile}
\alias{getDataFlat.ResidualFile}
\alias{getDataFlat.WeightsFile}
\alias{getDataPairs}
\alias{getDataPairs.AllelicCrosstalkCalibration}
\alias{getDataSet}
\alias{getDataSet.FirmaModel}
\alias{getDataSet.QualityAssessmentModel}
\alias{getDescription.AffymetrixFileSet}
\alias{getDeviceResolution}
\alias{getDeviceResolution.default}
\alias{getDimension}
\alias{getDimension.AffymetrixCdfFile}
\alias{getExclCells}
\alias{getExclCells.DChipQuantileNormalization}
\alias{getExclCells.QuantileNormalization}
\alias{getField}
\alias{getField.AffymetrixTsvFile}
\alias{getFields.SnpInformation}
\alias{getFields.UflSnpInformation}
\alias{getFile}
\alias{getFile.GenericDataFileSet}
\alias{getFileClass}
\alias{getFileClass.FirmaSet}
\alias{getFileFormat}
\alias{getFileFormat.AffymetrixCdfFile}
\alias{getFileFormat.AffymetrixCelFile}
\alias{getFilename}
\alias{getFilename.Arguments}
\alias{getFilenameExtension}
\alias{getFilenameExtension.AromaUflFile}
\alias{getFilenameExtension.AromaUgpFile}
\alias{getFilenameExtension.AromaUnitTabularBinaryFile}
\alias{getFiles}
\alias{getFiles.GenericDataFileSet}
\alias{getFileSet}
\alias{getFileSet.AffymetrixFileSetReporter}
\alias{getFileSetClass}
\alias{getFileSetClass.FirmaModel}
\alias{getFileSize.GenericDataFile}
\alias{getFileSize.GenericDataFileSet}
\alias{getFileType}
\alias{getFirmaScores}
\alias{getFirstCellIndices}
\alias{getFirstCellIndices.AffymetrixCdfFile}
\alias{getFirstCellPerUnitIndices}
\alias{getFirstCellPerUnitIndices.ExonChipEffectSet}
\alias{getFitFunction}
\alias{getFitFunction.SmoothMultiarrayModel}
\alias{getFitFunction.SmoothRmaModel}
\alias{getFitFunction.SmoothSaModel}
\alias{getFitUnitFunction}
\alias{getFitUnitFunction.CnPlm}
\alias{getFitUnitFunction.FirmaModel}
\alias{getFitUnitFunction.ProbeLevelModel}
\alias{getFitUnitFunction.UnitModel}
\alias{getFragmentLengths}
\alias{getFragmentLengths.AffymetrixTsvFile}
\alias{getFragmentLengths.SnpInformation}
\alias{getFragmentStarts}
\alias{getFragmentStarts.SnpInformation}
\alias{getFragmentStarts.UflSnpInformation}
\alias{getFragmentStops}
\alias{getFragmentStops.SnpInformation}
\alias{getFragmentStops.UflSnpInformation}
\alias{getFullName}
\alias{getFullName.AffymetrixCelSet}
\alias{getFullName.AffymetrixCelSetTuple}
\alias{getFullName.AffymetrixFileSetReporter}
\alias{getFullName.ChromosomalModel}
\alias{getFullName.CnagCfhSet}
\alias{getFullName.Explorer}
\alias{getFullName.QualityAssessmentModel}
\alias{getFullNames}
\alias{getFullNames.AffymetrixCelSetTuple}
\alias{getFullNames.ChipEffectSetTuple}
\alias{getFullNames.ChromosomalModel}
\alias{getFullNames.ChromosomeExplorer}
\alias{getFullNames.CopyNumberSegmentationModel}
\alias{getFullNames.GenericDataFileSet}
\alias{getGc}
\alias{getGc.AffymetrixTsvFile}
\alias{getGenome}
\alias{getGenome.ChromosomalModel}
\alias{getGenomeData}
\alias{getGenomeData.ChromosomalModel}
\alias{getGenomeFile}
\alias{getGenomeFile.ChromosomalModel}
\alias{getGenomeInformation}
\alias{getGenomeVersion}
\alias{getGenomeVersion.AromaUgpFile}
\alias{getHeader.GenericTabularFile}
\alias{getHeaderLength}
\alias{getHeaderLength.CnagCfhFile}
\alias{getHeaderV3}
\alias{getHeaderV3.AffymetrixCelFile}
\alias{getIdentifier}
\alias{getIdentifier.AffymetrixCelFile}
\alias{getIdentifier.AffymetrixCelSet}
\alias{getIdentifier.AffymetrixFileSet}
\alias{getIdentifier.CnagCfhFile}
\alias{getIdentifier.CnagCfhSet}
\alias{getImage}
\alias{getImage.ResidualFile}
\alias{getImage.WeightsFile}
\alias{getIncludePath}
\alias{getIncludePath.Explorer}
\alias{getIndexToRowMap}
\alias{getIndexToRowMap.AffymetrixProbeTabFile}
\alias{getInputDataSet}
\alias{getIntensities}
\alias{getLabel.AffymetrixFile}
\alias{getLabel.Model}
\alias{getListOfCdfs}
\alias{getListOfCdfs.AffymetrixCelSetTuple}
\alias{getListOfCdfs.ArrayExplorer}
\alias{getListOfCdfs.ChromosomalModel}
\alias{getListOfCdfs.CopyNumberSegmentationModel}
\alias{getListOfChipEffects}
\alias{getListOfChipEffects.ChromosomalModel}
\alias{getListOfChipEffects.CopyNumberSegmentationModel}
\alias{getListOfChipEffectSets}
\alias{getListOfChipEffectSets.ChromosomalModel}
\alias{getListOfChipEffectSets.CopyNumberSegmentationModel}
\alias{getListOfGenomeInformations}
\alias{getListOfGenomeInformations.ChromosomalModel}
\alias{getListOfReferences}
\alias{getListOfReferences.CopyNumberSegmentationModel}
\alias{getListOfReporters}
\alias{getListOfReporters.ArrayExplorer}
\alias{getListOfSets}
\alias{getListOfSets.AffymetrixCelSetTuple}
\alias{getLociFields}
\alias{getLociFields.RawCopyNumbers}
\alias{getLog2Ratios}
\alias{getLog2Ratios.CopyNumberSegmentationModel}
\alias{getMainPath}
\alias{getMainPath.AffymetrixFileSetReporter}
\alias{getMainPath.Explorer}
\alias{getMergeFunction}
\alias{getMergeFunction.ChipEffectGroupMerge}
\alias{getMergeFunction.SnpChipEffectGroupMerge}
\alias{getMergeGroups}
\alias{getMergeGroups.ExonChipEffectSet}
\alias{getMergeStrands}
\alias{getMergeStrands.SnpChipEffectSet}
\alias{getModel}
\alias{getMonoCell}
\alias{getMonoCell.AffymetrixCdfFile}
\alias{getName.AffymetrixCelSetTuple}
\alias{getName.ChromosomalModel}
\alias{getName.CopyNumberSegmentationModel}
\alias{getName.FirmaModel}
\alias{getName.QualityAssessmentModel}
\alias{getNameOfInput}
\alias{getNameOfInput.ArrayExplorer}
\alias{getNameOfInput.ChromosomeExplorer}
\alias{getNameOfInput.Explorer}
\alias{getNames}
\alias{getNames.AffymetrixCelSetTuple}
\alias{getNames.ChromosomalModel}
\alias{getNames.ChromosomeExplorer}
\alias{getNames.CopyNumberSegmentationModel}
\alias{getOutputDataSet}
\alias{getOutputDataSet.ChipEffectTransform}
\alias{getOutputDataSet.FragmentLengthNormalization}
\alias{getOutputDataSet.GcContentNormalization}
\alias{getOutputFiles}
\alias{getOutputFiles.Transform}
\alias{getOutputIdentifier}
\alias{getOutputIdentifier.QuantileNormalization}
\alias{getOutputTuple}
\alias{getOutputTuple.SmoothMultiarrayModel}
\alias{getPackageName}
\alias{getParameters.AllelicCrosstalkCalibration}
\alias{getParameters.BackgroundCorrection}
\alias{getParameters.ChipEffectFile}
\alias{getParameters.ChipEffectGroupMerge}
\alias{getParameters.ChipEffectSet}
\alias{getParameters.CnChipEffectFile}
\alias{getParameters.DChipQuantileNormalization}
\alias{getParameters.ExonChipEffectFile}
\alias{getParameters.FragmentLengthNormalization}
\alias{getParameters.GcContentNormalization}
\alias{getParameters.GcRmaBackgroundCorrection}
\alias{getParameters.Model}
\alias{getParameters.OpticalBackgroundCorrection}
\alias{getParameters.ParameterCelFile}
\alias{getParameters.ProbeAffinityFile}
\alias{getParameters.QuantileNormalization}
\alias{getParameters.ResidualFile}
\alias{getParameters.ResidualSet}
\alias{getParameters.RmaBackgroundCorrection}
\alias{getParameters.ScaleNormalization}
\alias{getParameters.SnpChipEffectFile}
\alias{getParameters.SnpChipEffectGroupMerge}
\alias{getParameters.Transform}
\alias{getParameters.WeightsFile}
\alias{getParameters.WeightsSet}
\alias{getParametersAsString}
\alias{getParametersAsString.ChipEffectFile}
\alias{getParametersAsString.ChipEffectSet}
\alias{getParametersAsString.Model}
\alias{getParametersAsString.ProbeAffinityFile}
\alias{getParametersAsString.ResidualFile}
\alias{getParametersAsString.ResidualSet}
\alias{getParametersAsString.Transform}
\alias{getParametersAsString.WeightsFile}
\alias{getParametersAsString.WeightsSet}
\alias{getParameterSet}
\alias{getParameterSet.AvgPlm}
\alias{getParameterSet.CnPlm}
\alias{getParameterSet.Model}
\alias{getParameterSet.ProbeLevelModel}
\alias{getParameterSet.RmaPlm}
\alias{getParameterSet.SnpPlm}
\alias{getParentPath}
\alias{getParentPath.ChromosomalModel}
\alias{getPath.AffymetrixCelSetReporter}
\alias{getPath.AffymetrixFileSetReporter}
\alias{getPath.ChromosomalModel}
\alias{getPath.ChromosomeExplorer}
\alias{getPath.Explorer}
\alias{getPath.QualityAssessmentModel}
\alias{getPathnames}
\alias{getPatterns}
\alias{getPatterns.SampleAnnotationFile}
\alias{getPcuTheta}
\alias{getPcuTheta.ChromosomalModel}
\alias{getPhysicalPositions}
\alias{getPhysicalPositions.RawCopyNumbers}
\alias{getPlasqTypes}
\alias{getPlm}
\alias{getPlm.FirmaModel}
\alias{getPlm.QualityAssessmentModel}
\alias{getPloidy}
\alias{getPloidy.AffymetrixCelFile}
\alias{getPosition.AffymetrixTsvFile}
\alias{getPositionChipTypeUnit}
\alias{getPositionChipTypeUnit.ChromosomalModel}
\alias{getPositionChipTypeUnit.CopyNumberSegmentationModel}
\alias{getPositions}
\alias{getProbeAffinities}
\alias{getProbeAffinities.ExonRmaPlm}
\alias{getProbeAffinityFile}
\alias{getProbeAffinityFile.AffinePlm}
\alias{getProbeAffinityFile.AvgPlm}
\alias{getProbeAffinityFile.CnPlm}
\alias{getProbeAffinityFile.ExonRmaPlm}
\alias{getProbeAffinityFile.MbeiPlm}
\alias{getProbeAffinityFile.RmaPlm}
\alias{getProbeAffinityFile.SnpPlm}
\alias{getRawCnData}
\alias{getRawCnData.CopyNumberSegmentationModel}
\alias{getReadArguments}
\alias{getReadArguments.GenericTabularFile}
\alias{getRectangle}
\alias{getRectangle.AffymetrixCelFile}
\alias{getReferenceFiles}
\alias{getReferenceFiles.CopyNumberSegmentationModel}
\alias{getRefSetTuple}
\alias{getRefSetTuple.CopyNumberSegmentationModel}
\alias{getRegions}
\alias{getRegions.CopyNumberSegmentationModel}
\alias{getRegions.profileCGH}
\alias{getReportPath}
\alias{getReportPath.ChromosomalModel}
\alias{getReportSet}
\alias{getReportSet.AffymetrixFileSetReporter}
\alias{getReportSet.SpatialReporter}
\alias{getResidualFileClass}
\alias{getResidualFileClass.ResidualSet}
\alias{getResiduals}
\alias{getResidualSet}
\alias{getResidualSet.ProbeLevelModel}
\alias{getResidualSetClass}
\alias{getResidualSetClass.ProbeLevelModel}
\alias{getRestructor}
\alias{getRestructor.AffymetrixCdfFile}
\alias{getRootPath}
\alias{getRootPath.AffymetrixFileSetReporter}
\alias{getRootPath.ChipEffectTransform}
\alias{getRootPath.ChromosomalModel}
\alias{getRootPath.Explorer}
\alias{getRootPath.FirmaModel}
\alias{getRootPath.ProbeLevelModel}
\alias{getRootPath.ProbeLevelTransform}
\alias{getRootPath.QualityAssessmentModel}
\alias{getRootPath.SmoothMultiarrayModel}
\alias{getRootPath.Transform}
\alias{getRootPath.TransformReport}
\alias{getSetTag}
\alias{getSetTag.ChromosomalModel}
\alias{getSetTuple}
\alias{getSetTuple.ArrayExplorer}
\alias{getSetTuple.ChromosomalModel}
\alias{getSetTuple.ChromosomeExplorer}
\alias{getSetTuple.CopyNumberSegmentationModel}
\alias{getSiblings}
\alias{getSnpInformation}
\alias{getSnpInformation.AffymetrixCdfFile}
\alias{getSnpNames}
\alias{getSubsetToAvg}
\alias{getSubsetToAvg.AllelicCrosstalkCalibration}
\alias{getSubsetToAvg.QuantileNormalization}
\alias{getSubsetToAvg.ScaleNormalization}
\alias{getSubsetToFit}
\alias{getSubsetToFit.FragmentLengthNormalization}
\alias{getSubsetToFit.GcContentNormalization}
\alias{getSubsetToUpdate}
\alias{getSubsetToUpdate.BackgroundCorrection}
\alias{getSubsetToUpdate.DChipQuantileNormalization}
\alias{getSubsetToUpdate.QuantileNormalization}
\alias{getSubsetToUpdate.ScaleNormalization}
\alias{getTableOfArrays}
\alias{getTags}
\alias{getTags.AffymetrixCelSetTuple}
\alias{getTags.ChromosomalModel}
\alias{getTags.FirmaModel}
\alias{getTags.QualityAssessmentModel}
\alias{getTagsOfInput}
\alias{getTagsOfInput.ArrayExplorer}
\alias{getTagsOfInput.ChromosomeExplorer}
\alias{getTagsOfInput.Explorer}
\alias{getTargetDistribution}
\alias{getTargetDistribution.QuantileNormalization}
\alias{getTargetDistributionIdentifier}
\alias{getTargetDistributionIdentifier.QuantileNormalization}
\alias{getTargetDistributionPathname}
\alias{getTargetDistributionPathname.QuantileNormalization}
\alias{getTargetFunction}
\alias{getTargetFunction.FragmentLengthNormalization}
\alias{getTargetFunction.GcContentNormalization}
\alias{getTemplatePath}
\alias{getTemplatePath.Explorer}
\alias{getTimestamp}
\alias{getTimestamp.CnagCfhFile}
\alias{getTimestamps}
\alias{getTimestamps.AffymetrixCelSet}
\alias{getTimestamps.CnagCfhSet}
\alias{getTuple}
\alias{getUnitGroupCellMap}
\alias{getUnitGroupCellMap.AffymetrixCdfFile}
\alias{getUnitIndices}
\alias{getUnitIntensities}
\alias{getUnitNames}
\alias{getUnitsAt}
\alias{getUnitsAt.AromaUgpFile}
\alias{getUnitSizes}
\alias{getUnitsOnChromosome}
\alias{getUnitsOnChromosome.GenomeInformation}
\alias{getUnitsOnChromosome.UgpGenomeInformation}
\alias{getWeights}
\alias{getWeightsFileClass}
\alias{getWeightsFileClass.WeightsSet}
\alias{getWeightsSet}
\alias{getWeightsSet.ProbeLevelModel}
\alias{getWeightsSetClass}
\alias{getWeightsSetClass.ProbeLevelModel}
\alias{getXAM}
\alias{getXY}
\alias{getXY.RawCopyNumbers}
\alias{getYY}
\alias{getYY.TransformReport}
\alias{hasAttributeXY}
\alias{hasAttributeXY.AffymetrixCelFile}
\alias{hasTag}
\alias{hasTag.GenericDataFile}
\alias{hasTag.GenericDataFileSet}
\alias{hasTags}
\alias{hasTags.GenericDataFile}
\alias{hasTags.GenericDataFileSet}
\alias{highlight}
\alias{highlight.AffymetrixCelFile}
\alias{highlight.ProbeLevelModel}
\alias{identifyCells}
\alias{identifyCells.AffymetrixCdfFile}
\alias{image270}
\alias{importFrom}
\alias{importFrom.AromaUnitTabularBinaryFile}
\alias{importFromAffymetrixNetAffxCsvFile}
\alias{importFromAffymetrixNetAffxCsvFile.AromaUflFile}
\alias{importFromAffymetrixNetAffxCsvFile.AromaUgpFile}
\alias{importFromAffymetrixNetAffxCsvFile.AromaUnitTabularBinaryFile}
\alias{importFromAffymetrixTabularFile}
\alias{importFromAffymetrixTabularFile.AromaUgpFile}
\alias{importFromAffymetrixTabularFile.AromaUnitTabularBinaryFile}
\alias{importFromDChip}
\alias{importFromDChipGenomeInformation}
\alias{importFromDChipGenomeInformation.AromaUnitTabularBinaryFile}
\alias{importFromGenericTabularFile}
\alias{importFromGenericTabularFile.AromaUgpFile}
\alias{importFromGenericTabularFile.AromaUnitTabularBinaryFile}
\alias{importFromGenomeInformation}
\alias{importFromGenomeInformation.AromaUgpFile}
\alias{importFromGenomeInformation.AromaUnitTabularBinaryFile}
\alias{importFromTable}
\alias{importFromTable.FileMatrix}
\alias{indexOf}
\alias{indexOf.GenericDataFileSet}
\alias{indexOfArrays}
\alias{indexOfUnits}
\alias{indexOfUnits.AromaUnitTabularBinaryFile}
\alias{inferParameters}
\alias{inferParameters.CnChipEffectSet}
\alias{inferParameters.SnpChipEffectSet}
\alias{isDuplicated}
\alias{isFile.GenericDataFile}
\alias{isPm}
\alias{isSnpChip}
\alias{lapply.AromaTabularBinaryFile}
\alias{lapplyInChunks}
\alias{lapplyInChunks.numeric}
\alias{library}
\alias{lines.CopyNumberOutliers}
\alias{lines.CopyNumberRegions}
\alias{lines.RawCopyNumbers}
\alias{locallyUnique}
\alias{locallyUnique.default}
\alias{log2abs}
\alias{log2neg}
\alias{log2pos}
\alias{matchPatterns}
\alias{matchPatterns.SampleAnnotationFile}
\alias{mergeByCommonTails}
\alias{mergeGroups}
\alias{mergeGroups.ChipEffectFile}
\alias{mergeStrands}
\alias{mergeStrands.CnChipEffectFile}
\alias{mergeStrands.SnpChipEffectFile}
\alias{migrateTool}
\alias{migrateTool.CopyNumberSegmentationModel}
\alias{nbrOfArrays}
\alias{nbrOfArrays.AffymetrixCelSetReporter}
\alias{nbrOfArrays.QualityAssessmentModel}
\alias{nbrOfArrays.TransformReport}
\alias{nbrOfCells}
\alias{nbrOfCells.AffymetrixCdfFile}
\alias{nbrOfCells.AffymetrixCelFile}
\alias{nbrOfCells.CnagCfhFile}
\alias{nbrOfChipTypes}
\alias{nbrOfChipTypes.ArrayExplorer}
\alias{nbrOfColumns}
\alias{nbrOfColumns.AffymetrixCdfFile}
\alias{nbrOfColumns.AromaTabularBinaryFile}
\alias{nbrOfEnzymes}
\alias{nbrOfEnzymes.AromaUflFile}
\alias{nbrOfFiles}
\alias{nbrOfLoci}
\alias{nbrOfLoci.RawCopyNumbers}
\alias{nbrOfQcUnits}
\alias{nbrOfQcUnits.AffymetrixCdfFile}
\alias{nbrOfRegions}
\alias{nbrOfRegions.CopyNumberOutliers}
\alias{nbrOfRegions.CopyNumberRegions}
\alias{nbrOfRows}
\alias{nbrOfRows.AffymetrixCdfFile}
\alias{nbrOfRows.AromaTabularBinaryFile}
\alias{nbrOfSnps}
\alias{nbrOfSnps.CnagCfhFile}
\alias{nbrOfUnits}
\alias{nbrOfUnits.AffymetrixCdfFile}
\alias{nbrOfUnits.AromaUnitTabularBinaryFile}
\alias{nbrOfUnits.GenomeInformation}
\alias{nbrOfUnits.SnpInformation}
\alias{nbrOfUnits.UflSnpInformation}
\alias{ncol.AromaTabularBinaryFile}
\alias{newPlot}
\alias{newPlot.CopyNumberSegmentationModel}
\alias{normalizeGenotypeCone}
\alias{normalizeGenotypeCone.matrix}
\alias{normalizeQuantileSpline}
\alias{nrow.AromaTabularBinaryFile}
\alias{onFitAddGenotypeCalls}
\alias{onFitAddGenotypeCalls.default}
\alias{onFitAddGenotypeCalls.GladModel}
\alias{plasqGaussian}
\alias{plot.CopyNumberSegmentationModel}
\alias{plot.profileCGH}
\alias{plot.RawCopyNumbers}
\alias{plotBasepair}
\alias{plotBasepair.AllelicCrosstalkCalibration}
\alias{plotImage}
\alias{plotMvsX}
\alias{plotNuse}
\alias{plotNuse.QualityAssessmentModel}
\alias{plotProfile2}
\alias{plotProfile2.profileCGH}
\alias{plotRawCNs}
\alias{plotRawCNs.profileCGH}
\alias{plotRle}
\alias{plotRle.QualityAssessmentModel}
\alias{plotXYCurve}
\alias{plotXYCurve.TransformReport}
\alias{plotXYCurveLog2}
\alias{plotXYCurveLog2.TransformReport}
\alias{points.RawCopyNumbers}
\alias{pointsRawCNs}
\alias{pointsRawCNs.default}
\alias{pointsRawCNs.profileCGH}
\alias{predict.lowess}
\alias{print.GenericSummary}
\alias{print.Interface}
\alias{process}
\alias{range.AffymetrixCelFile}
\alias{range.AffymetrixCelSet}
\alias{range.CnagCfhFile}
\alias{range.CnagCfhSet}
\alias{RawCopyNumbers}
\alias{read10K}
\alias{read10K.DChipSnpInformation}
\alias{read250K}
\alias{read250K.DChipSnpInformation}
\alias{read250KHg17}
\alias{read250KHg17.DChipGenomeInformation}
\alias{read50K}
\alias{read50K.DChipSnpInformation}
\alias{read50KHg17}
\alias{read50KHg17.DChipGenomeInformation}
\alias{readCdfGroupStrands}
\alias{readCfhHeader}
\alias{readCfhHeader.default}
\alias{readCfhUnits}
\alias{readCfhUnits.default}
\alias{readCfnHeader}
\alias{readCfnHeader.default}
\alias{readCfnUnits}
\alias{readCfnUnits.default}
\alias{readChecksum}
\alias{readChecksum.GenericDataFile}
\alias{readData}
\alias{readData.AffymetrixCsvGenomeInformation}
\alias{readData.AffymetrixTsvFile}
\alias{readData.AromaTabularBinaryFile}
\alias{readData.AromaUflFile}
\alias{readData.AromaUgpFile}
\alias{readData.DChipGenomeInformation}
\alias{readData.DChipSnpInformation}
\alias{readData.GenericTabularFile}
\alias{readData.GenomeInformation}
\alias{readData.SampleAnnotationFile}
\alias{readData.SnpInformation}
\alias{readData.UflSnpInformation}
\alias{readData.UgpGenomeInformation}
\alias{readDataUnitChromosomePosition}
\alias{readDataUnitChromosomePosition.AffymetrixNetAffxCsvFile}
\alias{readDataUnitFragmentLength}
\alias{readDataUnitFragmentLength.AffymetrixNetAffxCsvFile}
\alias{readFooter}
\alias{readFooter.AromaTabularBinaryFile}
\alias{readGenomeWideSNP}
\alias{readGenomeWideSNP.DChipGenomeInformation}
\alias{readHeader.AromaTabularBinaryFile}
\alias{readHeader.GenericTabularFile}
\alias{readMouse430}
\alias{readMouse430.DChipGenomeInformation}
\alias{readMouse430KenHardwired}
\alias{readMouse430KenHardwired.DChipGenomeInformation}
\alias{readPlasqTypes}
\alias{readPlasqTypes.AffymetrixCdfFile}
\alias{readTableInternal}
\alias{readTableInternal.GenomeInformation}
\alias{readTableInternal.SnpInformation}
\alias{readUnits}
\alias{readUnits.AffymetrixCelSet}
\alias{readUnits.ChipEffectFile}
\alias{readUnits.ChipEffectSet}
\alias{readUnits.CnagCfhSet}
\alias{readUnits.CnChipEffectFile}
\alias{readUnits.ExonChipEffectFile}
\alias{readUnits.FirmaFile}
\alias{readUnits.FirmaSet}
\alias{readUnits.ParameterCelFile}
\alias{readUnits.ProbeAffinityFile}
\alias{readUnits.ResidualFile}
\alias{readUnits.ResidualSet}
\alias{readUnits.SnpChipEffectFile}
\alias{readUnits.WeightsFile}
\alias{readUnits.WeightsSet}
\alias{renameTo}
\alias{renameTo.GenericDataFile}
\alias{renameToUpperCaseExt}
\alias{renameToUpperCaseExt.GenericDataFile}
\alias{reorder.GenericDataFileSet}
\alias{require}
\alias{rescale}
\alias{rescale.AllelicCrosstalkCalibration}
\alias{restruct}
\alias{restruct.AffymetrixCdfFile}
\alias{rgbTransform}
\alias{rgbTransform.Image}
\alias{rmaSummary}
\alias{rowMads}
\alias{rowSds}
\alias{rowVars}
\alias{SampleAnnotationFile}
\alias{SampleAnnotationSet}
\alias{sapply}
\alias{sapply.default}
\alias{sapply.GenericDataFileSet}
\alias{seq.TransformReport}
\alias{setAlias}
\alias{setAlias.AffymetrixCelSetTuple}
\alias{setAlias.ArrayExplorer}
\alias{setAlias.ChromosomalModel}
\alias{setAlias.CopyNumberSegmentationModel}
\alias{setAlias.GenericDataFile}
\alias{setAlias.TransformReport}
\alias{setArrays}
\alias{setAttribute}
\alias{setAttribute.AffymetrixFile}
\alias{setAttributes}
\alias{setAttributes.AffymetrixFile}
\alias{setAttributesBy}
\alias{setAttributesBy.AffymetrixCelSet}
\alias{setAttributesBy.AffymetrixFile}
\alias{setAttributesBySampleAnnotationFile}
\alias{setAttributesBySampleAnnotationFile.AffymetrixCelSet}
\alias{setAttributesBySampleAnnotationSet}
\alias{setAttributesBySampleAnnotationSet.AffymetrixCelSet}
\alias{setAttributesByTags}
\alias{setAttributesByTags.AffymetrixCelFile}
\alias{setAttributesByTags.AffymetrixFile}
\alias{setAttributesByTags.AromaTabularBinaryFile}
\alias{setAttributeXY}
\alias{setAttributeXY.AffymetrixCelFile}
\alias{setBandwidth}
\alias{setBandwidth.SmoothMultiarrayModel}
\alias{setCdf}
\alias{setColorMaps}
\alias{setColorMaps.ArrayExplorer}
\alias{setColorMaps.SpatialReporter}
\alias{setCombineAlleles}
\alias{setCombineAlleles.CnChipEffectSet}
\alias{setCombineAlleles.CnPlm}
\alias{setCombineAlleles.CnProbeAffinityFile}
\alias{setCustomFindCdf}
\alias{setCytoband}
\alias{setCytoband.ChromosomeExplorer}
\alias{setDecodeFunction}
\alias{setDecodeFunction.ParameterCelFile}
\alias{setEncodeFunction}
\alias{setEncodeFunction.ParameterCelFile}
\alias{setGenome}
\alias{setGenome.ChromosomalModel}
\alias{setLabel.AffymetrixFile}
\alias{setLabel.Model}
\alias{setMergeFunction}
\alias{setMergeFunction.ChipEffectGroupMerge}
\alias{setMergeGroups}
\alias{setMergeGroups.ExonChipEffectSet}
\alias{setMergeGroups.ExonProbeAffinityFile}
\alias{setMergeGroups.ExonRmaPlm}
\alias{setMergeStrands}
\alias{setMergeStrands.SnpChipEffectSet}
\alias{setMergeStrands.SnpPlm}
\alias{setMergeStrands.SnpProbeAffinityFile}
\alias{setRestructor}
\alias{setTags}
\alias{setTags.AffymetrixCelSetTuple}
\alias{setTags.Transform}
\alias{setup}
\alias{setup.AffymetrixFileSetReporter}
\alias{setup.ArrayExplorer}
\alias{setup.ChromosomeExplorer}
\alias{setup.Explorer}
\alias{SmoothSaModel}
\alias{smoothScatterMvsA}
\alias{smoothWRMA}
\alias{smoothWRMA.matrix}
\alias{smoothWSA}
\alias{smoothWSA.matrix}
\alias{sort.RawCopyNumbers}
\alias{splitByCommonTails}
\alias{splitInChunks}
\alias{splitInChunks.numeric}
\alias{sqrtabs}
\alias{sqrtneg}
\alias{sqrtpos}
\alias{stext}
\alias{stext.default}
\alias{stextChipType}
\alias{stextChipType.AffymetrixCdfFile}
\alias{stextLabel}
\alias{stextLabel.AffymetrixFile}
\alias{stextLabels}
\alias{stextLabels.AffymetrixFile}
\alias{stextSize}
\alias{stextSize.AffymetrixFile}
\alias{stringTree}
\alias{stringTree.character}
\alias{subset.AromaTabularBinaryFile}
\alias{summary.AromaTabularBinaryFile}
\alias{summary.RawCopyNumbers}
\alias{testAttributes}
\alias{testAttributes.AffymetrixFile}
\alias{transformAffine}
\alias{transformAffine.AffymetrixCelFile}
\alias{transformAffine.AffymetrixCelSet}
\alias{translateColumnNames}
\alias{translateColumnNames.AffymetrixTabularFile}
\alias{translateColumnNames.GenericTabularFile}
\alias{UflSnpInformation}
\alias{updateData}
\alias{updateData.AromaTabularBinaryFile}
\alias{updateDataColumn}
\alias{updateDataColumn.AromaTabularBinaryFile}
\alias{updateDataFlat}
\alias{updateDataFlat.ChipEffectFile}
\alias{updateDataFlat.FirmaFile}
\alias{updateDataFlat.ResidualFile}
\alias{updateDataFlat.WeightsFile}
\alias{updateOnChipTypeJS}
\alias{updateOnChipTypeJS.ArrayExplorer}
\alias{updateOnLoadJS}
\alias{updateOnLoadJS.ArrayExplorer}
\alias{updateSamplesFile}
\alias{updateUnits}
\alias{updateUnits.ChipEffectFile}
\alias{updateUnits.ChipEffectSet}
\alias{updateUnits.FirmaFile}
\alias{updateUnits.FirmaSet}
\alias{updateUnits.ParameterCelFile}
\alias{updateUnits.ProbeAffinityFile}
\alias{updateUnits.ResidualFile}
\alias{updateUnits.ResidualSet}
\alias{updateUnits.WeightsFile}
\alias{updateUnits.WeightsSet}
\alias{uses}
\alias{uses.Interface}
\alias{validateChecksum}
\alias{validateChecksum.GenericDataFile}
\alias{verify}
\alias{verify.AffymetrixCsvGenomeInformation}
\alias{verify.AffymetrixTsvFile}
\alias{verify.DChipGenomeInformation}
\alias{verify.DChipSnpInformation}
\alias{verify.GenericTabularFile}
\alias{verify.UflSnpInformation}
\alias{verify.UgpGenomeInformation}
\alias{whatDataType}
\alias{whatDataType.default}
\alias{write}
\alias{writeChecksum}
\alias{writeChecksum.GenericDataFile}
\alias{writeGraphs}
\alias{writeGraphs.ChromosomeExplorer}
\alias{writeImage}
\alias{writeImage.ResidualFile}
\alias{writeImage.WeightsFile}
\alias{writeImageCombined}
\alias{writeImageCombined.TransformReport}
\alias{writeImages}
\alias{writeImages.SpatialReporter}
\alias{writeImages.TransformReport}
\alias{writeRegions}
\alias{writeRegions.ChromosomeExplorer}
\alias{writeRegions.CopyNumberSegmentationModel}
\alias{writeRegions.GladModel}
\alias{writeRegions.profileCGH}
\alias{writeWig}
\alias{writeWig.CnChipEffectSet}
\alias{yellow.colors}
\alias{asMatrixOfFiles}
\alias{asMatrixOfFiles.AffymetrixCelSetTuple}
\alias{getMatrixChipEffectFiles}
\alias{getMatrixChipEffectFiles.CopyNumberSegmentationModel}
\alias{getReferenceSetTuple}
\alias{getReferenceSetTuple.CopyNumberSegmentationModel}
\alias{isPaired}
\alias{isPaired.CopyNumberSegmentationModel}
\alias{findPngDevice}
\alias{findPngDevice.default}
\alias{fitWHLAPLM}
\alias{fitWHLAPLM.matrix}
\alias{fitWHRCModel}
\alias{fitWHRCModel.matrix}
\alias{fitWLAPLM}
\alias{fitWLAPLM.matrix}
\alias{fitWRCModel}
\alias{fitWRCModel.matrix}
\alias{getAsteriskTags.HetLogAddPlm}
\alias{getAsteriskTags.UnitModel}
\alias{getTags.UnitModel}
\alias{plotAxesLayers}
\alias{plotAxesLayers.CopyNumberSegmentationModel}
\alias{plotChromosomesLayers}
\alias{plotChromosomesLayers.CopyNumberSegmentationModel}
\alias{plotCopyNumberRegionLayers}
\alias{plotCopyNumberRegionLayers.CopyNumberSegmentationModel}
\alias{plotCytobandLayers}
\alias{plotCytobandLayers.CopyNumberSegmentationModel}
\alias{plotFitLayers}
\alias{plotFitLayers.CopyNumberSegmentationModel}
\alias{plotGridHorizontalLayers}
\alias{plotGridHorizontalLayers.CopyNumberSegmentationModel}
\alias{plotRawCopyNumbers}
\alias{plotRawCopyNumbers.CopyNumberSegmentationModel}
\alias{writeAxesLayers}
\alias{writeAxesLayers.ChromosomeExplorer}
\alias{writeCopyNumberRegionLayers}
\alias{writeCopyNumberRegionLayers.ChromosomeExplorer}
\alias{writeCytobandLayers}
\alias{writeCytobandLayers.ChromosomeExplorer}
\alias{writeGridHorizontalLayers}
\alias{writeGridHorizontalLayers.ChromosomeExplorer}
\alias{writeRawCopyNumberLayers}
\alias{writeRawCopyNumberLayers.ChromosomeExplorer}
\alias{as.character.CopyNumberChromosomalModel}
\alias{calculateChromosomeStatistics.CopyNumberChromosomalModel}
\alias{clearCache.CopyNumberChromosomalModel}
\alias{clearCache.UgpGenomeInformation}
\alias{extractRawCopyNumbers.CopyNumberChromosomalModel}
\alias{getAsteriskTags.RawCopyNumberModel}
\alias{getFragmentLengths.UflSnpInformation}
\alias{getMatrixChipEffectFiles.CopyNumberChromosomalModel}
\alias{getRawCnData.CopyNumberChromosomalModel}
\alias{getReferenceSetTuple.CopyNumberChromosomalModel}
\alias{getRefSetTuple.CopyNumberChromosomalModel}
\alias{getReportPathPattern}
\alias{getReportPathPattern.Explorer}
\alias{getRootPath.SmoothRmaModel}
\alias{getSampleLayerName}
\alias{getSampleLayerName.Explorer}
\alias{getSampleLayerPrefix}
\alias{getSampleLayerPrefix.Explorer}
\alias{getSetTag.RawCopyNumberModel}
\alias{getSubname}
\alias{getSubname.Explorer}
\alias{getTags.CopyNumberSegmentationModel}
\alias{getTargetFunctions}
\alias{getTargetFunctions.FragmentLengthNormalization}
\alias{isPaired.CopyNumberChromosomalModel}
\alias{nbrOfEnzymes.SnpInformation}
\alias{nbrOfEnzymes.UflSnpInformation}
\alias{newPlot.CopyNumberChromosomalModel}
\alias{plotAxesLayers.CopyNumberChromosomalModel}
\alias{plotChromosomesLayers.CopyNumberChromosomalModel}
\alias{plotCytobandLayers.CopyNumberChromosomalModel}
\alias{plotFitLayers.CopyNumberChromosomalModel}
\alias{plotGridHorizontalLayers.CopyNumberChromosomalModel}
\alias{plotRawCopyNumbers.CopyNumberChromosomalModel}
\alias{plotSampleLayers}
\alias{plotSampleLayers.CopyNumberChromosomalModel}
\alias{rowCollapse}
\alias{setReportPathPattern}
\alias{setReportPathPattern.Explorer}
\alias{setSubname}
\alias{setSubname.Explorer}
\alias{splitByReportPathPattern}
\alias{splitByReportPathPattern.Explorer}
\alias{doCRMA}
\alias{doCRMA.default}
\alias{rescaleByAll}
\alias{rescaleByAll.AllelicCrosstalkCalibration}
\alias{rescaleByGroups}
\alias{rescaleByGroups.AllelicCrosstalkCalibration}
\alias{getAsteriskTags.AllelicCrosstalkCalibration}
\alias{clearCache.AllelicCrosstalkCalibration}
\alias{getAsteriskTags.FragmentLengthNormalization}
\alias{getAsteriskTags.AffineCnPlm}
\alias{getAsteriskTags.AffineSnpPlm}
\alias{getAsteriskTags.AvgCnPlm}
\alias{getAsteriskTags.AvgSnpPlm}
\alias{getAsteriskTags.ExonRmaPlm}
\alias{getAsteriskTags.FirmaModel}
\alias{getAsteriskTags.HetLogAddCnPlm}
\alias{getAsteriskTags.HetLogAddSnpPlm}
\alias{getAsteriskTags.MbeiCnPlm}
\alias{getAsteriskTags.MbeiSnpPlm}
\alias{getAsteriskTags.RmaCnPlm}
\alias{getAsteriskTags.RmaSnpPlm}
\alias{getAsteriskTags.AffymetrixFileSetReporter}
\alias{getAsteriskTags.ArrayExplorer}
\alias{getAsteriskTags.Explorer}
\alias{setGenomeInformation}
\alias{setGenomeInformation.AffymetrixCdfFile}
\alias{setSnpInformation}
\alias{setSnpInformation.AffymetrixCdfFile}
\alias{isCompatibleWithCdf}
\alias{isCompatibleWithCdf.GenomeInformation}
\alias{isCompatibleWithCdf.SnpInformation}
\alias{isCompatibleWithCdf.UflSnpInformation}
\alias{isCompatibleWithCdf.UgpGenomeInformation}
\alias{nbrOfUnits.UgpGenomeInformation}
\alias{clearCache.AromaUnitTabularBinaryFile}
\alias{getAsteriskTags.QualityAssessmentModel}
\alias{getChipType.UflSnpInformation}
\alias{getChipType.UgpGenomeInformation}
\alias{patch}
\alias{patch.AromaAffymetrix}
\alias{update.AromaAffymetrix}
\alias{summaryOfUnits}
\alias{summaryOfUnits.AromaUflFile}
\alias{AffymetrixAptSummaryFile}
\alias{AffymetrixCnChpFile}
\alias{allocateFromCdf.AromaUfcFile}
\alias{AromaUfcFile}
\alias{as.character.AffymetrixCnChpFile}
\alias{calculateAverages}
\alias{calculateAverages.FragmentEquivalentClassNormalization}
\alias{clearCache.AffymetrixCnChpFile}
\alias{clearCache.FragmentEquivalentClassNormalization}
\alias{clone.AffymetrixCnChpFile}
\alias{colnames.AromaUfcFile}
\alias{findByCdf2}
\alias{findByCdf2.default}
\alias{fromFile.AffymetrixCnChpFile}
\alias{getAromaUfcFile}
\alias{getAromaUfcFile.FragmentEquivalentClassNormalization}
\alias{getArrayNames}
\alias{getArrayNames.AffymetrixAptSummaryFile}
\alias{getAsteriskTags.FragmentEquivalentClassNormalization}
\alias{getCdf.AffymetrixCnChpFile}
\alias{getCdf.FragmentEquivalentClassNormalization}
\alias{getChipType.AffymetrixAptSummaryFile}
\alias{getChipType.AffymetrixCnChpFile}
\alias{getData.AffymetrixCnChpFile}
\alias{getExclusiveUnitSubsets}
\alias{getExclusiveUnitSubsets.FragmentEquivalentClassNormalization}
\alias{getFileFormat.AffymetrixCnChpFile}
\alias{getFilenameExtension.AromaUfcFile}
\alias{getHeader.AffymetrixCnChpFile}
\alias{getOrderedFragmentPairMap}
\alias{getOrderedFragmentPairMap.AromaUfcFile}
\alias{getOrderedFragmentPairs}
\alias{getOrderedFragmentPairs.AromaUfcFile}
\alias{getOutputDataSet.FragmentEquivalentClassNormalization}
\alias{getParameter.AffymetrixAptSummaryFile}
\alias{getParameters.FragmentEquivalentClassNormalization}
\alias{getProbesetIds}
\alias{getProbesetIds.AffymetrixAptSummaryFile}
\alias{getQuantificationScale}
\alias{getQuantificationScale.AffymetrixAptSummaryFile}
\alias{getRawData}
\alias{getRawData.AffymetrixCnChpFile}
\alias{getReadArguments.AffymetrixAptSummaryFile}
\alias{getSubsetToFit.FragmentEquivalentClassNormalization}
\alias{getTargetAverages}
\alias{getTargetAverages.FragmentEquivalentClassNormalization}
\alias{getTimeStamp}
\alias{getTimeStamp.AffymetrixAptSummaryFile}
\alias{getTimestamp.AffymetrixCnChpFile}
\alias{getUnitReadMap}
\alias{getUnitReadMap.AffymetrixCnChpFile}
\alias{hasUnitReadMap}
\alias{hasUnitReadMap.AffymetrixCnChpFile}
\alias{importFromAffymetrixTabularFile.AromaUfcFile}
\alias{importFromApt}
\alias{nbrOfArrays.AffymetrixAptSummaryFile}
\alias{nbrOfEnzymes.AromaUfcFile}
\alias{normalizeOneArrayVector}
\alias{normalizeOneArrayVector.FragmentEquivalentClassNormalization}
\alias{readArrays}
\alias{readArrays.AffymetrixAptSummaryFile}
\alias{readData.AromaUfcFile}
\alias{readHeader.AffymetrixAptSummaryFile}
\alias{readProbesetIds}
\alias{readProbesetIds.AffymetrixAptSummaryFile}
\alias{readRawFooter}
\alias{readRawFooter.AromaTabularBinaryFile}
\alias{setCdf.AffymetrixCnChpFile}
\alias{setUnitReadMap}
\alias{setUnitReadMap.AffymetrixCnChpFile}
\alias{translateColumnNames.AffymetrixAptSummaryFile}
\alias{writeFooter}
\alias{writeFooter.AromaTabularBinaryFile}
\alias{writeRawFooter}
\alias{writeRawFooter.AromaTabularBinaryFile}
\alias{as.character.DChipDcpFile}
\alias{as.DChipDcpSet}
\alias{byName.DChipDcpSet}
\alias{dim.DChipDcpFile}
\alias{findByName.DChipDcpSet}
\alias{fromFile.DChipDcpFile}
\alias{fromFiles.DChipDcpSet}
\alias{fromName.DChipDcpSet}
\alias{getCalls}
\alias{getCalls.DChipDcpFile}
\alias{getExcludes}
\alias{getExcludes.DChipDcpFile}
\alias{getFileFormat.DChipDcpFile}
\alias{getFullName.DChipDcpSet}
\alias{getHeader.DChipDcpFile}
\alias{getNormalizedIntensities}
\alias{getNormalizedIntensities.DChipDcpFile}
\alias{getRawIntensities}
\alias{getRawIntensities.DChipDcpFile}
\alias{getThetas}
\alias{getThetas.DChipDcpFile}
\alias{getThetasAB}
\alias{getThetasAB.DChipDcpFile}
\alias{getThetaStds}
\alias{getThetaStds.DChipDcpFile}
\alias{hasMbeiData}
\alias{hasMbeiData.DChipDcpFile}
\alias{hasNormalizedData}
\alias{hasNormalizedData.DChipDcpFile}
\alias{nbrOfCells.DChipDcpFile}
\alias{nbrOfUnits.DChipDcpFile}
\alias{readDcp}
\alias{readDcpHeader}
\alias{extractDataFrame}
\alias{getUnitGroupNamesFromUgcMap}
\alias{getUnitGroupNamesFromUgcMap.AffymetrixCdfFile}
\alias{as.character.AromaUnitTabularBinaryFile}
\alias{listToXml}
\alias{listToXml.list}
\alias{xmlToList}
\alias{xmlToList.character}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
