%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% The AffymetrixDataFile class
\alias{addTransform}
\alias{addTransform.AffymetrixDataFile}
\alias{createOptimalReadMap}
\alias{createOptimalReadMap.AffymetrixDataFile}
\alias{fromFile}
\alias{getApdMap}
\alias{getApdMap.AffymetrixDataFile}
\alias{getChipType}
\alias{getChipType.AffymetrixApdFile}
\alias{getProbes}
\alias{getProbes.AffymetrixDataFile}
\alias{hasApdMap}
\alias{readUnits}

% The AffymetrixApdFile class
\alias{getFileType.AffymetrixApdFile}
\alias{readUnits.AffymetrixApdFile}

% The AffymetrixCelFile class
\alias{getFileType.AffymetrixCelFile}

% The AffymetrixDataset class
\alias{as.AffymetrixDataset}
\alias{compactApdMap}
\alias{compactApdMap.AffymetrixDataset}
\alias{createOptimalReadMap.AffymetrixDataset}
\alias{getProbes.AffymetrixDataset}
\alias{findAverageQuantile}
\alias{findAverageQuantile.AffymetrixDataset}
\alias{getApdMap.AffymetrixDataset}
\alias{getFileType}
\alias{getFileType.AffymetrixDataset}

\alias{newInstance.Object}
\alias{getSnpNames}

% The ApdMap class
\alias{as.character.ApdMap}
\alias{equals.ApdMap}
\alias{findApdMap.ApdMap}
\alias{fromCdf}
\alias{fromCdf.ApdMap}
\alias{fromMapType}
\alias{fromMapType.ApdMap}
\alias{getChipType.ApdMap}
\alias{getMapType.ApdMap}
\alias{getReadMap}
\alias{getReadMap.ApdMap}
\alias{getWriteMap}
\alias{getWriteMap.ApdMap}
\alias{read}
\alias{read.ApdMap}
\alias{write}
\alias{write.ApdMap}

% The AffymetrixModel class
\alias{as.character.AffymetrixModel}
\alias{findUnitsDone}
\alias{findUnitsDone.AffymetrixModel}
\alias{fit}
\alias{fit.AffymetrixModel}
\alias{fromAffymetrixDataset}
\alias{fromAffymetrixDataset.AffymetrixModel}
\alias{getDataset}
\alias{getDataset.AffymetrixModel}
\alias{getFit}
\alias{getFit.AffymetrixModel}
\alias{getPath.AffymetrixModel}
\alias{nbrOfArrays}
\alias{nbrOfArrays.AffymetrixModel}
\alias{nbrOfProbes.AffymetrixModel}
\alias{nbrOfProbesets}
\alias{nbrOfProbesets.AffymetrixModel}

\alias{colMeans}
\alias{colSums}
\alias{doTransform}
\alias{findCdf}
\alias{findCdf.default}
\alias{fitQuantileNormFcn}
\alias{fromFiles}
\alias{getFilename}
\alias{getMapType}
\alias{getMapType.AffymetrixApdFile}
\alias{getMapType.AffymetrixCelFile}
\alias{getNames}
\alias{getProbeIndices}
\alias{getProbeIntensities}
\alias{getProbesetIntensities}
\alias{getProbesetNames}
\alias{getProbesets}
\alias{getProbesets.AffymetrixDataFile}
\alias{hasReadMap}
\alias{hasTransforms}
\alias{hasTransforms.AffymetrixDataFile}
\alias{isTransforming}
\alias{isTransforming.AffymetrixDataFile}
\alias{lapply}
\alias{lapply.default}
\alias{nbrOfProbes}
\alias{nbrOfProbes.AffymetrixApdFile}
\alias{nbrOfProbes.AffymetrixCelFile}
\alias{nbrOfSnps}
\alias{readCdfUnitsWriteMap}
\alias{readCdfUnitsWriteMap.default}
\alias{readIntensities}
\alias{readIntensities.AffymetrixApdFile}
\alias{readIntensities.AffymetrixCelFile}
\alias{setApdMap}
\alias{setTransform}
\alias{setTransform.AffymetrixDataFile}
\alias{setTransformStatus}
\alias{setTransformStatus.AffymetrixDataFile}
\alias{stringTree}
\alias{stringTree.character}
\alias{transformOff}
\alias{transformOff.AffymetrixDataFile}
\alias{transformOn}
\alias{transformOn.AffymetrixDataFile}
\alias{transformProbeSignals}
\alias{transformProbeSignals.AffymetrixDataFile}
\alias{write.default}
\alias{writeApd.AffymetrixDataFile}
% Calibration and normalization functions

% Miscellaneous statistical functions

% Other missing docs
\alias{[.AffymetrixCelFile}
\alias{[.AffymetrixCelSet}
\alias{[.GdasAnnotationFile}
\alias{[.GdasAnnotationSet}
\alias{[[.AffymetrixCelFile}
\alias{[[.AffymetrixCelSet}
\alias{annotateMvsA}
\alias{annotateMvsA.AffymetrixCelFile}
\alias{applyToUnitIntensities}
\alias{applyToUnitIntensities.AffymetrixCelSet}
\alias{as.AffymetrixCelSet}
\alias{as.AffymetrixFileSet}
\alias{as.character.AffymetrixCdfFile}
\alias{as.character.Interface}
\alias{as.character.UnitGroupsModel}
\alias{calcMvsA}
\alias{calcMvsA.AffymetrixCelFile}
\alias{clearCache.AffymetrixFileSet}
\alias{clearData}
\alias{clone.AffymetrixCelFile}
\alias{clone.AffymetrixCelSet}
\alias{clone.AffymetrixFile}
\alias{clone.AffymetrixFileSet}
\alias{compare}
\alias{convert}
\alias{copyFile}
\alias{copyFile.AffymetrixFile}
\alias{createFrom}
\alias{createFrom.AffymetrixCelFile}
\alias{createMonoCell}
\alias{createMonoCell.AffymetrixCdfFile}
\alias{createParamCdf}
\alias{createParamCdf.ChipEffectFile}
\alias{decode}
\alias{decode.ParameterCelFile}
\alias{decodeUnit}
\alias{decodeUnit.ParameterCelFile}
\alias{encode}
\alias{encode.ParameterCelFile}
\alias{encodeUnit}
\alias{encodeUnit.ParameterCelFile}
\alias{equals.AffymetrixCdfFile}
\alias{extend.Interface}
\alias{findUnitsTodo}
\alias{findUnitsTodo.ChipEffectFile}
\alias{findUnitsTodo.ChipEffectSet}
\alias{forDataSet}
\alias{forDataSet.GdasAnnotationSet}
\alias{fromChipType}
\alias{fromChipType.GdasAnnotationFile}
\alias{fromChipType.GdasAnnotationSet}
\alias{fromDataFile}
\alias{fromDataFile.ChipEffectFile}
\alias{fromDataSet}
\alias{fromDataSet.ChipEffectSet}
\alias{fromFile.AffymetrixFile}
\alias{fromFiles.AffymetrixCelSet}
\alias{fromFiles.GdasAnnotationSet}
\alias{GdasAnnotationFile}
\alias{GdasAnnotationSet}
\alias{getAverageFile}
\alias{getAverageFile.AffymetrixCelSet}
\alias{getAverageFile.ChipEffectSet}
\alias{getBlockPositions}
\alias{getBlockPositions.GdasAnnotationFile}
\alias{getCdf}
\alias{getCdf.GdasAnnotationSet}
\alias{getCellIndices}
\alias{getCellIndices.ChipEffectFile}
\alias{getCellIndices.ChipEffectSet}
\alias{getCellIndices.CnChipEffectFile}
\alias{getCellIndices.CnPlm}
\alias{getCellIndices.CnProbeAffinityFile}
\alias{getCellIndices.ProbeAffinityFile}
\alias{getCellIndices.ProbeLevelModel}
\alias{getCellIndices.SnpChipEffectFile}
\alias{getCellIndices.SnpPlm}
\alias{getCellIndices.SnpProbeAffinityFile}
\alias{getChipEffectFileClass}
\alias{getChipEffectFileClass.ChipEffectSet}
\alias{getChipEffectFileClass.CnChipEffectSet}
\alias{getChipEffectFileClass.SnpChipEffectSet}
\alias{getChipEffects}
\alias{getChipEffects.CnPlm}
\alias{getChipEffects.SnpPlm}
\alias{getChipEffectSetClass}
\alias{getChipEffectSetClass.CnPlm}
\alias{getChipEffectSetClass.ProbeLevelModel}
\alias{getChipEffectSetClass.SnpPlm}
\alias{getChipType.AffymetrixCdfFile}
\alias{getChipType.GdasAnnotationSet}
\alias{getCombineAlleles}
\alias{getCombineAlleles.CnChipEffectSet}
\alias{getDataSet}
\alias{getDimension}
\alias{getDimension.AffymetrixCdfFile}
\alias{getFile}
\alias{getFile.AffymetrixFileSet}
\alias{getFileSize.AffymetrixFile}
\alias{getFileSize.AffymetrixFileSet}
\alias{getFirstCellIndices}
\alias{getFirstCellIndices.AffymetrixCdfFile}
\alias{getFitFunction}
\alias{getFitUnitFunction}
\alias{getFitUnitFunction.CnPlm}
\alias{getFitUnitFunction.ProbeLevelModel}
\alias{getHeader}
\alias{getIntensities}
\alias{getLabel.AffymetrixFile}
\alias{getLabel.UnitGroupsModel}
\alias{getMergeStrands}
\alias{getMergeStrands.SnpChipEffectSet}
\alias{getName.AffymetrixCelSet}
\alias{getName.AffymetrixFileSet}
\alias{getName.UnitGroupsModel}
\alias{getParameter}
\alias{getParameter.ParameterFile}
\alias{getParameters}
\alias{getParameters.ParameterFile}
\alias{getParameterSet}
\alias{getParameterSet.UnitGroupsModel}
\alias{getPathnames}
\alias{getProbeAffinities}
\alias{getProbeAffinities.AffinePlm}
\alias{getProbeAffinities.CnPlm}
\alias{getProbeAffinities.MbeiPlm}
\alias{getProbeAffinities.RmaPlm}
\alias{getProbeAffinities.SnpPlm}
\alias{getRectangle}
\alias{getRectangle.AffymetrixCelFile}
\alias{getRestructor}
\alias{getRestructor.AffymetrixCdfFile}
\alias{getUnitIndices}
\alias{getUnitIndices.GdasAnnotationSet}
\alias{getUnitIntensities}
\alias{getUnitNames}
\alias{getUnitSizes}
\alias{highlight}
\alias{highlight.AffymetrixCelFile}
\alias{highlight.ProbeLevelModel}
\alias{identifyCells}
\alias{identifyCells.AffymetrixCdfFile}
\alias{image270}
\alias{indexOf}
\alias{isPm}
\alias{isPm.AffymetrixCdfFile}
\alias{join}
\alias{join.GdasAnnotationSet}
\alias{nbrOfCells}
\alias{nbrOfCells.AffymetrixCdfFile}
\alias{nbrOfCells.AffymetrixCelFile}
\alias{nbrOfColumns}
\alias{nbrOfColumns.AffymetrixCdfFile}
\alias{nbrOfFiles}
\alias{nbrOfParameters}
\alias{nbrOfParameters.ParameterFile}
\alias{nbrOfQcUnits}
\alias{nbrOfQcUnits.AffymetrixCdfFile}
\alias{nbrOfRows}
\alias{nbrOfRows.AffymetrixCdfFile}
\alias{nbrOfUnits}
\alias{nbrOfUnits.AffymetrixCdfFile}
\alias{normalizeQuantiles}
\alias{plotMvsPosition}
\alias{plotMvsPosition.ProbeLevelModel}
\alias{plotMvsX}
\alias{print.GenericSummary}
\alias{print.Interface}
\alias{range.AffymetrixCelFile}
\alias{range.AffymetrixCelSet}
\alias{readBlock}
\alias{readBlock.GdasAnnotationFile}
\alias{readColumnName}
\alias{readColumnName.GdasAnnotationFile}
\alias{readData}
\alias{readData.GdasAnnotationFile}
\alias{readHeader.GdasAnnotationFile}
\alias{readUnits.AffymetrixCdfFile}
\alias{readUnits.AffymetrixCelSet}
\alias{readUnits.ChipEffectFile}
\alias{readUnits.ChipEffectSet}
\alias{readUnits.ParameterCelFile}
\alias{readUnits.ProbeAffinityFile}
\alias{restruct}
\alias{restruct.AffymetrixCdfFile}
\alias{rowSds}
\alias{rowVars}
\alias{select}
\alias{select.GdasAnnotationSet}
\alias{setCdf}
\alias{setCombineAlleles}
\alias{setCombineAlleles.CnChipEffectSet}
\alias{setCombineAlleles.CnPlm}
\alias{setCombineAlleles.CnProbeAffinityFile}
\alias{setDecodeFunction}
\alias{setDecodeFunction.ParameterCelFile}
\alias{setEncodeFunction}
\alias{setEncodeFunction.ParameterCelFile}
\alias{setLabel.AffymetrixFile}
\alias{setLabel.UnitGroupsModel}
\alias{setMergeStrands}
\alias{setMergeStrands.SnpChipEffectSet}
\alias{setMergeStrands.SnpPlm}
\alias{setMergeStrands.SnpProbeAffinityFile}
\alias{setRestructor}
\alias{setup}
\alias{setup.UnitGroupsModel}
\alias{smoothScatterMvsA}
\alias{stext}
\alias{stext.default}
\alias{stextChipType}
\alias{stextChipType.AffymetrixCelFile}
\alias{stextLabel}
\alias{stextLabel.AffymetrixFile}
\alias{stextLabels}
\alias{stextLabels.AffymetrixFile}
\alias{transformAffine}
\alias{transformAffine.AffymetrixCelFile}
\alias{updateUnits}
\alias{updateUnits.ChipEffectFile}
\alias{updateUnits.ChipEffectSet}
\alias{updateUnits.ParameterCelFile}
\alias{updateUnits.ProbeAffinityFile}
\alias{uses}
\alias{uses.Interface}
\alias{whatDataType}
\alias{whatDataType.default}
\alias{writeSpatial}
\alias{writeSpatial.ProbeAffinityFile}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
