%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{[.AffymetrixCelFile}
\alias{[.AffymetrixCelSet}
\alias{[.AromaGenomePositionFile}
\alias{[.GdasAnnotationFile}
\alias{[.GdasAnnotationSet}
\alias{[[.AffymetrixCelFile}
\alias{[[.AffymetrixCelSet}
\alias{addColorMap}
\alias{addColorMap.ArrayExplorer}
\alias{addColorMap.SpatialReporter}
\alias{addIncludes}
\alias{addIncludes.Explorer}
\alias{addIndexFile}
\alias{addIndexFile.ChromosomeExplorer}
\alias{addIndexFile.Explorer}
\alias{AffymetrixCsvGenomeInformation}
\alias{annotateMvsA}
\alias{annotateMvsA.AffymetrixCelFile}
\alias{append}
\alias{append.AffymetrixCelSet}
\alias{append.AffymetrixFileSet}
\alias{append.default}
\alias{appendFiles}
\alias{appendFiles.AffymetrixFileSet}
\alias{apply}
\alias{apply.default}
\alias{apply.SampleAnnotationFile}
\alias{applyToUnitIntensities}
\alias{applyToUnitIntensities.AffymetrixCelSet}
\alias{AptProbesetGenotype}
\alias{AromaAffymetrix}
\alias{AromaCgpFile}
\alias{AromaGenomePositionFile}
\alias{AromaUgpFile}
\alias{as.AffymetrixCelSet}
\alias{as.AffymetrixFileSet}
\alias{as.character.AffymetrixCdfFile}
\alias{as.character.AffymetrixCelFile}
\alias{as.character.AffymetrixCelSetReporter}
\alias{as.character.AffymetrixCelSetTuple}
\alias{as.character.AffymetrixFileSetReporter}
\alias{as.character.AffymetrixPowerTools}
\alias{as.character.AffymetrixProbeTabFile}
\alias{as.character.AptProbesetGenotype}
\alias{as.character.ArrayExplorer}
\alias{as.character.ChipEffectFile}
\alias{as.character.ChipEffectSet}
\alias{as.character.ChromosomeExplorer}
\alias{as.character.CrlmmModel}
\alias{as.character.Explorer}
\alias{as.character.FirmaModel}
\alias{as.character.GenomeInformation}
\alias{as.character.GenotypeCallFile}
\alias{as.character.GladModel}
\alias{as.character.Interface}
\alias{as.character.Model}
\alias{as.character.PlatformDesign}
\alias{as.character.ProbeAffinityFile}
\alias{as.character.QualityAssessmentModel}
\alias{as.character.ResidualFile}
\alias{as.character.ResidualSet}
\alias{as.character.SnpInformation}
\alias{as.character.SpatialReporter}
\alias{as.character.Transform}
\alias{as.character.TransformReport}
\alias{as.character.WeightsFile}
\alias{as.character.WeightsSet}
\alias{as.GenotypeCallSet}
\alias{backtransformGenotypeCone}
\alias{backtransformGenotypeCone.matrix}
\alias{bgAdjustGcrma}
\alias{bgAdjustOptical}
\alias{bgAdjustRma}
\alias{byPath}
\alias{byPath.AffymetrixCelSetTuple}
\alias{calculateBaseline}
\alias{calculateGsbParameters}
\alias{calculateResiduals}
\alias{calculateResiduals.ProbeLevelModel}
\alias{calculateResidualSet}
\alias{calculateResidualSet.ProbeLevelModel}
\alias{calculateTargetDistribution}
\alias{calculateTargetDistribution.QuantileNormalization}
\alias{calculateWeights}
\alias{calculateWeights.ExonRmaPlm}
\alias{calculateWeights.FirmaModel}
\alias{calculateWeights.ProbeLevelModel}
\alias{calibrateAllelicCrosstalk}
\alias{cdfMergeGroups}
\alias{ChipEffectSetTuple}
\alias{classifyAsXorXX}
\alias{cleanToRealChipType}
\alias{cleanToRealChipType.AffymetrixCdfFile}
\alias{clearCache.AffymetrixCdfFile}
\alias{clearCache.AffymetrixCelFile}
\alias{clearCache.AffymetrixCelSet}
\alias{clearCache.AffymetrixFileSet}
\alias{clearCache.AffymetrixProbeTabFile}
\alias{clearCache.ChipEffectFile}
\alias{clearCache.ChipEffectSet}
\alias{clearCache.ExonChipEffectFile}
\alias{clearCache.FirmaFile}
\alias{clearCache.FirmaModel}
\alias{clearCache.FirmaSet}
\alias{clearCache.GdasAnnotationFile}
\alias{clearCache.GenomeInformation}
\alias{clearCache.ParameterCelFile}
\alias{clearCache.PcrProbeCalibration}
\alias{clearCache.ProbeAffinityFile}
\alias{clearCache.ProbeLevelModel}
\alias{clearCache.QualityAssessmentFile}
\alias{clearCache.ResidualFile}
\alias{clearCache.ResidualSet}
\alias{clearCache.SnpChipEffectFile}
\alias{clearCache.SnpInformation}
\alias{clearCache.Transform}
\alias{clearCache.TransformReport}
\alias{clearCache.WeightsFile}
\alias{clearCache.WeightsSet}
\alias{clearData}
\alias{clearFileCache}
\alias{clearFileCache.PlatformDesign}
\alias{clone.AffymetrixCelFile}
\alias{clone.AffymetrixCelSet}
\alias{clone.AffymetrixFile}
\alias{clone.AffymetrixFileSet}
\alias{cna}
\alias{cna.AffymetrixCelSet}
\alias{colMeans}
\alias{colSums}
\alias{compare}
\alias{compareChecksum}
\alias{compareChecksum.AffymetrixFile}
\alias{computeAffinities}
\alias{convert}
\alias{convertTable}
\alias{convertTable.default}
\alias{convertTxtFilesToXdr}
\alias{convertTxtFilesToXdr.AptProbesetGenotype}
\alias{convertUnits}
\alias{copyFile}
\alias{copyFile.AffymetrixFile}
\alias{create}
\alias{create.AromaGenomePositionFile}
\alias{createBlankSet}
\alias{createBlankSet.AffymetrixCelSet}
\alias{createFrom}
\alias{createFromCdf}
\alias{createFromCdf.AromaCgpFile}
\alias{createFromCdf.AromaUgpFile}
\alias{createFromCrlmmFile}
\alias{createFromCrlmmFile.GenotypeCallSet}
\alias{createFromFeatureSet}
\alias{createFromFeatureSet.AffymetrixCelSet}
\alias{createFromGenomeInformation}
\alias{createFromGenomeInformation.AromaUgpFile}
\alias{createMonoCell}
\alias{createParamCdf}
\alias{createParamCdf.ChipEffectFile}
\alias{createParamCdf.FirmaFile}
\alias{dataFrame}
\alias{dataFrame.default}
\alias{decode}
\alias{decode.ParameterCelFile}
\alias{display}
\alias{display.ChromosomeExplorer}
\alias{drawCytoBand}
\alias{drawCytoBand.default}
\alias{encode}
\alias{encode.ParameterCelFile}
\alias{extend.Interface}
\alias{extractAffyBatch}
\alias{extractMatrix}
\alias{extractSnpFeatureSet}
\alias{extractSnpQSet}
\alias{findByChipType}
\alias{findByChipType.AffymetrixCsvGenomeInformation}
\alias{findByChipType.AffymetrixProbeTabFile}
\alias{findByChipType.AromaUgpFile}
\alias{findByChipType.DChipGenomeInformation}
\alias{findByChipType.DChipSnpInformation}
\alias{findByChipType.GdasAnnotationFile}
\alias{findByName}
\alias{findByName.AffymetrixCelSet}
\alias{findByName.ChipEffectSet}
\alias{findCdf.patch}
\alias{findSAFs}
\alias{findSAFs.SampleAnnotationSet}
\alias{findUnitsTodo}
\alias{findUnitsTodo.ChipEffectFile}
\alias{findUnitsTodo.ChipEffectSet}
\alias{findUnitsTodo.ExonChipEffectFile}
\alias{findUnitsTodo.ExonChipEffectSet}
\alias{findUnitsTodo.ExonRmaPlm}
\alias{findUnitsTodo.FirmaFile}
\alias{findUnitsTodo.FirmaSet}
\alias{findUnitsTodo.QualityAssessmentFile}
\alias{findUnitsTodo.ResidualFile}
\alias{findUnitsTodo.ResidualSet}
\alias{findUnitsTodo.WeightsFile}
\alias{findUnitsTodo.WeightsSet}
\alias{fit}
\alias{fit.PcrProbeCalibration}
\alias{fitGenotypeCone}
\alias{fitGlad}
\alias{fitGlad.CnChipEffectSet}
\alias{fitOne}
\alias{fitPlasqUnit}
\alias{fitPlasqUnit.matrix}
\alias{fitQuantileNormFcn}
\alias{fixSearchPath}
\alias{fixSearchPath.AromaAffymetrix}
\alias{forDataSet}
\alias{forDataSet.GdasAnnotationSet}
\alias{fromCdf}
\alias{fromCdf.AffymetrixProbeTabFile}
\alias{fromCdf.AromaUgpFile}
\alias{fromChipType}
\alias{fromChipType.AffymetrixCsvGenomeInformation}
\alias{fromChipType.AffymetrixProbeTabFile}
\alias{fromChipType.AromaUgpFile}
\alias{fromChipType.GdasAnnotationFile}
\alias{fromChipType.GdasAnnotationSet}
\alias{fromCleanChipType}
\alias{fromCleanChipType.AffymetrixCdfFile}
\alias{fromDataFile}
\alias{fromDataFile.ChipEffectFile}
\alias{fromDataFile.FirmaFile}
\alias{fromDataFile.ResidualFile}
\alias{fromDataFile.WeightsFile}
\alias{fromDataSet}
\alias{fromDataSet.ChipEffectSet}
\alias{fromDataSet.FirmaSet}
\alias{fromDataSet.GenomeInformation}
\alias{fromDataSet.ResidualSet}
\alias{fromDataSet.SnpInformation}
\alias{fromDataSet.WeightsSet}
\alias{fromFile}
\alias{fromFile.AffymetrixFile}
\alias{fromFile.GenotypeCallXdrFile}
\alias{fromFiles}
\alias{fromFiles.AffymetrixCelSet}
\alias{fromFiles.ChipEffectSet}
\alias{fromFiles.CnChipEffectSet}
\alias{fromFiles.ExonChipEffectSet}
\alias{fromFiles.FirmaSet}
\alias{fromFiles.GdasAnnotationSet}
\alias{fromFiles.GenotypeCallXdrSet}
\alias{fromFiles.ResidualSet}
\alias{fromFiles.SnpChipEffectSet}
\alias{fromFiles.WeightsSet}
\alias{fromName}
\alias{fromName.AffymetrixCelSet}
\alias{fromPath}
\alias{fromPath.SampleAnnotationFile}
\alias{fromPath.SampleAnnotationSet}
\alias{gcrmaSummary}
\alias{GdasAnnotationFile}
\alias{GdasAnnotationSet}
\alias{GenericSummary}
\alias{getAlias}
\alias{getAlias.AffymetrixCelSetTuple}
\alias{getAlias.GladModel}
\alias{getAlias.TransformReport}
\alias{getAlleleAB}
\alias{getAlleleAB.PlatformDesign}
\alias{getAllelePairs}
\alias{getAllelePairUnitSets}
\alias{getAlleleProbePairs}
\alias{getAlleleProbePairs2}
\alias{getAlleleProbePairs2.AffymetrixCdfFile}
\alias{getAm}
\alias{getAM}
\alias{getAm.AffymetrixCelFile}
\alias{getAnnotations}
\alias{getAnnotations.PlatformDesign}
\alias{getAptPathnames}
\alias{getAptPathnames.AptProbesetGenotype}
\alias{getArrays}
\alias{getArraysOfInput}
\alias{getArraysOfInput.ArrayExplorer}
\alias{getArraysOfInput.ChromosomeExplorer}
\alias{getAsteriskTag}
\alias{getAsteriskTag.AffymetrixCelSetTuple}
\alias{getAsteriskTag.GladModel}
\alias{getAsteriskTag.SnpChipEffectGroupMerge}
\alias{getAsteriskTag.Transform}
\alias{getAsteriskTags}
\alias{getAsteriskTags.AffymetrixFileSetReporter}
\alias{getAsteriskTags.ArrayExplorer}
\alias{getAsteriskTags.Explorer}
\alias{getAttribute}
\alias{getAttribute.AffymetrixFile}
\alias{getAttributes}
\alias{getAttributes.AffymetrixFile}
\alias{getAttributeXY}
\alias{getAttributeXY.AffymetrixCelFile}
\alias{getAverage}
\alias{getAverage.AffymetrixCelSet}
\alias{getAverageAsinh}
\alias{getAverageAsinh.AffymetrixCelSet}
\alias{getAverageFile}
\alias{getAverageFile.ChipEffectSet}
\alias{getAverageFile.CnChipEffectSet}
\alias{getAverageFile.ExonChipEffectSet}
\alias{getAverageFile.ResidualSet}
\alias{getAverageFile.SnpChipEffectSet}
\alias{getAverageFile.WeightsSet}
\alias{getAverageLog}
\alias{getAverageLog.AffymetrixCelSet}
\alias{getBlockPositions}
\alias{getBlockPositions.GdasAnnotationFile}
\alias{getCachedFields}
\alias{getCachedFields.PlatformDesign}
\alias{getCachePath.PlatformDesign}
\alias{getCalculateResidualsFunction}
\alias{getCalculateResidualsFunction.ProbeLevelModel}
\alias{getCalculateResidualsFunction.RmaPlm}
\alias{getCdf}
\alias{getCdf.AffymetrixProbeTabFile}
\alias{getCdf.AptProbesetGenotype}
\alias{getCdf.AromaUgpFile}
\alias{getCdf.ChipEffectGroupMerge}
\alias{getCdf.FirmaModel}
\alias{getCdf.FragmentLengthNormalization}
\alias{getCdf.GdasAnnotationSet}
\alias{getCdf.GenotypeCallXdrFile}
\alias{getCdf.PlatformDesign}
\alias{getCdf.QualityAssessmentModel}
\alias{getCdf.TransformReport}
\alias{getCellIndices}
\alias{getCellIndices.ChipEffectFile}
\alias{getCellIndices.ChipEffectSet}
\alias{getCellIndices.CnChipEffectFile}
\alias{getCellIndices.CnPlm}
\alias{getCellIndices.CnProbeAffinityFile}
\alias{getCellIndices.ExonChipEffectFile}
\alias{getCellIndices.ExonProbeAffinityFile}
\alias{getCellIndices.ExonRmaPlm}
\alias{getCellIndices.FirmaFile}
\alias{getCellIndices.FirmaSet}
\alias{getCellIndices.PlatformDesign}
\alias{getCellIndices.ProbeAffinityFile}
\alias{getCellIndices.ProbeLevelModel}
\alias{getCellIndices.ResidualFile}
\alias{getCellIndices.ResidualSet}
\alias{getCellIndices.SnpChipEffectFile}
\alias{getCellIndices.SnpPlm}
\alias{getCellIndices.SnpProbeAffinityFile}
\alias{getCellIndices.WeightsFile}
\alias{getCellIndices.WeightsSet}
\alias{getCellMap}
\alias{getCellMap.FirmaFile}
\alias{getCellMap.ResidualFile}
\alias{getCellMap.WeightsFile}
\alias{getCellsAt}
\alias{getCellsAt.AromaCgpFile}
\alias{getChecksum}
\alias{getChecksum.AffymetrixFile}
\alias{getChipEffectFileClass}
\alias{getChipEffectFileClass.ChipEffectSet}
\alias{getChipEffectFileClass.CnChipEffectSet}
\alias{getChipEffectFileClass.ExonChipEffectSet}
\alias{getChipEffectFileClass.SnpChipEffectSet}
\alias{getChipEffectFiles}
\alias{getChipEffectFiles.GladModel}
\alias{getChipEffects}
\alias{getChipEffects.CnPlm}
\alias{getChipEffects.ExonRmaPlm}
\alias{getChipEffects.PlasqModel}
\alias{getChipEffects.QualityAssessmentModel}
\alias{getChipEffects.SnpPlm}
\alias{getChipEffectSet}
\alias{getChipEffectSetClass}
\alias{getChipEffectSetClass.CnPlm}
\alias{getChipEffectSetClass.ExonRmaPlm}
\alias{getChipEffectSetClass.ProbeLevelModel}
\alias{getChipEffectSetClass.SnpPlm}
\alias{getChipType}
\alias{getChipType.AffymetrixCdfFile}
\alias{getChipType.AffymetrixCelSetReporter}
\alias{getChipType.AffymetrixProbeTabFile}
\alias{getChipType.AromaGenomePositionFile}
\alias{getChipType.GdasAnnotationSet}
\alias{getChipType.PlatformDesign}
\alias{getChipTypes}
\alias{getChipTypes.GladModel}
\alias{getChromosome}
\alias{getChromosome.Arguments}
\alias{getChromosomes}
\alias{getChromosomes.Arguments}
\alias{getChromosomes.GenomeInformation}
\alias{getChromosomeStats}
\alias{getChromosomeStats.GenomeInformation}
\alias{getChrXFile}
\alias{getChrXFile.AptProbesetGenotype}
\alias{getColorMaps}
\alias{getColorMaps.ArrayExplorer}
\alias{getColorMaps.SpatialReporter}
\alias{getCombineAlleles}
\alias{getCombineAlleles.CnChipEffectSet}
\alias{getCommonListElements}
\alias{getCommonPrefix}
\alias{getCrlmmInfo}
\alias{getCrlmmInfo.PlatformDesign}
\alias{getData}
\alias{getData.AffymetrixCelSet}
\alias{getData.AffymetrixProbeTabFile}
\alias{getDataColumns}
\alias{getDataColumns.GenomeInformation}
\alias{getDataFile}
\alias{getDataFile.PlatformDesign}
\alias{getDataFiles}
\alias{getDataFiles.PlatformDesign}
\alias{getDataFlat}
\alias{getDataFlat.ChipEffectFile}
\alias{getDataFlat.FirmaFile}
\alias{getDataFlat.ResidualFile}
\alias{getDataFlat.WeightsFile}
\alias{getDataSet}
\alias{getDataSet.AptProbesetGenotype}
\alias{getDataSet.FirmaModel}
\alias{getDataSet.QualityAssessmentModel}
\alias{getDescription.AffymetrixFileSet}
\alias{getDescription.PlatformDesign}
\alias{getDeviceResolution}
\alias{getDeviceResolution.default}
\alias{getDimension}
\alias{getDimension.AffymetrixCdfFile}
\alias{getElementsAt}
\alias{getElementsAt.AromaGenomePositionFile}
\alias{getExternalCommand}
\alias{getExternalCommand.AptProbesetGenotype}
\alias{getFeatureInfo}
\alias{getFeatureInfo.PlatformDesign}
\alias{getFields.SnpInformation}
\alias{getFile}
\alias{getFile.AffymetrixFileSet}
\alias{getFileClass}
\alias{getFileClass.FirmaSet}
\alias{getFilename}
\alias{getFilename.Arguments}
\alias{getFilenameExtension}
\alias{getFilenameExtension.AromaCgpFile}
\alias{getFilenameExtension.AromaGenomePositionFile}
\alias{getFilenameExtension.AromaUgpFile}
\alias{getFiles}
\alias{getFiles.AffymetrixFileSet}
\alias{getFileSet}
\alias{getFileSet.AffymetrixFileSetReporter}
\alias{getFileSetClass}
\alias{getFileSetClass.FirmaModel}
\alias{getFileSize.AffymetrixFile}
\alias{getFileSize.AffymetrixFileSet}
\alias{getFileType}
\alias{getFirmaScores}
\alias{getFirstCellIndices}
\alias{getFirstCellIndices.AffymetrixCdfFile}
\alias{getFirstCellPerUnitIndices}
\alias{getFirstCellPerUnitIndices.ExonChipEffectSet}
\alias{getFitFunction}
\alias{getFitUnitFunction}
\alias{getFitUnitFunction.CnPlm}
\alias{getFitUnitFunction.FirmaModel}
\alias{getFitUnitFunction.PlasqModel}
\alias{getFitUnitFunction.ProbeLevelModel}
\alias{getFitUnitFunction.UnitModel}
\alias{getFragmentLengths}
\alias{getFragmentLengths.SnpInformation}
\alias{getFragmentStarts}
\alias{getFragmentStarts.SnpInformation}
\alias{getFragmentStops}
\alias{getFragmentStops.SnpInformation}
\alias{getFullName}
\alias{getFullName.AffymetrixCelSet}
\alias{getFullName.AffymetrixCelSetTuple}
\alias{getFullName.AffymetrixFileSetReporter}
\alias{getFullName.AptProbesetGenotype}
\alias{getFullName.Explorer}
\alias{getFullName.GenotypeCallSet}
\alias{getFullName.GladModel}
\alias{getFullName.QualityAssessmentModel}
\alias{getFullNames}
\alias{getFullNames.AffymetrixCelSetTuple}
\alias{getFullNames.AffymetrixFileSet}
\alias{getFullNames.ChipEffectSetTuple}
\alias{getFullNames.ChromosomeExplorer}
\alias{getFullNames.GladModel}
\alias{getGenomeInformation}
\alias{getGenomeVersion}
\alias{getGenomeVersion.AromaGenomePositionFile}
\alias{getGenotypeCallSet}
\alias{getGenotypeCallSet.AptProbesetGenotype}
\alias{getHeader.GenotypeCallFile}
\alias{getHeaderV3}
\alias{getHeaderV3.AffymetrixCelFile}
\alias{getIdentifier}
\alias{getIdentifier.AffymetrixCelFile}
\alias{getIdentifier.AffymetrixCelSet}
\alias{getIdentifier.AffymetrixFileSet}
\alias{getImage}
\alias{getImage.ResidualFile}
\alias{getImage.WeightsFile}
\alias{getIncludePath}
\alias{getIncludePath.Explorer}
\alias{getIndexToRowMap}
\alias{getIndexToRowMap.AffymetrixProbeTabFile}
\alias{getInputDataSet}
\alias{getIntensities}
\alias{getLabel.AffymetrixFile}
\alias{getLabel.Model}
\alias{getListOfCdfs}
\alias{getListOfCdfs.AffymetrixCelSetTuple}
\alias{getListOfCdfs.ArrayExplorer}
\alias{getListOfCdfs.GladModel}
\alias{getListOfChipEffects}
\alias{getListOfChipEffects.GladModel}
\alias{getListOfGenomeInformations}
\alias{getListOfGenomeInformations.GladModel}
\alias{getListOfReferences}
\alias{getListOfReferences.GladModel}
\alias{getListOfReporters}
\alias{getListOfReporters.ArrayExplorer}
\alias{getListOfSets}
\alias{getListOfSets.AffymetrixCelSetTuple}
\alias{getLog2Ratios}
\alias{getLog2Ratios.GladModel}
\alias{getMainPath}
\alias{getMainPath.AffymetrixFileSetReporter}
\alias{getMainPath.Explorer}
\alias{getMergeFunction}
\alias{getMergeFunction.ChipEffectGroupMerge}
\alias{getMergeFunction.SnpChipEffectGroupMerge}
\alias{getMergeGroups}
\alias{getMergeGroups.ExonChipEffectSet}
\alias{getMergeStrands}
\alias{getMergeStrands.SnpChipEffectSet}
\alias{getModel}
\alias{getMonoCell}
\alias{getMonoCell.AffymetrixCdfFile}
\alias{getName.AffymetrixCelSetTuple}
\alias{getName.AptProbesetGenotype}
\alias{getName.FirmaModel}
\alias{getName.GladModel}
\alias{getName.PlatformDesign}
\alias{getName.QualityAssessmentModel}
\alias{getNameOfInput}
\alias{getNameOfInput.ArrayExplorer}
\alias{getNameOfInput.ChromosomeExplorer}
\alias{getNameOfInput.Explorer}
\alias{getNames}
\alias{getNames.AffymetrixCelSetTuple}
\alias{getNames.ChromosomeExplorer}
\alias{getNames.GladModel}
\alias{getOligoSequences}
\alias{getOligoSequences.PlatformDesign}
\alias{getOligoToCelMap}
\alias{getOligoToCelMap.AffymetrixCdfFile}
\alias{getOutputDataSet}
\alias{getOutputDataSet.FragmentLengthNormalization}
\alias{getOutputFiles}
\alias{getOutputFiles.Transform}
\alias{getOutputIdentifier}
\alias{getOutputIdentifier.QuantileNormalization}
\alias{getPackageName}
\alias{getPackageName.default}
\alias{getPackageName.PlatformDesign}
\alias{getParameters.ChipEffectFile}
\alias{getParameters.ChipEffectGroupMerge}
\alias{getParameters.ChipEffectSet}
\alias{getParameters.CnChipEffectFile}
\alias{getParameters.ExonChipEffectFile}
\alias{getParameters.FragmentLengthNormalization}
\alias{getParameters.Model}
\alias{getParameters.PcrProbeCalibration}
\alias{getParameters.ProbeAffinityFile}
\alias{getParameters.QuantileNormalization}
\alias{getParameters.ResidualFile}
\alias{getParameters.ResidualSet}
\alias{getParameters.SnpChipEffectFile}
\alias{getParameters.SnpChipEffectGroupMerge}
\alias{getParameters.Transform}
\alias{getParameters.WeightsFile}
\alias{getParameters.WeightsSet}
\alias{getParametersAsString}
\alias{getParametersAsString.ChipEffectFile}
\alias{getParametersAsString.ChipEffectSet}
\alias{getParametersAsString.Model}
\alias{getParametersAsString.ProbeAffinityFile}
\alias{getParametersAsString.ResidualFile}
\alias{getParametersAsString.ResidualSet}
\alias{getParametersAsString.Transform}
\alias{getParametersAsString.WeightsFile}
\alias{getParametersAsString.WeightsSet}
\alias{getParameterSet}
\alias{getParameterSet.CnPlm}
\alias{getParameterSet.Model}
\alias{getParameterSet.ProbeLevelModel}
\alias{getParameterSet.RmaPlm}
\alias{getParameterSet.SnpPlm}
\alias{getPath.AffymetrixCelSetReporter}
\alias{getPath.AffymetrixFileSetReporter}
\alias{getPath.AptProbesetGenotype}
\alias{getPath.ChromosomeExplorer}
\alias{getPath.Explorer}
\alias{getPath.GladModel}
\alias{getPath.PlatformDesign}
\alias{getPath.QualityAssessmentModel}
\alias{getPathnames}
\alias{getPathnamesFile}
\alias{getPathnamesFile.AptProbesetGenotype}
\alias{getPatterns}
\alias{getPatterns.SampleAnnotationFile}
\alias{getPlasqTypes}
\alias{getPlatformDesignObject}
\alias{getPlatformDesignObject.AffymetrixCdfFile}
\alias{getPlatformDesignObject.PlatformDesign}
\alias{getPlm}
\alias{getPlm.FirmaModel}
\alias{getPlm.QualityAssessmentModel}
\alias{getPloidy}
\alias{getPloidy.AffymetrixCelFile}
\alias{getPmDesignMatrix}
\alias{getPmDesignMatrix.PcrProbeCalibration}
\alias{getPositions}
\alias{getProbeAffinities}
\alias{getProbeAffinities.AffinePlm}
\alias{getProbeAffinities.CnPlm}
\alias{getProbeAffinities.ExonRmaPlm}
\alias{getProbeAffinities.MbeiPlm}
\alias{getProbeAffinities.RmaPlm}
\alias{getProbeAffinities.SnpPlm}
\alias{getProbeAffinityFile}
\alias{getProbeEffects}
\alias{getProbeEffects.PlatformDesign}
\alias{getProbeParameters}
\alias{getProbeParameters.PlasqModel}
\alias{getReadMap}
\alias{getReadMap.PlatformDesign}
\alias{getRectangle}
\alias{getRectangle.AffymetrixCelFile}
\alias{getReferenceFiles}
\alias{getReferenceFiles.GladModel}
\alias{getReferenceName}
\alias{getReferenceName.GladModel}
\alias{getReferenceQuantiles}
\alias{getReferenceQuantiles.PlatformDesign}
\alias{getRegions}
\alias{getRegions.GladModel}
\alias{getRegions.profileCGH}
\alias{getReportPath}
\alias{getReportPath.GladModel}
\alias{getReportSet}
\alias{getReportSet.AffymetrixFileSetReporter}
\alias{getReportSet.SpatialReporter}
\alias{getResidualFileClass}
\alias{getResidualFileClass.ResidualSet}
\alias{getResiduals}
\alias{getResidualSet}
\alias{getResidualSet.ProbeLevelModel}
\alias{getResidualSetClass}
\alias{getResidualSetClass.ProbeLevelModel}
\alias{getRestructor}
\alias{getRestructor.AffymetrixCdfFile}
\alias{getRootPath}
\alias{getRootPath.AffymetrixFileSetReporter}
\alias{getRootPath.AptProbesetGenotype}
\alias{getRootPath.ChipEffectTransform}
\alias{getRootPath.Explorer}
\alias{getRootPath.FirmaModel}
\alias{getRootPath.GladModel}
\alias{getRootPath.PlasqModel}
\alias{getRootPath.ProbeLevelModel}
\alias{getRootPath.ProbeLevelTransform}
\alias{getRootPath.QualityAssessmentModel}
\alias{getRootPath.Transform}
\alias{getRootPath.TransformReport}
\alias{getSetTuple}
\alias{getSetTuple.ArrayExplorer}
\alias{getSetTuple.ChromosomeExplorer}
\alias{getSetTuple.GladModel}
\alias{getSiblings}
\alias{getSnpInformation}
\alias{getSnpInformation.AffymetrixCdfFile}
\alias{getSnpNames}
\alias{getSubsetToFit}
\alias{getSubsetToFit.FragmentLengthNormalization}
\alias{getSubsetToFit.PcrProbeCalibration}
\alias{getSubsetToUpdate}
\alias{getSubsetToUpdate.OligoQuantileNormalization}
\alias{getSubsetToUpdate.QuantileNormalization}
\alias{getTableOfArrays}
\alias{getTags}
\alias{getTags.AffymetrixCelSetTuple}
\alias{getTags.AptProbesetGenotype}
\alias{getTags.FirmaModel}
\alias{getTags.GladModel}
\alias{getTags.QualityAssessmentModel}
\alias{getTagsOfInput}
\alias{getTagsOfInput.ArrayExplorer}
\alias{getTagsOfInput.ChromosomeExplorer}
\alias{getTagsOfInput.Explorer}
\alias{getTargetDistribution}
\alias{getTargetDistribution.OligoQuantileNormalization}
\alias{getTargetDistribution.QuantileNormalization}
\alias{getTargetDistributionIdentifier}
\alias{getTargetDistributionIdentifier.QuantileNormalization}
\alias{getTargetDistributionPathname}
\alias{getTargetDistributionPathname.QuantileNormalization}
\alias{getTargetFunction}
\alias{getTargetFunction.FragmentLengthNormalization}
\alias{getTemplatePath}
\alias{getTemplatePath.Explorer}
\alias{getTimestamp}
\alias{getTimestamps}
\alias{getTimestamps.AffymetrixCelSet}
\alias{getTuple}
\alias{getUnitGroupCellMap}
\alias{getUnitGroupCellMap.AffymetrixCdfFile}
\alias{getUnitIndices}
\alias{getUnitIndices.GdasAnnotationSet}
\alias{getUnitIntensities}
\alias{getUnitNames}
\alias{getUnitNames.PlatformDesign}
\alias{getUnitsAt}
\alias{getUnitsAt.AromaUgpFile}
\alias{getUnitSizes}
\alias{getUnitsOnChromosome}
\alias{getUnitsOnChromosome.GenomeInformation}
\alias{getUnitsOnChromosome.PlatformDesign}
\alias{getWeights}
\alias{getWeightsFileClass}
\alias{getWeightsFileClass.WeightsSet}
\alias{getWeightsSet}
\alias{getWeightsSet.ProbeLevelModel}
\alias{getWeightsSetClass}
\alias{getWeightsSetClass.ProbeLevelModel}
\alias{getWriteMap}
\alias{getWriteMap.PlatformDesign}
\alias{getX}
\alias{getX.PlatformDesign}
\alias{getXAM}
\alias{getY}
\alias{getY.PlatformDesign}
\alias{getYY}
\alias{getYY.TransformReport}
\alias{hasAttributeXY}
\alias{hasAttributeXY.AffymetrixCelFile}
\alias{hasTag}
\alias{hasTag.AffymetrixFile}
\alias{hasTag.AffymetrixFileSet}
\alias{hasTags}
\alias{hasTags.AffymetrixFile}
\alias{hasTags.AffymetrixFileSet}
\alias{highlight}
\alias{highlight.AffymetrixCelFile}
\alias{highlight.ProbeLevelModel}
\alias{identifyCells}
\alias{identifyCells.AffymetrixCdfFile}
\alias{image270}
\alias{importFeatureSet}
\alias{importFeatureSet.AffymetrixCelFile}
\alias{importFromAffymetrixCsvFile}
\alias{importFromAffymetrixCsvFile.AromaCgpFile}
\alias{importFromCrlmmFile}
\alias{importFromDChip}
\alias{importFromGenomeInformation}
\alias{importFromGenomeInformation.AromaUgpFile}
\alias{indexOf}
\alias{indexOf.AffymetrixFileSet}
\alias{indexOfArrays}
\alias{indexOfElements}
\alias{indexOfElements.AromaGenomePositionFile}
\alias{indexOfElements.AromaUgpFile}
\alias{inferParameters}
\alias{inferParameters.CnChipEffectSet}
\alias{inferParameters.SnpChipEffectSet}
\alias{isDone.AptProbesetGenotype}
\alias{isDuplicated}
\alias{isInstalled}
\alias{isInstalled.PlatformDesign}
\alias{isLoaded.PlatformDesign}
\alias{isPm}
\alias{isPm.PlatformDesign}
\alias{isSnpChip}
\alias{join}
\alias{join.GdasAnnotationSet}
\alias{lapply}
\alias{lapply.default}
\alias{lapplyInChunks}
\alias{lapplyInChunks.numeric}
\alias{splitInChunks}
\alias{splitInChunks.numeric}
\alias{library}
\alias{load.PlatformDesign}
\alias{loadDataFile}
\alias{loadDataFile.PlatformDesign}
\alias{log2abs}
\alias{log2neg}
\alias{log2pos}
\alias{matchPatterns}
\alias{matchPatterns.SampleAnnotationFile}
\alias{mergeByCommonTails}
\alias{mergeGroups}
\alias{mergeGroups.ChipEffectFile}
\alias{mergeStrands}
\alias{mergeStrands.CnChipEffectFile}
\alias{mergeStrands.SnpChipEffectFile}
\alias{nbrOfArrays}
\alias{nbrOfArrays.AffymetrixCelSetReporter}
\alias{nbrOfArrays.AptProbesetGenotype}
\alias{nbrOfArrays.QualityAssessmentModel}
\alias{nbrOfArrays.TransformReport}
\alias{nbrOfCells}
\alias{nbrOfCells.AffymetrixCdfFile}
\alias{nbrOfCells.AffymetrixCelFile}
\alias{nbrOfCells.AromaCgpFile}
\alias{nbrOfChipTypes}
\alias{nbrOfChipTypes.ArrayExplorer}
\alias{nbrOfColumns}
\alias{nbrOfColumns.AffymetrixCdfFile}
\alias{nbrOfElements}
\alias{nbrOfElements.AromaGenomePositionFile}
\alias{nbrOfFiles}
\alias{nbrOfQcUnits}
\alias{nbrOfQcUnits.AffymetrixCdfFile}
\alias{nbrOfRows}
\alias{nbrOfRows.AffymetrixCdfFile}
\alias{nbrOfSnps}
\alias{nbrOfUnits}
\alias{nbrOfUnits.AffymetrixCdfFile}
\alias{nbrOfUnits.AromaUgpFile}
\alias{nbrOfUnits.GenomeInformation}
\alias{nbrOfUnits.SnpInformation}
\alias{normalizeGenotypeCone}
\alias{normalizeGenotypeCone.matrix}
\alias{normalizeQuantileSpline}
\alias{plasqGaussian}
\alias{plot.GladModel}
\alias{plot.profileCGH}
\alias{plotFit}
\alias{plotFit.PcrProbeCalibration}
\alias{plotImage}
\alias{plotMvsPosition}
\alias{plotMvsPosition.ProbeLevelModel}
\alias{plotMvsX}
\alias{plotNuse}
\alias{plotNuse.QualityAssessmentModel}
\alias{plotProfile2}
\alias{plotProfile2.profileCGH}
\alias{plotRle}
\alias{plotRle.QualityAssessmentModel}
\alias{plotXYCurve}
\alias{plotXYCurve.TransformReport}
\alias{plotXYCurveLog2}
\alias{plotXYCurveLog2.TransformReport}
\alias{predict.lowess}
\alias{print.GenericSummary}
\alias{print.Interface}
\alias{process}
\alias{process.AptProbesetGenotype}
\alias{process.PcrProbeCalibration}
\alias{range.AffymetrixCelFile}
\alias{range.AffymetrixCelSet}
\alias{read10K}
\alias{read10K.DChipSnpInformation}
\alias{read250K}
\alias{read250K.DChipSnpInformation}
\alias{read250KHg17}
\alias{read250KHg17.DChipGenomeInformation}
\alias{read50K}
\alias{read50K.DChipSnpInformation}
\alias{read50KHg17}
\alias{read50KHg17.DChipGenomeInformation}
\alias{readBlock}
\alias{readBlock.GdasAnnotationFile}
\alias{readCdfGroupStrands}
\alias{readChecksum}
\alias{readChecksum.AffymetrixFile}
\alias{readColumnName}
\alias{readColumnName.GdasAnnotationFile}
\alias{readData}
\alias{readData.AffymetrixCsvGenomeInformation}
\alias{readData.AromaGenomePositionFile}
\alias{readData.DChipGenomeInformation}
\alias{readData.DChipSnpInformation}
\alias{readData.GdasAnnotationFile}
\alias{readData.GenomeInformation}
\alias{readData.SampleAnnotationFile}
\alias{readData.SnpInformation}
\alias{readHeader.AromaGenomePositionFile}
\alias{readHeader.GdasAnnotationFile}
\alias{readLog}
\alias{readLog.AptProbesetGenotype}
\alias{readMouse430}
\alias{readMouse430.DChipGenomeInformation}
\alias{readPlasqTypes}
\alias{readPlasqTypes.AffymetrixCdfFile}
\alias{readTableInternal}
\alias{readTableInternal.GenomeInformation}
\alias{readTableInternal.SnpInformation}
\alias{readUnits}
\alias{readUnits.AffymetrixCelSet}
\alias{readUnits.ChipEffectFile}
\alias{readUnits.ChipEffectSet}
\alias{readUnits.CnChipEffectFile}
\alias{readUnits.ExonChipEffectFile}
\alias{readUnits.FirmaFile}
\alias{readUnits.FirmaSet}
\alias{readUnits.GenotypeCallSet}
\alias{readUnits.GenotypeCallXdrFile}
\alias{readUnits.ParameterCelFile}
\alias{readUnits.ProbeAffinityFile}
\alias{readUnits.ResidualFile}
\alias{readUnits.ResidualSet}
\alias{readUnits.SnpChipEffectFile}
\alias{readUnits.WeightsFile}
\alias{readUnits.WeightsSet}
\alias{reorder.AffymetrixFileSet}
\alias{require}
\alias{restruct}
\alias{restruct.AffymetrixCdfFile}
\alias{rgbTransform}
\alias{rgbTransform.Image}
\alias{rmaSummary}
\alias{rowSds}
\alias{rowVars}
\alias{rowMads}
\alias{SampleAnnotationFile}
\alias{SampleAnnotationSet}
\alias{sapply}
\alias{sapply.AffymetrixFileSet}
\alias{sapply.default}
\alias{select}
\alias{select.GdasAnnotationSet}
\alias{seq.TransformReport}
\alias{setAlias}
\alias{setAlias.AffymetrixCelSetTuple}
\alias{setAlias.GladModel}
\alias{setAlias.TransformReport}
\alias{setArrays}
\alias{setAttribute}
\alias{setAttribute.AffymetrixFile}
\alias{setAttributes}
\alias{setAttributes.AffymetrixFile}
\alias{setAttributesBy}
\alias{setAttributesBy.AffymetrixCelSet}
\alias{setAttributesBy.AffymetrixFile}
\alias{setAttributesBySampleAnnotationFile}
\alias{setAttributesBySampleAnnotationFile.AffymetrixCelSet}
\alias{setAttributesBySampleAnnotationSet}
\alias{setAttributesBySampleAnnotationSet.AffymetrixCelSet}
\alias{setAttributesByTags}
\alias{setAttributesByTags.AffymetrixCelFile}
\alias{setAttributesByTags.AffymetrixFile}
\alias{setAttributeXY}
\alias{setAttributeXY.AffymetrixCelFile}
\alias{setCdf}
\alias{setCdf.GenotypeCallFile}
\alias{setColorMaps}
\alias{setColorMaps.ArrayExplorer}
\alias{setColorMaps.SpatialReporter}
\alias{setCombineAlleles}
\alias{setCombineAlleles.CnChipEffectSet}
\alias{setCombineAlleles.CnPlm}
\alias{setCombineAlleles.CnProbeAffinityFile}
\alias{setCustomFindCdf}
\alias{setCytoband}
\alias{setCytoband.ChromosomeExplorer}
\alias{setDecodeFunction}
\alias{setDecodeFunction.ParameterCelFile}
\alias{setEncodeFunction}
\alias{setEncodeFunction.ParameterCelFile}
\alias{setLabel.AffymetrixFile}
\alias{setLabel.Model}
\alias{setMergeFunction}
\alias{setMergeFunction.ChipEffectGroupMerge}
\alias{setMergeGroups}
\alias{setMergeGroups.ExonChipEffectSet}
\alias{setMergeGroups.ExonProbeAffinityFile}
\alias{setMergeGroups.ExonRmaPlm}
\alias{setMergeStrands}
\alias{setMergeStrands.SnpChipEffectSet}
\alias{setMergeStrands.SnpPlm}
\alias{setMergeStrands.SnpProbeAffinityFile}
\alias{setRestructor}
\alias{setTags}
\alias{setTags.AffymetrixCelSetTuple}
\alias{setTags.Transform}
\alias{setup}
\alias{setup.AffymetrixFileSetReporter}
\alias{setup.ArrayExplorer}
\alias{setup.ChromosomeExplorer}
\alias{setup.Explorer}
\alias{showAptVersion}
\alias{showAptVersion.AptProbesetGenotype}
\alias{showLog}
\alias{showLog.AptProbesetGenotype}
\alias{smoothScatterMvsA}
\alias{splitByCommonTails}
\alias{sqrtabs}
\alias{sqrtneg}
\alias{sqrtpos}
\alias{stext}
\alias{stext.default}
\alias{stextChipType}
\alias{stextChipType.AffymetrixCdfFile}
\alias{stextLabel}
\alias{stextLabel.AffymetrixFile}
\alias{stextLabels}
\alias{stextLabels.AffymetrixFile}
\alias{stextSize}
\alias{stextSize.AffymetrixFile}
\alias{stringTree}
\alias{stringTree.character}
\alias{subset.AromaGenomePositionFile}
\alias{testAttributes}
\alias{testAttributes.AffymetrixFile}
\alias{transformAffine}
\alias{transformAffine.AffymetrixCelFile}
\alias{transformAffine.AffymetrixCelSet}
\alias{unload.PlatformDesign}
\alias{updateData}
\alias{updateData.AromaGenomePositionFile}
\alias{updateDataFlat}
\alias{updateDataFlat.ChipEffectFile}
\alias{updateDataFlat.FirmaFile}
\alias{updateDataFlat.ResidualFile}
\alias{updateDataFlat.WeightsFile}
\alias{updateSamplesFile}
\alias{updateOnChipTypeJS.ArrayExplorer}
\alias{updateOnLoadJS.ArrayExplorer}
\alias{updateOnChipTypeJS}
\alias{updateOnLoadJS}
\alias{updateUnits}
\alias{updateUnits.ChipEffectFile}
\alias{updateUnits.ChipEffectSet}
\alias{updateUnits.FirmaFile}
\alias{updateUnits.FirmaSet}
\alias{updateUnits.GenotypeCallXdrFile}
\alias{updateUnits.ParameterCelFile}
\alias{updateUnits.ProbeAffinityFile}
\alias{updateUnits.ResidualFile}
\alias{updateUnits.ResidualSet}
\alias{updateUnits.WeightsFile}
\alias{updateUnits.WeightsSet}
\alias{uses}
\alias{uses.Interface}
\alias{validate.PlatformDesign}
\alias{validateChecksum}
\alias{validateChecksum.AffymetrixFile}
\alias{verify}
\alias{verify.AffymetrixCsvGenomeInformation}
\alias{verify.DChipGenomeInformation}
\alias{verify.DChipSnpInformation}
\alias{whatDataType}
\alias{whatDataType.default}
\alias{write}
\alias{writeChecksum}
\alias{writeChecksum.AffymetrixFile}
\alias{writeChrXFile}
\alias{writeChrXFile.AffymetrixPowerTools}
\alias{writeGraphs}
\alias{writeGraphs.ChromosomeExplorer}
\alias{writeImage}
\alias{writeImage.ResidualFile}
\alias{writeImage.WeightsFile}
\alias{writeImageCombined}
\alias{writeImageCombined.TransformReport}
\alias{writeImages}
\alias{writeImages.SpatialReporter}
\alias{writeImages.TransformReport}
\alias{writeRegions}
\alias{writeRegions.ChromosomeExplorer}
\alias{writeRegions.GladModel}
\alias{writeRegions.profileCGH}
\alias{writeWig}
\alias{writeWig.CnChipEffectSet}
\alias{yellow.colors}
\alias{addExclCells}
\alias{addExclCells.DChipQuantileNormalization}
\alias{AptCopyNumberPipeline}
\alias{calculateParametersGsb}
\alias{excludeChrXFromFit}
\alias{excludeChrXFromFit.DChipQuantileNormalization}
\alias{getAlias.AffymetrixFile}
\alias{getAlias.ArrayExplorer}
\alias{getBaseline}
\alias{getBaseline.CnChipEffectSet}
\alias{getBaseline.SnpChipEffectSet}
\alias{getExclCells}
\alias{getExclCells.DChipQuantileNormalization}
\alias{getParameters.BackgroundCorrection}
\alias{getParameters.GcRmaBackgroundCorrection}
\alias{getParameters.OpticalBackgroundCorrection}
\alias{getParameters.RmaBackgroundCorrection}
\alias{getSubsetToUpdate.BackgroundCorrection}
\alias{readMouse430KenHardwired}
\alias{readMouse430KenHardwired.DChipGenomeInformation}
\alias{setAlias.AffymetrixFile}
\alias{setAlias.ArrayExplorer}
\alias{BrlmmModel}
\alias{as.character.AptCopyNumberPipeline}
\alias{convertTxtFilesToXdr.AptCopyNumberPipeline}
\alias{getAptPathnames.AptCopyNumberPipeline}
\alias{getCdf.AptCopyNumberPipeline}
\alias{getChipEffectSet.CnPlm}
\alias{getChipEffectSet.SnpPlm}
\alias{getChrXFile.AptCopyNumberPipeline}
\alias{getDataSet.AptCopyNumberPipeline}
\alias{getExternalCommand.AptCopyNumberPipeline}
\alias{getFullName.AptCopyNumberPipeline}
\alias{getGenotypeCallSet.AptCopyNumberPipeline}
\alias{getName.AptCopyNumberPipeline}
\alias{getPath.AptCopyNumberPipeline}
\alias{getPathnamesFile.AptCopyNumberPipeline}
\alias{getProbeAffinityFile.CnPlm}
\alias{getProbeAffinityFile.SnpPlm}
\alias{getRootPath.AptCopyNumberPipeline}
\alias{getTags.AptCopyNumberPipeline}
\alias{isDone.AptCopyNumberPipeline}
\alias{nbrOfArrays.AptCopyNumberPipeline}
\alias{process.AptCopyNumberPipeline}
\alias{readLog.AptCopyNumberPipeline}
\alias{showAptVersion.AptCopyNumberPipeline}
\alias{showLog.AptCopyNumberPipeline}
\alias{as.character.character}
\alias{as.character.DChipQuantileNormalization}
\alias{as.complex.complex}
\alias{as.double.double}
\alias{as.integer.integer}
\alias{as.logical.logical}
\alias{getParameters.AllelicCrosstalkCalibration}
\alias{getParameters.DChipQuantileNormalization}
\alias{getProbeAffinityFile.AffinePlm}
\alias{getProbeAffinityFile.MbeiPlm}
\alias{getProbeAffinityFile.RmaPlm}
\alias{AffymetrixTsvFile}
\alias{clearCache.AffymetrixTsvFile}
\alias{exportToAPT}
\alias{findByChipType.AffymetrixTsvFile}
\alias{fromChipType.AffymetrixTsvFile}
\alias{getCdf.AffymetrixTsvFile}
\alias{getCdf.GcContentNormalization}
\alias{getChipType.AffymetrixTsvFile}
\alias{getData.AffymetrixTsvFile}
\alias{getField}
\alias{getField.AffymetrixTsvFile}
\alias{getFragmentLengths.AffymetrixTsvFile}
\alias{getGc}
\alias{getGc.AffymetrixTsvFile}
\alias{getOutputDataSet.GcContentNormalization}
\alias{getParameters.GcContentNormalization}
\alias{getPosition.AffymetrixTsvFile}
\alias{getSubsetToFit.GcContentNormalization}
\alias{getTargetFunction.GcContentNormalization}
\alias{importFromAPT}
\alias{importFromTable}
\alias{importFromTable.FileMatrix}
\alias{isEof}
\alias{isEof.connection}
\alias{readData.AffymetrixTsvFile}
\alias{verify.AffymetrixTsvFile}
\alias{[.CnagCfhFile}
\alias{[.CnagCfhSet}
\alias{[[.CnagCfhFile}
\alias{[[.CnagCfhSet}
\alias{append.CnagCfhSet}
\alias{as.character.CnagCfhFile}
\alias{as.CnagCfhSet}
\alias{clearCache.CnagCfhFile}
\alias{clearCache.CnagCfhSet}
\alias{clearCache.QuantileNormalization}
\alias{clone.CnagCfhFile}
\alias{clone.CnagCfhSet}
\alias{copyTo}
\alias{copyTo.AffymetrixFile}
\alias{copyTo.AffymetrixFileSet}
\alias{findByName.CnagCfhSet}
\alias{fromFiles.CnagCfhSet}
\alias{fromName.CnagCfhSet}
\alias{gaussianSmoothing}
\alias{gaussianSmoothing.matrix}
\alias{gaussianSmoothing.numeric}
\alias{getAsteriskTag.AffinePlm}
\alias{getAsteriskTag.MbeiPlm}
\alias{getAsteriskTag.Model}
\alias{getAsteriskTag.ProbeLevelModel}
\alias{getAsteriskTag.RmaPlm}
\alias{getAverage.CnagCfhSet}
\alias{getAverageAsinh.CnagCfhSet}
\alias{getAverageLog.CnagCfhSet}
\alias{getData.CnagCfhSet}
\alias{getFullName.CnagCfhSet}
\alias{getHeaderLength}
\alias{getHeaderLength.CnagCfhFile}
\alias{getIdentifier.CnagCfhFile}
\alias{getIdentifier.CnagCfhSet}
\alias{getParameters.ParameterCelFile}
\alias{getSubsetToAvg}
\alias{getSubsetToAvg.QuantileNormalization}
\alias{getSubsetToUpdate.DChipQuantileNormalization}
\alias{getTimestamp.CnagCfhFile}
\alias{getTimestamps.CnagCfhSet}
\alias{nbrOfCells.CnagCfhFile}
\alias{nbrOfSnps.CnagCfhFile}
\alias{range.CnagCfhFile}
\alias{range.CnagCfhSet}
\alias{readCfhHeader}
\alias{readCfhHeader.default}
\alias{readCfhUnits}
\alias{readCfhUnits.default}
\alias{readCfnHeader}
\alias{readCfnHeader.default}
\alias{readCfnUnits}
\alias{readCfnUnits.default}
\alias{readUnits.CnagCfhSet}
\alias{getChipEffectSet.ExonRmaPlm}
\alias{getExclCells.QuantileNormalization}
\alias{getParameters.ScaleNormalization}
\alias{getProbeAffinityFile.ExonRmaPlm}
\alias{getSubsetToAvg.ScaleNormalization}
\alias{getSubsetToUpdate.ScaleNormalization}
\alias{matrix.patch}
\alias{getFileFormat}
\alias{getFileFormat.AffymetrixCdfFile}
\alias{getFileFormat.AffymetrixCelFile}
\alias{byChipType}
\alias{byName}
\alias{byName.AffymetrixCelSet}
\alias{fromChipType.AffymetrixCdfFile}
\alias{normalizeAverage}
\alias{renameToUpperCaseExt}
\alias{renameToUpperCaseExt.AffymetrixFile}
\alias{readGenomeWideSNP}
\alias{readGenomeWideSNP.DChipGenomeInformation}
\alias{as.character.CopyNumberSegmentationModel}
\alias{getAlias.CopyNumberSegmentationModel}
\alias{getAsteriskTag.CbsModel}
\alias{getAsteriskTag.CopyNumberSegmentationModel}
\alias{getChipEffectFiles.CopyNumberSegmentationModel}
\alias{getChipTypes.CopyNumberSegmentationModel}
\alias{getFullName.CopyNumberSegmentationModel}
\alias{getFullNames.CopyNumberSegmentationModel}
\alias{getGenomeData}
\alias{getGenomeData.CopyNumberSegmentationModel}
\alias{getListOfCdfs.CopyNumberSegmentationModel}
\alias{getListOfChipEffects.CopyNumberSegmentationModel}
\alias{getListOfGenomeInformations.CopyNumberSegmentationModel}
\alias{getListOfReferences.CopyNumberSegmentationModel}
\alias{getLog2Ratios.CopyNumberSegmentationModel}
\alias{getName.CopyNumberSegmentationModel}
\alias{getNames.CopyNumberSegmentationModel}
\alias{getPath.CopyNumberSegmentationModel}
\alias{getRawCnData}
\alias{getRawCnData.CopyNumberSegmentationModel}
\alias{getReferenceFiles.CopyNumberSegmentationModel}
\alias{getReferenceName.CopyNumberSegmentationModel}
\alias{getRegions.CopyNumberSegmentationModel}
\alias{getReportPath.CopyNumberSegmentationModel}
\alias{getRootPath.CopyNumberSegmentationModel}
\alias{getSetTuple.CopyNumberSegmentationModel}
\alias{getTags.CopyNumberSegmentationModel}
\alias{plot.CbsModel}
\alias{plot.CopyNumberSegmentationModel}
\alias{setAlias.CopyNumberSegmentationModel}
\alias{writeRegions.CopyNumberSegmentationModel}
\alias{drawCnRegions}
\alias{drawCnRegions.profileCGH}
\alias{drawCnRegions.DNAcopy}
\alias{Undocumented}
\alias{code}
\alias{objects:}
\alias{applyRegions}
\alias{applyRegions.CopyNumberRegions}
\alias{as.character.CopyNumberRegions}
\alias{as.character.RawCopyNumbers}
\alias{as.data.frame.CopyNumberRegions}
\alias{as.data.frame.RawCopyNumbers}
\alias{CopyNumberRegions}
\alias{drawLevels}
\alias{drawLevels.CopyNumberRegions}
\alias{extractCNRs}
\alias{extractCNRs.default}
\alias{extractCopyNumberRegions}
\alias{extractCopyNumberRegions.DNAcopy}
\alias{extractCopyNumberRegions.profileCGH}
\alias{extractRawCNs}
\alias{extractRawCNs.default}
\alias{extractRawCopyNumbers}
\alias{extractRawCopyNumbers.DNAcopy}
\alias{extractRawCopyNumbers.profileCGH}
\alias{gaussianSmoothing.RawCopyNumbers}
\alias{getCNs}
\alias{getCNs.RawCopyNumbers}
\alias{getLociFields}
\alias{getLociFields.RawCopyNumbers}
\alias{getPhysicalPositions}
\alias{getPhysicalPositions.RawCopyNumbers}
\alias{getXY}
\alias{getXY.RawCopyNumbers}
\alias{lines.CopyNumberRegions}
\alias{lines.RawCopyNumbers}
\alias{nbrOfLoci}
\alias{nbrOfLoci.RawCopyNumbers}
\alias{nbrOfRegions}
\alias{nbrOfRegions.CopyNumberRegions}
\alias{newPlot}
\alias{newPlot.profileCGH}
\alias{plot.RawCopyNumbers}
\alias{plotRawCNs}
\alias{plotRawCNs.profileCGH}
\alias{points.RawCopyNumbers}
\alias{pointsRawCNs}
\alias{pointsRawCNs.profileCGH}
\alias{RawCopyNumbers}
\alias{sort.RawCopyNumbers}
\alias{summary.RawCopyNumbers}
\alias{digest2}
\alias{applyRegions.CopyNumberOutliers}
\alias{applyRows}
\alias{applyRows.CopyNumberRegions}
\alias{as.character.CopyNumberOutliers}
\alias{as.data.frame.CopyNumberOutliers}
\alias{CopyNumberOutliers}
\alias{drawCytoband}
\alias{drawCytoband.CopyNumberSegmentationModel}
\alias{drawCytoband.default}
\alias{drawCytoband.profileCGH}
\alias{drawCytoband2}
\alias{drawCytoband2.default}
\alias{drawExtraAnnotations}
\alias{drawExtraAnnotations.default}
\alias{drawExtraAnnotations.profileCGH}
\alias{drawLevels.CopyNumberOutliers}
\alias{extractCopyNumberOutliers}
\alias{extractCopyNumberOutliers.DNAcopy}
\alias{extractCopyNumberOutliers.profileCGH}
\alias{getSetTag}
\alias{getSetTag.CopyNumberSegmentationModel}
\alias{lines.CopyNumberOutliers}
\alias{nbrOfRegions.CopyNumberOutliers}
\alias{newPlot.CopyNumberSegmentationModel}
\alias{onFitAddGenotypeCalls}
\alias{onFitAddGenotypeCalls.default}
\alias{onFitAddGenotypeCalls.GladModel}
\alias{pointsRawCNs.default}
\alias{migrateTool}
\alias{migrateTool.CopyNumberSegmentationModel}
\alias{calculateChromosomeStatistics}
\alias{calculateChromosomeStatistics.CopyNumberSegmentationModel}
\alias{getDataPairs}
\alias{getDataPairs.AllelicCrosstalkCalibration}
\alias{getSubsetToAvg.AllelicCrosstalkCalibration}
\alias{rescale}
\alias{rescale.AllelicCrosstalkCalibration}
\alias{plotBasepair}
\alias{plotBasepair.AllelicCrosstalkCalibration}
\alias{getAsteriskTag.AvgPlm}
\alias{getCalculateResidualsFunction.AvgPlm}
\alias{getParameterSet.AvgPlm}
\alias{getProbeAffinityFile.AvgPlm}
\alias{AffymetrixCsvFile}
\alias{AffymetrixNetAffxCsvFile}
\alias{AffymetrixTabularFile}
\alias{as.character.AffymetrixTabularFile}
\alias{findByChipType.AffymetrixCsvFile}
\alias{findByChipType.AffymetrixNetAffxCsvFile}
\alias{findByChipType.AffymetrixTabularFile}
\alias{findByChipType.UgpGenomeInformation}
\alias{fromChipType.AffymetrixTabularFile}
\alias{getAromaUgpFile}
\alias{getAromaUgpFile.UgpGenomeInformation}
\alias{getChromosomes.UgpGenomeInformation}
\alias{getData.UgpGenomeInformation}
\alias{getDataColumns.UgpGenomeInformation}
\alias{getHeader.AffymetrixTabularFile}
\alias{getReadArguments}
\alias{getReadArguments.AffymetrixTabularFile}
\alias{importFromAffymetrixNetAffxCsvFile}
\alias{importFromAffymetrixNetAffxCsvFile.AromaUgpFile}
\alias{readData.AffymetrixTabularFile}
\alias{readData.UgpGenomeInformation}
\alias{readDataUnitChromosomePosition}
\alias{readDataUnitChromosomePosition.AffymetrixNetAffxCsvFile}
\alias{readHeader.AffymetrixTabularFile}
\alias{verify.AffymetrixTabularFile}
\alias{verify.UgpGenomeInformation}
\alias{[.AromaAnnotationFile}
\alias{[[.AromaAnnotationFile}
\alias{[<-.AromaAnnotationFile}
\alias{AromaAnnotationFile}
\alias{AromaUflFile}
\alias{as.character.AromaAnnotationFile}
\alias{colMeans.AromaAnnotationFile}
\alias{colMeans.default}
\alias{colMedians}
\alias{colMedians.AromaAnnotationFile}
\alias{colnames.AffymetrixTabularFile}
\alias{colStats}
\alias{colStats.AromaAnnotationFile}
\alias{colSums.AromaAnnotationFile}
\alias{colSums.default}
\alias{create.AromaAnnotationFile}
\alias{createFromCdf.AromaUflFile}
\alias{dim.AromaAnnotationFile}
\alias{findByChipType.AromaUflFile}
\alias{findByChipType.UflSnpInformation}
\alias{fromChipType.AromaUflFile}
\alias{fromChipType.UflSnpInformation}
\alias{getAromaUflFile}
\alias{getAromaUflFile.UflSnpInformation}
\alias{getCdf.AromaGenomePositionFile}
\alias{getCdf.AromaUflFile}
\alias{getChipType.AromaUflFile}
\alias{getColClasses}
\alias{getColClasses.AromaAnnotationFile}
\alias{getData.UflSnpInformation}
\alias{getDataColumns.UflSnpInformation}
\alias{getFields.UflSnpInformation}
\alias{getFragmentStarts.UflSnpInformation}
\alias{getFragmentStops.UflSnpInformation}
\alias{importFromAffymetrixNetAffxCsvFile.AromaUflFile}
\alias{lapply.AromaAnnotationFile}
\alias{nbrOfUnits.AromaUflFile}
\alias{nbrOfUnits.UflSnpInformation}
\alias{ncol.AromaAnnotationFile}
\alias{nrow.AromaAnnotationFile}
\alias{range.AromaUflFile}
\alias{readData.AromaAnnotationFile}
\alias{readData.UflSnpInformation}
\alias{readDataUnitFragmentLength}
\alias{readDataUnitFragmentLength.AffymetrixNetAffxCsvFile}
\alias{readHeader.AromaAnnotationFile}
\alias{subset.AromaAnnotationFile}
\alias{summary.AromaAnnotationFile}
\alias{UflSnpInformation}
\alias{updateData.AromaAnnotationFile}
\alias{updateDataColumn}
\alias{updateDataColumn.AromaAnnotationFile}
\alias{verify.UflSnpInformation}
\alias{readData.AromaUflFile}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
