%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelFile.PLOT.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{plotImage.AffymetrixCelFile}
\alias{plotImage.AffymetrixCelFile}
\alias{AffymetrixCelFile.plotImage}
\alias{plotImage,AffymetrixCelFile-method}

\title{Displays a spatial plot of a CEL file}

\description{
 Displays a spatial plot of a CEL file.
}

\usage{
\method{plotImage}{AffymetrixCelFile}(this, ...)
}

\arguments{
  \item{...}{Arguments passed to \code{\link[aroma.affymetrix:getImage.AffymetrixCelFile]{*getImage}()}.}
}

\value{
  Returns (invisibly) the image displayed.
}

\author{Ken Simpson, Henrik Bengtsson}

\seealso{
  For more information see \code{\link{AffymetrixCelFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
