%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AllelicCrosstalkCalibration.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.AllelicCrosstalkCalibration}
\alias{process.AllelicCrosstalkCalibration}
\alias{AllelicCrosstalkCalibration.process}
\alias{process,AllelicCrosstalkCalibration-method}

\title{Calibrates the data set}

\description{
 Calibrates the data set.
}

\usage{
\method{process}{AllelicCrosstalkCalibration}(this, ..., force=FALSE, verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{force}{If \code{\link[base:logical]{TRUE}}, data already calibrated is re-calibrated,
      otherwise not.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link[base]{double}} \code{\link[base]{vector}}.
}

\seealso{
  For more information see \code{\link{AllelicCrosstalkCalibration}}.
}
\keyword{internal}
\keyword{methods}
