%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BackgroundCorrection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.BackgroundCorrection}
\alias{process.BackgroundCorrection}
\alias{BackgroundCorrection.process}
\alias{process,BackgroundCorrection-method}

\title{Processes the data set}

\description{
 Processes the data set.
}

\usage{
\method{process}{BackgroundCorrection}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{double}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{BackgroundCorrection}}.
}
\keyword{internal}
\keyword{methods}
