%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericReporter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.GenericReporter}
\alias{process.GenericReporter}
\alias{GenericReporter.process}
\alias{process,GenericReporter-method}

\title{Generates report}

\description{
 Generates report.
}

\usage{
\method{process}{GenericReporter}(...)
}

\arguments{
  \item{...}{See subclasses.}
}

\value{
 Returns nothing.
}

\seealso{
  For more information see \code{\link{GenericReporter}}.
}
\keyword{internal}
\keyword{methods}
