%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  OpticalBackgroundCorrection.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.OpticalBackgroundCorrection}
\alias{process.OpticalBackgroundCorrection}
\alias{OpticalBackgroundCorrection.process}
\alias{process.OpticalBackgroundCorrection}
\alias{process,OpticalBackgroundCorrection-method}


\title{Performs background correction}

\description{
 Performs background correction.
}

\usage{\method{process}{OpticalBackgroundCorrection}(this, ..., force=FALSE, verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns a \code{\link[base]{double}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{OpticalBackgroundCorrection}}.
}
\keyword{internal}
\keyword{methods}
