%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AffymetrixCelFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setCdf.AffymetrixCelFile}
\alias{setCdf.AffymetrixCelFile}
\alias{AffymetrixCelFile.setCdf}
\alias{setCdf,AffymetrixCelFile-method}

\title{Sets the CDF structure for this CEL file}

\description{
 Sets the CDF structure for this CEL file.
}

\usage{
\method{setCdf}{AffymetrixCelFile}(this, cdf, ..., .checkArgs=TRUE)
}

\arguments{
  \item{cdf}{An \code{\link{AffymetrixCdfFile}} object.}
  \item{...}{Not used.}
  \item{.checkArgs}{(Internal) If \code{\link[base:logical]{FALSE}}, arguments are not validated.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[aroma.affymetrix:getCdf.AffymetrixCelFile]{*getCdf}()}.
  For more information see \code{\link{AffymetrixCelFile}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
